### 2018-07-12 #236 @jackwasey
submission: CRAN pacakge icd
Classifies medical diagnostic codes into disease comorbidity groups, calculates risk scores, and converts between types of ICD code, and decodes them to plain language.

```
Package: icd
Title: Tools for Working with ICD-9 and ICD-10 Codes, and
    Finding Comorbidities
Version: 3.2.1
    c(person(given = "Jack O.",
             family = "Wasey",
             email = "jack@jackwasey.com",
      person(given = "William",
             family = "Murphy",
             role = "ctb",
             email = "WMurphy@eatright.org",
             comment = "Van Walraven scores"),
      person(given = "Anobel",
             family = "Odisho",
             role = "ctb",
             email = "anobel.odisho@ucsf.edu",
             comment = "Hierarchical Condition Codes"),
      person(given = "Vitaly",
             family = "Druker",
             role = "ctb",
             email = "vdruker@gmail.com",
             comment = "AHRQ CCS"),
      person(given = "Ed",
             family = "Lee",
             role = "ctb",
             comment = "explain codes in table format"),
      person(given = "Kevin",
             family = "Ushey",
             role = "ctb",
             comment = "Code adapted for fast factor generation"),
      person(given = "R Core Team",
             role = c("ctb", "cph"),
             comment = "m4 macro for OpenMP detection in configure"))
Maintainer: Jack O. Wasey <jack@jackwasey.com>
Description: Calculate comorbidities, Charlson scores, perform
    fast and accurate validation, conversion, manipulation, filtering and
    comparison of ICD-9 and ICD-10 codes. This package enables a work flow
    from raw lists of ICD codes in hospital databases to comorbidities.
    ICD-9 and ICD-10 comorbidity mappings from Quan (Deyo and Elixhauser
    versions), Elixhauser and AHRQ included.  Common ambiguities and code
    formats are handled.
    R (>= 3.4),
    icd.data
    checkmate (>= 1.7.0),
    Rcpp (>= 0.12.3),
    utils
    roxygen2 (>= 5.0.0),
    rticles,
    tinytex,
    xml2
    Rcpp (>= 0.12.3),
    RcppEigen,
    knitr
Classification/ACM-2012: Social and professional topics~Medical
    records, Applied computing~Health care information systems, Applied
    computing~Health informatics, Applied computing~Bioinformatics
Copyright: See file (inst/)COPYRIGHTS
LazyDataCompression: xz

```

https://github.com/jackwasey/icd

data munging: although data extraction is built in, the core functionality is the comorbidity calculation which is data munging.

health and medical science researchers, public health, epidemiology, global health

No other package manipulates ICD codes. There are some overlapping comorbidity calculation packages. 'comorbidity' is a new package. It is much slower than icd to produce the same results (see JSS vignette in 'icd'), and is more demanding of the user in setting up data in the format required (e.g. 'icd' guesses field names, and structure of ICD codes, making conversions automatically when possible). More problematically, 'comorbidity' is very generous in accepting and classifying invalid ICD codes into comorbidities. This has been thought through carefully over years in 'icd' and with feedback from the many users. The JSS article submission vignette touches on the decisions made around this.

'medicalrisk' has similar deficits to 'comorbidity'. 'pccc' is specifically focussed on comorbities of one type (pediatric complex chronic conditions), and does these also much slower than 'icd' (which also offers PCCC calculations), despite using Rcpp and parallelization.

'icd' has a carefully thought out algorithm to do this quickly which is unique in the R package ecosystem. In addition, 'icd' calculates comorbidities of more kinds, (CSS and HCC, from contributed code). 'icd' uses S3 classes to be extensible to different national and international variations of the ICD coding scheme from the WHO.

NA

- [x] does not violate the Terms of Service of any service it interacts with.  (NA)

- [x] Do you intend for this package to go on CRAN?  (it already is)
- [x] NO - a JSS article has been submitted] Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)?
- [x] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? No.

functions are named as follows:

 -  icd9_... for functions which operate on ICD-9 codes
 - icd10_... for functions which operate on ICD-10 codes
 - functions which operate on either (determined based on input) are not prefixed. This decision was taken because of Hadley Wickham's recommendation not to use library, but to write the external package namespace each time, so I expect users might prefer to use, e.g. `icd::comorbid_ahrq(my_patients)`. In any case, the names of functions do not overlap with any well known functions. Previously `explain()` was changed to `explain_code()` for this reason.

NA

### other

There is an accompanying package called 'icd.data' which contains the rarely changing, larger data sets required for 'icd' to function. It is also on CRAN and is very simple. Should I submit a separate onboarding issue for this pacakge? Thanks.
