### 2018-07-12 eb3876b
update devel version in readme and drop code coverage from setup code.
