### 2018-07-12 291a9d1
drop unused code, simplify peripheral internal documentation. Try to stop Rcpp generating spurious changes when compileAttributes called on linux vs macos by using full roxygen tags in Rcpp comments.
