### 2018-07-12 #236 @jackwasey
Thanks for getting back so quickly.

1. I'm not 100% sure what you mean, but I can say that the scores are standardized and used very widely through medical research over many years. There are no novel calculations in the package, just accurate and validated implementations of the existing standards. These are all referenced in the documentation.
2. Yes, the vignette "Efficient ..." is submitted to JSS and under review.
