### 2018-07-13 #236 @maelle
Thanks for your answers @jackwasey!

Your package seems in scope but it cannot be reviewed simultaneously at the two venues. It's up to you to have `icd` reviewed first at JSS or here. Note that our reviews might result in request for changes, even for the stuff reviewed/requested by JSS. With that in mind, do you want to put this submission on hold?
