### 2018-07-15 #236 @maelle
Ok, thanks, good luck with the JSS submission!

If our reviews result in big changes, we have [some guidance about that](https://ropensci.github.io/dev_guide/evolution.html).
