### 2018-07-19 bdad118
travis, and trivial vignette change
