### 2018-07-19 35e20b3
travis - reverted to sudo false, simplified, compilation flags for faster compilation
