### 2018-07-19 997d60f
travis compiler flags syntax
