### 2018-07-19 b66c933
travis CRAN env flags [ci skip]
