+++ .travis.yml
@@ -35,7 +35,9 @@ before_install:
 
 env:
   global:
-    - R_CHECK_CRAN_INCOMING_=FALSE _R_CHECK_FORCE_SUGGESTS_=FALSE
+    - _R_CHECK_CRAN_INCOMING_=FALSE
+    - _R_CHECK_CRAN_INCOMING_REMOTE_=FALSE
+    - _R_CHECK_FORCE_SUGGESTS_=FALSE
     - LINTR_COMMENT_BOT=false
     - ICD_COVERAGE=false
     - OMP_NUM_THREADS=2
