### 2018-07-19 fc808ea
simplify travis build matrix [ci skip]
