### 2018-07-19 5143944
try to get covr to recognize an actually covered code path in factor_split_rcpp
