### 2018-07-20 15e35c3
expand test cases for core categorization when input data is empty.
