+++ tests/testthat/test-categorize.R
@@ -12,11 +12,29 @@ test_that("categorize result is empty", {
                nrow(rdf <- categorize(one_icd10_pt[0, ],
                                       map = icd10_map_ahrq,
                                       id_name = "visit",
-                                      code_name = "icd"))
+                                      code_name = "icd",
+                                      return_df = TRUE))
   )
-  expect_identical(colnames(rdf), names(icd10_map_ahrq))
+  expect_identical(c("visit", names(icd10_map_ahrq)), colnames(rdf))
 })
 
+test_that("categorize_simple result is empty", {
+  expect_equal(0,
+               nrow(res <- categorize_simple(one_icd10_pt[0, ],
+                                             map = icd10_map_ahrq,
+                                             id_name = "visit",
+                                             code_name = "icd"))
+  )
+  expect_identical(colnames(res), names(icd10_map_ahrq))
+  expect_equal(0,
+               nrow(rdf <- categorize_simple(one_icd10_pt[0, ],
+                                             map = icd10_map_ahrq,
+                                             id_name = "visit",
+                                             code_name = "icd",
+                                             return_df = TRUE))
+  )
+  expect_identical(c("visit", names(icd10_map_ahrq)), colnames(rdf))
+})
 test_that("binary instead of logical can be returned", {
   expect_identical(
     categorize(one_icd10_pt,
@@ -48,46 +66,43 @@ test_that("binary instead of logical can be returned", {
   )
 })
 
-test_that("different ID types are preserved", {
-  pts_int <- pts_num <- pts_chr <- random_icd10_pts
+test_that("different ID types are preserved w data frame return", {
+  pts_int <- pts_num <- pts_chr <- pts_fac <- random_icd10_pts
+  empty_i <- empty_n <- empty_c <- empty_f <- random_icd10_pts[NULL, ]
   pts_int$visit_id <- as.integer(pts_int$visit_id)
   pts_num$visit_id <- as.numeric(pts_num$visit_id)
   pts_chr$visit_id <- as_char_no_warn(pts_chr$visit_id)
-  ress <- lapply(
-    list(pts_int, pts_num, pts_chr),
-    function(x) categorize(x,
-                           map = icd10_map_elix,
-                           id_name = "visit_id",
-                           code_name = "icd10",
-                           return_df = TRUE,
-                           preserve_visit_id_type = TRUE))
-  expect_equal(class(ress[[1]][[1]]), "integer")
-  expect_equal(class(ress[[2]][[1]]), "numeric")
-  expect_equal(class(ress[[3]][[1]]), "character")
-  ress_noid <- lapply(ress, `[`, -1L)
-  expect_identical(ress_noid[[1]], ress_noid[[2]])
-  expect_identical(ress_noid[[1]], ress_noid[[3]])
-})
-
-test_that("different ID types are preserved (simple)", {
-  pts_int <- pts_num <- pts_chr <- random_icd10_pts
-  pts_int$visit_id <- as.integer(pts_int$visit_id)
-  pts_num$visit_id <- as.numeric(pts_num$visit_id)
-  pts_chr$visit_id <- as_char_no_warn(pts_chr$visit_id)
-  ress <- lapply(
-    list(pts_int, pts_num, pts_chr),
-    function(x) categorize_simple(x,
-                                  map = icd10_map_elix,
-                                  id_name = "visit_id",
-                                  code_name = "icd10",
-                                  return_df = TRUE,
[…]
