+++ R/RcppExports.R
@@ -1,6 +1,22 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
+}
+
+icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
+}
+
+icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
+}
+
+icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
+}
+
 #' Convert \code{mjr} and \code{mnr} vectors to single code
 #'
 #' In debug mode, will check that \code{mjr} and \code{mnr} are same length.
@@ -23,22 +39,6 @@ icd9MajMinToDecimal <- function(mjr, mnr) {
 
 icd9AppendMinors <- function(m, mnr, isShort) {
     invisible(.Call(`_icd_icd9AppendMinors`, m, mnr, isShort))
-}
-
-icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
-}
-
-icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
-}
-
-icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
-}
-
-icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
 }
 
 #' Set ICD short-form diagnosis code attribute
@@ -218,6 +218,10 @@ comorbidMatMul <- function(icd9df, icd9Mapping, visitId, icd9Field, threads = 8L
     .Call(`_icd_comorbidMatMul`, icd9df, icd9Mapping, visitId, icd9Field, threads, chunk_size, omp_chunk_size)
 }
 
+icd9MajMinToCode_alt_Old <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_Old`, mjr, mnr, isShort)
+}
+
 icd9PartsToShort <- function(parts) {
     .Call(`_icd_icd9PartsToShort`, parts)
 }
@@ -271,8 +275,8 @@ get_major.icd9 <- function(x, short_code) {
     .Call(`_icd_icd9GetMajor`, x, short_code)
 }
 
-icd9MajMinToCode_alt_Old <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_Old`, mjr, mnr, isShort)
+fastIntToString_alt_Std <- function(x) {
+    .Call(`_icd_fastIntToString_alt_Std`, x)
 }
 
 #' @title Convert integers to strings as quickly as possible
@@ -284,10 +288,6 @@ icd9MajMinToCode_alt_Old <- function(mjr, mnr, isShort) {
[…]

[…]
