### 2018-07-20 be7a3a7
Return empty data frames which match the input when splitting. Dropped new-fangled Rcpp things not yet on CRAN.
