### 2018-07-20 1f46849
metadata stuff, travis, pkgdown
