+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: icd
-Title: Tools for Working with ICD-9 and ICD-10 Codes, and
-    Finding Comorbidities
+Title: Comorbidity Calculations and Tools for ICD-9 and ICD-10
+    Codes
 Version: 3.2.1
 Authors@R: 
     c(person(given = "Jack O.",
@@ -35,13 +35,13 @@ Authors@R:
              role = c("ctb", "cph"),
              comment = "m4 macro for OpenMP detection in configure"))
 Maintainer: Jack O. Wasey <jack@jackwasey.com>
-Description: Calculate comorbidities, Charlson scores, perform
-    fast and accurate validation, conversion, manipulation, filtering and
-    comparison of ICD-9 and ICD-10 codes. This package enables a work flow
-    from raw lists of ICD codes in hospital databases to comorbidities.
-    ICD-9 and ICD-10 comorbidity mappings from Quan (Deyo and Elixhauser
-    versions), Elixhauser and AHRQ included.  Common ambiguities and code
-    formats are handled.
+Description: Calculate comorbidities, Charlson and van Walraven
+    scores, perform fast and accurate validation, conversion,
+    manipulation, filtering and comparison of ICD-9 and ICD-10 codes. This
+    package enables a work flow from raw lists of ICD codes in hospital
+    databases to comorbidities.  ICD-9 and ICD-10 comorbidity mappings
+    from Quan (Deyo and Elixhauser versions), Elixhauser and AHRQ
+    included.  Common ambiguities and code formats are handled.
 License: GPL-3
 URL: https://jackwasey.github.io/icd/
 BugReports: https://github.com/jackwasey/icd/issues
