### 2018-07-20 4957dfe
huge purge of old comorbidity calculation methods and associated tests and documentation, leaving just the matrix method. They were not simply interchangable, so not worth keeping for benchmarking. Also make travis do just coverage, not run tests gratuitously
