+++ src/RcppExports.cpp
@@ -245,39 +245,6 @@ RcppExport SEXP _icd_setShortDiag(SEXP xSEXP, SEXP valueSEXP) {
     UNPROTECT(1);
     return rcpp_result_gen;
 }
-// factorSplit
-List factorSplit(const List& df, const Nullable<CharacterVector>& relevant, const String& id_name, const String& code_name);
-static SEXP _icd_factorSplit_try(SEXP dfSEXP, SEXP relevantSEXP, SEXP id_nameSEXP, SEXP code_nameSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::traits::input_parameter< const List& >::type df(dfSEXP);
-    Rcpp::traits::input_parameter< const Nullable<CharacterVector>& >::type relevant(relevantSEXP);
-    Rcpp::traits::input_parameter< const String& >::type id_name(id_nameSEXP);
-    Rcpp::traits::input_parameter< const String& >::type code_name(code_nameSEXP);
-    rcpp_result_gen = Rcpp::wrap(factorSplit(df, relevant, id_name, code_name));
-    return rcpp_result_gen;
-END_RCPP_RETURN_ERROR
-}
-RcppExport SEXP _icd_factorSplit(SEXP dfSEXP, SEXP relevantSEXP, SEXP id_nameSEXP, SEXP code_nameSEXP) {
-    SEXP rcpp_result_gen;
-    {
-        Rcpp::RNGScope rcpp_rngScope_gen;
-        rcpp_result_gen = PROTECT(_icd_factorSplit_try(dfSEXP, relevantSEXP, id_nameSEXP, code_nameSEXP));
-    }
-    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
-    if (rcpp_isInterrupt_gen) {
-        UNPROTECT(1);
-        Rf_onintr();
-    }
-    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
-    if (rcpp_isError_gen) {
-        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
-        UNPROTECT(1);
-        Rf_error(CHAR(rcpp_msgSEXP_gen));
-    }
-    UNPROTECT(1);
-    return rcpp_result_gen;
-}
 // categorize_rcpp
 List categorize_rcpp();
 static SEXP _icd_categorize_rcpp_try() {
@@ -319,108 +286,6 @@ BEGIN_RCPP
     rcpp_result_gen = Rcpp::wrap(icd10cmChildrenDefined(x, icd10cm2016, nc));
     return rcpp_result_gen;
 END_RCPP
-}
-// icd9ComorbidShortCpp
-Rcpp::LogicalVector icd9ComorbidShortCpp(const SEXP& icd9df, const Rcpp::List& icd9Mapping, const std::string visitId, const std::string icd9Field, const int threads, const int chunk_size, const int omp_chunk_size);
-static SEXP _icd_icd9ComorbidShortCpp_try(SEXP icd9dfSEXP, SEXP icd9MappingSEXP, SEXP visitIdSEXP, SEXP icd9FieldSEXP, SEXP threadsSEXP, SEXP chunk_sizeSEXP, SEXP omp_chunk_sizeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::traits::input_parameter< const SEXP& >::type icd9df(icd9dfSEXP);
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type icd9Mapping(icd9MappingSEXP);
[…]
