### 2018-07-20 50cebdb
documentation mistake and more clean-up
