+++ tests/testthat/test-score.R
@@ -93,7 +93,7 @@ test_that("Charlson score", {
   expect_identical(charlson(mydff,
                             return_df = TRUE,
                             stringsAsFactors = TRUE,
-                            isShort = FALSE),
+                            short_code = FALSE),
                    structure(list(visit_id = factor(c("a", "b", "c")),
                                   Charlson = c(1, 1, 6)),
                              .Names = c("visit_id", "Charlson"),
@@ -105,7 +105,7 @@ test_that("Charlson score", {
   expect_equivalent(charlson(mydfff,
                              return_df = TRUE,
                              stringsAsFactors = FALSE,
-                             isShort = FALSE),
+                             short_code = FALSE),
                     structure(list(v = c("a", "b", "c"),
                                    Charlson = c(1, 1, 6)),
                               .Names = c("v", "Charlson"),
