+++ man/count_codes.Rd
@@ -68,7 +68,7 @@ first column is used.
   count_codes(mydf, return_df = TRUE)
   count_codes(mydf)
 
-  cmb <- icd9_comorbid_quan_deyo(mydf, isShort = FALSE, return_df = TRUE)
+  cmb <- icd9_comorbid_quan_deyo(mydf, short_code = FALSE, return_df = TRUE)
   count_comorbid(cmb)
 
   wide <- data.frame(visit_name = c("r", "s", "t"),
