### 2018-07-20 111a475
make the deprecated functions more noisy, clean up some deprecated flags to pass tests
