+++ man-roxygen/isShort.R
@@ -1,3 +1,4 @@
 #' @param isShort Deprecated. Single logical value which determines whether the
 #'   ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
-#'   form. Where reasonable, this is guessed from the input data.
+#'   form. Where reasonable, this is guessed from the input data. Use
+#'   \code{short_code} instead.
