### 2018-07-21 1c5b789
R's 'methods' package needs to be imported for some reason, to avoid a gratuitous warning on R CMD check. Unclear what change made this a new necessity. Looking into: `R_DEFAULT_PACKAGES=NULL R -e 'tools:::.check_packages_used(package = "icd")'`
