+++ tools/build_check_quick.sh
@@ -9,7 +9,11 @@ function finish {
 trap finish EXIT
 cp -r "${ICD_HOME:-$HOME/rprojects/icd}" "$tmpd"
 pushd "$tmpd"
-R CMD build --no-build-vignettes --no-manual --resave-data=no icd
+R CMD build \
+  --no-build-vignettes \
+  --no-manual \
+  --resave-data=no \
+  icd
 
 # try to unset debug flag, so ccache caches the results regardless of original path,
 # or configure ccache to do this
@@ -18,7 +22,7 @@ R CMD build --no-build-vignettes --no-manual --resave-data=no icd
 # https://cran.r-project.org/doc/manuals/r-release/R-ints.html#Tools
 # R_MAKEVARS_USER="$HOME/.R/Makevars.mac.quick" \
 R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
-  MAKEFLAGS="-j16" \
+  MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   ICD_TEST_SLOW=false \
   ICD_TEST_BUILD_DATA=false \
   ICD_TEST_DEPRECATED=false \
@@ -34,5 +38,7 @@ R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
   _R_CHECK_TESTS_NLINES_=0 \
   _R_CHECK_USE_INSTALL_LOG_=FALSE \
   _R_CHECK_VIGNETTES_NLINES_=0 \
-  R CMD check --no-build-vignettes "$(ls -t $tmpd/icd*.tar.gz | head -1)"
+  R CMD check \
+    --no-build-vignettes \
+    "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
