+++ tools/build_check_vquick.sh
@@ -17,7 +17,9 @@ R CMD build \
 
 # for all environment variable options see here:
 # https://cran.r-project.org/doc/manuals/r-release/R-ints.html#Tools
-MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
+
+R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
+  MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   ICD_TEST_SLOW=false \
   ICD_TEST_BUILD_DATA=false \
   ICD_TEST_DEPRECATED=false \
@@ -34,5 +36,9 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_USE_INSTALL_LOG_=FALSE \
   _R_CHECK_VIGNETTES_NLINES_=0 \
   R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
- R CMD check --no-build-vignettes --no-manual --ignore-vignettes --no-install "$(ls -t /tmp/icd*.tar.gz | head -1)"
+ R CMD check \
+   --no-build-vignettes \
+   --no-manual \
+   --ignore-vignettes \
+   --no-install "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
