### 2018-07-21 93d26a7
Rcpp inserts a "methods::" so methods must be in DESCRIPTION. Scythed out more dead code. Cover corner cases in refactoring.
