+++ R/factor_nosort.R
@@ -82,8 +82,10 @@ factor_nosort_rcpp <- function(x, levels, na.rm = FALSE) {
 #' @md
 #' @keywords internal manip
 refactor <- function(x, levels, na.rm = FALSE, exclude_na = TRUE) {
-  checkmate::assert_factor(x)
-  checkmate::assert_character(levels)
+  assert_factor(x)
+  assert_character(levels)
+  assert_flag(na.rm)
+  assert_flag(exclude_na)
   if (na.rm)
     refactor_narm_worker(x, levels)
   else
