+++ src/comorbidMatMul.cpp
@@ -1,9 +1,9 @@
 // [[Rcpp::depends(RcppEigen)]]
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::plugins(cpp11)]]
-#include "config.h"                     // for valgrind, CXX11 etc
-#include "local.h"                     // for ICD_OPENMP
-#include "icd_types.h"                 // for ComorbidOut, VecVecInt, VecVec...
+#include "icd_types.h"
+#include "config.h"
+#include "local.h"
 #include "comorbidMatMul.h"
 #include "fastIntToString.h"
 #include <algorithm>                   // for binary_search, copy
@@ -58,7 +58,8 @@ public:
 
 // TODO: make exportable version of get relevant? // [[Rcpp::export]]
 
-// nocov start
+// # nocov start
+#ifdef ICD_DEBUG
 void printCornerMap(DenseMap x) {
   DEBUG("Map matrix:");
   if (x.rows() >= 8 && x.cols() >= 32) {
@@ -85,7 +86,6 @@ void printCornerSparse(PtsSparse x) {
   else
     Rcpp::Rcout << dense << std::endl;
 }
-#ifdef ICD_DEBUG
 #define PRINTCORNERMAP(x) Rcpp::Rcout << #x << ": "; printCornerMap(x);
 #define PRINTCORNERSP(x) Rcpp::Rcout << #x << ": "; printCornerSparse(x);
 #define ICD_ASSIGN(row,col) mat(row, col) = true; // bounds check
@@ -94,7 +94,7 @@ void printCornerSparse(PtsSparse x) {
 #define PRINTCORNERSP(x) ((void)0);
 #define ICD_ASSIGN(row,col) mat.coeffRef(row, col) = true;
 #endif
-// nocov end
+// # nocov end
 
 // This is a complicated function
 //
@@ -199,99 +199,6 @@ void buildVisitCodesSparseSimple(const RObject& visits,
   visMat.conservativeResize(visitIds.size(), relevantSize);
   DEBUG("Built visMt, rows:" << visMat.rows() << ", cols: " << visMat.cols());
   PRINTCORNERSP(visMat);
-}
-
-void buildVisitCodesSparse(const SEXP& icd9df,
-                           const std::string& visitId,
-                           const std::string& icd9Field,
-                           PtsSparse& visit_codes_sparse,
-                           VecStr& visitIds // will have to get this from sparse matrix at end, but needed?
-) {
-  SEXP icds = PROTECT(getRListOrDfElement(icd9df, icd9Field.c_str())); // this is a factor
-  SEXP vsexp = PROTECT(getRListOrDfElement(icd9df, visitId.c_str()));
-  IntegerVector icd9dfFactor = as<IntegerVector>(icds);
-  CV factorLevels = icd9dfFactor.attr("levels");
-  R_xlen_t numUniqueCodes = factorLevels.length();
-
-  //char* lastVisit;
-  //std::strcpy(lastVisit, "JJ94967295JJ"); // random string
-  std::string lastVisit = "JJ94967295JJ94967295JJ"; // random long string TODO: test consequences of going over length.
-
-  int vlen = Rf_length(icds); // same as length of vsexp
-  // make an unordered set for quick check for duplicates while building list of unique visit ids
-  VisLk vis_lookup;
-  vis_lookup.reserve(vlen);
-  // also maintain list of (ordered as first encountered) visit ids
-  visitIds.resize(vlen); // resize and trim at end, as alternative to reserve
-  int n;
-
-  VecVecIntSz vcdb_max_idx = -1; // we increment immediately to zero as first index
-  VecVecIntSz vcdb_new_idx;
[…]
