### 2018-07-21 5e5d3f9
fix deprecated test and travis tweak
