### 2018-07-21 e48989a
trim down travis builds for speed
