### 2018-07-21 29580c7
travis, namespace tweak
