+++ NAMESPACE
@@ -353,6 +353,7 @@ importFrom(magrittr,"%<>%")
 importFrom(magrittr,"%>%")
 importFrom(magrittr,extract2)
 importFrom(magrittr,set_names)
+importFrom(methods,setLoadAction)
 importFrom(stats,aggregate)
 importFrom(stats,setNames)
 importFrom(utils,head)
