+++ .travis.yml
@@ -63,10 +63,10 @@ matrix:
   allow_failures:
   - r: oldrel
     r_build_args: --no-build-vignettes --no-manual --resave-data=no
-    r_check_args: --no-install -no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
+    r_check_args: -no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
   - r: release
     r_build_args: --no-build-vignettes --no-manual --resave-data=no
-    r_check_args: --no-install -no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
+    r_check_args: -no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
     env:
     - ICD_COVERAGE=true
     - ICD_TEST_DEPRECATED=true
