### 2018-07-30 d0acc57
redocument with new roxygen (better concepts), and new Rcpp longjump code
