+++ man/is_valid.Rd
@@ -133,4 +133,5 @@ Leading zeroes in the decimal form are not ambiguous. Although
 \url{http://www.stata.com/users/wgould/icd9/icd9.hlp}
   \url{http://www.sascommunity.org/wiki/Validate_the_format_of_ICD-9_codes}
 }
+\concept{ICD9 validation}
 \keyword{internal}
