### 2018-07-30 #236 @maelle
@rcih Could you please ask your question in the repo of the package i.e. https://github.com/jackwasey/icd/issues ? Thank you.
