### 2018-07-31 fe6ca8d
set package encoding as UTF-8. tidy
