+++ man/icd9_parse_ahrq_ccs.Rd
@@ -25,10 +25,10 @@ Readmission metric)
 }
 \examples{
 \dontrun{
-  icd:::icd9_parse_ahrq_ccs(single = TRUE, save_data = FALSE, offline = FALSE)
-  icd:::icd9_parse_ahrq_ccs(single = TRUE, save_data = FALSE, offline = TRUE)
-  icd:::icd9_parse_ahrq_ccs(single = FALSE, save_data = FALSE, offline = FALSE)
-  icd:::icd9_parse_ahrq_ccs(single = FALSE, save_data = FALSE, offline = TRUE)
+  icd:::icd9_parse_ahrq_ccs(single = TRUE, offline = FALSE)
+  icd:::icd9_parse_ahrq_ccs(single = TRUE, offline = TRUE)
+  icd:::icd9_parse_ahrq_ccs(single = FALSE, offline = FALSE)
+  icd:::icd9_parse_ahrq_ccs(single = FALSE, offline = TRUE)
 }
 }
 \keyword{internal}
