+++ DESCRIPTION
@@ -76,4 +76,5 @@ Classification/ACM-2012: Social and professional topics~Medical
 Copyright: See file (inst/)COPYRIGHTS
 LazyData: true
 LazyDataCompression: xz
+Encoding: UTF-8
 RoxygenNote: 6.1.0
