### 2018-07-31 c8366f9
final RcppExports order churn now https://github.com/RcppCore/Rcpp/pull/878 is on CRAN [ci skip]
