+++ tests/testthat/test-comorbid-calcs-icd10.R
@@ -199,7 +199,7 @@ test_that("ICD-10 comorbidities from uranium", {
 })
 
 test_that("Alcoholic fatty liver", {
-  dat <- data.frame(id=1:3, icd10 = c('K700','K7030','K709'))
-  res <- icd10_comorbid_elix(dat)
-  expect_true(all(res$Alcohol))
+  dat <- data.frame(id = 1:3, icd10 = c("K700", "K7030", "K709"))
+  res <- icd10_comorbid_quan_elix(dat)
+  expect_true(all(res[, "Alcohol"]))
 })
