### 2018-08-07 d9325bc
remove spurious Eigen compiler warnings by modifying Makevars.in [ci skip]
