### 2018-08-08 c9f6260
Makevars can't ignore warnings, even if spurious, for CRAN
