+++ src/Makevars.in
@@ -3,6 +3,7 @@
 CXX_STD = CXX11
 
 PKG_CFLAGS = @OPENMP_CFLAGS_IN_PKG@
-PKG_CXXFLAGS = -I. @OPENMP_CFLAGS_IN_PKG@ -DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS -Wno-ignored-attributes
+PKG_CXXFLAGS = -I. @OPENMP_CFLAGS_IN_PKG@ -DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS
+S -Wno-ignored-attributes
 # @NO_CLANG_PRAGMA_WARNING@
 PKG_LIBS = @OPENMP_CFLAGS_IN_PKG@ @LIBS@
