### 2018-08-08 21c5004
typo in Makevars.in
