+++ CONTRIBUTING.md
@@ -1,9 +1,17 @@
 # Contributing to `icd`
 
-I'm delighted to welcome code, documentation, tests and bug report from the many package users. I only get occasional feedback on how people are using the package, so feel free to email me or write soemthing in github. If you want it to work differently, or do new things, then the best way is to file a [github issue](https://github.com/jackwasey/icd/issues): there are probably others who have the same or similar needs, and they, or I, may have already thought about them.
+I'm delighted to welcome code, documentation, tests and bug report from the many `icd` package users. I get intermittent feedback on how people are using the package, so feel free to email me or write soemthing in github. If you want it to work differently, or do new things, then the best way is to file a [github issue](https://github.com/jackwasey/icd/issues): there are probably others who have the same or similar needs, and they, or I, may have already thought about them.
 
-If you find that an ICD code doesn't give the description you expect, this is an issue! If some code runs too slowly to be useful, this is also an issue. As people run more and more real world data through `icd`, more problems will undoubtedly pop up, and may be trivial to fix, so please share these experiences with other users by making [github issues](https://github.com/jackwasey/icd/issues).
+## Issues
+
+If you find that an ICD code doesn't give the comorbidity or description you expect, this is an issue! If some code runs too slowly to be useful, this is also an issue. As people run more and more real world data through `icd`, more problems will undoubtedly pop up, and may be trivial to fix, so please share these experiences with other users by making [github issues](https://github.com/jackwasey/icd/issues).
 
 The [github issues](https://github.com/jackwasey/icd/issues) also sketch out future plans and suggestions for the package.
 
+## Coding standards
+
+Contributions should follow the coding standard seen elsewhere in the package, which is defined by both the [general recommendations](http://adv-r.had.co.nz/Style.html) of Hadley Wickham, and the `.lintr` file in the project root directory. I'd prefer that new dependencies aren't included unless absolutely necessary.
+
 Thanks!
+
+Jack Wasey
