+++ R/RcppExports.R
@@ -1,22 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
-}
-
-icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
-}
-
-icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
-}
-
-icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
-}
-
 #' Convert \code{mjr} and \code{mnr} vectors to single code
 #'
 #' In debug mode, will check that \code{mjr} and \code{mnr} are same length.
@@ -39,6 +23,22 @@ icd9MajMinToDecimal <- function(mjr, mnr) {
 
 icd9AppendMinors <- function(m, mnr, isShort) {
     invisible(.Call(`_icd_icd9AppendMinors`, m, mnr, isShort))
+}
+
+icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
+}
+
+icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
+}
+
+icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
+}
+
+icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
 }
 
 #' Set ICD short-form diagnosis code attribute
@@ -75,12 +75,11 @@ icd10cm_children_defined_cpp <- function(x, icd10cm2016, nc) {
     .Call(`_icd_icd10cmChildrenDefined`, x, icd10cm2016, nc)
 }
 
-#' @title Internal function to simplify a comorbidity map by only including codes
-#' which are parents, or identical to, a given list of codes.
-#' @description
-#' Specifically, this is useful for ICD-10 codes where there are a huge number
-#' of possible codes, but we do not want to make a comorbidity map with such a
-#' large number of codes in it.
+#' @title Internal function to simplify a comorbidity map by only including
+#'   codes which are parents, or identical to, a given list of codes.
+#' @description Specifically, this is useful for ICD-10 codes where there are a
+#' huge number of possible codes, but we do not want to make a comorbidity map
+#' with such a large number of codes in it.
 #' @param x Character vector (not factor)
 #' @template mapping
 #' @template visit_name
@@ -127,10 +126,6 @@ simplify_map_lex <- function(pt_codes, map) {
 #' @keywords internal array algebra
 comorbidMatMulSimple <- function(icd9df, icd9Mapping, visitId, icd9Field) {
     .Call(`_icd_comorbidMatMulSimple`, icd9df, icd9Mapping, visitId, icd9Field)
-}
[…]
