+++ src/RcppExports.cpp
@@ -10,56 +10,6 @@
 
 using namespace Rcpp;
 
-// icd9MajMinToCode_alt_PrePadded
-CV icd9MajMinToCode_alt_PrePadded(const CV mjr, const CV mnr, bool isShort);
-RcppExport SEXP _icd_icd9MajMinToCode_alt_PrePadded(SEXP mjrSEXP, SEXP mnrSEXP, SEXP isShortSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const CV >::type mjr(mjrSEXP);
-    Rcpp::traits::input_parameter< const CV >::type mnr(mnrSEXP);
-    Rcpp::traits::input_parameter< bool >::type isShort(isShortSEXP);
-    rcpp_result_gen = Rcpp::wrap(icd9MajMinToCode_alt_PrePadded(mjr, mnr, isShort));
-    return rcpp_result_gen;
-END_RCPP
-}
-// icd9MajMinToCode_alt_Std
-VecStr icd9MajMinToCode_alt_Std(const VecStr& mjr, const VecStr& mnr, bool isShort);
-RcppExport SEXP _icd_icd9MajMinToCode_alt_Std(SEXP mjrSEXP, SEXP mnrSEXP, SEXP isShortSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const VecStr& >::type mjr(mjrSEXP);
-    Rcpp::traits::input_parameter< const VecStr& >::type mnr(mnrSEXP);
-    Rcpp::traits::input_parameter< bool >::type isShort(isShortSEXP);
-    rcpp_result_gen = Rcpp::wrap(icd9MajMinToCode_alt_Std(mjr, mnr, isShort));
-    return rcpp_result_gen;
-END_RCPP
-}
-// icd9MajMinToShort_alt_Std
-VecStr icd9MajMinToShort_alt_Std(const VecStr& mjr, const VecStr& mnr);
-RcppExport SEXP _icd_icd9MajMinToShort_alt_Std(SEXP mjrSEXP, SEXP mnrSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const VecStr& >::type mjr(mjrSEXP);
-    Rcpp::traits::input_parameter< const VecStr& >::type mnr(mnrSEXP);
-    rcpp_result_gen = Rcpp::wrap(icd9MajMinToShort_alt_Std(mjr, mnr));
-    return rcpp_result_gen;
-END_RCPP
-}
-// icd9MajMinToShortSingle_alt_Std
-VecStr icd9MajMinToShortSingle_alt_Std(const Str& mjr, const VecStr& mnr);
-RcppExport SEXP _icd_icd9MajMinToShortSingle_alt_Std(SEXP mjrSEXP, SEXP mnrSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const Str& >::type mjr(mjrSEXP);
-    Rcpp::traits::input_parameter< const VecStr& >::type mnr(mnrSEXP);
-    rcpp_result_gen = Rcpp::wrap(icd9MajMinToShortSingle_alt_Std(mjr, mnr));
[…]
