### 2018-08-10 614e598
remove all openmp and configure code, no longer needed now Eigen deals with multithreading. don't export headers for simpler includes.
