+++ src/categorize.cpp
@@ -15,9 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with icd. If not, see <http://www.gnu.org/licenses/>.
 
-// [[Rcpp::interfaces(r, cpp)]]
 #include "local.h"
-#include "config.h"                             // for ICD_VALGRIND
 #include "util.h"
 #include <stdlib.h>
 #include <math.h>                              // for floor
