### 2018-08-10 fcd0910
[ci skip] drop more old parallel openmp debug code
