### 2018-08-10 a3afaad
fixed an internal documentation URL discovered only on winbuilder checks. fix minor compiler warning.
