### 2018-08-10 fe8b955
spelling, removed old code comments
