+++ .aspell/defaults.R
@@ -1,5 +1,13 @@
-vignettes <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"), personal = "words.pws")
-Rd_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"), personal = "words.pws")
-R_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"), personal = "words.pws")
-C_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"), personal = "words.pws")
+vignettes <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"),
+                  personal = "words.pws",
+                  dictionaries = c("en_stats", "icdwords"))
+Rd_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"),
+                 personal = "words.pws",
+                 dictionaries = c("en_stats", "icdwords"))
+R_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"),
+                personal = "words.pws",
+                dictionaries = c("en_stats", "icdwords"))
+C_files <- list(control = c("--master=en_US", "--add-extra-dicts=en_GB"),
+                personal = "words.pws",
+                dictionaries = c("en_stats", "icdwords"))
 
