+++ man/generate_spelling.Rd
@@ -4,11 +4,9 @@
 \alias{generate_spelling}
 \title{Generate spelling exceptions}
 \usage{
-generate_spelling(out = ".aspell/words.pws", save_data = TRUE)
+generate_spelling(save_data = TRUE)
 }
 \arguments{
-\item{out}{character scalar with path to save the \code{aspell} word list}
-
 \item{save_data}{single logical value, if \code{TRUE} will save the generated
 data}
 }
