+++ tests/testthat/test-spell.R
@@ -18,6 +18,6 @@
 context("dictionary compiles for spelling check")
 
 test_that("dictionary created", {
-  words <- generate_spelling(out = f, save_data = FALSE)
+  words <- generate_spelling(save_data = FALSE)
   expect_character(words)
 })
