### 2018-08-10 cda4fde
drop deprecated version of spell check
