### 2018-08-11 f62b43f
[ci skip] static analysis mini script update
