+++ tools/scan-build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -euo pipefail
 IFS=$'\n\t'
-tmpd=$(mktemp -d /tmp/icdquickcheck.XXXXXXXXXXX)
+tmpd=$(mktemp -d /tmp/icdscanbuild.XXXXXXXXXXX)
 function finish {
 #	  rm -rf "$tmpd"
   echo "Finished with $tmpd"
@@ -11,7 +11,7 @@ cp -r "${ICD_HOME:-$HOME/rprojects/icd}" "$tmpd"
 pushd "$tmpd"
 mkdir "library"
 # or try --configure-args="CXXFLAGS=-O0"
-R_MAKEVARS_USER="$HOME/.R/Makevars.clang.scan-build" \
+R_MAKEVARS_USER="$HOME/.R/Makevars.scan-build" \
   MAKEFLAGS="-j1" \
   R CMD INSTALL \
     --build \
@@ -22,6 +22,5 @@ R_MAKEVARS_USER="$HOME/.R/Makevars.clang.scan-build" \
     --install-tests \
     --no-docs \
     --no-data \
-    --resave-data=no \
     icd
 popd
