### 2018-08-11 add3ff7
remove package startup message. pre-cran docs. remove standalone stuff.
