+++ vignettes/introduction.Rmd
@@ -1,11 +1,9 @@
 ---
 title: 'Quick Introduction to icd'
 author: "Jack O. Wasey"
-date: "8 May, 2018"
+date: "`r Sys.Date()`"
 output:
   html_document:
-    toc: yes
-  pdf_document:
     toc: yes
 vignette: >
   %\VignetteIndexEntry{Quick Introduction to icd}
@@ -56,6 +54,7 @@ patients_icd9 <- data.frame(
 ## Quick Start
 
 ```{r quickstart}
+library(icd)
 head(uranium_pathology, 10)
 head(comorbid_charlson(uranium_pathology))
 comorbid_charlson(uranium_pathology, return_df = TRUE)[1:5, 1:5]
