+++ vignettes/Ranges.Rmd
@@ -1,11 +1,9 @@
 ---
 title: "Ranges of ICD codes"
 author: "Jack O. Wasey"
-date: "8 May, 2018"
+date: "`r Sys.Date()`"
 output:
   html_document:
-    toc: yes
-  pdf_document:
     toc: yes
 vignette: >
   %\VignetteIndexEntry{Ranges of ICD codes}
