+++ man/icd9_map_ahrq.Rd
@@ -12,9 +12,9 @@
 \url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
 }
 \description{
-This mapping of comorbidities to ICD-9 codes is derived directly from
-SAS code provided by AHRQ, and translated into this R data structure.
-This is a revision of the Elixhauser system, notably excluding
+These mappings of comorbidities to ICD-9 and ICD-10 codes are derived
+directly from SAS code provided by AHRQ then translated into this R data
+structure. This is a revision of the Elixhauser system, notably excluding
 cardiac arrythmia.
 }
 \seealso{
