### 2018-08-12 7ab64f7
fix compiler warnings for long vs int type on all platforms. ignore some vignette building output
