### 2018-08-12 40e3020
cran comments; don't ignore NEWS.md now it is supported. [ci skip]
