### 2018-08-13 ea01418
dropped superfluous release questions [ci skip]
