### 2018-08-13 d0aea6e
drop random patient generation code, moved to new package 'patients'. Drop vestiges of openmp, makevars.win in line with makevars, changed URL which was giving certificate warning on CRAN's checks.
