+++ R/RcppExports.R
@@ -1,22 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
-}
-
-icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
-}
-
-icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
-}
-
-icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
-    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
-}
-
 #' Convert \code{mjr} and \code{mnr} vectors to single code
 #'
 #' In debug mode, will check that \code{mjr} and \code{mnr} are same length.
@@ -39,6 +23,22 @@ icd9MajMinToDecimal <- function(mjr, mnr) {
 
 icd9AppendMinors <- function(m, mnr, isShort) {
     invisible(.Call(`_icd_icd9AppendMinors`, m, mnr, isShort))
+}
+
+icd9MajMinToCode_alt_PrePadded <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_PrePadded`, mjr, mnr, isShort)
+}
+
+icd9MajMinToCode_alt_Std <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_Std`, mjr, mnr, isShort)
+}
+
+icd9MajMinToShort_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShort_alt_Std`, mjr, mnr)
+}
+
+icd9MajMinToShortSingle_alt_Std <- function(mjr, mnr) {
+    .Call(`_icd_icd9MajMinToShortSingle_alt_Std`, mjr, mnr)
 }
 
 #' Set ICD short-form diagnosis code attribute
@@ -128,10 +128,6 @@ comorbidMatMulSimple <- function(icd9df, icd9Mapping, visitId, icd9Field) {
     .Call(`_icd_comorbidMatMulSimple`, icd9df, icd9Mapping, visitId, icd9Field)
 }
 
-icd9MajMinToCode_alt_Old <- function(mjr, mnr, isShort) {
-    .Call(`_icd_icd9MajMinToCode_alt_Old`, mjr, mnr, isShort)
-}
-
 icd9PartsToShort <- function(parts) {
     .Call(`_icd_icd9PartsToShort`, parts)
 }
@@ -185,8 +181,8 @@ get_major.icd9 <- function(x, short_code) {
     .Call(`_icd_icd9GetMajor`, x, short_code)
 }
 
-fastIntToString_alt_Std <- function(x) {
-    .Call(`_icd_fastIntToString_alt_Std`, x)
+icd9MajMinToCode_alt_Old <- function(mjr, mnr, isShort) {
+    .Call(`_icd_icd9MajMinToCode_alt_Old`, mjr, mnr, isShort)
 }
 
 #' @title Convert integers to strings as quickly as possible
@@ -198,6 +194,10 @@ fastIntToString_alt_Std <- function(x) {
 #' @keywords internal manip
 fastIntToStringRcpp <- function(x) {
[…]
