### 2018-08-13 4ee9424
remove all alternative code paths, VignetteDepends added for each vignette, more cross-linking in vignettes
