+++ tests/testthat/test-util.R
@@ -138,16 +138,6 @@ test_that("chapter to desc range works for icd10", {
   )
 })
 
-test_that("good random ICD-9 codes are made", {
-  rs <- icd9RandomShort(1000)
-  expect_true(all(nchar(rs) >= 2))
-  expect_true(all(nchar(rs) <= 5))
-  # basic regex checks here, not full vaildation
-  expect_true(any(grepl(pattern = "^[0-9]{1,5}", rs)))
-  expect_true(any(grepl(pattern = "^V.+", rs)))
-  expect_true(any(grepl(pattern = "^E.+", rs)))
-})
-
 test_that("title case works", {
   expect_identical(to_title_case("jack"), "Jack")
   expect_identical(to_title_case("jack wasey"), "Jack Wasey")
