+++ src/fastIntToString_alt_Std.cpp
@@ -1,14 +1,0 @@
-#include <vector>
-#include <string>
-
-// [[Rcpp::export]]
-std::vector<std::string> fastIntToString_alt_Std(std::vector<int> x) {
-  std::vector<std::string>::size_type len = x.size();
-  std::vector<std::string> out(len);
-  char buffer[64];
-  for (std::vector<double>::size_type i = 0; i != len; ++i) {
-    sprintf(buffer, "%u", x[i]);
-    out[i] = buffer;
-  }
-  return out;
-}

[…]
