### 2018-08-13 7ee98ce
remove all alternative code paths, VignetteDepends added for each vignette, more cross-linking in vignettes
