### 2018-08-13 56b0f9e
more test code for cpp code coverage, bump description version
