+++ src/test-cpp.cpp
@@ -109,14 +109,44 @@ context("MajMin to code") {
 }
 
 context("add leading zeroes to major") {
-  test_that("when major len is 0, result is empty") {
+  test_that("when major len is 0, result is empty, non-std") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("") == NA_CHARACTER);
+  }
+  test_that("when major len is 0, result is empty, std") {
     expect_true(icd9AddLeadingZeroesMajorSingleStd("") == "");
+  }
+  test_that("when major len is 2, result is ok") {
+    expect_true(icd9AddLeadingZeroesMajorSingleStd("01") == "001");
   }
   test_that("E code with two char major works") {
     expect_true(icd9AddLeadingZeroesMajorSingleStd("E1") == "E001");
   }
   test_that("E code with three char major works") {
     expect_true(icd9AddLeadingZeroesMajorSingleStd("E01") == "E001");
+  }
+  test_that("0") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("0") == "000");
+    expect_true(icd9AddLeadingZeroesMajorSingle("0") == "000");
+  }
+  test_that("00") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("00") == "000");
+    expect_true(icd9AddLeadingZeroesMajorSingle("00") == "000");
+  }
+  test_that("01") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("01") == "001");
+    expect_true(icd9AddLeadingZeroesMajorSingleStd("01") == "001");
+  }
+  test_that("V1") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("V1") == "V01");
+    expect_true(icd9AddLeadingZeroesMajorSingleStd("V1") == "V01");
+  }
+  test_that("E2") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("E2") == "E002");
+    expect_true(icd9AddLeadingZeroesMajorSingleStd("E2") == "E002");
+  }
+  test_that("1234 too long major") {
+    expect_true(icd9AddLeadingZeroesMajorSingle("1234") == NA_CHARACTER);
+    expect_true(icd9AddLeadingZeroesMajorSingleStd("1234") == "");
   }
 }
 
