### 2018-08-13 203431a
more tests, bump badge devel ver
