+++ tests/testthat/test-valid.R
@@ -226,6 +226,30 @@ test_that("detect valid short codes", {
   expect_false(is_valid.icd9(short_code = TRUE, "11.22")) # am not expecting decimal points
 })
 
+test_that("detect valid short codes directly", {
+  expect_equal(icd9_is_valid_short_n(character()), logical())
+  expect_error(icd9_is_valid_short_n(list(1230, c(12323, 2323), c("nonesnseses"))))
+  expect_true(icd9_is_valid_short_n("0"))
+  expect_equal(
+    icd9_is_valid_short_n(c("0", "00", "000", "0000", "00000")),
+    c(TRUE, TRUE, TRUE, TRUE, TRUE))
+  expect_true(icd9_is_valid_short_n("12345"))
+  expect_true(icd9_is_valid_short_n("12 "))
+  expect_error(icd9_is_valid_short_n(1))
+  expect_equal(icd9_is_valid_short_n(c("99999", "1")), c(TRUE, TRUE))
+  expect_equal(icd9_is_valid_short_n(c("1", "100", "222.22")), c(TRUE, TRUE, FALSE))
+  expect_equal(
+    icd9_is_valid_short_n(as.factor(c("1", "100", "222.22", "100", "1.1"))),
+    c(TRUE, TRUE, FALSE, TRUE, FALSE))
+  expect_true(icd9_is_valid_short_v("V1")) # not zero-padded, but not ambiguous.
+
+  expect_false(icd9_is_valid_short_n("jericho"))
+  expect_false(icd9_is_valid_short_n(""))
+  expect_false(icd9_is_valid_short_n("123456"))
+  expect_false(icd9_is_valid_short_n("10.1"))
+  expect_false(icd9_is_valid_short_n("11.22")) # am not expecting decimal points
+})
+
 test_that("validate short form V codes - invalid codes", {
   expect_false(is_valid.icd9(short_code = TRUE, "V"))
   expect_false(is_valid.icd9(short_code = TRUE, "VV"))
@@ -233,11 +257,19 @@ test_that("validate short form V codes - invalid codes", {
   expect_false(is_valid.icd9(short_code = TRUE, " V0"))
   expect_false(is_valid.icd9(short_code = TRUE, "V00000"))
   expect_false(is_valid.icd9(short_code = TRUE, "V123456"))
+  expect_false(icd9_is_valid_short_v("V"))
+  expect_false(icd9_is_valid_short_v("VV"))
+  expect_false(icd9_is_valid_short_v("Vbog"))
+  expect_false(icd9_is_valid_short_v(" V0"))
+  expect_false(icd9_is_valid_short_v("V00000"))
+  expect_false(icd9_is_valid_short_v("V123456"))
 })
 
 test_that("validate short form V codes - valid codes", {
   expect_true(is_valid.icd9(short_code = TRUE, "V234"))
   expect_true(is_valid.icd9(short_code = TRUE, " V45"))
+  expect_true(icd9_is_valid_short_v("V234"))
+  expect_true(icd9_is_valid_short_v(" V45"))
 })
 
 test_that("valid short form E codes - invalid input", {
@@ -289,6 +321,80 @@ test_that("valid short form E codes - invalid input", {
   expect_false(is_valid.icd9(short_code = TRUE, ".E999.1"))
   expect_false(is_valid.icd9(short_code = TRUE, ".E9991"))
   expect_false(is_valid.icd9(short_code = TRUE, "E98765"))
+})
+
+test_that("valid short form E codes - invalid input", {
+  # call the E validation directly for these completely non-E codes
[…]
