+++ R/zzz.R
@@ -20,11 +20,11 @@
 .onAttach <- function(libname, pkgname) {
   if (system.file(package = "icd9") != "")
     packageStartupMessage(paste(
-      "The 'icd9' package is now deprecated, and should be removed to avoid conflicts",
-      "with 'icd'. The 'icd' package up to version 2.1 contains tested versions",
-      "of all the deprecated function names which overlap with those in the old",
-      "'icd9' package, e.g. 'icd9ComorbidAhrq'. It is strongly recommended to run the command:",
-      "remove.packages(\"icd9\")"))
+      "The 'icd9' package is now deprecated, and should be removed to avoid",
+      "conflicts with 'icd'. The 'icd' package up to version 2.1 contains",
+      "tested versions of all the deprecated function names which overlap with",
+      "those in the old 'icd9' package, e.g. 'icd9ComorbidAhrq'. It is",
+      "strongly recommended to run the command: remove.packages(\"icd9\")"))
 }
 
 .onUnload <- function(libpath) {
@@ -55,6 +55,6 @@ release_questions <- function() {
       resembles a CRAN maintainers environment?",
     # final manual check:
     "Are all NOTES from R CMD check documented in cran-comments.md",
-    "Have all unnecessary files been ignored in built archive? Look in the final built archive before submitting to CRAN.")
+    "Have all unnecessary files been ignored in built archive?")
 }
 # nocov end
