### 2018-08-15 638dd13
more introduction in readme. tiny JSS bib update. [ci skip]
