+++ vignettes/icd.bib
@@ -301,6 +301,7 @@ The ICD-10 coding algorithms developed in this study to define AMI comorbidities
   doi = {10.1186/1472-6963-6-161},
   issn = {1472-6963},
   journal = {BMC Health Services Research},
+  keywords = {icd,todo},
   month = dec,
   pages = {161},
   pmcid = {PMC1781440},
