### 2018-08-15 8fe38b8
drop unnecessary includes, tweak Makevars*
