+++ src/refactor.cpp
@@ -1,0 +1,234 @@
+#include <Rcpp.h>
+#include "local.h"
+//#include "util.h"
+
+using namespace Rcpp;
+
+//' @describeIn factor_nosort \pkg{Rcpp} implementation, requiring character
+//' vector inputs only, no argument checking.
+//' @keywords internal manip
+// [[Rcpp::export(factor_nosort_rcpp_worker)]]
+IntegerVector factorNoSort(const CharacterVector& x,
+                           const CharacterVector& levels,
+                           const bool na_rm) {
+  // TODO: use new factor code? but this is fine.
+  IntegerVector out = match(x, levels);
+  out.attr("levels") = (CharacterVector) levels;
+  out.attr("class") = "factor";
+  if (!na_rm)
+    return out;
+  return(out[!is_na(out)]);
+}
+
+// TODO someday: can the following be done using the R_StringHash global cache
+// instead of making a new hash table to do the integer matching?
+
+//' @title Re-generate a factor with new levels, without doing string matching
+//' @description This is called by an R wrapper. There is an `na.rm` version,
+//' too.
+//' @md
+//' @keywords internal manip
+// [[Rcpp::export(refactor_worker)]]
+Rcpp::IntegerVector refactor(const IntegerVector& x, const CV& new_levels, bool exclude_na) {
+  TRACE_UTIL("Refactoring");
+  IntegerVector f(x.size()); // too many if we are dropping NA values.
+  CharacterVector lx = x.attr("levels");
+  DEBUG_UTIL_VEC(x);
+  DEBUG_UTIL_VEC(lx);
+  DEBUG_UTIL_VEC(new_levels);
+  if (lx.isNULL()) Rcpp::stop("icd codes must be in a factor");
+  CV no_na_lx;
+  CV no_na_new_levels;
+  //bool any_na_lx = false;
+  LogicalVector which_na_new_levels = is_na(new_levels);
+  // TODO: get the position of the NA in new_levels for insertion later if needed.
+  LogicalVector which_na_old_levels = is_na(lx);
+  if (exclude_na) {
+    DEBUG_UTIL("Dropping NA in input factor levels");
+    no_na_lx = lx[!which_na_old_levels];
+  } else {
+    DEBUG_UTIL("Not looking for NA in the input data levels");
+    no_na_lx = lx;
+  }
+  if (exclude_na) {
+    DEBUG_UTIL("Dropping NA in input levels");
+    no_na_new_levels = new_levels[!which_na_new_levels];
+  } else {
+    DEBUG_UTIL("Not dropping NA in input levels");
+    no_na_new_levels = new_levels;
+  }
+  DEBUG_UTIL_VEC(no_na_lx);
+  DEBUG_UTIL_VEC(no_na_new_levels);
+  if (no_na_new_levels.size() == 0) {
+    DEBUG("no_na_new_levels is empty");
+    f = Rcpp::rep(NA_INTEGER, x.size());
+    f.attr("levels") = CV::create();
+    f.attr("class") = "factor";
+    return f;
+  }
+  IntegerVector new_level_old_idx = Rcpp::match(no_na_lx, no_na_new_levels);
+  DEBUG_UTIL_VEC(new_level_old_idx);
+  R_xlen_t fsz = x.size();
+  DEBUG_UTIL("fsz = " << fsz);
+  LogicalVector matched_na_level(fsz, false);
+  R_xlen_t fi = 0;
+  R_xlen_t i;
+  for (i = 0; i < fsz; ++i) {
+    TRACE_UTIL("refactor considering i: " << i << ", x[i]: " << x[i] << ", "
[…]
