+++ src/cutil.h
@@ -1,17 +1,25 @@
-/*
- * util.h
- *
- *  Created on: Mar 7, 2015
- *      Author: jack
- */
+// Copyright (C) 2014 - 2018  Jack O. Wasey
+//
+// This file is part of icd.
+//
+// icd is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// icd is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with icd. If not, see <http://www.gnu.org/licenses/>.
 
 #ifndef CUTIL_H_
 #define CUTIL_H_
 
-#include <R.h>
 #include <Rinternals.h>
-#include <string.h>
 
 SEXP getRListOrDfElement(SEXP list, const char *str);
 
-#endif // CUTIL_H_
+#endif /* CUTIL_H_ */
