### 2018-08-20 c724c71
[ci skip] use VignetteEncoding instead of including the older encoding latex extenion in vignettes, rebuild bilbiography without cruft. Add icd.data in VignetteDepends in vignettes.
