+++ vignettes/Charlson_and_other_scores.Rmd
@@ -9,7 +9,7 @@ vignette: >
   %\VignetteIndexEntry{Charlson and Van Walraven scores}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteDepends{magrittr, icd.data}
-  \usepackage[utf8]{inputenc}
+  %\VignetteEncoding{UTF-8}
 ---
 <!--
 Copyright (C) 2014 - 2018  Jack O. Wasey
