+++ vignettes/all.bib
@@ -1,82 +1,55 @@
 
 @Misc{		  AgencyforHealthcareResearchandQuality_ClinicalClassificationsSoftware_2012,
+  author	= {{Agency for Healthcare Research {and} Quality}},
+  language	= {en-us},
+  note		= {Accessed 2018-04-24; Agency for Healthcare Research and
+		  Quality, Rockville, MD, USA},
   title		= {Clinical {{Classifications Software}} for {{ICD}}-10
 		  {{Data}}},
-  language	= {en-us},
   url		= {http://www.ahrq.gov/research/data/hcup/icd10usrgd.html},
-  author	= {{Agency for Healthcare Research {and} Quality}},
-  year		= {2012},
-  note		= {Accessed 2018-04-24; Agency for Healthcare Research and
-		  Quality, Rockville, MD, USA}
+  year		= {2012}
 }
 
 @Misc{		  AgencyforHealthcareResearchandQuality_Elixhausercomorbiditysoftware_2018,
-  type		= {Text},
-  title		= {Elixhauser Comorbidity Software for {{ICD}}-10-{{CM}}
-		  Healthcare Cost and Utilization Project},
-  url		= {https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity\_icd10.jsp},
   author	= {{Agency for Healthcare Research {and} Quality}},
   month		= mar,
-  year		= {2018},
   note		= {Accessed 2018-04-23; Agency for Healthcare Research and
-		  Quality, Rockville, MD, USA}
+		  Quality, Rockville, MD, USA},
+  title		= {Elixhauser Comorbidity Software for {{ICD}}-10-{{CM}}
+		  Healthcare Cost and Utilization Project},
+  type		= {Text},
+  url		= {https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity\_icd10.jsp},
+  year		= {2018}
 }
 
 @Article{	  charlson_new_1987,
+  author	= {Charlson, Mary E. and Pompei, Peter and Ales, Kathy L. and
+		  MacKenzie, C. Ronald},
+  doi		= {10.1016/0021-9681(87)90171-8},
+  issn		= {0021-9681},
+  journal	= {Journal of Chronic Diseases},
+  number	= {5},
+  pages		= {373-383},
+  shorttitle	= {A New Method of Classifying Prognostic Comorbidity in
+		  Longitudinal Studies},
   title		= {A New Method of Classifying Prognostic Comorbidity in
 		  Longitudinal Studies: {{Development}} and Validation},
   volume	= {40},
-  issn		= {0021-9681},
-  shorttitle	= {A New Method of Classifying Prognostic Comorbidity in
-		  Longitudinal Studies},
-  doi		= {10.1016/0021-9681(87)90171-8},
-  abstract	= {The objective of this study was to develop a prospectively
-		  applicable method for classifying comorbid conditions which
-		  might alter the risk of mortality for use in longitudinal
-		  studies. A weighted index that takes into account the
-		  number and the seriousness of comorbid disease was
-		  developed in a cohort of 559 medical patients. The 1-yr
-		  mortality rates for the different scores were: ``0'', 12\%
-		  (181); ``1\textendash{}2'', 26\% (225);
-		  ``3\textendash{}4'', 52\% (71); and ``$\geqslant$ 5'', 85\%
-		  (82). The index was tested for its ability to predict risk
-		  of death from comorbid disease in the second cohort of 685
-		  patients during a 10-yr follow-up. The percent of patients
-		  who died of comorbid disease for the different scores were:
-		  ``0'', 8\% (588); ``1'', 25\% (54); ``2'', 48\% (25); ``
-		  $\geqslant$ 3'', 59\% (18). With each increased level of
[…]
