+++ vignettes/icd.bib
@@ -7,9 +7,6 @@
   url = {http://www.ahrq.gov/research/data/hcup/icd10usrgd.html},
   year = {2012}
 }
-% == BibTeX quality report for AgencyforHealthcareResearchandQuality_ClinicalClassificationsSoftware_2012:
-% Missing required field 'howpublished'
-% ? Title looks like it was stored in title-case in Zotero
 
 @misc{AgencyforHealthcareResearchandQuality_Elixhausercomorbiditysoftware_2018,
   author = {{Agency for Healthcare Research {and} Quality}},
@@ -20,11 +17,8 @@
   url = {https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp},
   year = {2018}
 }
-% == BibTeX quality report for AgencyforHealthcareResearchandQuality_Elixhausercomorbiditysoftware_2018:
-% Missing required field 'howpublished'
 
 @article{charlson_new_1987,
-  abstract = {The objective of this study was to develop a prospectively applicable method for classifying comorbid conditions which might alter the risk of mortality for use in longitudinal studies. A weighted index that takes into account the number and the seriousness of comorbid disease was developed in a cohort of 559 medical patients. The 1-yr mortality rates for the different scores were: ``0'', 12\% (181); ``1\textendash{}2'', 26\% (225); ``3\textendash{}4'', 52\% (71); and ``$\geqslant$ 5'', 85\% (82). The index was tested for its ability to predict risk of death from comorbid disease in the second cohort of 685 patients during a 10-yr follow-up. The percent of patients who died of comorbid disease for the different scores were: ``0'', 8\% (588); ``1'', 25\% (54); ``2'', 48\% (25); `` $\geqslant$ 3'', 59\% (18). With each increased level of the comorbidity index, there were stepwise increases in the cumulative mortality attributable to comorbid disease (log rank $\chi$2 = 165; p \&lt; 0.0001). In this longer follow-up, age was also a predictor of mortality (p \&lt; 0.001). The new index performed similarly to a previous system devised by Kaplan and Feinstein. The method of classifying comorbidity provides a simple, readily applicable and valid method of estimating risk of death from comorbid disease for use in longitudinal studies. Further work in larger populations is still required to refine the approach because the number of patients with any given condition in this study was relatively small.},
   author = {Charlson, Mary E. and Pompei, Peter and Ales, Kathy L. and MacKenzie, C. Ronald},
   doi = {10.1016/0021-9681(87)90171-8},
   issn = {0021-9681},
@@ -47,11 +41,8 @@
   url = {https://www.cms.gov/},
   year = {2018}
 }
-% == BibTeX quality report for CMS_CentersMedicareMedicaid_2018:
-% Missing required field 'howpublished'
 
 @article{diamond_genetic_2012,
-  abstract = {This paper presents genetic matching, a method of multivariate matching that uses an evolutionary search algorithm to determine the weight each covariate is given. Both propensity score matching and matching based on Mahalanobis distance are limiting cases of this method. The algorithm makes transparent certain issues that all matching methods must confront. We present simulation studies that show that the algorithm improves covariate balance and that it may reduce bias if the selection on observables assumption holds. We then present a reanalysis of a number of data sets in the LaLonde (1986) controversy.},
   author = {Diamond, Alexis and Sekhon, Jasjeet S.},
   doi = {10.1162/REST_a_00318},
   issn = {0034-6535},
@@ -61,11 +52,6 @@
   title = {Genetic {{Matching}} for {{Estimating Causal Effects}}: {{A General Multivariate Matching Method}} for {{Achieving Balance}} in {{Observational Studies}}},
[…]

[…]
