### 2018-08-21 96f55b8
better makefile for building replication code [ci skip]
