+++ benchmarks/icd-JSS3447-replication/bench-versus-dput-3-9801a7af6f79.R
@@ -1,0 +1,5 @@
+structure(list(datarows = c(10, 100, 1000), icd = c(0.00253339600000047, 
+0.00294909299999979, 0.00375528100000011), comorbidity = c(0.00799111649999951, 
+0.0185132645000028, 0.120547388000002), medicalrisk = c(0.00379940100000198, 
+0.0115843309999981, 0.0866471950000012)), row.names = c(NA, -3L
+), class = "data.frame")
