### 2018-08-22 f4a3ebd
[ci skip] replication makefile now works when bash not default shell for make
