+++ benchmarks/icd-JSS3447-replication/Makefile
@@ -1,12 +1,15 @@
 # this should work on unices and Windows
 HOST := $(shell hostname)
 RES3 := $(shell Rscript -e 'cat(paste0(paste("bench-versus-dput", 3, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES5 := $(shell Rscript -e 'cat(paste0(paste("bench-versus-dput", 5, Sys.info()["nodename"], sep = "-"), ".R"))')
 RES7 := $(shell Rscript -e 'cat(paste0(paste("bench-versus-dput", 7, Sys.info()["nodename"], sep = "-"), ".R"))')
 NAME := icd-JSS3447-replication
 ARTICLE := Efficient_calculation_of_comorbidities_from_medical_codes
 README := README-replication
 #MAKEFILE_PATH := $(abspath $(lastword $(MAKEFILE_LIST)))
 #PARENT_DIR := $(shell dirname $(MAKEFILE_PATH))
+
+replicate: bench
 
 all: quickbench extras
 
@@ -16,7 +19,9 @@ extras: purl readme tar
 
 # replication of benchmark results in icd JSS submission
 $(ARTICLE).R:
-	Rscript -e 'knitr::purl(system.file("doc", "$(ARTICLE).Rmd", package = "icd"))'
+	@echo 'icd' must be installed for the extraction of replication code,
+	@echo or 'knitr::purl' should be called directly from the source tree.
+	Rscript -e 'knitr::purl(system.file("doc", "$(ARTICLE).Rmd", package = "icd"), output = "$(ARTICLE).R")'
 
 $(README).html:
 	Rscript -e 'rmarkdown::render("$(README).Rmd")'
@@ -39,9 +44,9 @@ $(RES7):
 
 quickbench: $(RES3) bench-versus.R
 
-bench: $(RES5)
+bench: $(RES5) bench-versus.R
 
-fullbench: $(RES7)
+fullbench: $(RES7) bench-versus.R
 
 tar: $(NAME).tar.gz
 
@@ -50,9 +55,10 @@ $(NAME).tar.gz:
 	$(eval ICDTMP := $(shell mktemp -d))
 	@mkdir "$(ICDTMP)/$(NAME)"
 	cp * $(ICDTMP)/$(NAME); \
-	pushd $(ICDTMP); \
+	olddir="$$PWD"; \
+	cd $(ICDTMP); \
 	COPYFILE_DISABLE=1 tar -cvz -f $(NAME).tar.gz --exclude=".*" --exclude=$(NAME).tar.gz --exclude="bench-versus-dput*.R" --exclude=bench-versus-result\*.rds $(NAME); \
-	popd; cp $(ICDTMP)/$(NAME).tar.gz .;	rm -rf "$(ICDTMP)"
+	cd "$$olddir";	cp $(ICDTMP)/$(NAME).tar.gz "$$olddir"; rm -rf "$(ICDTMP)"
 
 clean: cleantar
 	rm -f $(ARTICLE).R
@@ -64,4 +70,4 @@ cleantar:
 cleanbench:
 	rm -f bench-versus-result* bench-versus-dput*
 
-.PHONY: bench quickbench fullbench deps all full extras purl tar cleantar cleanbench clean readme
+.PHONY: bench quickbench fullbench deps all full extras purl tar cleantar cleanbench clean readme replicate
