### 2018-08-22 0c760f5
updated to work in Docker WIP [ci skip]
