+++ benchmarks/icd-JSS3447-replication/install-dependencies.R
@@ -1,14 +1,15 @@
 # Ensure the user has a CRAN repository to install benchmark dependencies
 repos_new <- repos_old <- getOption("repos")
 if (length(repos_old) == 0) {
-  repos_new[["CRAN"]] = "https://cloud.r-project.org/"
+	  repos_new[["CRAN"]] = "https://cloud.r-project.org/"
   options(repos = repos_new)
 }
-on.exit(options(repos = repos_old), add = TRUE)
 libPaths_old <- .libPaths()
-on.exit(.libPaths(libPaths_old), add = TRUE)
-icd_bench_lib <- "icd-bench-lib"
+icd_bench_lib <- file.path(getwd(), "icd-bench-lib")
+dir.create(icd_bench_lib, showWarnings = FALSE)
 .libPaths(icd_bench_lib)
+message("Using the following libPaths:")
+print(.libPaths())
 
 # icd should already have required this in package Depends, but let's make sure
 if (!require("icd.data")) install.packages("icd.data")
@@ -19,8 +20,8 @@ if (!require("magrittr")) install.packages("magrittr")
 if (!require("Rcpp")) install.packages("Rcpp")
 
 # some suggested dependencies needed for building replication materials
-if (!require("rmarkdown")) install.packages("Rcpp")
-if (!require("knitr")) install.packages("Rcpp")
+if (!require("rmarkdown")) install.packages("rmarkdown")
+if (!require("knitr")) install.packages("knitr")
 
 # The following are specifically for benchmark and not pacakge dependencies
 if (!require("comorbidity")) install.packages("comorbidity")
@@ -28,3 +29,12 @@ if (!require("medicalrisk")) install.packages("medicalrisk")
 if (!require("R.cache")) install.packages("R.cache")
 if (!require("bench")) install.packages("bench")
 if (!require("tidyr")) install.packages("tidyr")
+
+if (!require("icd")) {
+	  yn <- readline("'icd' not installed. Installing from CRAN? (y/n)")
+  if (tolower(yn) == "y") install.packages("icd")
+}
+
+.libPaths(libPaths_old)
+options(repos = repos_old)
+
