### 2018-08-22 9ffc07b
wip docker and make for jss replication
