+++ benchmarks/icd-JSS3447-replication/Dockerfile
@@ -1,8 +1,7 @@
-FROM rocker/rstudio:latest
+FROM rocker/tidyverse:latest
 COPY . /
-RUN echo 'options(repos = c(CRAN = "https://cran.rstudio.com"))' >> /usr/local/lib/R/etc/Rprofile.site 
+RUN echo 'options(repos = c(CRAN = "https://cran.rstudio.com"))' >> /usr/local/lib/R/etc/Rprofile.site
 RUN mkdir ~/.R
-RUN echo 'CXX11FLAGS+=-march=native -flto -pipe -wno-ignored-attributes' >> ~/.R/Makeflags
-RUN MAKEFLAGS=-j8 Rscript install-dependencies.R
-CMD MAKEFLAGS=-j8 make
-
+RUN echo 'CXXFLAGS+=-march=native -flto -pipe -wno-ignored-attributes' >> ~/.R/Makevars
+RUN MAKEFLAGS=-j4 Rscript install-dependencies.R
+CMD MAKEFLAGS=-j4 make
