### 2018-08-22 b171616
improve dependencies for JSS replication [ci skip]
