### 2018-08-23 5dfd139
workflow for JSS replication with 'checkpoint' package, WIP. [ci skip]
