+++ benchmarks/icd-JSS3447-replication/checkpoint.R
@@ -1,6 +1,7 @@
 icd_checkpoint <- function(dt_str = "2018-08-23") {
   if (!requireNamespace("checkpoint")) install.packages("checkpoint")
   old_lib_paths <- .libPaths()
+  dir.create(".checkpoint", showWarnings = FALSE)
   checkpoint::checkpoint(dt_str,
                          #project = system.file(package = "icd"),
                          R.version = NULL,
@@ -11,3 +12,4 @@ icd_checkpoint <- function(dt_str = "2018-08-23") {
                          )
   old_lib_paths
 }
+
