### 2018-08-25 1a04cdd
checkpoint dir check cross platform for jss repl [ci skip]
