### 2018-08-28 17fc14b
Update some user input validation as checkmate package has changed behavior: https://github.com/mllg/checkmate/issues/146 Fixes #160
