+++ R/print.icd_comorbidity_map.R
@@ -34,7 +34,7 @@
 #' }
 #' @export
 print.comorbidity_map <- function(x, ..., n_comorbidities = 7, n_codes = 7) {
-  assert_list(x, min.len = 1, names = "unique", any.missing = FALSE, types = "character")
+  assert_list(x, min.len = 1, names = "unique", types = "character")
   assert_int(n_comorbidities)
   assert_int(n_codes)
 
