+++ R/categorize.R
@@ -61,7 +61,7 @@ categorize_simple <- function(x, map, id_name, code_name,
                               comorbid_fun = comorbidMatMulSimple) {
   assert_data_frame(x, min.cols = 2, col.names = "unique")
   class(x) <- "data.frame"
-  assert_list(map, any.missing = FALSE, min.len = 1, names = "unique")
+  assert_list(map, min.len = 1, names = "unique")
   assert(check_string(id_name), check_null(id_name))
   assert(check_string(code_name), check_null(code_name))
   stopifnot(id_name %in% names(x))
