### 2018-08-28 844b71f
complete fix for #161 by regenerating documentation
