+++ benchmarks/icd-JSS3447-replication/Dockerfile
@@ -5,9 +5,16 @@ COPY install-dependencies.R .
 COPY checkpoint.R .
 RUN mkdir ~/.R
 RUN mkdir .checkpoint
-RUN echo 'CXXFLAGS+=-march=native -O0 -pipe -w' > ~/.R/Makevars
-RUN echo 'CXX11FLAGS+=-march=native -O0 -pipe -w' >> ~/.R/Makevars
-RUN MAKEFLAGS=-j8 Rscript install-dependencies.R
+RUN echo 'CFLAGS=-w' > ~/.R/Makevars
+RUN echo 'CXXFLAGS=-w' >> ~/.R/Makevars
+RUN echo 'CXX11FLAGS=-w' >> ~/.R/Makevars
+#RUN echo 'CFLAGS=-march=native -O0 -pipe -w' > ~/.R/Makevars
+#RUN echo 'CXXFLAGS=-march=native -O0 -pipe -w' >> ~/.R/Makevars
+#RUN echo 'CXX11FLAGS=-march=native -O0 -pipe -w' >> ~/.R/Makevars
+#RUN echo 'CXX14FLAGS=-march=native -O0 -pipe -w' >> ~/.R/Makevars
+#RUN Rscript install-dependencies.R
 COPY bench-versus.R .
 COPY Makefile .
-CMD MAKEFLAGS=-j8 make
+RUN make deps
+CMD make 
+
