### 2018-08-28 28425e2
drop checkpoint for replication due to checkpoint bug [ci skip]
