+++ benchmarks/icd-JSS3447-replication/Makefile
@@ -1,22 +1,17 @@
 # this should work on unices and Windows, but PATH to R is not set on Windows in command window by default
 R_HOME := $(shell R RHOME)
-VANILLA := 
 RSCRIPT := $(shell echo '$(R_HOME)/bin/Rscript')
 HOST := $(shell hostname)
-RES3 := $(shell "$(RSCRIPT)" "$(VANILLA)" -e 'cat(paste0(paste("bench-versus-dput", 3, Sys.info()["nodename"], sep = "-"), ".R"))')
-RES5 := $(shell "$(RSCRIPT)" "$(VANILLA)" -e 'cat(paste0(paste("bench-versus-dput", 5, Sys.info()["nodename"], sep = "-"), ".R"))')
-RES7 := $(shell "$(RSCRIPT)" "$(VANILLA)" -e 'cat(paste0(paste("bench-versus-dput", 7, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES3 := $(shell $(RSCRIPT) -e 'cat(paste0(paste("bench-versus-dput", 3, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES5 := $(shell $(RSCRIPT) -e 'cat(paste0(paste("bench-versus-dput", 5, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES5 := $(shell $(RSCRIPT) -e 'cat(paste0(paste("bench-versus-dput", 6, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES7 := $(shell $(RSCRIPT) -e 'cat(paste0(paste("bench-versus-dput", 7, Sys.info()["nodename"], sep = "-"), ".R"))')
+RES8 := $(shell $(RSCRIPT) -e 'cat(paste0(paste("bench-versus-dput", 8, Sys.info()["nodename"], sep = "-"), ".R"))')
 NAME := icd-JSS3447-replication
 ARTICLE := Efficient_calculation_of_comorbidities_from_medical_codes
 README := README-replication
-#MAKEFILE_PATH := $(abspath $(lastword $(MAKEFILE_LIST)))
-#PARENT_DIR := $(shell dirname $(MAKEFILE_PATH))
-
-replicate: bench
 
 all: quickbench extras
-
-full: fullbench extras
 
 extras: purl readme tar
 
@@ -24,10 +19,10 @@ extras: purl readme tar
 $(ARTICLE).R:
 	@echo 'icd' must be installed for the extraction of replication code,
 	@echo or 'knitr::purl' should be called directly from the source tree.
-	"$(RSCRIPT)" "$(VANILLA)" -e 'knitr::purl(system.file("doc", "$(ARTICLE).Rmd", package = "icd"), output = "$(ARTICLE).R")'
+	$(RSCRIPT) -e 'knitr::purl(system.file("doc", "$(ARTICLE).Rmd", package = "icd"), output = "$(ARTICLE).R")'
 
 $(README).html:
-	"$(RSCRIPT)" "$(VANILLA)" -e 'rmarkdown::render("$(README).Rmd")'
+	$(RSCRIPT) -e 'rmarkdown::render("$(README).Rmd")'
 
 readme: $(README).html
 
@@ -41,22 +36,37 @@ deps: .checkpoint/2018-08-23
 	$(RSCRIPT) install-dependencies.R
 
 $(RES3): deps
-	"$(RSCRIPT)" "$(VANILLA)" bench-versus.R 3
+	$(RSCRIPT) bench-versus.R 3
 
 $(RES5): deps
-	"$(RSCRIPT)" "$(VANILLA)" bench-versus.R 5
+	$(RSCRIPT) bench-versus.R 5
+
+$(RES5): deps
[…]
