+++ benchmarks/icd-JSS3447-replication/bench-versus.R
@@ -1,9 +1,6 @@
 source("install-dependencies.R")
-#source("checkpoint.R")
-
-#old_lib_paths <- icd_checkpoint()
 args <- commandArgs(trailingOnly = TRUE)
-n_order <- 5L
+n_order <- 3L
 if (length(args) > 1L) stop("Only one argument is accepted, which is the order",
                             " of magitude of the biggest benchmark")
 if (length(args) == 1L)
@@ -45,13 +42,33 @@ get_ten_million_icd9_pts <- function() {
   invisible(ten_million_random_pts)
 }
 
+get_pts <- function(n = 1e6, dz_per_pt = 5, use_cache = FALSE) {
+  if (use_cache) {
+    pts <- R.cache::loadCache(key = list("pts_real_diags", n, dz_per_pt),
+                              suffix = "icd.Rcache")
+    if (!is.null(pts)) return(pts)
+  }
+  set.seed(1441)
+  diags <- sample(icd.data::icd9cm_hierarchy$code, size = n, replace = TRUE)
+  pts <- data.frame(
+    visit_id = as.character(floor(seq_len(n) / dz_per_pt)),
+    code = diags,
+    stringsAsFactors = FALSE
+  )
+  R.cache::saveCache(pts,
+                     key = list("pts_real_diags", n, dz_per_pt),
+                     suffix = "icd.Rcache")
+  pts
+}
+
 medicalrisk_fix <- function(pts_mr) {
   names(pts_mr) <- c("id", "icd9cm")
   pts_mr$icd9cm <- paste("D", pts_mr$icd9cm, sep = "")
   pts_mr
 }
 
-ten_million_random_pts <- get_ten_million_icd9_pts()
+#ten_million_random_pts <- get_ten_million_icd9_pts()
+ten_million_random_pts <- get_pts(1e7, dz_per_pt = 5, use_cache = TRUE)
 n <- 10^(1L:n_order)
 bres <- bench::press(n = n, {
   pts <- ten_million_random_pts[seq_len(n), ]
@@ -78,7 +95,10 @@ saveRDS(bres,
 )
 # now take the medians and make suitable for the article:
 res <- tidyr::spread(bres[c(1, 2, 5)], expression, median)
-names(res) <- c("datarows", "icd", "comorbidity", "medicalrisk")
+# name order is not deterministic!
+names(res) <- c("datarows",
+                sub("([^:]*).*", "\\1", names(res)[-1]))
+res[c("datarows", "icd", "comorbidity", "medicalrisk")]
 res$icd <- as.numeric(res$icd)
 res$comorbidity <- as.numeric(res$comorbidity)
 res$medicalrisk <- as.numeric(res$medicalrisk)
@@ -93,5 +113,3 @@ dput(res,
        ".R")
 )
 options(old_opt_dml)
-#checkpoint::unCheckpoint(old_lib_paths)
-
