### 2018-08-30 8bf68a4
[ci skip] replication make looks good and simpler
