+++ benchmarks/icd-JSS3447-replication/bench-versus.R
@@ -1,6 +1,14 @@
 source("install-dependencies.R")
-args <- commandArgs(trailingOnly = TRUE)
+# these control default behavior
+# n is total number of rows of data
+# dz_per_pt is number of disease codes per patient
+# total number of patients is n / dz_per_pt
+#
+# N.b., changing these numbers will interfere with Makefile knowing what to do.
 n_order <- 3L
+dz_per_pt <- 20L
+
+args <- commandArgs(trailingOnly = TRUE)
 if (length(args) > 1L) stop("Only one argument is accepted, which is the order",
                             " of magitude of the biggest benchmark")
 if (length(args) == 1L)
@@ -10,55 +18,28 @@ message("Running benchmarks, with biggest synthetic data set having 10^",
 if (n_order > 5L)
   warning("Depending on hardware, running these benchmarks with 10^6 or more",
           " rows may take hours.")
+#
+# get_ten_million_icd9_pts <- function() {
+#   ten_million_random_pts <-
+#     R.cache::loadCache(key = list("ten_million_random_pts"),
+#                        suffix = "icd.Rcache")
+#   if (is.null(ten_million_random_pts)) {
+#     ten_million_random_pts <- generate_pts(1e7)
+#     R.cache::saveCache(ten_million_random_pts,
+#                        key = list("ten_million_random_pts"),
+#                        suffix = "icd.Rcache")
+#   }
+#   invisible(ten_million_random_pts)
+# }
 
-# include generation functions for reproducibility
-generate_random_short_icd9 <- function(n = 50000)
-  as.character(floor(stats::runif(min = 1, max = 99999, n = n)))
-
-generate_pts <- function(num_patients, dz_per_patient = 20,
-                         n = num_patients, np = dz_per_patient) {
-  set.seed(1441)
-  pts <- round(n / np)
-  data.frame(
-    visit_id = as.character(sample(seq(1, pts), replace = TRUE, size = n)),
-    code = fun(n),
-    poa = as.factor(
-      sample(x = c("Y", "N", "n", "n", "y", "X", "E", "", NA),
-             replace = TRUE, size = n)),
-    stringsAsFactors = FALSE
-  )
-}
-
-get_ten_million_icd9_pts <- function() {
-  ten_million_random_pts <-
-    R.cache::loadCache(key = list("ten_million_random_pts"),
-                       suffix = "icd.Rcache")
-  if (is.null(ten_million_random_pts)) {
-    ten_million_random_pts <- generate_pts(1e7)
-    R.cache::saveCache(ten_million_random_pts,
-                       key = list("ten_million_random_pts"),
-                       suffix = "icd.Rcache")
-  }
-  invisible(ten_million_random_pts)
-}
-
-get_pts <- function(n = 1e6, dz_per_pt = 5, use_cache = FALSE) {
-  if (use_cache) {
-    pts <- R.cache::loadCache(key = list("pts_real_diags", n, dz_per_pt),
-                              suffix = "icd.Rcache")
-    if (!is.null(pts)) return(pts)
-  }
+get_pts <- function(n = 1e6, dz_per_pt = 20) {
   set.seed(1441)
   diags <- sample(icd.data::icd9cm_hierarchy$code, size = n, replace = TRUE)
-  pts <- data.frame(
+  data.frame(
     visit_id = as.character(floor(seq_len(n) / dz_per_pt)),
     code = diags,
     stringsAsFactors = FALSE
   )
-  R.cache::saveCache(pts,
[…]
