+++ benchmarks/icd-JSS3447-replication/install-dependencies.R
@@ -7,7 +7,6 @@ for (p in c(
   "testthat",
   "knitr",
   "profmem",
-  "R.cache",
   "Rcpp",
   "RcppEigen",
   "tidyr",
