### 2018-09-02 2de047e
[ci skip] repl don't store actual gigantic results in result data frame, avoid memory errors when doing huge benchmarks
