### 2018-09-03 057bcb3
[ci skip] clean up makefile a little for repl
