### 2018-09-04 06f4b07
repl [ci skip] bench package uses too much memory workaround
