### 2018-09-04 223e473
[ci skip] repl code for many iterations without memory overload
