### 2018-09-04 602531e
[ci skip] repl install deps as function
