### 2018-09-04 3ceb8d6
[ci skip] repl don't assume CRAN repo set correctly.
