+++ benchmarks/icd-JSS3447-replication/bench-versus.R
@@ -17,7 +17,7 @@ source("install-dependencies.R")
 # total number of patients is n / dz_per_pt
 #
 # N.b., changing these numbers will interfere with Makefile knowing what to do.
-n_order <- 3L
+n_order_default <- 3L
 n_order_big = 6L # cut-off for only doing one iteration
 dz_per_pt <- 20L
 
@@ -31,10 +31,10 @@ args <- commandArgs(trailingOnly = TRUE)
 if (length(args) > 1L) stop("Only one argument is accepted, which is the order",
                             " of magitude of the biggest benchmark")
 if (length(args) == 1L)
-  n_order <- as.integer(args[1])
+  n_order_default <- as.integer(args[1])
 message("Running benchmarks, with biggest synthetic data set having 10^",
-        n_order, " rows.")
-if (n_order > 5L)
+        n_order_default, " rows.")
+if (n_order_default > 5L)
   warning("Depending on hardware, running these benchmarks with 10^6 or more",
           " rows may take hours.")
 
@@ -54,7 +54,7 @@ medicalrisk_fix <- function(pts_mr) {
   pts_mr
 }
 
-bench_small <- function() {
+bench_small <- function(n_order = n_order_default) {
   n = 10L^(1L:min(n_order, n_order_big - 1))
   bench::press(n = n, {
     pts <- get_pts(n, dz_per_pt = dz_per_pt)
@@ -73,7 +73,7 @@ bench_small <- function() {
   })
 }
 
-time_big <- function() {
+time_big <- function(n_order = n_order_default) {
   n = 10L^(n_order_big:n_order)
   res <- data.frame(datarows = integer(),
                     icd = numeric(),
@@ -98,34 +98,40 @@ time_big <- function() {
   res
 }
 
-get_bench_filename <- function(prefix, suffix, use_date = FALSE) {
+get_bench_filename <- function(prefix, suffix, use_date = FALSE,
+                               n_order = n_order_default) {
   paste0(
     paste(prefix, "n", n_order, "dz", dz_per_pt,
           Sys.info()["nodename"], ifelse(use_date, Sys.Date(), ""), sep = "-"),
     ".", suffix)
 }
 
-get_bench_short_filename <- function(prefix, suffix) {
+get_bench_short_filename <- function(prefix, suffix,
+                                     n_order = n_order_default) {
   paste0(
     paste(prefix, "n", n_order, # "dz", dz_per_pt,
           sep = "-"),
     ".", suffix)
 }
 
-bres <- bench_small()
-# now take the medians and make suitable for the article:
-res <- tidyr::spread(bres[c("expression", "n", "median")], expression, median)
-# name order is not deterministic!
-names(res) <- c("datarows",
-                sub("local\\(([^:]*).*", "\\1", names(res)[-1]))
-res <- res[c("datarows", "icd", "comorbidity", "medicalrisk")]
-res$icd <- as.numeric(res$icd)
-res$comorbidity <- as.numeric(res$comorbidity)
-res$medicalrisk <- as.numeric(res$medicalrisk)
-res <- as.data.frame(res)
-# add the timings for the very long-running computations (which bench only does
-# once anyway, and which are dominated by the computations themselves)
-res <- rbind(res, time_big())
+bench_versus <- function(n_order = n_order_default) {
+  bres <- bench_small(n_order)
+  # now take the medians and make suitable for the article:
+  res <- tidyr::spread(bres[c("expression", "n", "median")], expression, median)
+  # name order is not deterministic!
+  names(res) <- c("datarows",
+                  sub("local\\(([^:]*).*", "\\1", names(res)[-1]))
+  res <- res[c("datarows", "icd", "comorbidity", "medicalrisk")]
[…]
