### 2018-09-04 012447a
[ci skip] repl document cutoff calculation for comorbidity
