### 2018-09-05 deb19fc
update JSS submission with repeat benchmark results, also fixing some caption text; make clear there is no ambiguity in repl code [ci skip]
