### 2018-09-05 f7aff91
return invisibly after installing repl deps [ci skip]
