### 2018-09-06 2b5d184
Makefile now builds everything ready for submission [ci skip]
