### 2018-10-03 ab565e3
fix Catch cpp tests due to a symbol change and different signature for refactor
