### 2018-10-25 770ffef
replication materials for JSS, lock to version of comorbidity and medicalrisk used in paper. Allow interrupts in long/wide conversions. Add 'touch' to benchmarks, initial look for smaller data 17x slower than icd, and faster than others. Try making the v short benchmark in travis.
