### 2018-10-25 fa6781a
just test the mini benchmarks for one build and try to do mini-benchmarks in appveyor
