### 2018-10-26 69591d2
try to find benchmarks in appveyor. travis bash syntax for benchmark
