### 2018-10-26 f6e4a06
don't buildignore benchmarks - may not be needed as we did a git pull in the first place for travis and appveyor.
