+++ appveyor.yml
@@ -23,6 +23,7 @@ before_build:
   - ls -aR
   - rm .Rinstignore  # do install everything for testing
   - sed -i'' -e '/.*[Dd]ata.*/d' .Rbuildignore # do include data for testing, some data not in github though
+  - sed -i'' -e '/\benchmark/d' .Rbuildignore # do include benchmarks
   - Rscript -e "install.packages(c('devtools', 'Rcpp'), repos = 'https://cloud.r-project.org/'); Rcpp::compileAttributes()" # make sure Rcpp generated code is made, if not checked in to repo
 
 build_script:
