### 2018-10-26 3f6b792
try again to make appveyor do mini-benchmark. need reshape2 for benchmark, not reshape
