### 2018-10-26 450a53e
appveyor experimenting for mini benchmark
