### 2018-10-27 8e1b7c9
[ci skip] appveyor experimenting for mini benchmark
