+++ appveyor.yml
@@ -66,7 +66,7 @@ after_build:
   - ls -aR
   - ps: |
       if ($Env:MINI_BENCH -eq "yes") {
-        cd /c/projects/icd/benchmarks/icd-JSS3447-replication
+        cd benchmarks/icd-JSS3447-replication
         make
       }
 
