+++ appveyor.yml
@@ -66,7 +66,7 @@ after_build:
   - ps: |
       if ($Env:MINI_BENCH -eq "yes") {
         cd benchmarks/icd-JSS3447-replication
-        make
+        make 2>&1
         cd ..
         }
 
