### 2018-10-27 5b98047
[ci skip] appveyor runs minibenchmark now
