### 2018-10-28 250102a
Makefile tweak for replication materials for JSS
