### 2018-10-28 f44bc33
Makefile tweak for replication materials for JSS, and include replicaation materials generation in appveyor and travis
