### 2018-10-28 998b2b7
[ci skip] appveyor needs knitr for building replication materials
