### 2018-10-29 f8f71aa
[ci skip] don't build replmat on appveyor
