+++ appveyor.yml
@@ -65,7 +65,7 @@ after_build:
   - pwd
   - IF NOT "%MINI_BENCH%"=="" cd benchmarks/icd-JSS3447-replication
   - IF NOT "%MINI_BENCH%"=="" make bench3
-  - IF NOT "%MINI_BENCH%"=="" make replmat
+#  - IF NOT "%MINI_BENCH%"=="" make replmat # needs to call Rscript on appveyor path.
   - IF NOT "%MINI_BENCH%"=="" cd ../..
 
 artifacts:
