### 2018-11-16 25475cb
No longer using any fast factor code from Kevin Ushey.
