### 2018-11-16 c66ca04
distinguish JSS references from package documentation references in bib files; better file names for vignettes. No Ushey code remains. [ci skip]
