+++ vignettes/icdpkg.bib
@@ -1,0 +1,334 @@
+
+@online{AgencyforHealthcareResearchandQuality_ClinicalClassificationsSoftware_2012,
+  author = {Agency for Healthcare Research {and} Quality},
+  date = {2012},
+  langid = {american},
+  note = {Accessed 2018-04-24; Agency for Healthcare Research and Quality, Rockville, MD, USA},
+  title = {Clinical {{Classifications Software}} for {{ICD}}-10 {{Data}}},
+  url = {http://www.ahrq.gov/research/data/hcup/icd10usrgd.html}
+}
+% == BibLateX quality report for AgencyforHealthcareResearchandQuality_ClinicalClassificationsSoftware_2012:
+% ? Title looks like it was stored in title-case in Zotero
+
+@online{AgencyforHealthcareResearchandQuality_Elixhausercomorbiditysoftware_2018,
+  author = {Agency for Healthcare Research {and} Quality},
+  date = {2018-03},
+  note = {Accessed 2018-04-23; Agency for Healthcare Research and Quality, Rockville, MD, USA},
+  title = {Elixhauser Comorbidity Software for {{ICD}}-10-{{CM}} Healthcare Cost and Utilization Project},
+  type = {Text},
+  url = {https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
+}
+% == BibLateX quality report for AgencyforHealthcareResearchandQuality_Elixhausercomorbiditysoftware_2018:
+% Unexpected field 'type'
+
+@article{charlson_new_1987,
+  author = {Charlson, Mary E. and Pompei, Peter and Ales, Kathy L. and MacKenzie, C. Ronald},
+  date = {1987},
+  doi = {10.1016/0021-9681(87)90171-8},
+  issn = {0021-9681},
+  journaltitle = {Journal of Chronic Diseases},
+  number = {5},
+  pages = {373-383},
+  shortjournal = {Journal of Chronic Diseases},
+  shorttitle = {A New Method of Classifying Prognostic Comorbidity in Longitudinal Studies},
+  title = {A New Method of Classifying Prognostic Comorbidity in Longitudinal Studies: {{Development}} and Validation},
+  volume = {40}
+}
+
+@online{CMS_CentersMedicareMedicaid_2018,
+  author = {CMS},
+  date = {2018-11-06},
+  journaltitle = {Research, Statistics, Data \& Systems},
+  langid = {american},
+  note = {Accessed 2018-06-11; U.S. Centers for Medicare \& Medicaid Services. 7500 Security Boulevard, Baltimore, MD 21244, USA},
+  title = {Centers for {{Medicare}} and {{Medicaid Services}}: {{Research}}, Statistics, Data \& Systems},
+  url = {https://www.cms.gov/}
+}
+% == BibLateX quality report for CMS_CentersMedicareMedicaid_2018:
+% Unexpected field 'journaltitle'
+
+@article{diamond_genetic_2012,
+  author = {Diamond, Alexis and Sekhon, Jasjeet S.},
+  date = {2012-10-10},
+  doi = {10.1162/REST_a_00318},
+  issn = {0034-6535},
+  journaltitle = {Review of Economics and Statistics},
+  shortjournal = {Review of Economics and Statistics},
+  shorttitle = {Genetic {{Matching}} for {{Estimating Causal Effects}}},
+  title = {Genetic {{Matching}} for {{Estimating Causal Effects}}: {{A General Multivariate Matching Method}} for {{Achieving Balance}} in {{Observational Studies}}}
+}
[…]
