+++ DESCRIPTION
@@ -26,11 +26,7 @@ Authors@R:
       person(given = "Ed",
              family = "Lee",
              role = "ctb",
-             comment = "explain codes in table format"),
-      person(given = "Kevin",
-             family = "Ushey",
-             role = "ctb",
-             comment = "Code adapted for fast factor generation"))
+             comment = "explain codes in table format"))
 Maintainer: Jack O. Wasey <jack@jackwasey.com>
 Description: Calculate comorbidities, Charlson and van Walraven
     scores, perform fast and accurate validation, conversion,
