### 2018-11-16 655d772
tweaked vignettes, added NEWS
