+++ vignettes/icdpkg.bib
@@ -47,19 +47,6 @@
 % == BibLateX quality report for CMS_CentersMedicareMedicaid_2018:
 % Unexpected field 'journaltitle'
 
-@article{diamond_genetic_2012,
-  author = {Diamond, Alexis and Sekhon, Jasjeet S.},
-  date = {2012-10-10},
-  doi = {10.1162/REST_a_00318},
-  issn = {0034-6535},
-  journaltitle = {Review of Economics and Statistics},
-  shortjournal = {Review of Economics and Statistics},
-  shorttitle = {Genetic {{Matching}} for {{Estimating Causal Effects}}},
-  title = {Genetic {{Matching}} for {{Estimating Causal Effects}}: {{A General Multivariate Matching Method}} for {{Achieving Balance}} in {{Observational Studies}}}
-}
-% == BibLateX quality report for diamond_genetic_2012:
-% ? Title looks like it was stored in title-case in Zotero
-
 @article{Eddelbeuttel_FastElegantNumerical_2013,
   author = {Eddelbeuttel, Dirk and Bates, Douglas},
   date = {2013},
@@ -175,22 +162,6 @@
   urldate = {2018-07-09}
 }
 
-@inproceedings{Jarvi_AlgorithmSpecializationGeneric_2006,
-  author = {Järvi, Jaakko and Gregor, Douglas and Willcock, Jeremiah and Lumsdaine, Andrew and Siek, Jeremy},
-  booktitle = {Proceedings of the 27th {{ACM SIGPLAN Conference}} on {{Programming Language Design}} and {{Implementation}}},
-  date = {2006},
-  doi = {10.1145/1133981.1134014},
-  isbn = {978-1-59593-320-1},
-  location = {{New York, NY, USA}},
-  pages = {272--282},
-  publisher = {{ACM}},
-  series = {PLDI '06},
-  shorttitle = {Algorithm {{Specialization}} in {{Generic Programming}}},
-  title = {Algorithm {{Specialization}} in {{Generic Programming}}: {{Challenges}} of {{Constrained Generics}} in {{C}}++}
-}
-% == BibLateX quality report for Jarvi_AlgorithmSpecializationGeneric_2006:
-% ? Title looks like it was stored in title-case in Zotero
-
 @article{Lee_Predictingmortalitypatients_2003,
   author = {Lee, Douglas S. and Austin, Peter C. and Rouleau, Jean L. and Liu, Peter P. and Naimark, David and Tu, Jack V.},
   date = {2003-11-19},
@@ -257,22 +228,6 @@
 % == BibLateX quality report for quan_updating_2011:
 % 'issn': not a valid ISSN
 % ? Title looks like it was stored in title-case in Zotero
-
-@article{simpao_big_2015,
-  author = {Simpao, A. F. and Ahumada, L. M. and Rehman, M. A.},
-  date = {2015-01-26},
-  doi = {10.1093/bja/aeu552},
-  eprint = {25627395},
-  eprinttype = {pmid},
-  issn = {0007-0912, 1471-6771},
-  journaltitle = {British Journal of Anaesthesia},
-  langid = {english},
-  pages = {552},
-  shortjournal = {Br. J. Anaesth.},
-  title = {Big Data and Visual Analytics in Anaesthesia and Health Care}
-}
-% == BibLateX quality report for simpao_big_2015:
-% 'issn': not a valid ISSN
 
 @article{So_ICD10codingalgorithms_2006,
   author = {So, Lawrence and Evans, Dewey and Quan, Hude},
