+++ README.md
@@ -18,6 +18,7 @@ along with icd. If not, see <http:#www.gnu.org/licenses/>.
 -->
 <!-- README.md is generated from README.Rmd. Please edit that file and render with rmarkdown::render("README.Rmd")
 -->
+
 icd
 ===
 
@@ -85,11 +86,46 @@ Features
 -   used, tested and benchmarked against other comorbidity calculators
     on hardware from laptops to big servers
 
-Install
--------
+Examples
+--------
+
+See also the vignettes and examples embedded in the help for each
+function for more. Here’s a taste:
 
 ``` r
-install.packages("icd")
+# install.packages("icd")
+library(icd)
+
+# Typical diagnostic code data, with many-to-many relationship
+patient_data
+#>   visit_id  icd9
+#> 1     1000 40201
+#> 2     1000  2258
+#> 3     1000  7208
+#> 4     1000 25001
+#> 5     1001 34400
+#> 6     1001  4011
+#> 7     1002  4011
+#> 8     1000  <NA>
+
+# get comorbidities using Quan's application of Deyo's Charlson comorbidity groups
+comorbid_charlson(patient_data)
+#>         MI   CHF   PVD Stroke Dementia Pulmonary Rheumatic   PUD LiverMild
+#> 1000 FALSE  TRUE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
+#> 1001 FALSE FALSE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
+#> 1002 FALSE FALSE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
+#>         DM  DMcx Paralysis Renal Cancer LiverSevere  Mets   HIV
+#> 1000  TRUE FALSE     FALSE FALSE  FALSE       FALSE FALSE FALSE
+#> 1001 FALSE FALSE      TRUE FALSE  FALSE       FALSE FALSE FALSE
+#> 1002 FALSE FALSE     FALSE FALSE  FALSE       FALSE FALSE FALSE
+
+# or go straight to the Charlson scores:
+charlson(patient_data)
+#> 1000 1001 1002 
+#>    2    2    0
+
+# for more examples, see this and other vignettes
+vignette("introduction", package = "icd")
 ```
 
 Relevance
@@ -144,65 +180,27 @@ ICD-10, but comorbidities can be generated from older ICD-9 codes and
 newer ICD-10 codes in parallel, and the comorbidities can then be
 compared.
 
-Examples
---------
-
-See also the vignettes and examples embedded in the help for each
-function for more. Here’s a taste:
-
-``` r
-library(icd)
-
-patient_data
-#>   visit_id  icd9  poa
-#> 1     1000 40201    Y
-#> 2     1000  2258 <NA>
-#> 3     1000  7208    N
-#> 4     1000 25001    Y
-#> 5     1001 34400    X
-#> 6     1001  4011    Y
-#> 7     1002  4011    E
-
-# get comorbidities using Quan's application of Deyo's Charlson comorbidity groups
-comorbid_charlson(patient_data)
-#>         MI   CHF   PVD Stroke Dementia Pulmonary Rheumatic   PUD LiverMild
-#> 1000 FALSE  TRUE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
-#> 1001 FALSE FALSE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
-#> 1002 FALSE FALSE FALSE  FALSE    FALSE     FALSE     FALSE FALSE     FALSE
[…]
