### 2018-11-17 2ac33f6
add clean print icd code functions and bump version
