+++ man/icd_attr_clean.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/attr.R
+\name{icd_attr_clean}
+\alias{icd_attr_clean}
+\title{Remove any attributes set by 'icd'}
+\usage{
+icd_attr_clean(x)
+}
+\description{
+Remove any attributes set by 'icd'
+}
+\keyword{internal}
