+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: icd
 Title: Comorbidity Calculations and Tools for ICD-9 and ICD-10
     Codes
-Version: 3.2.2
+Version: 3.3
 Authors@R: 
     c(person(given = "Jack O.",
              family = "Wasey",
