+++ NAMESPACE
@@ -91,6 +91,8 @@ S3method(is_valid_major,default)
 S3method(is_valid_major,icd10)
 S3method(is_valid_major,icd9)
 S3method(print,comorbidity_map)
+S3method(print,icd10)
+S3method(print,icd9)
 S3method(print,icd_comorbidity_map)
 S3method(short_to_decimal,default)
 S3method(short_to_decimal,icd10)
