### 2018-11-17 1a7b6fe
update links in NEWS, and a simple build script
