+++ tools/check_cran.sh
@@ -15,7 +15,7 @@ ${ICD_HOME:-$HOME/rprojects/icd}/tools/build.sh
 # https://cran.r-project.org/doc/manuals/r-release/R-ints.html#Tools
 #R_MAKEVARS_USER="$HOME/.R/Makevars.clang" \
 MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
-  _R_CHECK_ALWAYS_LOG_VIGNETTE_OUTPUT_=TRUE \
+  #_R_CHECK_ALWAYS_LOG_VIGNETTE_OUTPUT_=TRUE \
   _R_CHECK_USE_INSTALL_LOG_=TRUE \
   R CMD check --as-cran "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
