### 2018-11-18 ae8e0eb
[ci skip] cran comments
