### 2018-11-18 f8f18fa
start cpp tests of Relevant and MapPlus; move classes to header; bump version
