+++ src/comorbidMatMul.cpp
@@ -9,6 +9,8 @@
 #include "refactor.h"
 #include <string>
 #include <cstring>
+#include "relevant.h"
+#include "mapplus.h"
 
 /*
  MAKEFLAGS=-j16 R -e 'devtools::load_all(); icd9_comorbid_ahrq(ahrq_test_dat)'
@@ -17,40 +19,6 @@
  */
 
 using namespace Rcpp;
-
-class Relevant {
-public:
-  const List& src_map;
-  const CV relevant;
-  IHS hash;
-  CV keys;
-  Relevant(const List& map, const CV& codes) : src_map(map), relevant(findRelevant(codes)), hash(IHS(relevant).fill()), keys(hash.keys()) {}
-
-  CV findRelevant(const CV& codes) {
-    std::unordered_set<std::string> r;
-    r.reserve(100 * src_map.size());
-    IHS codeHash(codes);
-    codeHash.fill();
-    DEBUG_VEC(codes);
-    DEBUG_VEC(codeHash.keys());
-    // TODO codes may be factor, if so, return levels
-    for (CV cmb : src_map) {
-      for (auto code : cmb) {
-        // IndexHash contains causes compiler warning, so compare directly:
-        if ((int) codeHash.get_index(code) != NA_INTEGER) {
-          TRACE("Pushing back " << code);
-          //r.push_back(((String) code).get_cstring());
-          r.insert(((String) code).get_cstring());
-        }
-      }
-    }
-    return(wrap(r)); // or keep as STL container?
-  }
-
-  R_xlen_t size() { return relevant.size(); }
-}; // Relevant
-
-// TODO: make exportable version of get relevant? // [[Rcpp::export]]
 
 // # nocov start
 #ifdef ICD_DEBUG
@@ -104,11 +72,12 @@ void printCornerSparse(PtsSparse x) {
 // If the icd codes are character, we can construct the indexhash, then return a
 // new factor. This is basically what factorNosort does, but I need the hash
 // internals for lookups, so can't use it directly here.
-void buildVisitCodesSparseSimple(const RObject& visits,
-                                 const RObject& codes, // todo handle factor in parent function
-                                 Relevant& rh,
-                                 PtsSparse& visMat,
-                                 VecStr& visitIds // get this from sparse matrix at end, but needed?
+void buildVisitCodesSparseSimple(
+    const RObject& visits,
+    const RObject& codes, // todo handle factor in parent function
+    Relevant& rh,
+    PtsSparse& visMat, // output
+    VecStr& visitIds // output: can get this from sparse matrix at end? Needed?
 ) {
   TRACE("Starting build of visit matrix");
   assert(Rf_length(visits) == Rf_length(codes));
@@ -175,8 +144,8 @@ void buildVisitCodesSparseSimple(const RObject& visits,
   // now we have rows and columns, just make the triplets and insert.
   for (R_xlen_t i = 0; i != rows.size(); ++i) {
     if (IntegerVector::is_na(cols[i])) {
-     TRACE("dropping NA col value");
-    continue;
+      TRACE("dropping NA col value");
+      continue;
     }
     TRACE("adding triplet at R idx:" << rows[i] << ", " << cols[i]);
     visTriplets.push_back(Triplet(rows[i] - 1, cols[i] - 1, true));
@@ -190,43 +159,39 @@ void buildVisitCodesSparseSimple(const RObject& visits,
   PRINTCORNERSP(visMat);
 }
 
-class MapPlus {
-public:
-  MapPlus(const List& icd9Mapping, const Relevant& rh);
-  void buildMatrix();
-  List map; // consider ListOf<IntegerVector>
-  DenseMap mat;
-  R_xlen_t rows() { return mat.rows(); }
[…]
