### 2018-11-20 93542f5
unit tests, code prep for direct wide data processing
