+++ src/relevant.h
@@ -20,17 +20,20 @@ public:
     src_map(map),
     relevant(findRelevant(codes)),
     hash(IHS(relevant).fill()),
-    keys(hash.keys())
-    //  allCodesSet,
-    //  r()
-  {
-    //allCodesSet = US();
-    //r = US();
+    keys(hash.keys()) {
+  }
+  Relevant(const List& map, const List& data, CV code_fields) :
+    src_map(map),
+    relevant(findRelevant(data, code_fields)),
+    hash(IHS(relevant).fill()),
+    keys(hash.keys()) {
   }
   void buildCodeSetCV(const CV& codes);
   void buildCodeSet(const SEXP& codes);
+  CV findRelevant();
   CV findRelevant(const SEXP& codes);
   //CV findRelevant(const DataFrame& data, CV code_fields);
+  CV findRelevant(const List& data, CV code_fields);
   R_xlen_t size() { return relevant.size(); }
 }; // Relevant
 
