### 2018-11-20 9e4b362
separate headers for comorbidMatMul and MapPlus
