### 2018-11-20 9ca3251
removed deprecated icd9Field and visitId. fleshing out comorbidmatmul wide version
