+++ src/longToWide.cpp
@@ -16,6 +16,7 @@
 // along with icd. If not, see <http://www.gnu.org/licenses/>.
 
 #include "icd_types.h"                  // for VecStr, VecVecStr, CV
+#include "local.h"
 #include <string.h>                     // for strcmp
 #include <algorithm>                    // for find, fill
 #include <string>                       // for string, operator==
@@ -26,137 +27,90 @@
 #endif
 
 extern "C" {
-#include "cutil.h"                            // for getRListOrDfElement
+#include "cutil.h"
 }
 
 CV raggedToWide(const VecVecStr& ragged, int max_per_pt,
-                const VecStr &visitIds) {
+                const VecStr &ids) {
 #ifdef ICD_DEBUG_TRACE
-  Rcpp::Rcout << "visitIds = ";
-  // printIt(visitIds); // broken, not sure why.
+  Rcpp::Rcout << "ids = ";
+  // printIt(ids); // broken, not sure why.
 #endif
   VecStr::size_type distinct_visits = ragged.size();
-  CV out(distinct_visits * max_per_pt, NA_STRING); // optionally default empty strings? NA? User can do this for now.
+  // optionally default empty strings? NA? User can do this for now.
+  CV out(distinct_visits * max_per_pt, NA_STRING);
 #ifdef ICD_DEBUG
-  if (distinct_visits == 0) {
-    Rcpp::Rcout << "no visits. returning blank data\n";
-    return CV::create();
-  }
-  if (distinct_visits != visitIds.size()) {
-    Rcpp::Rcout << "visit and ragged sizes differ. visits = " << visitIds.size() <<
-      ", ragged size = " << distinct_visits << ": returning blank data\n";
-    return CV::create();
-  }
+  assert(distinct_visits > 0)
+    assert(distinct_visits == ids.size())
 #endif
-  for (VecVecStr::size_type row_it = 0; row_it < distinct_visits; ++row_it) {
-    const VecStr& this_row = ragged[row_it];
-    VecStr::size_type this_row_len = this_row.size();
-    for (VecStr::size_type col_it = 0; col_it < this_row_len; ++col_it) {
-      // write in row major format, but this means transpose needed later
-      VecVecStr::size_type out_idx = row_it + (distinct_visits * col_it);
-      out[out_idx] = this_row[col_it];
-    } // end inner for
-    Rcpp::checkUserInterrupt();
-  } // end outer for
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "writing dimensions\n";
-#endif
-  // set dimensions in reverse (row major for parallel step)
-  out.attr("dim") = Rcpp::Dimension(distinct_visits, max_per_pt);
-#ifdef ICD_DEBUG
-  Rcpp::Rcout << "writing labels\n";
-#endif
+    for (VecVecStr::size_type row_it = 0; row_it < distinct_visits; ++row_it) {
+      const VecStr& this_row = ragged[row_it];
+      VecStr::size_type this_row_len = this_row.size();
+      for (VecStr::size_type col_it = 0; col_it < this_row_len; ++col_it) {
+        // write in row major format, but this means transpose needed later
+        VecVecStr::size_type out_idx = row_it + (distinct_visits * col_it);
+        out[out_idx] = this_row[col_it];
+      } // end inner for
+      Rcpp::checkUserInterrupt();
+    } // end outer for
+    // set dimensions in reverse (row major for parallel step)
+    out.attr("dim") = Rcpp::Dimension(distinct_visits, max_per_pt);
   CV nonames;
-  rownames(out) = Rcpp::wrap(visitIds);
[…]
