### 2018-11-21 d913bc6
WIP moving away from IndexHash to unordered_map
