+++ src/test-cpp-matmul.cpp
@@ -6,6 +6,7 @@
 #include "appendMinor.h"
 #include "convert.h"
 #include "refactor.h"
+#include "comorbidMatMul.h"
 #include "relevant.h"
 #include "mapplus.h"
 
@@ -25,11 +26,35 @@ context("matmul cpp") {
   List map = List::create(_["X"] = v1,
                           _["Y"] = v2);
   List map_v1 = List::create(_["V"] = v1,
-                          _["W"] = v1);
+                             _["W"] = v1);
   List map_v12 = List::create(
     _["A"] = CharacterVector::create("V10", "V20"),
     _["B"] = CharacterVector::create("V12", "V40")
   );
+  DataFrame df1 = DataFrame::create(
+    _["id"] = CharacterVector::create("1001", "1002"),
+    _["dx0"] = v1,
+    _["dx1"] = v2,
+    _["stringsAsFactors"] = false);
+  DataFrame df2 = DataFrame::create(
+    _["id"] = CharacterVector::create("1001", "1002"),
+    _["dx0"] = v1,
+    _["dx1"] = v2,
+    _["dx2"] = CharacterVector::create(NA_STRING, NA_STRING),
+    _["dx2"] = CharacterVector::create(NA_STRING, "100"),
+    _["stringsAsFactors"] = false);
+  DataFrame dff1 = DataFrame::create(
+    _["id"] = CharacterVector::create("1001", "1002"),
+    _["dx0"] = v1,
+    _["dx1"] = v2,
+    _["stringsAsFactors"] = true);
+  DataFrame dff2 = DataFrame::create(
+    _["id"] = CharacterVector::create("1001", "1002"),
+    _["dx0"] = v1,
+    _["dx1"] = v2,
+    _["dx2"] = CharacterVector::create(NA_STRING, NA_STRING),
+    _["dx2"] = CharacterVector::create(NA_STRING, "100"),
+    _["stringsAsFactors"] = true);
   test_that("make a relevant object with character vectors") {
     Relevant r1(map, v1);
     expect_true(r1.relevant.size() == v1.size());
@@ -46,10 +71,9 @@ context("matmul cpp") {
   }
 
   test_that("make a relevant object with some cols of data frame") {
-    DataFrame df = DataFrame::create(_["dx0"] = v1, _["dx1"] = v2);
-    Relevant rdf0(map_v1, df, "dx0");
-    Relevant rdf1(map_v1, df, "dx1");
-    Relevant rdf01(map_v1, df, CV::create("dx0", "dx1"));
+    Relevant rdf0(map_v1, df1, "dx0");
+    Relevant rdf1(map_v1, df1, "dx1");
+    Relevant rdf01(map_v1, df1, CV::create("dx0", "dx1"));
     expect_true(rdf0.keys.size() == v1.size());
     expect_true(rdf1.keys.size() == 0);
   }
@@ -80,5 +104,24 @@ context("matmul cpp") {
     MapPlus m1(map, r1);
     MapPlus m2(map, r2);
   }
+
+  test_that("do a wide comorbid matrix multiplication") {
+    comorbidMatMulWide(df1, map, "id", CharacterVector::create("dx0", "dx1"));
+    //  comorbidMatMulWide(dff1, map, "id", CharacterVector::create("dx0", "dx1"));
+  }
+
+  test_that("use rel unordered map test lookup") {
+    Relevant r12(map, df1, CharacterVector::create("dx0", "dx1"));
+    RelMap::iterator got;
+    got = r12.rel.find("V10");
+    expect_true(got != r12.rel.end());
+    got = r12.rel.find("V11");
+    expect_true(got != r12.rel.end());
+    got = r12.rel.find("V12");
[…]
