+++ src/icd_types.h
@@ -40,6 +40,8 @@ typedef VecInt NewOutPt;
 typedef std::vector<NewOutPt> NewOut;
 typedef VecVecInt::iterator NewOutIt;
 typedef std::unordered_map<std::string, VecInt::size_type> VisLk;
+typedef std::unordered_map<std::string, int> RelMap;
+typedef std::pair<std::string, int> RelPair;
 typedef std::unordered_set<std::string> icd_set;
 typedef std::unordered_set<std::string> US;
 typedef Rcpp::CharacterVector CV;

[…]
