### 2018-11-21 829025d
WIP updating col name guessing to accept wide data [ci skip]
