+++ R/comorbid.R
@@ -210,9 +210,10 @@ icd9_comorbid <- function(x, map, visit_name = NULL, icd_name = NULL,
   assert_data_frame(x, min.cols = 2, col.names = "unique")
   assert_list(map, min.len = 1, names = "unique")
   assert(check_string(visit_name), check_null(visit_name))
-  assert(check_string(icd_name), check_null(icd_name))
+  #assert(check_string(icd_name), check_null(icd_name))
   visit_name <- get_visit_name(x, visit_name)
-  icd_name <- get_icd_name(x, icd_name)
+  #icd_name <- get_icd_name(x, icd_name)
+  stopifnot(is.character(icd_name) || is.null(icd_name))
   assert_string(visit_name)
   assert_string(icd_name)
   assert_flag(short_code)
