+++ R/util.R
@@ -155,7 +155,7 @@ str_pair_match <- function(string, pattern, pos, swap = FALSE, ...) {
     assert_integerish(pos, len = 2L, lower = 1L, any.missing = FALSE)
   res <- lapply(string,
                 function(x) unlist(
-                  regmatches(x = x, m = regexec(pattern = pattern, text = x, ...))
+                  regmatches(x, m = regexec(pattern = pattern, text = x, ...))
                 )[-1]
   )
   res <- res[vapply(res, function(x) length(x) != 0, logical(1))]
@@ -225,8 +225,11 @@ get_visit_name.matrix <- function(x, visit_name = NULL)
 #' @param icd_name usually \code{NULL} but if specified, will be checked it is
 #'   valid (i.e. a character vector of length one, which is indeed a name of one
 #'   of \code{x}'s columns) and returned unchanged
+#' @param multi If \code{TRUE}, allow multiple ICD field names to be returned.
 #' @keywords internal
-get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE, defined_codes = FALSE) {
+get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE,
+                         defined_codes = FALSE, multi = FALSE) {
+  # TODO: change this error message, no longer true!
   if (is.icd_wide_data(x))
     stop("Unable to infer the name of a single ICD field name from wide data, which has multiple ICD fields. ",
          "Comorbidity calculations require 'long' format data, 'wide' data should be converted to 'long' ",
@@ -234,45 +237,24 @@ get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE, defined_codes =
          "If the data is indeed 'long' format, remove the class 'icd_wide_data' and ",
          "use 'as.icd_long_data' to set the correct class. See '?icd_long_data' for help.")
   if (!is.null(icd_name)) {
-    assert_string(icd_name)
     stopifnot(icd_name %in% names(x))
     return(icd_name)
   }
-  guesses <- c("icd.?(9|10)", "icd.?(9|10).?Code", "icd",
-               "diagnos", "diag.?code", "diag", "i(9|10)", "code")
-  assert_data_frame(x, min.cols = 1, col.names = "named")
-  # if one column exactly has a class like icd9, then we're done.
-  cls <- lapply(x, class)
-  class_cols <- unlist(lapply(cls, function(cl) sum(cl %in% icd_version_classes)))
-  one_class_col <- which(class_cols >= 1)
-  if (length(unname(one_class_col)) == 1)
-    return(names(x)[one_class_col])
-  # zero or multiple columns matched classes, and icd_name not already set:
-  for (guess in guesses) {
-    guess_matched <- grep(guess, names(x), ignore.case = TRUE, value = TRUE)
-    if (length(guess_matched) == 1) {
-      icd_name <- guess_matched
-      break
-    } else if (length(guess_matched) > 1) {
-      warning(paste("more than one column name matched while guessing icd_name. Using first one.",
-                    "To specify exactly which column, use icd_name=\"myicdcol\" or set the class of",
-                    "the column, e.g. x$myicdcol <- as.icd10cm(x$myicdcol) .",
-                    "Matched column names are:", guess_matched))
-      icd_name <- guess_matched[1]
-    }
-  }
-  if (is.null(icd_name))
+  icd_name <- guess_icd_col_by_name(x, valid_codes = valid_codes,
+                        defined_codes = defined_codes)
+  if (is.null(icd_name)) {
[…]
