### 2018-11-21 f4c81b2
most of 'wide' conversion done.
