+++ src/relevant.h
@@ -27,9 +27,9 @@ public:
   // construct from a dataframe with given columns
   Relevant(const List& map,
            const List& data,
-           CV code_fields) :
+           CV code_names) :
     src_map(map),
-    relevant(findRelevant(data, code_fields)),
+    relevant(findRelevant(data, code_names)),
     rel(findRel(relevant)),
     hash(IHS(relevant).fill()),
     keys(hash.keys()) {
@@ -50,7 +50,7 @@ public:
   CV findRelevant();
   CV findRelevant(const SEXP& codes);
   //CV findRelevant(const DataFrame& data, CV code_fields);
-  CV findRelevant(const List& data, CV code_fields);
+  CV findRelevant(const List& data, CV code_names);
   RelMap findRel(const CharacterVector x);
 }; // Relevant
 
