### 2018-11-21 81d5b4b
WIP working out some refactor corner cases [ci skip]
