+++ src/local.h
@@ -26,7 +26,7 @@ extern "C" {
 #include "cutil.h"
 #include <cstdlib>
 }
-#define ICD_DEBUG
+// #define ICD_DEBUG
 // #define ICD_DEBUG_TRACE
 // #define ICD_DEBUG_UTIL
 // #define ICD_DEBUG_SETUP
