### 2018-11-22 e053546
added failing test cases for NA in refactoring levels [ci skip]
