+++ tests/testthat/test-refactor.R
@@ -103,76 +103,228 @@ test_that("new factor has empty levels when necessary", {
     refactor(f, levels = NA, na.rm = FALSE, exclude_na = FALSE),
     factor(NA, exclude = NULL))
   for (narm in c(TRUE, FALSE)) {
-    refactor(NA, NA, na.rm = narm)
-    refactor(factor(NA), NA, na.rm = narm)
-    refactor(factor(NA), "a", na.rm = narm)
-    refactor(factor(NA, levels = NA), c("a", NA), na.rm = narm)
-    refactor(factor(NA), c("a", NA), na.rm = narm)
-    refactor(factor("a"), c("a", NA), na.rm = narm)
-    refactor(factor("a"), NA, na.rm = narm)
+    expect_error(info = paste("narm = ", narm), refactor(NA, NA, na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor(NA), NA, na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor(NA), "a", na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor(NA, levels = NA), c("a", NA), na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor(NA), c("a", NA), na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor("a"), c("a", NA), na.rm = narm))
+    expect_error(regexp = NA, info = paste("narm = ", narm),
+                 refactor(factor("a"), NA, na.rm = narm))
   }
 })
 
 test_that("NA anywhere in middle of input levels ok", {
   # todo: duplicate target levels should fail or drop silenetly?
   v <- c("a", "b", NA)
-  f_perms <- expand.grid(v, v, v, length(v))
+  f_perms <- unname(expand.grid(v, v, v))
   l_perms <- unique(apply(f_perms, 1, unique))
   for (i in nrow(f_perms)) { # orig values
     for (ii in seq_along(l_perms)) { # new levels
       for (iii in seq_along(l_perms)) { # old levels
-        w <- factor(f_perms[i, ], levels = l_perms[[iii]], exclude = NULL)
-        l <- l_perms[[ii]]
-        expect_error(regexp = NA,
-                     refactor(w, l, na.rm = TRUE, exclude_na = TRUE),
-                     info = paste("w = ", paste(w, collapse = ", "),
-                                  "; l = ", paste(l, collapse = ", ")))
-        expect_error(regexp = NA,
-                     refactor(w, l, na.rm = FALSE, exclude_na = TRUE),
-                     info = paste("w = ", paste(w, collapse = ", "),
-                                  "; l = ", paste(l, collapse = ", ")))
-        expect_error(regexp = NA,
[…]
