+++ R/hcc.R
@@ -78,7 +78,7 @@ comorbid_hcc_worker <- function(x,
                                 icd_name) {
   assert_data_frame(x, min.cols = 3, col.names = "unique")
   assert(check_string(visit_name), check_null(visit_name))
-  assert(check_string(icd_name), check_null(icd_name))
+  assert(check_character(icd_name), check_null(icd_name))
   visit_name <- get_visit_name(x, visit_name)
   icd_name <- get_icd_name(x, icd_name)
   assert_string(date_name)
