+++ src/refactor.cpp
@@ -24,40 +24,37 @@ IntegerVector factorNoSort(const CharacterVector& x,
 
 //' @title Re-generate a factor with new levels, without doing string matching
 //' @description This is called by an R wrapper. There is an `na.rm` version,
-//' too.
+//'   too. Some work simply to mirror behavior of `base::factor`, e.g. when a
+//'   level is not available, but NA level is available, NA is inserted into the
+//'   integer vector, not an index to the NA level.
 //' @md
 //' @keywords internal manip
 // [[Rcpp::export(refactor_worker)]]
 Rcpp::IntegerVector refactor(const IntegerVector& x, const CV& new_levels,
                              bool exclude_na) {
-  TRACE("Refactoring");
+  TRACE("Refactoring, keeping NAs");
   bool isValid = factorIsValid(x);
 #ifdef ICD_DEBUG
   assert(isValid);
 #endif
   if (!isValid) Rcpp::stop("input is not a valid factor");
-  IntegerVector f(x.size()); // too many if we are dropping NA values.
+  IntegerVector f(x.size());
   CharacterVector lx = x.attr("levels");
   DEBUG_VEC(new_levels);
-  CV no_na_lx;
   CV no_na_new_levels;
-  //bool any_na_lx = false;
-  LogicalVector which_na_new_levels = is_na(new_levels);
-  // TODO: get the position of the NA in new_levels for insertion later if
-  // needed.
-  LogicalVector which_na_old_levels = is_na(lx);
+  LogicalVector is_na_old_levels = is_na(lx);
+  LogicalVector is_na_new_levels = is_na(new_levels);
+  DEBUG_VEC(is_na_old_levels);
+  DEBUG_VEC(is_na_new_levels);
+  if (sum(is_na_new_levels) > 1)
+    Rcpp::stop("New levels have multiple NA values, which is disallowed.");
   if (exclude_na) {
-    DEBUG("Dropping NA in input factor levels");
-    no_na_lx = lx[!which_na_old_levels];
     DEBUG("Dropping NA in input levels");
-    no_na_new_levels = new_levels[!which_na_new_levels];
+    no_na_new_levels = new_levels[!is_na_new_levels];
   } else {
-    DEBUG("Not excluding NA from input factor levels");
-    no_na_lx = lx;
     DEBUG("Not dropping NA from target levels");
     no_na_new_levels = new_levels;
   }
-  DEBUG_VEC(no_na_lx);
   DEBUG_VEC(no_na_new_levels);
   if (no_na_new_levels.size() == 0) {
     DEBUG("no_na_new_levels is empty");
@@ -67,73 +64,55 @@ Rcpp::IntegerVector refactor(const IntegerVector& x, const CV& new_levels,
     return f;
   }
   // Rcpp::match is 1-indexed
-  IntegerVector new_level_old_idx = Rcpp::match(no_na_lx, no_na_new_levels);
+  IntegerVector new_level_old_idx = Rcpp::match(lx, no_na_new_levels);
   DEBUG_VEC(new_level_old_idx);
   R_xlen_t fsz = x.size();
   DEBUG("fsz = " << fsz);
   LogicalVector matched_na_level(fsz, false);
-  R_xlen_t fi = 0;
   R_xlen_t i;
+  bool new_na_level = is_true(any(is_na_new_levels));
+  bool old_na_level = is_true(any(is_na_old_levels));
+  int new_na_level_idx = which_max(is_na_new_levels);
+  int old_na_level_idx = which_max(is_na_old_levels);
   for (i = 0; i < fsz; ++i) {
-    TRACE("refactor considering i: " << i << ", x[i]: " << x[i] << ", "
-                                     << "fi: " << fi);
[…]
