### 2018-11-24 7004d89
passes tests after refactor work
