+++ man/refactor_worker.Rd
@@ -11,7 +11,9 @@ refactor_narm_worker(x, new_levels)
 }
 \description{
 This is called by an R wrapper. There is an \code{na.rm} version,
-too.
+too. Some work simply to mirror behavior of \code{base::factor}, e.g. when a
+level is not available, but NA level is available, NA is inserted into the
+integer vector, not an index to the NA level.
 }
 \section{Functions}{
 \itemize{
