### 2018-11-24 9fa84ba
minor tweaks for check
