+++ R/convert.R
@@ -157,10 +157,6 @@ wide_to_long <- function(x,
 #' @param min_width, single integer, if specified, writes out this many columns
 #'   even if no patients have that many codes. Must be greater than or equal to
 #'   the maximum number of codes per patient.
-#' @param return_df single logical value, if \code{TRUE}, return a data frame
-#'   with a field for the visit_name. This may be more convenient, but the
-#'   default of \code{FALSE} gives the more natural return data of a matrix with
-#'   row names being the visit IDs from \code{visit_name}s.
 #' @examples
 #'   longdf <- data.frame(visit_name = c("a", "b", "b", "c"),
 #'     icd9 = c("441", "4424", "443", "441"))
