+++ NAMESPACE
@@ -357,6 +357,7 @@ importFrom(magrittr,extract2)
 importFrom(magrittr,set_names)
 importFrom(methods,setLoadAction)
 importFrom(stats,aggregate)
+importFrom(stats,reshape)
 importFrom(stats,setNames)
 importFrom(utils,head)
 importFrom(utils,read.csv)
