### 2018-11-24 43b4f08
docs, mostly vignettes. dropped 'unique_ids' argument, default preclean to FALSE, since most data does not need this.
