+++ vignettes/PCCC.Rmd
@@ -38,24 +38,17 @@ The PCCC comorbidities may be derived from both diagnostic and procedure codes. 
 ```{r headdat}
 head(dat)
 ```
-The data is in 'wide' format. Currently `icd` requires it be in 'long' format, which can be achieved as follows:
-```{r widelong}
-dx <- wide_to_long(dat[1:3])
-pcs <- wide_to_long(dat[c(1, 4:5)], icd_name = "pcs", icd_regex = "icd_pcs")
-head(dx)
-head(pcs)
-```
-Now the PCCC comorbidity classes can be computed and summarized in a plot:
+The data is in 'wide' format. `icd` no longer needs the data to be in 'long' format. The PCCC comorbidity classes can be computed directly from the source data and summarized in a plot. Here we calculate the PCCC from ICD diagnostic codes and ICD procedure codes separately, and logically combine them to get the result.
 ```{r calc}
-pccc_dx <- comorbid_pccc_dx(dx)
-pccc_pcs <- icd10_comorbid_pccc_pcs(pcs, 
-                                    visit_name = "id", icd_name = "pcs")
+pccc_dx <- comorbid_pccc_dx(dat)
+pccc_pcs <- icd10_comorbid_pccc_pcs(dat,
+                                    icd_name = c("icd_pcs1", "icd_pcs2"))
 res <- pccc_dx | pccc_pcs
 ```
 
 ```{r pcccbarplot, fig.asp=0.5, echo=FALSE}
-barplot(sort(colSums(res), decreasing = TRUE), 
-        ylim = c(0, 70),
+barplot(sort(colSums(res), decreasing = TRUE),
+        ylim = c(0, signif(max(colSums(res)), 1)),
         ylab = "count", las = 2)
 ```
 
