### 2018-11-24 b4709e0
remove some wide/long conversions
