+++ tests/testthat/test-condense.R
@@ -241,12 +241,7 @@ test_that("condense single major and its children", {
   expect_equal(explain_code.icd9(children.icd9("391", short_code = TRUE, defined = TRUE)), rheum_fever)
 })
 
-vermont_dx %>%
-  wide_to_long  %>%
-  extract2("icd_code")  %>%
-  sort_icd.icd9(short_code = TRUE) %>%
-  unique %>%
-  head(10) -> vdat
+vdat <- unique(vermont_dx$DX1)[1:10]
 
 test_that("condense a factor of codes instead of character vector", {
   # this is not a condensable list
