### 2018-11-24 bb74f78
fix strange test and trivial compiler warning
