+++ tests/testthat/test-comorbid-maps-icd9.R
@@ -433,7 +433,7 @@ test_that("ICD-9 codes from manually specified Elixhauser mapping exist", {
 
 test_that("github #34 - short and long custom map give different results", {
   mydf <- data.frame(visit_id = c("a", "b", "b", "c"),
-                     icd9 = c("1", "010", "10", "20"))
+                     icd9 = c("001", "010", "010", "020"))
   mymaps <- list(jack = c("1", "2", "3"), alf = c("010", "20"))
   mymapd <- lapply(mymaps, icd:::short_to_decimal.icd9)
   expect_identical(
