### 2018-11-24 3acbaf7
tidy up another minor test
