### 2018-11-25 e97a083
remove old code, now that wide calculation can do long and wide data. PCCC benchmark code much tidier, and now much faster for 'icd' with wide NEDS-type input.
