### 2018-11-25 01617d6
clean up neds generation to pass cran check
