### 2018-11-25 d417e3e
drop C++ long to wide conversion in favor of simpler base R conversion.
