### 2018-11-25 8875443
allow disable of factor validation - saves significant time for known-good big input data. default currently to do validation.
