+++ src/comorbidMatMul.h
@@ -25,6 +25,7 @@ void buildVisitCodesSparseWide(
         const DataFrame& data,
         const std::string id_name,
         const CV code_names,
+        const bool validate,
         Relevant& rh,
         // output
         PtsSparse& visMat,
@@ -35,4 +36,5 @@ LogicalMatrix comorbidMatMulWide(
         const DataFrame& data,
         const List& map,
         const std::string id_name,
-        const CV code_name);
+        const CV code_name,
+        const bool validate = false);
