### 2018-11-25 0809dd0
register changed functions, PCCC benchmarking tweaks
