+++ benchmarks/bench-pccc.R
@@ -2,7 +2,11 @@
 # ln -sf ~/.R/Makevars.valgrind Makevars
 # ~/icd/tools/install_full.sh
 # R --vanilla -d "valgrind --tool=callgrind --instr-atstart=no"
+
+# Run with less garbage collection. e.g.
 #
+# R_GC_MEM_GROW=3 R -e 'devtools::load_all(); pts <- icd:::generate_neds_pts(1e7, icd10 = TRUE); gprofiler::profile(icd::comorbid_pccc_dx(pts), out_format="pdf", out_filename = "/tmp/out7-icd10.pdf")'
+
 library(icd)
 library(dplyr)
 library(magrittr)
@@ -10,16 +14,23 @@ library(R.cache)
 library(pccc)
 # simulating NEDS (icd-9) db for size (also need to simulate width, which is 15+4 columns!)
 n_neds <- 28584301L
-if (!exists("divisor"))
-  divisor <- as.integer(readline(prompt = "Enter a divisor: "))
+if (!exists("divisor")) {
+  divisor <- if (interactive())
+    as.integer(readline(prompt = "Enter a divisor: "))
+  else
+    10
+}
 n <- n_neds / divisor
 key = list("bench-pccc-wide", n, icd10 = FALSE, ncol = 20L)
 dat_wide_str <- R.cache::loadCache(key)
 if (is.null(dat_wide_str)) {
   dat_wide_str <-
-    generate_neds_pts(n = n, ncol = 20L, icd10 = FALSE, verbose = TRUE)
+    icd:::generate_neds_pts(n = n, ncol = 20L, icd10 = FALSE, verbose = TRUE)
   R.cache::saveCache(dat_wide_str, key)
 }
+# See https://stat.ethz.ch/R-manual/R-devel/library/base/html/Memory.html for
+# possible ways to reduce frequency or agressiveness of the garbage collection.
+
 message("Benchmarking for ", n, " rows of simulated NEDS data.")
 message("icd:")
 icdtm2 <- proc.time()
