+++ src/registration.c
@@ -45,7 +45,7 @@ extern SEXP _icd_icd9ShortToPartsCpp(SEXP, SEXP);
 extern SEXP _icd_inFast(SEXP, SEXP);
 extern SEXP _icd_matchFast(SEXP, SEXP);
 extern SEXP _icd_refactor(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _icd_refactor_narm(SEXP, SEXP);
+extern SEXP _icd_refactor_narm(SEXP, SEXP, SEXP);
 extern SEXP _icd_setDecimalDiag(SEXP, SEXP);
 extern SEXP _icd_setShortDiag(SEXP, SEXP);
 extern SEXP _icd_simplifyMapLexicographic(SEXP, SEXP);
@@ -94,7 +94,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_icd_inFast",                             (DL_FUNC) &_icd_inFast,                             2},
     {"_icd_matchFast",                          (DL_FUNC) &_icd_matchFast,                          2},
     {"_icd_refactor",                           (DL_FUNC) &_icd_refactor,                           4},
-    {"_icd_refactor_narm",                      (DL_FUNC) &_icd_refactor_narm,                      2},
+    {"_icd_refactor_narm",                      (DL_FUNC) &_icd_refactor_narm,                      3},
     {"_icd_setDecimalDiag",                     (DL_FUNC) &_icd_setDecimalDiag,                     2},
     {"_icd_setShortDiag",                       (DL_FUNC) &_icd_setShortDiag,                       2},
     {"_icd_simplifyMapLexicographic",           (DL_FUNC) &_icd_simplifyMapLexicographic,           2},
