### 2018-11-26 a55a5a6
add lifecycle badge and drop weekly downloads badge [ci skip]
