+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.2.1
-pkgdown: 1.1.0
-pkgdown_sha: ~
+pkgdown: 1.2.0.9000
+pkgdown_sha: c980e84a573614087c6e151d5e6e19a22748b761
 articles:
   CMS-HCC: CMS-HCC.html
   ICD-10: ICD-10.html
