### 2018-11-27 26839ab
tidy up internal documentation for pkgdown [ci skip]
