+++ man/print.comorbidity_map.Rd
@@ -30,3 +30,4 @@ print.list(icd9_map_ahrq)
 print(list(icd9_map_ahrq))
 }
 }
+\keyword{internal}
