### 2018-11-27 e49e051
travis and pkgdown build tool tweak
