+++ tools/publish.sh
@@ -15,7 +15,7 @@ function finish {
 trap finish EXIT
 
 #Rscript --vanilla --default-packages=jwutil -e 'jwutil::reqinst("pkgdown"); pkgdown::build_site()'
-Rscript --vanilla -e 'jwutil::reqinst("pkgdown"); pkgdown::build_site()'
+Rscript --vanilla -e 'jwutil::reqinst("pkgdown"); devtools::load_all(); pkgdown::build_site()'
 mkdir -p "$GH_PAGES"
 cp -r docs/* "$GH_PAGES" 
 git -C "$GH_PAGES" status
