### 2018-11-27 bee8250
weaning from magrittr
