+++ R/parse-ahrq-ccs.R
@@ -77,12 +77,7 @@ icd9_parse_ahrq_ccs <- function(single = TRUE, save_data = FALSE,
   assert_flag(save_data)
   assert_flag(offline)
   ahrq_ccs <- icd9_fetch_ahrq_ccs(single = single, offline = offline)
-  clean_icd9 <- function(x){
-    x %>%
-      trimws() %>%
-      as.icd9() %>%
-      as.short_diag()
-  }
+  clean_icd9 <- function(x) as.short_diag(as.icd9(trimws(x)))
   resort_lvls <- function(x) {
     # Function to reorder numbers of CCS
     lvls_names <- names(x)
@@ -104,22 +99,20 @@ icd9_parse_ahrq_ccs <- function(single = TRUE, save_data = FALSE,
     if (lvls_has_empty) lvls <- c(lvls, " ")
     x[lvls]
   }
+  rsrt <- function(x) comorbidity_map(resort_lvls(x))
   if (!single) {
-    ahrq_df <- read.csv(ahrq_ccs$file_path, quote = "'\"", colClasses = "character")
-    lvl1 <- tapply(ahrq_df[["ICD.9.CM.CODE"]], ahrq_df[["CCS.LVL.1"]], clean_icd9) %>%
-      resort_lvls %>%
-      comorbidity_map
-    lvl2 <- tapply(ahrq_df[["ICD.9.CM.CODE"]], ahrq_df[["CCS.LVL.2"]], clean_icd9) %>%
-      resort_lvls %>%
-      comorbidity_map
-    lvl3 <- tapply(ahrq_df[["ICD.9.CM.CODE"]], ahrq_df[["CCS.LVL.3"]], clean_icd9) %>%
-      resort_lvls %>%
-      comorbidity_map
-    lvl4 <- tapply(ahrq_df[["ICD.9.CM.CODE"]], ahrq_df[["CCS.LVL.4"]], clean_icd9) %>%
-      resort_lvls %>%
-      comorbidity_map
-    icd9_map_multi_ccs <- list(lvl1 = lvl1, lvl2 = lvl2, lvl3 = lvl3, lvl4 = lvl4)
-
+    ahrq_df <- read.csv(ahrq_ccs$file_path, quote = "'\"",
+                        colClasses = "character")
+    lvl1 <- rsrt(tapply(ahrq_df[["ICD.9.CM.CODE"]],
+                        ahrq_df[["CCS.LVL.1"]], clean_icd9))
+    lvl2 <- rsrt(tapply(ahrq_df[["ICD.9.CM.CODE"]],
+                        ahrq_df[["CCS.LVL.2"]], clean_icd9))
+    lvl3 <- rsrt(tapply(ahrq_df[["ICD.9.CM.CODE"]],
+                        ahrq_df[["CCS.LVL.3"]], clean_icd9))
+    lvl4 <- rsrt(tapply(ahrq_df[["ICD.9.CM.CODE"]],
+                        ahrq_df[["CCS.LVL.4"]], clean_icd9))
+    icd9_map_multi_ccs <- list(lvl1 = lvl1, lvl2 = lvl2,
+                               lvl3 = lvl3, lvl4 = lvl4)
     if (save_data)
       save_in_data_dir("icd9_map_multi_ccs")
     out <- icd9_map_multi_ccs
@@ -127,13 +120,10 @@ icd9_parse_ahrq_ccs <- function(single = TRUE, save_data = FALSE,
     ahrq_df <- read.csv(ahrq_ccs$file_path, quote = "'\"",
                         colClasses = "character", skip = 1)
     icd9_map_single_ccs <-
[…]
