### 2018-11-28 a99cea1
add recommended repostatus badge, tweak ignores [ci skip]
