### 2018-11-28 b3199a8
drop magrittr as an Import dependency, still used in vignettes; much linting, mostly for line length closer to 80
