+++ tests/testthat/test-is.R
@@ -57,9 +57,12 @@ test_that("'is' works for factors", {
 
 test_that("'is' mixed values, factors and vectors", {
   v <- c("E100.1", "V234", "12", "V34.2", "61523", "10.2", "E9991", " V45XX ")
-  expect_equal(icd9_is_n(v), c(FALSE, FALSE, TRUE, FALSE, TRUE, TRUE, FALSE, FALSE))
-  expect_equal(icd9_is_v(v), c(FALSE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE))
-  expect_equal(icd9_is_e(v), c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE))
+  expect_equal(icd9_is_n(v),
+               c(FALSE, FALSE, TRUE, FALSE, TRUE, TRUE, FALSE, FALSE))
+  expect_equal(icd9_is_v(v),
+               c(FALSE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE))
+  expect_equal(icd9_is_e(v),
+               c(TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE))
 })
 
 test_that("ICD-10 majors identified", {
