### 2018-11-28 80c2063
unbreak some linting errors
