+++ R/manip.R
@@ -29,8 +29,8 @@ icd9_extract_alpha_numeric <- function(x) {
   # alpha part in first column, and numeric part in the second
   t(
     vapply(
-      as_char_no_warn(
-        str_match_all(x, pattern = "([VvEe]?)([[:digit:].]+)")),
+      str_match_all(as_char_no_warn(x),
+                    pattern = "([VvEe]?)([[:digit:].]+)"),
       FUN = function(y) matrix(data = y[2:3], nrow = 1, ncol = 2),
       FUN.VALUE = c(NA_character_, NA_character_)))
 }
