### 2018-11-28 583153f
code coverage found a minor bug
