### 2018-11-28 6abeefd
cover refactor corner cases, better code coverage
