### 2018-11-28 a358b64
test and fix finding bad elements in comorbidity map
