### 2018-11-28 758fb54
permit integer id field, drop redundant code
