+++ tests/testthat/test-comorbid-calcs-icd10.R
@@ -120,7 +120,7 @@ test_that("cmb from a ICD-10, no infinite recursion with Elix", {
 })
 
 test_that("ICD-10 map reduction is sane", {
-  uranium_short_codes <- as_char_or_levels(decimal_to_short.icd10(uranium_pathology$icd10))
+  uranium_short_codes <- as_char_no_warn(decimal_to_short.icd10(uranium_pathology$icd10))
   red_map <- simplify_map_lex(uranium_short_codes, icd10_map_ahrq)
   # the map should not have its original contents, but only those codes which
   # were in the pt data. Converse is not true, as some patient codes do not fall
