+++ tests/testthat/test-guess.R
@@ -99,3 +99,8 @@ test_that("guess and update version", {
 test_that("guess uranium codes are long", {
   expect_false(guess_short(uranium_pathology$icd10))
 })
+
+test_that("we should allow integer ids which get passed to guessing code", {
+  pts <- data.frame(id = 1:20, weirdname = rep("0932", 20))
+  expect_no_error(icd9_comorbid_ahrq(pts))
+})
