+++ man/as_char_no_warn.Rd
@@ -2,12 +2,9 @@
 % Please edit documentation in R/as_char_no_warn.R
 \name{as_char_no_warn}
 \alias{as_char_no_warn}
-\alias{as_char_or_levels}
 \title{convert to character vector without warning}
 \usage{
 as_char_no_warn(x)
-
-as_char_or_levels(x)
 }
 \arguments{
 \item{x}{vector, typically numeric or a factor}
@@ -18,10 +15,4 @@ character vector
 \description{
 convert to character vector without warning
 }
-\section{Functions}{
-\itemize{
-\item \code{as_char_or_levels}: If the argument is a factor, return the levels
-instead
-}}
-
 \keyword{internal}
