### 2018-11-28 a22f494
cover more sort code with tests
