+++ tests/testthat/test-sort.R
@@ -32,8 +32,16 @@ test_that("sorting of icd9 object", {
 })
 
 test_that("sorting of icd10 object", {
-  j <- icd10cm(c("Bad disease" = "I119", "Another bad disease" = "I110"))
+  a <- c("Bad disease" = "I119", "Another bad disease" = "I110")
+  j <- icd10cm(a)
   k <- sort_icd(j)
   expect_identical(j[2], k[1])
   expect_identical(j[1], k[2])
+  expect_identical(sort_icd(a), unclass(k))
+})
+
+test_that("warn if NA when ordering ICD-9 codes", {
+  expect_warning(icd9_order_short(c("a", NA)))
+  expect_identical(
+    expect_warning(icd9_order_short(NA)), character())
 })
