### 2018-11-29 ec88a48
simplifying code and documentation a little, fix a broken test
