+++ NAMESPACE
@@ -61,8 +61,6 @@ S3method(get_valid,character)
 S3method(get_valid,icd10)
 S3method(get_valid,icd10cm)
 S3method(get_valid,icd9)
-S3method(get_visit_name,data.frame)
-S3method(get_visit_name,matrix)
 S3method(guess_version,character)
 S3method(guess_version,data.frame)
 S3method(guess_version,factor)
