### 2018-12-01 29928c2
lint away sapply
