### 2018-12-01 d222180
vapply correctly for a data-build test
