+++ R/parse.R
@@ -68,8 +68,7 @@ update_everything <- function() {
 # quick sanity checks - full tests of x in test-parse.R
 icd9cm_hierarchy_sanity <- function(x) {
   stopifnot(all(is_valid.icd9(x[["code"]], short_code = TRUE)))
-  if (!any(vapply(x, is.na, logical(1))))
-    return()
+  if (!any(vapply(x, is.na, logical(nrow(x))))) return()
   print(colSums(vapply(x, is.na, logical(1))))
   print(x[which(is.na(x$major)), ])
   print(x[which(is.na(x$three_digit)), ])
