### 2018-12-16 cbd87ad
readme references the vignettes by links, updated CONTRIBUTING. repostatus active badge
