+++ README.Rmd
@@ -29,8 +29,8 @@ along with icd. If not, see <http:#www.gnu.org/licenses/>.
 -->
 ```{r, echo = FALSE}
 suppressPackageStartupMessages({
-  library(icd)
-  library(magrittr, wwarn.conflicts = FALSE)
+  library(icd, warn.conflicts = FALSE)
+  library(magrittr, warn.conflicts = FALSE)
 })
 knitr::opts_chunk$set(
   collapse = TRUE,
@@ -99,8 +99,31 @@ comorbid_charlson(patient_data)
 
 # or go straight to the Charlson scores:
 charlson(patient_data)
+```
 
-# for more examples, see this and other vignettes
+## How to get help
+
+Look at the help files for details and examples of almost every function in this package. There are several vignettes showing the main features:
+
+ - [Introduction](introduction.html)
+ - [Charlson scores](charlson-scores.html)
+ - [Examples using ICD-10 codes](ICD-10.html)
+ - [CMS Hierarchical Condition Codes (HCC)](CMS-HCC.html)
+ - [Pediatric Complex Chronic Conditions (PCCC)](PCCC.html)
+ - [Working with ICD code ranges](ranges.html)
+ - [Comparing comorbidity maps](compare-maps.html)
+ - [Paper detailing efficient matrix method of comorbidities](efficiency.html)
+
+Many users have emailed me directly for help, and I'll do what I can, but it is often better to examine or add to the list of [issues](https://github.com/jackwasey/icd) so we can help each other. Advanced users may look at the [source code](https://github.com/jackwasey/icd), particularly the extensive test suite which exercises all the key functions.
+
+``` r
+?comorbid
+?comorbid_hcc
+?explain_code
+?is_valid
+
+# first show the list
+vignette(package = "icd")
 vignette("introduction", package = "icd")
 ```
 
@@ -120,35 +143,10 @@ ICD-9 codes are not numbers, and great care is needed when matching individual c
 
 ICD-10 has a somewhat simpler format, with consistent use of a letter, then two alphanumeric characters. However, especially for ICD-10-CM, there are a multitude of qualifiers, e.g. specifying recurrence, laterality, which vastly increase the number of possible codes. This package recognizes validity of codes by syntax alone, or whether the codes appear in a canonical list. The current ICD-10-CM master list is the 2016 set. There is no capability of converting between ICD-9 and ICD-10, but comorbidities can be generated from older ICD-9 codes and newer ICD-10 codes in parallel, and the comorbidities can then be compared.
 
-## How to get help
-
-Look at the help files for details and examples of almost every function in this package. There are several vignettes showing the main features. Many users have emailed me directly for help, and I'll do what I can, but it is often better to examine or add to the list of [issues](https://github.com/jackwasey/icd) so we can help each other. Advanced users may look at the [source code](https://github.com/jackwasey/icd), particularly the extensive test suite which exercises all the key functions.
-
-```{r how to get further examples, echo = TRUE, eval = FALSE}
-?comorbid
-?comorbid_hcc
-?explain_code
-?is_valid
-
-# first show the list
-vignette(package = "icd")
-vignette("pccc", package = "icd")
-```
-
 ## Development version
 
 The latest version is available in [github icd](https://github.com/jackwasey/icd), and can be installed with:
-```{r show install, eval = FALSE, echo = TRUE}
-    install.packages("devtools")
-    devtools::install_github("jackwasey/icd")
-```
-
-## Contributing and Building
-
[…]
