### 2018-12-27 83d34aa
try to avoid dependency on package being loaded, not just having namespace.
