### 2018-12-30 087e6b8
no need for special package registration, Rcpp can make it happen
