+++ src/registration.c
@@ -1,114 +1,0 @@
-#include <R.h>
-#include <Rinternals.h>
-#include <stdlib.h> // for NULL
-#include <R_ext/Rdynload.h>
-
-/* FIXME: 
-   Check these declarations against the C/Fortran source code.
-*/
-
-/* .Call calls */
-extern SEXP _icd_categorize_rcpp();
-extern SEXP _icd_comorbidMatMulWide(SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _icd_factorIsValid(SEXP);
-extern SEXP _icd_factorNoSort(SEXP, SEXP, SEXP);
-extern SEXP _icd_fastIntToStringRcpp(SEXP);
-extern SEXP _icd_guessShortCompleteCpp(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _icd_guessShortPlusFactorCpp(SEXP, SEXP);
-extern SEXP _icd_icd10cmChildrenDefined(SEXP, SEXP, SEXP);
-extern SEXP _icd_icd10DecimalToPartsCpp(SEXP, SEXP);
-extern SEXP _icd_icd10ShortToPartsCpp(SEXP, SEXP);
-extern SEXP _icd_icd9_is_e_cpp(SEXP);
-extern SEXP _icd_icd9_is_n_cpp(SEXP);
-extern SEXP _icd_icd9_is_v_cpp(SEXP);
-extern SEXP _icd_icd9AddLeadingZeroes(SEXP, SEXP);
-extern SEXP _icd_icd9AddLeadingZeroesMajor(SEXP);
-extern SEXP _icd_icd9AddLeadingZeroesMajorSingle(SEXP);
-extern SEXP _icd_icd9AddLeadingZeroesMajorSingleStd(SEXP);
-extern SEXP _icd_icd9ChildrenCpp(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _icd_icd9ChildrenDecimalCpp(SEXP, SEXP, SEXP);
-extern SEXP _icd_icd9ChildrenShort(SEXP, SEXP, SEXP);
-extern SEXP _icd_icd9ChildrenShortUnordered(SEXP, SEXP, SEXP);
-extern SEXP _icd_icd9DecimalToPartsCpp(SEXP, SEXP);
-extern SEXP _icd_icd9DecimalToShort(SEXP);
-extern SEXP _icd_icd9ExpandMinor(SEXP, SEXP);
-extern SEXP _icd_icd9GetMajor(SEXP, SEXP);
-extern SEXP _icd_icd9MajMinToCode(SEXP, SEXP, SEXP);
-extern SEXP _icd_icd9MajMinToDecimal(SEXP, SEXP);
-extern SEXP _icd_icd9MajMinToParts(SEXP, SEXP);
-extern SEXP _icd_icd9MajMinToShort(SEXP, SEXP);
-extern SEXP _icd_icd9OrderCpp(SEXP);
-extern SEXP _icd_icd9PartsToDecimal(SEXP);
-extern SEXP _icd_icd9PartsToShort(SEXP);
-extern SEXP _icd_icd9ShortToDecimal(SEXP);
-extern SEXP _icd_icd9ShortToPartsCpp(SEXP, SEXP);
-extern SEXP _icd_inFast(SEXP, SEXP);
-extern SEXP _icd_matchFast(SEXP, SEXP);
-extern SEXP _icd_refactor(SEXP, SEXP, SEXP, SEXP);
-extern SEXP _icd_refactor_narm(SEXP, SEXP, SEXP);
-extern SEXP _icd_setDecimalDiag(SEXP, SEXP);
-extern SEXP _icd_setShortDiag(SEXP, SEXP);
[…]
