+++ src/RcppExports.cpp
@@ -552,3 +552,61 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+
+RcppExport SEXP run_testthat_tests();
+
+static const R_CallMethodDef CallEntries[] = {
+    {"_icd_icd9MajMinToCode", (DL_FUNC) &_icd_icd9MajMinToCode, 3},
+    {"_icd_icd9MajMinToShort", (DL_FUNC) &_icd_icd9MajMinToShort, 2},
+    {"_icd_icd9MajMinToDecimal", (DL_FUNC) &_icd_icd9MajMinToDecimal, 2},
+    {"_icd_setDecimalDiag", (DL_FUNC) &_icd_setDecimalDiag, 2},
+    {"_icd_setShortDiag", (DL_FUNC) &_icd_setShortDiag, 2},
+    {"_icd_categorize_rcpp", (DL_FUNC) &_icd_categorize_rcpp, 0},
+    {"_icd_icd10cmChildrenDefined", (DL_FUNC) &_icd_icd10cmChildrenDefined, 3},
+    {"_icd_simplifyMapLexicographic", (DL_FUNC) &_icd_simplifyMapLexicographic, 2},
+    {"_icd_comorbidMatMulWide", (DL_FUNC) &_icd_comorbidMatMulWide, 5},
+    {"_icd_icd9PartsToShort", (DL_FUNC) &_icd_icd9PartsToShort, 1},
+    {"_icd_icd9PartsToDecimal", (DL_FUNC) &_icd_icd9PartsToDecimal, 1},
+    {"_icd_icd9MajMinToParts", (DL_FUNC) &_icd_icd9MajMinToParts, 2},
+    {"_icd_icd9ShortToPartsCpp", (DL_FUNC) &_icd_icd9ShortToPartsCpp, 2},
+    {"_icd_icd10ShortToPartsCpp", (DL_FUNC) &_icd_icd10ShortToPartsCpp, 2},
+    {"_icd_icd9DecimalToPartsCpp", (DL_FUNC) &_icd_icd9DecimalToPartsCpp, 2},
+    {"_icd_icd10DecimalToPartsCpp", (DL_FUNC) &_icd_icd10DecimalToPartsCpp, 2},
+    {"_icd_icd9ShortToDecimal", (DL_FUNC) &_icd_icd9ShortToDecimal, 1},
+    {"_icd_icd9DecimalToShort", (DL_FUNC) &_icd_icd9DecimalToShort, 1},
+    {"_icd_icd9GetMajor", (DL_FUNC) &_icd_icd9GetMajor, 2},
+    {"_icd_fastIntToStringRcpp", (DL_FUNC) &_icd_fastIntToStringRcpp, 1},
+    {"_icd_guessShortPlusFactorCpp", (DL_FUNC) &_icd_guessShortPlusFactorCpp, 2},
+    {"_icd_guessShortCompleteCpp", (DL_FUNC) &_icd_guessShortCompleteCpp, 4},
+    {"_icd_icd9_is_n_cpp", (DL_FUNC) &_icd_icd9_is_n_cpp, 1},
+    {"_icd_icd9_is_v_cpp", (DL_FUNC) &_icd_icd9_is_v_cpp, 1},
+    {"_icd_icd9_is_e_cpp", (DL_FUNC) &_icd_icd9_is_e_cpp, 1},
+    {"_icd_icd9AddLeadingZeroesMajorSingle", (DL_FUNC) &_icd_icd9AddLeadingZeroesMajorSingle, 1},
+    {"_icd_icd9AddLeadingZeroesMajorSingleStd", (DL_FUNC) &_icd_icd9AddLeadingZeroesMajorSingleStd, 1},
[…]
