### 2018-12-31 1cccbc9
clean up some convert interface functions; clean up C++ code, mostly dropping Rcpp namespace references
