### 2019-01-03 8044b77
tidying tools, ignore old bibliography location
