+++ tools/undebug.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-ICD_HOME=${ICD_HOME:-"$HOME/rprojects/icd"}
-
-"$ICD_HOME"/tools/flagoff.sh ICD_DEBUG
-
-# flip on ICD_DEBUG
-
