### 2019-01-11 bd25648
add function to plot summaries of comorbidities, update news before 3.4 release.
