### 2019-01-11 b19fdcd
export and improve plot_comorbid
