+++ NAMESPACE
@@ -312,6 +312,7 @@ export(is_billable)
 export(is_defined)
 export(is_valid)
 export(long_to_wide)
+export(plot_comorbid)
 export(poa_choices)
 export(short_to_decimal)
 export(short_to_parts)
