### 2019-01-11 17c608e
add plot to readme, ignore plot test output
