### 2019-01-20 1a31f03
big refactor, moving most data and pre-processing to icd.data package, and moving icd.data down to a Suggest in DESCRIPTION. Simplified and removed some documentation.
