### 2019-01-21 3b2f97d
moving around more code and data to separate icd.data; linting line length and removing copyright notices on each file, dropping more internal documentation from the actual manual.
