+++ R/generate_mappings.R
@@ -1,22 +1,3 @@
-# Copyright (C) 2014 - 2018  Jack O. Wasey
-#
-# This file is part of icd.
-#
-# icd is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# icd is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with icd. If not, see <http:#www.gnu.org/licenses/>.
-
-# some ICD-9 and ICD-10 mappings (others generated from SAS code parsing)
-
 #nocov start
 
 #' Generate Elixhauser comorbidities
@@ -31,10 +12,11 @@ icd9_generate_map_elix <- function(save_data = TRUE) {
     chf = c("398.91", "402.11", "402.91", "404.11", "404.13", "404.91",
             "404.93", "428.0" %i9da% "428.9"),
     arrhythmia = c("426.1", "426.11", "426.13", "426.2" %i9da% "426.53",
-                   "426.6" %i9da% "426.89", "427.0", "427.2", "427.31", "427.60",
-                   "427.9", "785", "V45.0", "V53.3"),
+                   "426.6" %i9da% "426.89", "427.0", "427.2", "427.31",
+                   "427.60", "427.9", "785", "V45.0", "V53.3"),
     valve = c("93.20" %i9da% "93.24", "394.0" %i9da% "397.1",
-              "424.0" %i9da% "424.91", "746.3" %i9da% "746.6", "V42.2", "V43.3"),
+              "424.0" %i9da% "424.91", "746.3" %i9da% "746.6",
+              "V42.2", "V43.3"),
     pulm.circ = c("416.0" %i9da% "416.9", " 417.9"),
     pvd = c("440.0" %i9da% "440.9", "441.2", "441.4", "441.7", "441.9",
             "443.1" %i9da% "443.9", "447.1", "557.1", "557.9", "V43.4"),
@@ -95,58 +77,88 @@ icd9_generate_map_elix <- function(save_data = TRUE) {
 #' @keywords internal
 icd10_generate_map_elix <- function(save_data = TRUE) {
   icd10_map_elix <- list(
-    chf = c("I099", "I110", "I130", "I132", "I255", "I420", "I425", "I426", "I427", "I428", "I429",
+    chf = c("I099", "I110", "I130", "I132", "I255", "I420", "I425", "I426",
+            "I427", "I428", "I429",
             "I43", "I50", "P290"),
-    arrhythmia = c("I441", "I442", "I443", "I456", "I459", "I47", "I48", "I49", "R000", "R001",
+    arrhythmia = c("I441", "I442", "I443", "I456", "I459", "I47", "I48", "I49",
+                   "R000", "R001",
                    "R008", "T821", "Z450", "Z950"),
-    valve = c("A520", "I05", "I06", "I07", "I08", "I091", "I098", "I34", "I35", "I36", "I37", "I38",
+    valve = c("A520", "I05", "I06", "I07", "I08", "I091", "I098", "I34", "I35",
+              "I36", "I37", "I38",
               "I39", "Q230", "Q231", "Q232", "Q233", "Z952", "Z953", "Z954"),
[…]
