+++ R/z-deprecated.R
@@ -1,515 +1,626 @@
-# git diff -U0 HEAD~4 -- NAMESPACE | grep icd_ | sed 's/.*(\(.*\)\(icd_\)\(.*\)).*/\1\2\3 <- function(...) { very_soft_deprecate("\1\3"); \1\3(...) }/' | tr , .
-
 #nocov start
 
-very_soft_deprecate <- function(fun, ...) {
+d <- function(fun, ...) {
   message("'icd_' prefixed functions have been replaced with either ",
           "ICD-9 or ICD-10 specific functions, which have an ",
           "'icd9_' or 'icd10_' prefix, or lost their prefix altogether. ")
 }
 
 #' @export
-`[[.icd_comorbidity_map` <- function(...) { very_soft_deprecate("[[.comorbidity_map"); `[[.comorbidity_map`(...) }
+`[[.icd_comorbidity_map` <-
+  function(...) { d("[[.comorbidity_map"); `[[.comorbidity_map`(...) }
 
 #' @rdname charlson
 #' @template deprecated-icd3
 #' @export
-icd_charlson <- function(...) { very_soft_deprecate("charlson"); charlson(...) }
+icd_charlson <-
+  function(...) { d("charlson"); charlson(...) }
 
 #' @rdname charlson
 #' @export
-icd_charlson.data.frame <- function(...) { very_soft_deprecate("charlson.data.frame"); charlson.data.frame(...) }
+icd_charlson.data.frame <-
+  function(...) { d("charlson.data.frame"); charlson.data.frame(...) }
 
 #' @rdname children
 #' @template deprecated-icd3
 #' @export
-icd_children.character <- function(...) { very_soft_deprecate("children.character"); children.character(...) }
+icd_children.character <-
+  function(...) { d("children.character"); children.character(...) }
 
 #' @rdname children
 #' @export
-icd_children.icd10 <- function(...) { very_soft_deprecate("children.icd10"); children.icd10(...) }
+icd_children.icd10 <-
+  function(...) { d("children.icd10"); children.icd10(...) }
 
 #' @rdname children
 #' @export
-icd_children.icd10cm <- function(...) { very_soft_deprecate("children.icd10cm"); children.icd10cm(...) }
+icd_children.icd10cm <-
+  function(...) { d("children.icd10cm"); children.icd10cm(...) }
 
 #' @rdname children
 #' @export
-icd_children.icd9 <- function(...) { very_soft_deprecate("children.icd9"); children.icd9(...) }
+icd_children.icd9 <-
+  function(...) { d("children.icd9"); children.icd9(...) }
 
 #' @rdname condense
 #' @template deprecated-icd3
 #' @export
-icd_condense.character <- function(...) { very_soft_deprecate("condense.character"); condense.character(...) }
+icd_condense.character <-
+  function(...) { d("condense.character"); condense.character(...) }
 
 #' @rdname condense
 #' @export
-icd_condense.icd9 <- function(...) { very_soft_deprecate("condense.icd9"); condense.icd9(...) }
+icd_condense.icd9 <-
+  function(...) { d("condense.icd9"); condense.icd9(...) }
 
 #' @rdname decimal_to_parts
 #' @export
 #' @keywords internal
-icd_decimal_to_parts.icd9 <- function(...) { very_soft_deprecate("decimal_to_parts.icd9"); decimal_to_parts.icd9(...) }
+icd_decimal_to_parts.icd9 <-
+  function(...) { d("decimal_to_parts.icd9"); decimal_to_parts.icd9(...) }
 
 #' @rdname decimal_to_short
 #' @export
 #' @keywords internal
[…]
