+++ tests/testthat/test-class.R
@@ -168,13 +168,14 @@ test_that("subsetting data frame works", {
   expect_true(is.icd9(icd.data::vermont_dx[[1, "DX9"]]))
   expect_true(is.icd9cm(icd.data::vermont_dx[[1, "DX12"]]))
   # columns
-  expect_is(icd.data::vermont_dx[6], c("icd9cm", "icd9", "data.frame")) # not necessarily wide anymore...
+  expect_is(icd.data::vermont_dx[6], c("icd9cm", "icd9", "data.frame"))
   expect_is(icd.data::vermont_dx[[6]], c("icd9cm", "icd9", "character"))
 })
 
 test_that("data frame subsetting doesn't incorrectly set class on columns", {
   expect_numeric(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"])
-  expect_false(inherits(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"], "icd9"))
+  expect_false(
+    inherits(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"], "icd9"))
   expect_equal(lapply(pts_invalid_mix[c(TRUE, TRUE, TRUE), ], class),
                structure(list(visit_id = "numeric", icd9 = c("icd9", "character"
                ), poa = "factor"), .Names = c("visit_id", "icd9", "poa"))
