+++ tests/testthat/test-condense.R
@@ -1,20 +1,3 @@
-# Copyright (C) 2014 - 2018  Jack O. Wasey
-#
-# This file is part of icd.
-#
-# icd is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# icd is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with icd. If not, see <http:#www.gnu.org/licenses/>.
-
 context("condense ranges of codes to parents and orphans")
 
 test_that("five digit to five digit code range condenses", {
@@ -74,15 +57,19 @@ test_that("condense ranges which do consense", {
         "V1221", info = paste("or1,2 are: ", or1, ", ", or2))
     }
   }
-  expect_equal(condense.icd9(children.icd9("V12", short_code = TRUE, defined = TRUE),
-                             defined = TRUE), "V12")
-  expect_equal(condense.icd9(children.icd9("V12", short_code = TRUE, defined = FALSE),
-                             defined = FALSE), "V12")
+  expect_equal(
+    condense.icd9(children.icd9("V12", short_code = TRUE, defined = TRUE),
+                  defined = TRUE), "V12")
+  expect_equal(
+    condense.icd9(children.icd9("V12", short_code = TRUE, defined = FALSE),
+                  defined = FALSE), "V12")
 })
 
 test_that("condense ranges that don't condense at all", {
-  expect_equal(sort(condense.icd9(c("1230", "1232", "1236"), defined = FALSE)), c("1230", "1232", "1236"))
-  expect_equal(sort(condense.icd9(c("1230", "1232", "1236"), defined = TRUE)), c("1230", "1232", "1236"))
+  expect_equal(sort(condense.icd9(c("1230", "1232", "1236"), defined = FALSE)),
+               c("1230", "1232", "1236"))
+  expect_equal(sort(condense.icd9(c("1230", "1232", "1236"), defined = TRUE)),
+               c("1230", "1232", "1236"))
   # missing 10009
   expect_equal(sort(condense.icd9(c("1000", as.character(10000:10008)),
                                   defined = FALSE)),
@@ -127,7 +114,8 @@ test_that("condense short range", {
   expect_equal(explain_code.icd9(othersalmonella), "Other salmonella infections")
 
   expect_equal(condense.icd9(short_code = TRUE, othersalmonella, defined = TRUE), "003")
-  expect_warning(res <- condense.icd9(short_code = TRUE, othersalmonella, defined = FALSE), regexp = NA)
+  expect_warning(
+    res <- condense.icd9(short_code = TRUE, othersalmonella, defined = FALSE), regexp = NA)
   expect_equal(res, othersalmonella)
   # missing this leaf node, we can't condense at all
   expect_equal(condense.icd9(short_code = TRUE, othersalmonella[-3], defined = TRUE),
@@ -142,11 +130,15 @@ test_that("condense short range", {
   expect_equal_no_icd(sort(children.icd9(short_code = TRUE, x = "001", defined = TRUE)),
                       c("001", "0010", "0011", "0019"))
 
-  expect_equal(condense.icd9(short_code = TRUE,
-                             children.icd9(short_code = TRUE, "00320", defined = TRUE), defined = TRUE), "00320")
+  expect_equal(
+    condense.icd9(short_code = TRUE,
+                  children.icd9(short_code = TRUE, "00320", defined = TRUE),
[…]
