+++ R/generate_pccc.R
@@ -143,33 +143,48 @@ icd9_generate_map_pccc_dx <- function(save_data = TRUE) {
 icd9_generate_map_pccc_pcs <- function(save_data = TRUE) {
   icd9_map_pccc_orig_pc <- list(
     neuromusc = c(
-      "0152", "0153", "0221", "0222", "0231", "0232", "0233", "0234", "0235", "0239", "0241",
+      "0152", "0153", "0221", "0222", "0231",
+      "0232", "0233", "0234", "0235", "0239", "0241",
       "0242", "0293", "0371", "0372", "0379", "0393", "0397", "0492"),
     cvd = c(
-      "0050", "0051", "0053", "0054", "0055", "0057", "1751", "1752", "3581", "3582", "3583",
-      "3584", "3741", "3751", "3752", "3753", "3754", "3755", "3760", "3761", "3763", "3765", "3766",
-      "3767", "3768", "3771", "3772", "3774", "3776", "3779", "3780", "3781", "3782", "3783", "3785",
-      "3786", "3787", "3789", "3794", "3795", "3796", "3797", "3798", "3981", "3982", "3983", "3984",
+      "0050", "0051", "0053", "0054", "0055", "0057",
+      "1751", "1752", "3581", "3582", "3583",
+      "3584", "3741", "3751", "3752", "3753", "3754",
+      "3755", "3760", "3761", "3763", "3765", "3766",
+      "3767", "3768", "3771", "3772", "3774", "3776",
+      "3779", "3780", "3781", "3782", "3783", "3785",
+      "3786", "3787", "3789", "3794", "3795", "3796",
+      "3797", "3798", "3981", "3982", "3983", "3984",
       "3985", "8945", "8946", "8947", "8948", "8949"),
     respiratory = c(
-      "303", "304", "3121", "3129", "3141", "3174", "3241", "3249", "3250", "3259", "3321",
+      "303", "304", "3121", "3129", "3141",
+      "3174", "3241", "3249", "3250", "3259", "3321",
       "3350", "3351", "3352", "336", "3485", "9655", "9723"),
     renal = c(
-      "3895", "3927", "3942", "3993", "3994", "3995", "5498", "5502", "5503", "5504", "5512",
-      "5551", "5552", "5553", "5554", "5561", "5569", "5593", "5594", "5597", "5641", "5642", "5651", "5652",
-      "5661", "5662", "5671", "5672", "5673", "5674", "5675", "5679", "5721", "5722", "5771", "5779", "5993",
+      "3895", "3927", "3942", "3993", "3994", "3995",
+      "5498", "5502", "5503", "5504", "5512",
+      "5551", "5552", "5553", "5554", "5561", "5569",
+      "5593", "5594", "5597", "5641", "5642", "5651", "5652",
+      "5661", "5662", "5671", "5672", "5673", "5674", "5675",
+      "5679", "5721", "5722", "5771", "5779", "5993",
       "5994", "8607", "9645", "9646", "9647"),
     gi = c(
-      "253", "254", "4210", "4211", "4242", "4281", "4311", "4319", "4391", "4399", "4412",
-      "4432", "4438", "4439", "4563", "4581", "4582", "4583", "4613", "4622", "4623", "4632", "4640", "4641",
-      "4643", "4697", "504", "5051", "5059", "526", "527", "5280", "5282", "5283", "5284", "5285", "5286",
[…]
