### 2019-01-21 1df87f9
move more code out to icd.data, export expand_range_major just so icd.data can use it. temporarily let travis and appveyor use githb version of icd.data until it is on CRAN
