+++ R/RcppExports.R
@@ -141,6 +141,7 @@ icd9_decimal_to_short_cpp <- function(x) {
 #'   part before the decimal, when a decimal point is used.
 #' @keywords internal manip
 #' @export
+#' @noRd
 get_major.icd9 <- function(x, short_code) {
     .Call(`_icd_icd9GetMajor`, x, short_code)
 }
