### 2019-01-21 de39026
appveyor and travis should temporarily use WHO branch of icd.data
