+++ appveyor.yml
@@ -23,9 +23,9 @@ before_build:
 #  - ls -aR
   - rm .Rinstignore  # do install everything for testing
   - sed -i'' -e '/.*[Dd]ata.*/d' .Rbuildignore # do include data for testing, some data not in github though
-  - Rscript -e "install.packages(c('devtools', 'Rcpp', 'knitr'), repos = 'https://cloud.r-project.org/'); Rcpp::compileAttributes()" # make sure Rcpp generated code is made, if not checked in to repo
+  - Rscript -e "install.packages(c('remotes', 'Rcpp', 'knitr'), repos = 'https://cloud.r-project.org/'); Rcpp::compileAttributes()" # make sure Rcpp generated code is made, if not checked in to repo
   # temporarily used while latest icd.data is not on github
-  - Rscript -e "devtools::install_github(\"jackwasey/icd.data\")"
+  - Rscript -e "remotes::install_github(\"jackwasey/icd.data\", ref = \"WHO\")"
 
 build_script:
   - travis-tool.sh install_deps
