+++ .travis.yml
@@ -23,7 +23,7 @@ r_packages:
 r_github_packages:
   - jimhester/lintr
   # for testing against icd.data which is not yet updated in CRAN
-  - jackwasey/icd.data
+  - jackwasey/icd.data@WHO
 
 cran: https://cloud.r-project.org
 
