### 2019-01-22 a3c09dc
fix vignettes with icd.data use; skeleton for WHO code vignette
