+++ vignettes/charlson-scores.Rmd
@@ -52,13 +52,14 @@ More complicated scoring systems may use lab values, patient demographic informa
 The Vermont data are actually discharge, not admission diagnoses, but can be used to demonstrate generating Charlson scores.
 ```{r vermont-charlson}
 head(icd.data::vermont_dx[1:10])
-summary(charlson(vermont_dx))
-head(charlson(vermont_dx))
-head(names(charlson(vermont_dx)))
+vch <- charlson(icd.data::vermont_dx)
+summary(vch)
+head(vch)
+head(names(vch))
 ```
 This default result is a numeric vector with the names (top numbers) as the patient identifiers. Those who like working with 'tidy' data frames can use:
 ```{r charlsondf}
-head(charlson(vermont_dx, return_df = TRUE))
+head(charlson(icd.data::vermont_dx, return_df = TRUE))
 ```
 
 Behind the scenes, _icd_ calculates the Charlson comorbidities for those ICD codes, applies the Charlson scoring system, and returns the Charlson score for each patient.
@@ -68,7 +69,7 @@ Behind the scenes, _icd_ calculates the Charlson comorbidities for those ICD cod
 The same principle can be used to calculate the Van Walraven score, which is the Charlson score counterpart for Elixhauser comorbidities.
 
 ```{r vermontvanwalraven}
-`Vermont Van Walraven Scores` <- van_walraven(vermont_dx)
+`Vermont Van Walraven Scores` <- van_walraven(icd.data::vermont_dx)
 hist(`Vermont Van Walraven Scores`)
 ```
 
