### 2019-01-22 e68dd28
remove last reference to vermont data in efficiency.rmd
