### 2019-01-22 86e465b
all vignettes compile
