+++ vignettes/PCCC.Rmd
@@ -17,11 +17,12 @@ knitr::opts_chunk$set(echo = TRUE)
 library("icd")
 n <- 10000
 set.seed(1441)
-dat <- data.frame(id = n + seq(n),
-                  icd_dx1 = sample(icd10cm2016$code, n, replace = TRUE),
-                  icd_dx2 = sample(icd10cm2016$code, n, replace = TRUE),
-                  icd_pcs1 = sample(icd10_pcs_2018$pcs, n, replace = TRUE),
-                  icd_pcs2 = sample(icd10_pcs_2018$pcs, n, replace = TRUE)
+dat <- data.frame(
+  id = n + seq(n),
+  icd_dx1 = sample(icd.data::icd10cm2016$code, n, replace = TRUE),
+  icd_dx2 = sample(icd.data::icd10cm2016$code, n, replace = TRUE),
+  icd_pcs1 = sample(icd.data::icd10_pcs_2018$pcs, n, replace = TRUE),
+  icd_pcs2 = sample(icd.data::icd10_pcs_2018$pcs, n, replace = TRUE)
 )
 ```
 
