+++ README.Rmd
@@ -101,7 +101,7 @@ comorbid_charlson(patient_data)
 charlson(patient_data)
 
 # plot summary of Uranium Cancer Registry sample data using AHRQ comorbidities
-plot_comorbid(uranium_pathology)
+plot_comorbid(icd.data::uranium_pathology)
 ```
 
 ## How to get help
