### 2019-01-25 bbb5984
begin tests of WHO ICD-10 explain
