+++ R/generate_sysdata.R
@@ -90,7 +90,7 @@ generate_sysdata <- function(save_data = TRUE) {
   sysdata_names <- c("icd9_short_n", "icd9_short_v", "icd9_short_e",
                      "icd9_short_n_defined", "icd9_short_v_defined", "icd9_short_e_defined",
                      "icd9_short_n_leaf", "icd9_short_v_leaf", "icd9_short_e_leaf",
-                     "icd9_sources", ".nc")
+                     ".nc")
   # we assume we are in the root of the package directory. Save to sysdata.rda
   # because these are probably not of interest to a user and would clutter an
   # already busy namespace.
