+++ tests/tests-build-data/test-build-sysdata.R
@@ -21,7 +21,7 @@ test_that("sysdata.rda is okay", {
   lknames <- c("icd9_short_n", "icd9_short_v", "icd9_short_e",
                "icd9_short_n_defined", "icd9_short_v_defined", "icd9_short_e_defined",
                "icd9_short_n_leaf", "icd9_short_v_leaf", "icd9_short_e_leaf",
-               "icd9_sources", ".nc")
+               ".nc")
 
   sysdat <- generate_sysdata(save_data = FALSE)
   expect_equal(names(sysdat), lknames)
