### 2019-01-25 8d0c252
remove most copyright headers of files. copyright status is unchanged [ci skip]
