+++ tests/testthat/test-ranges.R
@@ -259,7 +259,9 @@ test_that("major ranges", {
 test_that("range bugs", {
   expect_equal_no_icd("042" %i9s% "042", "042")
   expect_true("345" %nin% ("3420" %i9s% "3449"))
-  expect_identical("042.11" %i9da% "042.13", icd9(as.decimal_diag(c("042.11", "042.12", "042.13"))))
+  expect_identical(
+    "042.11" %i9da% "042.13",
+    icd9(as.decimal_diag(c("042.11", "042.12", "042.13"))))
 })
 
 # github issue #14
@@ -277,7 +279,8 @@ test_that("range doesn't include parent", {
   # ICD to comorbodity mappings. We might want to either include all possible
   # sub-codes, even if they are not (yet, or anymore) 'real'.
   expect_false("390" %in% ("389.9" %i9da% "390.1"))
-  # and if range definitely covers the higher level code, re-affirm it is still in there:
+  # and if range definitely covers the higher level code, re-affirm it is still
+  # in there:
   expect_true("390" %in% ("389.9" %i9d% "391.1"))
   expect_true("390" %in% ("389.9" %i9da% "391.1"))
 })
@@ -421,37 +424,49 @@ test_that("is_short ok with redundant children", {
 test_that("sorting char vectors", {
   expect_equal(sort_icd.icd9(short_code = TRUE, c("003", "002", "001", "003")),
                c("001", "002", "003", "003"))
-  # same with dispatch
   expect_equal(sort_icd(short_code = TRUE, c("003", "002", "001", "003")),
                c("001", "002", "003", "003"))
-
-  expect_equal(sort_icd.icd9(c("001", "V02", "V01", "E003"), short_code = TRUE),
+  expect_equal(sort_icd.icd9(c("001", "V02", "V01", "E003"),
+                             short_code = TRUE),
                c("001", "V01", "V02", "E003"))
-  expect_equal(sort_icd.icd9(short_code = TRUE, c("0032", "0288", "0019", "0031")),
-               c("0019", "0031", "0032", "0288"))
-  expect_equal(sort_icd.icd9(c("V251", "V25", "E0039", "E003"), short_code = TRUE),
+  expect_equal(sort_icd.icd9(
+    c("0032", "0288", "0019", "0031"), short_code = TRUE),
+    c("0019", "0031", "0032", "0288"))
+  expect_equal(sort_icd.icd9(c("V251", "V25", "E0039", "E003"),
+                             short_code = TRUE),
                c("V25", "V251", "E003", "E0039"))
-  expect_equal(sort_icd.icd9(c("V25.1", "V25", "E003.9", "E003"), short_code = FALSE),
+  expect_equal(sort_icd.icd9(c("V25.1", "V25", "E003.9", "E003"),
+                             short_code = FALSE),
                c("V25", "V25.1", "E003", "E003.9"))
-  expect_equal(sort_icd.icd9(short_code = FALSE,
-                             c("E1.1", "V2.2", "E001", "V02.1", "999.99", "88.8", "77")),
-               c("77", "88.8", "999.99", "V02.1", "V2.2", "E001", "E1.1"))
+  expect_equal(sort_icd.icd9(
+    short_code = FALSE,
+    c("E1.1", "V2.2", "E001", "V02.1", "999.99", "88.8", "77")),
+    c("77", "88.8", "999.99", "V02.1", "V2.2", "E001", "E1.1"))
 })
 
 test_that("sorting char factors", {
-  expect_equal(sort_icd.icd9(short_code = TRUE, factor(c("003", "002", "001", "003"))),
-               factor(c("001", "002", "003", "003")))
-  expect_equal(sort_icd.icd9(factor(c("001", "V02", "V01", "E003")), short_code = TRUE),
[…]
