+++ tools/docker/icd-check-docker/in_docker_base.sh
@@ -1,21 +1,4 @@
 #!/bin/bash
-# Copyright (C) 2014 - 2018  Jack O. Wasey
-#
-# This file is part of icd.
-#
-# icd is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# icd is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with icd. If not, see <http:#www.gnu.org/licenses/>.
-
 set -euo pipefail
 IFS=$'\n\t'
 
@@ -101,6 +84,5 @@ done
 
 
 # We need to build based on the directory name, not the package name. Don't build vignettes now: they get built during check anyway.
-# $R_CMD CMD build --no-build-vignettes $GITHUB_REPO 
+# $R_CMD CMD build --no-build-vignettes $GITHUB_REPO
 popd # out of /tmp
-

[…]
