### 2019-01-25 5b8aa97
cran comments, tweaks before submission
