+++ cran-comments.md
@@ -1,11 +1,14 @@
-Minor update, including a bug fix which could affect output accuracy.
+Please note that a pandoc version error prevents the automated CRAN checks from passing on some platforms. The Pandoc version dependency for building the vignettes is in the DESCRIPTION.
+
+This version of icd depends on version 1.1 of icd.data which I am submitting at the same time.
 
 # Test environments
-* Ubuntu 18.10 R 3.5.1, GCC 8.2, clang 7.0.0
-* Ubuntu 14.04 (Travis) R-devel, gcc
-* Windows Server 2012 R2 x64 (Appveyor), R devel 32 and 64 bit
-* MacOS High Sierra, brew clang 7.0.0
-* winbuilder fails due to pandoc version error
+
+  * Ubuntu 18.10 R 3.5.2, clang 7.0
+  * Ubuntu 14.04 (Travis) R-devel, gcc
+  * Windows Server 2012 R2 x64 (Appveyor), R devel 32 and 64 bit
+  * MacOS High Sierra, brew clang 7.0
+  * winbuilder fails due to pandoc version error
 
 # CRAN checks
  
@@ -18,8 +21,8 @@ Sometimes I get a URL download fail for the URL http://ftp.cdc.gov/pub/Health_St
 There is one note:
 
 * checking installed package size ... NOTE
-  installed size is 22.0Mb
-  sub-directories of 1Mb or more:
-    R     15.2Mb
-    doc    5.0Mb
-    libs   1.0Mb
+    installed size is  8.8Mb
+    sub-directories of 1Mb or more:
+      R      1.5Mb
+      doc    5.5Mb
+      libs   1.0Mb
