+++ .travis.yml
@@ -24,6 +24,7 @@ r_github_packages:
   - jimhester/lintr
   # for testing against icd.data which is not yet updated in CRAN
 #  - jackwasey/icd.data@WHO
+  - jackwasey/icd.data
 
 cran: https://cloud.r-project.org
 
