+++ appveyor.yml
@@ -24,8 +24,9 @@ before_build:
   - rm .Rinstignore  # do install everything for testing
   - sed -i'' -e '/.*[Dd]ata.*/d' .Rbuildignore # do include data for testing, some data not in github though
   - Rscript -e "install.packages(c('remotes', 'Rcpp', 'knitr'), repos = 'https://cloud.r-project.org/'); Rcpp::compileAttributes()" # make sure Rcpp generated code is made, if not checked in to repo
-  # temporarily used while latest icd.data is not on github
-#  - Rscript -e "remotes::install_github(\"jackwasey/icd.data\", ref = \"WHO\")"
+  # temporarily used while latest icd.data is not on CRAN
+  #- Rscript -e "remotes::install_github(\"jackwasey/icd.data\", ref = \"WHO\")"
+  - Rscript -e "remotes::install_github(\"jackwasey/icd.data\")"
 
 build_script:
   - travis-tool.sh install_deps
