### 2019-01-27 574b897
complete S3 classes for ICD-9 and ICD-10 WHO and CM-PC
