+++ R/class.R
@@ -1,37 +1,87 @@
-icd9_sub_classes <- c("icd9cm")
+icd9_sub_classes <- c("icd9cm", "icd9cm_pc", "icd9who")
 icd9_classes <- c(icd9_sub_classes, "icd9")
-icd10_sub_classes <- c("icd10cm", "icd10who")
+icd10_sub_classes <- c("icd10cm", "icd10cm_pc", "icd10who")
 icd10_classes <- c(icd10_sub_classes, "icd10")
 icd_version_classes <- c(icd9_classes, icd10_classes)
 icd_data_classes <- c("icd_long_data", "icd_wide_data")
 icd_other_classes <- c("comorbidity_map")
 icd_all_classes <- c(icd_version_classes, icd_data_classes, icd_other_classes)
 icd_system_classes <- c("data.frame", "list", "numeric", "character", "factor")
-icd_conflicts_with_icd9 <- function(x) inherits(x, icd10_classes)
-icd_conflicts_with_icd10 <- function(x) inherits(x, icd9_classes)
-icd_conflicts_with_icd9cm <- icd_conflicts_with_icd9
-icd_conflicts_with_icd10cm <- function(x) {
-  icd_conflicts_with_icd10(x) || inherits(x, "icd10who")
+
+#' Check for class conflict
+#' @param x Data to test
+#' @param do_stop logical, if `TRUE`, execution will stop with an error
+#' @keywords internal
+#' @noRd
+icd_conflicts <- function(x, do_stop = FALSE)
+  UseMethod("icd_conflicts")
+
+stop_conflict <- function(x)
+  stop("Cannot set requested class as the current data already has the ",
+       "following incompatible classes: ", paste(class(x), sep = ", "),
+       ". If you really wish to do this, use unclass, then set the desired ",
+       "icd class",
+       call. = FALSE)
+
+icd_conflicts.icd9 <- function(x, do_stop = FALSE) {
+  res <- inherits(x, icd10_classes)
+  if (res && do_stop) stop_conflict(x)
+  res
 }
-icd_conflicts_with_icd10who <- function(x) {
-  icd_conflicts_with_icd9(x) || inherits(x, "icd10cm")
+
+icd_conflicts.icd9cm <- function(x, do_stop = FALSE) {
+  res <- icd_conflicts.icd9(x, do_stop) ||
+    inherits(x, "icd9who") ||
+    inherits(x, "icd9cm_pc")
+  if (res && do_stop) stop_conflict(x)
+  res
 }
-icd_check_conflict_with_icd10 <- function(x)
-  if (icd_conflicts_with_icd10(x))
-    stop("Cannot set ICD-10 class when object already has an ICD-9 class")
-icd_check_conflict_with_icd9 <- function(x)
-  if (icd_conflicts_with_icd9(x))
-    stop("Cannot set ICD-9 class when object already has an ICD-10 class")
-# for now, but could be refined:
-icd_check_conflict_with_icd9cm <- icd_check_conflict_with_icd9
-icd_check_conflict_with_icd10cm <- function(x)
-  if (icd_conflicts_with_icd10cm(x))
-    stop("Cannot set ICD-10 WHO when data has an ICD-9 or ICD-10-CM class. ",
-         "If you really want to do this, remove the old class and try again.")
-icd_check_conflict_with_icd10who <- function(x)
-  if (icd_conflicts_with_icd10who(x))
-    stop("Cannot set ICD-10 WHO when data has an ICD-9 or ICD-10-CM class. ",
-         "If you really want to do this, remove the old class and try again.")
+
[…]
