+++ NAMESPACE
@@ -114,9 +114,11 @@ export(as.comorbidity_map)
 export(as.decimal_diag)
 export(as.icd10)
 export(as.icd10cm)
+export(as.icd10cm_pc)
 export(as.icd10who)
 export(as.icd9)
 export(as.icd9cm)
+export(as.icd9cm_pc)
 export(as.icd_comorbidity_map)
 export(as.icd_decimal_diag)
 export(as.icd_long_data)
@@ -303,8 +305,12 @@ export(is.comorbidity_map)
 export(is.decimal_diag)
 export(is.icd10)
 export(is.icd10cm)
+export(is.icd10cm_pc)
+export(is.icd10who)
 export(is.icd9)
 export(is.icd9cm)
+export(is.icd9cm_pc)
+export(is.icd9who)
 export(is.icd_comorbidity_map)
 export(is.icd_decimal_diag)
 export(is.icd_long_data)

[…]
