### 2019-01-28 277fdb8
fixed check warnings
