### 2019-01-28 dd5a426
tweak so it will check fine on previous release of R
