### 2019-01-28 0d4805b
WHO ICD-10 S3 methods for explain_table, children, etc.. reduce dependency on checkmate
