+++ NAMESPACE
@@ -12,9 +12,11 @@ S3method(charlson,data.frame)
 S3method(children,character)
 S3method(children,icd10)
 S3method(children,icd10cm)
+S3method(children,icd10who)
 S3method(children,icd9)
 S3method(children,icd9cm)
 S3method(children_defined,icd10cm)
+S3method(children_defined,icd10who)
 S3method(condense,character)
 S3method(condense,icd9)
 S3method(decimal_to_parts,character)
@@ -43,6 +45,7 @@ S3method(explain_code,list)
 S3method(explain_table,default)
 S3method(explain_table,icd10)
 S3method(explain_table,icd10cm)
+S3method(explain_table,icd10who)
 S3method(explain_table,icd9)
 S3method(explain_table,icd9cm)
 S3method(get_billable,default)
