### 2019-01-28 a520b8d
package should work with previous icd.data version; fixed attributes bug while working around R 3.3 compatibility. docs.
