+++ R/util.R
@@ -322,7 +322,9 @@ str_match_all <- function(string, pattern, ...) {
 
 # optional nice error message, could just fall back on icd.data::
 req_icd_data <- function() {
-  if (!requireNamespace("icd.data"))
+  if (requireNamespace("icd.data",
+                       versionCheck = list(version = 1.1,
+                                           op = ">=")))
     stop("Please install the 'icd.data' package to explain ICD codes.",
          call. = FALSE)
 
