+++ man/uranium_pathology.Rd
@@ -5,7 +5,7 @@
 \alias{uranium_pathology}
 \title{United States Transuranium & Uranium Registries}
 \description{
-This is an ICD-10 data set (not ICD-10-CM) with mortality from the United States
+an ICD-10 data set (not ICD-10-CM) with mortality from the United States
 Transuranium & Uranium Registries, published in the public domain.
 }
 \seealso{
