### 2019-01-29 e407ff4
export short_to_decimal and decimal_to_short; lots of documentation changes
