+++ R/convert.R
@@ -1,25 +1,6 @@
-#' @title Convert between and identify 'long' and 'wide' patient data formats
-#' @description Long and Wide Formats: As is common with many data sets, key
-#'   variables can be concentrated in one column or spread over several. Tools
-#'   format of clinical and administrative hospital data, we can perform the
-#'   conversion efficiently and accurately, while keeping some metadata about
-#'   the codes intact, e.g. whether they are ICD-9 or ICD-10.
-#' @details Long or wide format ICD data is expected to be in a data frame. It
-#'   does not carry any ICD classes at the top level, even if it only contains
-#'   one type of code, but its constituent columns may have a class specified,
-#'   e.g. 'icd9'.
-#' @param x \code{data.frame} or \code{matrix} to set class, or convert.
-#' @param ... arguments passed on to create a \code{data.frame}
-#' @description \code{icd_long_data} and \code{icd_wide_data} create
-#'   \code{data.frame}s using all the arguments, and sets the class, whereas
-#'   \code{as.icd_long_data} and \code{as.icd_wide_data} just set the class of
-#'   an existing \code{data.frame}.
-#' @name icd_long_data
-NULL
-
-#' Convert ICD9 codes between formats and structures.
+#' Convert ICD data between formats and structures.
 #'
-#' ICD-9 codes are represented in \emph{short} and \emph{decimal} forms. The
+#' ICD codes are represented in \emph{short} and \emph{decimal} forms. The
 #' short form has up to 5 digits, or V or E followed by up to four digits. The
 #' decimal form has a decimal point to delimit the top-level (henceforth
 #' \emph{major}) category, and the \emph{minor} part containing the subsidiary
@@ -31,7 +12,7 @@ NULL
 #'
 #' \pkg{icd} does not covert ICD-9 to ICD-10 codes yet.
 #'
-#' @family ICD code conversion
+#' @family ICD data conversion
 #' @name convert
 NULL
 
@@ -61,13 +42,12 @@ icd9_chapters_to_map <- function(x, defined = FALSE) {
 }
 
 #' Convert ICD data from wide to long format
-#'
-#' Reshaping data is a common task, and is made easier here by knowing more
-#' about the underlying structure of the data. This function wraps the
-#' \code{\link[stats]{reshape}} function with specific behavior and checks
-#' related to ICD codes. Empty strings and NA values will be dropped, and
-#' everything else kept. No validation of the ICD codes is done.
-#'
+#' @template wide_vs_long
+#' @details Reshaping data is a common task, and is made easier here by knowing
+#'   more about the underlying structure of the data. This function wraps the
+#'   \code{\link[stats]{reshape}} function with specific behavior and checks
+#'   related to ICD codes. Empty strings and NA values will be dropped, and
+#'   everything else kept. No validation of the ICD codes is done.
 #' @param x \code{data.frame} in wide format, i.e. one row per patient, and
 #'   multiple columns containing ICD codes, empty strings or NA.
 #' @template visit_name
@@ -87,7 +67,7 @@ icd9_chapters_to_map <- function(x, defined = FALSE) {
 #'   icd9_02 = c(NA, "443", NA)
 #'   )
 #' wide_to_long(widedf)
-#' @family ICD code conversion
+#' @family ICD data conversion
 #' @export
 wide_to_long <- function(x,
[…]
