+++ src/convert.cpp
@@ -175,9 +175,9 @@ CV icd9DecimalToShort(const CV& x) {
   return out;
 }
 
-//' @describeIn get_major Get major part of ICD-9 code, i.e. first three
-//'   digits of numeric or V code, or first four digits of E code. This is the
-//'   part before the decimal, when a decimal point is used.
+//' @describeIn get_major Get major part of ICD-9 code, i.e., first three digits
+//'   of numeric or \code{V} code, or first four digits of an \code{E} code.
+//'   This is the part before the decimal, when a decimal point is used.
 //' @keywords internal manip
 //' @export
 //' @noRd
