+++ NEWS.md
@@ -13,7 +13,7 @@
  * Accept integer identifiers in patient data
  * Default to not validating input factors: all factors generated by R are valid. This gives a significant performance increase for big data.
  * PCCC benchmarking improved (`icd` is about 10x faster than the `pccc` package)
- * Better documentation, including more links between documentation, removed internal documentation from public view
+ * Better documentation, including more links between documentation, removed internal documentation from public view. Added "Table One" example demonstration to README.
  * code cleaning, simplification and linting, e.g. replaced sapply with vapply throughout, and dropped custom wide-long-wide conversion C++ code in favour of base R.
 
 # Version 3.3
