### 2019-01-29 ba92cb2
JSS article no longer need to do wide to long conversion; more documentation clean-ups.
