+++ man-roxygen/warn.R
@@ -1,3 +1,3 @@
 #' @param warn single logical value, if \code{TRUE}, give warnings when
-#'   there is discrepancy between \code{onlyReal} being \code{TRUE} yet data
+#'   there is discrepancy between \code{defined} being \code{TRUE} yet data
 #'   containing undefined codes.

[…]
