### 2019-01-29 7dc8a3d
remove dependency on checkmate
