### 2019-01-29 bd7dc39
drop checkmate functions from namespace, and fix all tests that failed without checkmate
