+++ tests/testthat/test-filter-icd10.R
@@ -5,5 +5,5 @@ pts[1, "icd10_code"] <- "invalid"
 
 test_that("filter an invalid row", {
   expect_warning(res <- icd10_filter_invalid(pts), regexp = NA)
-  expect_data_frame(res, nrows = 1)
+  expect_is(res, "data.frame")
 })
