+++ tests/testthat/test-class.R
@@ -175,7 +175,7 @@ test_that("subsetting data frame works", {
 })
 
 test_that("data frame subsetting doesn't incorrectly set class on columns", {
-  expect_numeric(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"])
+  expect_true(is.numeric(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"]))
   expect_false(
     inherits(pts_invalid_mix[c(TRUE, TRUE, TRUE), "visit_id"], "icd9"))
   expect_equal(lapply(pts_invalid_mix[c(TRUE, TRUE, TRUE), ], class),
