+++ NAMESPACE
@@ -332,34 +332,6 @@ export(van_walraven)
 export(van_walraven_from_comorbid)
 export(wide_to_long)
 importFrom(Rcpp,cppFunction)
-importFrom(checkmate,assert)
-importFrom(checkmate,assert_character)
-importFrom(checkmate,assert_count)
-importFrom(checkmate,assert_data_frame)
-importFrom(checkmate,assert_environment)
-importFrom(checkmate,assert_factor)
-importFrom(checkmate,assert_flag)
-importFrom(checkmate,assert_int)
-importFrom(checkmate,assert_integer)
-importFrom(checkmate,assert_integerish)
-importFrom(checkmate,assert_list)
-importFrom(checkmate,assert_logical)
-importFrom(checkmate,assert_matrix)
-importFrom(checkmate,assert_scalar)
-importFrom(checkmate,assert_string)
-importFrom(checkmate,assert_vector)
-importFrom(checkmate,check_character)
-importFrom(checkmate,check_data_frame)
-importFrom(checkmate,check_factor)
-importFrom(checkmate,check_flag)
-importFrom(checkmate,check_matrix)
-importFrom(checkmate,check_null)
-importFrom(checkmate,check_string)
-importFrom(checkmate,expect_character)
-importFrom(checkmate,expect_data_frame)
-importFrom(checkmate,expect_logical)
-importFrom(checkmate,expect_numeric)
-importFrom(checkmate,expect_scalar_na)
 importFrom(graphics,barplot)
 importFrom(methods,setLoadAction)
 importFrom(stats,aggregate)
