### 2019-01-30 b713ab1
appveyor needs nhds; readme vignette code instead of direct links (which often fail)
