### 2019-01-30 2801513
[ci skip] cran comments
