### 2019-01-30 ee955d7
avoid using :: on a possibly non-existent function if icd.data is too old. Dedocument, drop slow/fast distinction as all are now fast enough.
