+++ tests/test-all.R
@@ -2,5 +2,4 @@
 library("icd")
 library("icd.data")
 library("testthat", warn.conflicts = FALSE, quietly = TRUE)
-if (!tolower(Sys.getenv("ICD_TEST_ALL")) %in% c("no", "false"))
-  testthat::test_check("icd")
+testthat::test_check("icd")
