### 2019-01-30 b95ea59
speed up our quick package checks more [ci skip]
