+++ tools/check_vquick.sh
@@ -19,6 +19,7 @@ R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
   _R_CHECK_ALL_NON_ISO_C_=TRUE \
   _R_CHECK_CODE_ASSIGN_TO_GLOBALENV_=TRUE \
   _R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE" \
+  _R_CHECK_COMPACT_DATA_=false \
   _R_CHECK_CRAN_INCOMING_=false \
   _R_CHECK_CRAN_INCOMING_REMOTE_=false \
   _R_CHECK_EXIT_ON_FIRST_ERROR_=TRUE \
