### 2019-02-02 5c98709
Improve heuristics for guessing "visit" identity column in patient data, avoiding false positive confusing "enc" in "emergENCy" for "ENCounter"
