### 2019-02-03 2b7a6ed
New vignette for custom comorbidity maps; fill out vignette for WHO codes; start using 'leaf' as better synonym for billable in is_defined; don't refer to icd package for data internally as this doesn't work without library.
