+++ R/comorbid.R
@@ -1,3 +1,18 @@
+utils::globalVariables(c(
+  "names_ahrq",
+  "names_ahrq_abbrev",
+  "names_ahrq_htn",
+  "names_ahrq_htn_abbrev",
+  "names_elix",
+  "names_elix_abbrev",
+  "names_elix_htn",
+  "names_elix_htn_abbrev",
+  "names_quan_elix",
+  "names_quan_elix_abbrev",
+  "names_quan_elix_htn",
+  "names_quan_elix_htn_abbrev",
+  "names_charlson"))
+
 #' Present-on-admission flags
 #'
 #' See \link{filter_poa} for more details.
@@ -250,29 +265,51 @@ icd10_comorbid_elix <- function(x, ..., abbrev_names = TRUE, hierarchy = TRUE) {
 
 #' @describeIn comorbid Quan's Elixhauser comorbidities for ICD-9 codes
 #' @export
-icd9_comorbid_quan_elix <- function(x, ..., abbrev_names = TRUE, hierarchy = TRUE) {
+icd9_comorbid_quan_elix <- function(x,
+                                    ...,
+                                    abbrev_names = TRUE,
+                                    hierarchy = TRUE
+) {
   cbd <- icd9_comorbid(x, map = icd::icd9_map_quan_elix, short_map = TRUE, ...)
   apply_hier_quan_elix(cbd, abbrev_names = abbrev_names, hierarchy = hierarchy)
 }
 
 #' @describeIn comorbid Quan's Elixhauser comorbidities for ICD-10 codes
 #' @export
-icd10_comorbid_quan_elix <- function(x, ..., abbrev_names = TRUE, hierarchy = TRUE) {
-  cbd <- icd10_comorbid(x, map = icd::icd10_map_quan_elix, short_map = TRUE, ...)
+icd10_comorbid_quan_elix <- function(x,
+                                     ...,
+                                     abbrev_names = TRUE,
+                                     hierarchy = TRUE
+) {
+  cbd <- icd10_comorbid(x,
+                        map = icd::icd10_map_quan_elix,
+                        short_map = TRUE,
+                        ...)
   apply_hier_quan_elix(cbd, abbrev_names = abbrev_names, hierarchy = hierarchy)
 }
 
 #' @describeIn comorbid Quan's Deyo (Charlson) comorbidities for ICD-9 codes
 #' @export
-icd9_comorbid_quan_deyo <- function(x, ..., abbrev_names = TRUE, hierarchy = TRUE) {
+icd9_comorbid_quan_deyo <- function(x,
+                                    ...,
+                                    abbrev_names = TRUE,
+                                    hierarchy = TRUE
+) {
   cbd <- icd9_comorbid(x, map = icd::icd9_map_quan_deyo, short_map = TRUE, ...)
   apply_hier_quan_deyo(cbd, abbrev_names = abbrev_names, hierarchy = hierarchy)
 }
 
 #' @describeIn comorbid Quan's Deyo (Charlson) comorbidities for ICD-10 codes
 #' @export
-icd10_comorbid_quan_deyo <- function(x, ..., abbrev_names = TRUE, hierarchy = TRUE) {
-  cbd <- icd10_comorbid(x, map = icd::icd10_map_quan_deyo, short_map = TRUE, ...)
+icd10_comorbid_quan_deyo <- function(x,
+                                     ...,
+                                     abbrev_names = TRUE,
+                                     hierarchy = TRUE
+) {
+  cbd <- icd10_comorbid(x,
+                        map = icd::icd10_map_quan_deyo,
[…]
