+++ man/icd10cm2016.Rd
@@ -16,7 +16,7 @@ There are annual revisions to this data. Currently, the 2016 edition is
 included.
 
 Format:  data frame, with columns for code, leaf status (0 or 1), short and
-  long descriptions.
+long descriptions.
 }
 \seealso{
 \code{\link[icd.data]{ls_icd_data}}
