+++ vignettes/who.Rmd
@@ -14,10 +14,14 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
+```{r setup, echo = FALSE}
 library(icd)
 library(icd.data)
 ```
+
+Surprinsgly, there are potential copyright restrictions on redistributing the WHO ICD-9 and ICD-10 definitions, which form the basis of ICD code systems around the world. These definitions are themselves also used as the internationally standard way to codify morbidity and mortality for public health.
+
+`icd` now includes the ability to work with these codes more easily via the companion `icd.data` package. The data is not available in the package so `icd.data` asks the user whether they wish to download the data from the WHO web site, whereupon it is made available just like any other R data, but the name `icd.data::icd10who2016`, which is the latest ICD-10 release from the WHO at the time of writing.
 
 ```{r}
 
