+++ DESCRIPTION
@@ -47,7 +47,7 @@ Imports:
     utils
 Suggests:
     graphics,
-    icd.data (>= 1.1),
+    icd.data,
     knitr,
     magrittr,
     nhds,
