+++ R/generate_sysdata.R
@@ -73,9 +73,7 @@ generate_sysdata <- function(save_data = TRUE) {
           paste(long_fns[long_fns != make.names(long_fns)]))
   message("non-portable short file names: ",
           paste(short_fns[short_fns != make.names(short_fns)]))
-  .nc <- nchar(icd.data::icd10cm2016[["code"]]) # nolint
   # minimal test here just to use variable names to avoid warnings!
-  stopifnot(all(!is.na(.nc)))
   stopifnot(length(icd9_short_n$vec) == length(icd9_short_n$env))
   stopifnot(length(icd9_short_v$vec) == length(icd9_short_v$env))
   stopifnot(length(icd9_short_e$vec) == length(icd9_short_e$env))
