### 2019-02-05 7ddc066
R CMD check fixes. Functions to allow access to icd.data _data_ when the package exists, but is not in search path.
