+++ R/explain-table.R
@@ -138,7 +138,7 @@ explain_table.icd10who <- function(x,
                                    warn = TRUE,
                                    ...
 ) {
-  explain_table_worker(x = x, hierarchy = icd.data::get_icd10who2016(),
+  explain_table_worker(x = x, hierarchy = get_get_icd10who2016(),
                        short_code = short_code, condense = condense,
                        brief = brief, warn = warn, ...)
 }
