+++ R/explain.R
@@ -152,8 +152,8 @@ explain_code.icd10who <- function(x, short_code = guess_short(x),
     x <- decimal_to_short.icd10(x)
   # this is a alow linear lookup, but usually only
   # "explaining" one or a few codes at a time.
-  icd.data::get_icd10who2016()[
-    icd.data::get_icd10who2016()[["code"]] %in% unique(as_char_no_warn(x)),
+  get_get_icd10who2016()[
+    get_get_icd10who2016()[["code"]] %in% unique(as_char_no_warn(x)),
     "desc"
     ]
 }
