+++ tools/build.sh
@@ -2,5 +2,5 @@
 set -euo pipefail
 IFS=$'\n\t'
 
-R CMD build --compact-vignettes=gs+qpdf "$@" "${ICD_HOME:-$HOME/rprojects/icd}"
+R CMD build --log --compact-vignettes=gs+qpdf "$@" "${ICD_HOME:-$HOME/rprojects/icd}"
 
