### 2019-02-05 2a8ed41
Importing icd.data is reasonable, and not slow.
