+++ DESCRIPTION
@@ -41,10 +41,10 @@ BugReports: https://github.com/jackwasey/icd/issues
 Depends:
     R (>= 3.4)
 Imports:
+    icd.data,
     Rcpp (>= 0.12.3)
 Suggests:
     graphics,
-    icd.data,
     knitr,
     magrittr,
     nhds,
