### 2019-02-05 08b19c6
better accommodation for older versions of icd.data
