+++ pkgdown.yml
@@ -7,6 +7,7 @@ articles:
   PCCC: PCCC.html
   charlson-scores: charlson-scores.html
   compare-maps: compare-maps.html
+  custom-maps: custom-maps.html
   efficiency: efficiency.html
   introduction: introduction.html
   ranges: ranges.html
