### 2019-02-06 bc6f06f
avert an error which only seems to show up in win-builder
