+++ tools/check_cran.sh
@@ -17,5 +17,7 @@ ${ICD_HOME:-$HOME/rprojects/icd}/tools/build.sh
 MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   #_R_CHECK_ALWAYS_LOG_VIGNETTE_OUTPUT_=TRUE \
   _R_CHECK_USE_INSTALL_LOG_=TRUE \
+  _R_CHECK_LENGTH_1_CONDITION_="verbose,abort" \
+  _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
   R CMD check --as-cran "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
