+++ tools/check_full.sh
@@ -30,5 +30,8 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_TESTS_NLINES_=0 \
   _R_CHECK_USE_INSTALL_LOG_=TRUE \
   _R_CHECK_VIGNETTES_NLINES_=0 \
+  _R_CHECK_PKG_SIZES_THRESHOLD_=50 \
+  _R_CHECK_LENGTH_1_CONDITION_="verbose,abort" \
+  _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
  R CMD check "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
