### 2019-02-06 ece6e9e
fix errors only seen on r-devel (3.6) check
