+++ tests/testthat/test-children-icd10.R
@@ -59,7 +59,6 @@ test_that("zero length ICD-10-CM children", {
   expect_empty_icd10cm_kids(c("%!^#&<>?,./", ""))
   expect_empty_icd10cm_kids(c("", ""))
   expect_empty_icd10cm_kids(character(0), has_warning = FALSE)
-
   expect_warning(children_defined(icd:::icd10cm(character(0))),
                  icd:::icd10cm(character(0)),
                  regexp = NA)
