+++ tools/check_quick.sh
@@ -33,8 +33,8 @@ R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
   _R_CHECK_USE_INSTALL_LOG_=FALSE \
   _R_CHECK_VIGNETTES_NLINES_=0 \
   _R_CHECK_PKG_SIZES_THRESHOLD_=50 \
-  _R_CHECK_LENGTH_1_CONDITION_="verbose,abort" \
-  _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
+  _R_CHECK_LENGTH_1_CONDITION_="verbose" \
+  _R_CHECK_LENGTH_1_LOGIC2_="verbose" \
   R CMD check \
     --no-build-vignettes \
     "$(ls -t $tmpd/icd*.tar.gz | head -1)"
