+++ tools/docker/in_docker_check.sh
@@ -2,6 +2,7 @@
 set -euo pipefail
 IFS=$'\n\t'
 
-source ~/in_docker_base.sh
-source ~/in_docker_get_icd.sh
-source ~/in_docker_build_check.sh
+# call from base: source /in_docker_ldpreload_asan.sh
+source /in_docker_base.sh
+source /in_docker_get_icd.sh
+source /in_docker_build_check.sh
