+++ tools/docker/rockicd2.sh
@@ -1,15 +1,14 @@
 #!/bin/bash
-
 set -euo pipefail
 IFS=$'\n\t'
 set -x
 #https://stackoverflow.com/questions/14719349/error-c-stack-usage-is-too-close-to-the-limit#14719448
 # C stack limit problems hopefully fixed by this. Can also be set to "unlimited" Default is 8192k on my pc.
 old_ulimit=$(ulimit -s)
-ulimit -s unlimited
+ulimit -s unlimited || true
 # verify with Cstack_info()
 function finish {
-	ulimit -s "$old_ulimit"
+	ulimit -s "$old_ulimit" || true
 }
 trap finish EXIT
 echo "Working directory: ${ICD_HOME:=$HOME/icd}"
