### 2019-02-08 0dd8449
refactored heuristics for guessing column types to better handle procedure codes.
