+++ R/util.R
@@ -177,17 +177,26 @@ get_visit_name.matrix <- function(x, visit_name = NULL)
 #'   of \code{x}'s columns) and returned unchanged
 #' @param multi If \code{TRUE}, allow multiple ICD field names to be returned.
 #' @keywords internal
-get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE,
-                         defined_codes = FALSE, multi = FALSE) {
+get_icd_dx_name <- function(
+  x,
+  icd_name = NULL,
+  valid_codes = TRUE,
+  defined_codes = FALSE,
+  multi = FALSE
+) {
   if (!is.null(icd_name)) {
     stopifnot(all(icd_name %in% names(x)))
     return(icd_name)
   }
   if (any(grepl(pattern = "poa", icd_name, ignore.case = TRUE)))
     warning("'POA' Present-on-arrival fields in 'icd_name'.")
-  icd_name <- guess_icd_col_by_class(x)
+  icd_name <- guess_icd_col_by_class(x, pattern = icd_dx_not_generic)
   if (!is.null(icd_name)) return(icd_name)
-  icd_name <- guess_icd_col_by_name(x, valid_codes = valid_codes,
+  icd_pc_name <- guess_icd_col_by_class(x, pattern = icd_pc_not_generic)
+  icd_generic <- guess_icd_col_by_class(x, pattern = c("icd9", "icd10"))
+  if (is.null(icd_pc_name) && !is.null(icd_name)) return(icd_generic)
+  icd_name <- guess_icd_col_by_name(x,
+                                    valid_codes = valid_codes,
                                     defined_codes = defined_codes)
   if (is.null(icd_name)) {
     icd_name <- character()
@@ -214,6 +223,25 @@ get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE,
   icd_name
 }
 
+get_icd_name <- get_icd_dx_name
+
+#' Guess the columns which contain ICD-9 or ICD10-CM procedure codes.
+#'
+#'  Will also guess procedure codes from other national ICD versions, when supported.
+#'  @keywords internal
+#'  @noRd
+get_icd_pc_name <- function(x, icd_name = NULL) {
+  if (!is.null(icd_name)) {
+    stopifnot(all(icd_name %in% names(x)))
+    return(icd_name)
+  }
+  if (any(grepl(pattern = "poa", icd_name, ignore.case = TRUE)))
+    warning("'POA' Present-on-arrival field name in 'icd_name'.")
+  icd_name <- guess_icd_col_by_class(x, pattern = icd_pc_not_generic)
+  if (!is.null(icd_name)) return(icd_name)
+  guess_icd_pc_col_by_name(x)
+}
+
 #' get candidate column(s) from wide or long data frame frame, using hints
 #' @examples
 #' wide_df <- data.frame(a = letters,
@@ -230,15 +258,33 @@ get_icd_name <- function(x, icd_name = NULL, valid_codes = TRUE,
 #' @return Zero, one or many names of columns likely to contain ICD codes based
 #'   on the column names.
 #' @keywords internal
-guess_icd_col_by_name <- function(x, valid_codes = TRUE,
-                                  defined_codes = FALSE) {
-  guesses <- c("icd.?(9|10)", "icd.?(9|10).?Code", "icd",
-               "diagnos", "diag.?code", "diag", "dx", "i(9|10)", "code")
-  assert_data_frame(x, min.cols = 1L, col.names = "named")
+guess_icd_col_by_name <- function(
+  x,
+  valid_codes = TRUE,
+  defined_codes = FALSE,
+  guesses = c("icd.?(9|10)",
+              "icd.?(9|10).?Code",
+              "icd",
[…]
