### 2019-02-08 6c74cd8
fixed documnetation error for R CMD check
