### 2019-02-09 d8683bf
Document and begin to make WHO vignette cover all international versions.
