### 2019-02-13 4f2c0ee
more hoops jumped through to avoid dependency on a newer version of icd.data
