+++ R/explain-table.R
@@ -120,23 +120,33 @@ explain_table.icd9cm <- function(x, short_code = guess_short(x),
 #' @author Ed Lee
 #' @export
 #' @keywords internal
-explain_table.icd10cm <- function(x, short_code = guess_short(x),
-                                  condense = FALSE, brief = TRUE,
-                                  warn = TRUE, ...) {
-  explain_table_worker(x = x, hierarchy = icd.data::icd10cm_active,
-                       short_code = short_code, condense = condense,
-                       brief = brief, warn = warn, ...)
+explain_table.icd10cm <- function(
+  x,
+  short_code = guess_short(x),
+  condense = FALSE,
+  brief = TRUE,
+  warn = TRUE, ...
+) {
+  i <- get_from_icd_data("icd10cm_active", alt = icd.data::icd10cm2016)
+  explain_table_worker(x = x,
+                       hierarchy = i,
+                       short_code = short_code,
+                       condense = condense,
+                       brief = brief,
+                       warn = warn,
+                       ...)
 }
 
 #' @describeIn explain_table explain character vector of ICD1-10-CM codes
 #' @export
 #' @keywords internal
-explain_table.icd10who <- function(x,
-                                   short_code = guess_short(x),
-                                   condense = FALSE,
-                                   brief = TRUE,
-                                   warn = TRUE,
-                                   ...
+explain_table.icd10who <- function(
+  x,
+  short_code = guess_short(x),
+  condense = FALSE,
+  brief = TRUE,
+  warn = TRUE,
+  ...
 ) {
   req_icd_data()
   explain_table_worker(x = x, hierarchy = icd.data::icd10who2016,
