+++ tests/testthat/test-build-sysdata.R
@@ -1,10 +1,15 @@
 context("sysdata")
 
 test_that("sysdata.rda is okay", {
-  lknames <- c("icd9_short_n", "icd9_short_v", "icd9_short_e",
-               "icd9_short_n_defined", "icd9_short_v_defined", "icd9_short_e_defined",
-               "icd9_short_n_leaf", "icd9_short_v_leaf", "icd9_short_e_leaf")
-
+  lknames <- c("icd9_short_n",
+               "icd9_short_v",
+               "icd9_short_e",
+               "icd9_short_n_defined",
+               "icd9_short_v_defined",
+               "icd9_short_e_defined",
+               "icd9_short_n_leaf",
+               "icd9_short_v_leaf",
+               "icd9_short_e_leaf")
   sysdat <- generate_sysdata(save_data = FALSE)
   expect_equal(names(sysdat), lknames)
 
