+++ man/icd10_chapters.Rd
@@ -16,7 +16,7 @@ U00-U49 - Provisional assignment of new diseases of uncertain etiology or
 emergency use U82-U85 - Resistance to antimicrobial and anti-neoplastic drugs
 
 Format:  list with chapter names stored in list names, each with two element
-named character vector with start and end codes.
+  named character vector with start and end codes.
 }
 \seealso{
 \code{\link[icd.data]{ls_icd_data}}

[…]
