### 2019-02-13 635dd72
tweak to clean R CMD check with icd.data 1.0
