+++ R/explain.R
@@ -117,7 +117,7 @@ explain_code.icd9cm <- function(
 
 get_from_icd_data <- function(name, alt = NULL) {
   out <- try(silent = TRUE, {
-    getFromNamespace(name, asNamespace("icd.data"))
+ utils::getFromNamespace(name, asNamespace("icd.data"))
   })
   if (!inherits(out, "try-error"))
     out
@@ -173,8 +173,9 @@ explain_code.icd10who <- function(x, short_code = guess_short(x),
     x <- decimal_to_short.icd10(x)
   # this is a alow linear lookup, but usually only
   # "explaining" one or a few codes at a time.
-  icd.data::icd10who2016[
-    icd.data::icd10who2016[["code"]] %in% unique(as_char_no_warn(x)),
+  i <- get_from_icd_data("icd10who2016")
+  i[
+    i[["code"]] %in% unique(as_char_no_warn(x)),
     "desc"
     ]
 }
