+++ NAMESPACE
@@ -348,6 +348,7 @@ importFrom(graphics,barplot)
 importFrom(stats,aggregate)
 importFrom(stats,reshape)
 importFrom(stats,setNames)
+importFrom(utils,getFromNamespace)
 importFrom(utils,head)
 importFrom(utils,read.csv)
 importFrom(utils,read.fwf)
