### 2019-02-13 e696932
tweak to clean R CMD check with icd.data 1.0, drop docker stuff, since rhub can now do this much better.
