+++ tools/docker/icd-check-docker/in_docker_build_check.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-IFS=$'\n\t'
-
-/in_docker_build.sh
-/in_docker_check.sh
-
