+++ tools/docker/icd-check-docker/in_docker_all.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-IFS=$'\n\t'
-
-# call from base: source /in_docker_ldpreload_asan.sh
-source /in_docker_base.sh
-source /in_docker_get_icd.sh
-source /in_docker_build.sh
-source /in_docker_check.sh

[…]
