### 2019-02-13 be3fef0
brief French-France ICD-10 example, and redocument based on icd.data 1.1 (icd re-exports some icd.data data)
