### 2019-02-13 c3a83f5
Better check we have data using with_icd10cm_version
