### 2019-02-13 f12f652
don't need this appveyor workaround anymore
