+++ appveyor.yml
@@ -34,7 +34,7 @@ build_script:
 
 environment:
   # temporarily work around a cache problem for old github version of icd.data
-  APPVEYOR_CACHE_SKIP_RESTORE: true
+  # APPVEYOR_CACHE_SKIP_RESTORE: true
   WARNINGS_ARE_ERRORS: 1
   USE_RTOOLS: true
   R_BUILD_ARGS: ""
