### 2019-02-13 455eeb7
stop re-exporting icd.data vermont and uranium data, ICD-10 vignette updated.
