### 2019-02-13 e2ce58f
drop uranium pathology and vermont documentation
