+++ man/uranium_pathology.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/datadocs.R
-\docType{data}
-\name{uranium_pathology}
-\alias{uranium_pathology}
-\title{United States Transuranium & Uranium Registries}
-\description{
-This is an ICD-10 data set with mortality from the United States Transuranium
-& Uranium Registries, published in the public domain.
-}
-\seealso{
-\code{\link[icd.data]{ls_icd_data}}
-}
-\keyword{datasets}
