+++ man/vermont_dx.Rd
@@ -1,34 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/datadocs.R
-\docType{data}
-\name{vermont_dx}
-\alias{vermont_dx}
-\title{Hospital discharge data from Vermont}
-\description{
-Anonymous data from public Vermont source for 2013
-}
-\details{
-Conditions of Release Release of public use data is subject to the following
-conditions, which the requestor agrees to upon accepting copies of the data:
-
-1. The data may not be used in any manner that attempts to or does identify,
-directly or indirectly, any individual patient or physician.
-
-2. The requestor agrees to incorporate the following, or a substantially
-similar, disclaimer in all reports or publications that include public use
-data: "Hospital discharge data for use in this study were supplied by the
-Vermont Association of Hospitals and Health Systems-Network Services
-Organization (VAHHS-NSO) and the Vermont Department of Banking, Insurance,
-Securities and Health Care Administration (BISHCA). All analyses,
-interpretations or conclusions based on these data are solely that of [the
-requestor]. VAHHS-NSO and BISHCA disclaim responsibility for any such
-analyses, interpretations or conclusions. In addition, as the data have been
-edited and processed by VAHHS-NSO, BISHCA assumes no responsibility for
-errors in the data due to coding or processing"
-
-Format:  CSV original, minimally processed into R data frame.
-}
-\seealso{
-\code{\link[icd.data]{ls_icd_data}}
-}
-\keyword{datasets}
