### 2019-02-13 9c146ce
final fix to vignettes, and man pages based on icd.data 1.1 (for last time)
