+++ man/icd9_chapters.Rd
@@ -5,7 +5,7 @@
 \alias{icd9_chapters}
 \alias{icd9_sub_chapters}
 \alias{icd9_majors}
-\title{ICD-9 chapters and sub_chapters}
+\title{ICD-9 chapters}
 \description{
 \code{icd9_chapters}, \code{icd9_chapters_sub} and \code{icd9_majors} contain
 mappings from the higher level descriptions of ICD-9 codes to the ranges of
