+++ vignettes/PCCC.Rmd
@@ -20,7 +20,7 @@ n <- 10000
 set.seed(1441)
 nm <- "code"
 i <- if (icd:::icd_data_ver_ok()) {
-  icd:::get_from_icd_data("icd10cm2018_pc", lazy = TRUE)
+  icd:::get_from_icd_data("icd10cm2016_pc", lazy = TRUE)
 } else {
   nm <- "pcs"
   icd:::get_from_icd_data("icd10_pcs_2016", lazy = TRUE)
