+++ vignettes/ICD-10.Rmd
@@ -111,10 +111,19 @@ icd10_map_quan_elix$HTNcx
 
 The AHRQ publishes an annually updated categorization of ICD-10-CM procedure codes into four classes, representing diagnostic and therapeutic procedures, each being either minor or major.
 
+```{r setuppcs, echo = FALSE}
+nm = "code"
+pcs <- if (icd:::icd_data_ver_ok()) {
+  icd:::get_from_icd_data("icd10cm2018_pc", lazy = TRUE)
+} else {
+  nm <- "pcs"
+  icd:::get_from_icd_data("icd10_pcs_2018", lazy = TRUE)
+}
+```
 ```{r icd10pcs}
 set.seed(1441)
 pts <- data.frame(id = sample(LETTERS, 10),
-                  pc = sample(icd10_pcs[["2018"]]$code, 10))
+                  pc = sample(pcs[[nm]], 10))
 res <- icd10_comorbid(pts,
                       map = icd10_map_ahrq_pcs, 
                       icd_name = "pc",
@@ -126,7 +135,7 @@ colSums(res)
 For more information on working with ICD-10 codes, see the introduction vignette, and function examples. E.g.
 ```
 ?comorbid
-?explain
+?explain_code
 ```
 
 # References
