### 2019-02-13 2255e72
class and explain functions for WHO and France
