+++ tests/testthat/test-class.R
@@ -292,3 +292,38 @@ test_that("can create NA valued ICD code types", {
   expect_true(is.na(as.icd10(NA)))
   expect_true(is.na(as.icd10cm(NA)))
 })
+test_that("icd9 procedure code classes look good", {
+  expect_false(icd_conflicts.icd9cm_pc(as.icd9cm_pc("0016073")))
+  expect_true(icd_conflicts.icd9cm_pc(as.icd9("A00")))
+  expect_true(icd_conflicts.icd9cm_pc(as.icd9cm("A00")))
+  expect_identical(icd9cm_pc("001"), as.icd9cm_pc("001"))
+  expect_error(as.icd9cm_pc(as.icd9cm("A00")))
+})
+test_that("icd10 procedure code classes look good", {
+  expect_false(icd_conflicts.icd10cm_pc(as.icd10cm_pc("0016073")))
+  expect_true(icd_conflicts.icd10cm_pc(as.icd10("A00")))
+  expect_true(icd_conflicts.icd10cm_pc(as.icd10cm("A00")))
+  expect_true(icd_conflicts.icd10cm_pc(as.icd10who("A00")))
+  expect_identical(icd10cm_pc("001"), as.icd10cm_pc("001"))
+  expect_true(is.icd10cm_pc(icd10cm_pc("001")))
+  expect_true(is.icd10cm_pc(as.icd10cm_pc("001")))
+  expect_error(as.icd10cm_pc(as.icd10cm("A00")))
+})
+test_that("icd10 WHO code classes look good", {
+  expect_false(icd_conflicts.icd10who(as.icd10who("A00")))
+  expect_true(icd_conflicts.icd10cm_pc(as.icd10("A00")))
+  expect_true(icd_conflicts.icd10who(as.icd10cm_pc("001")))
+  expect_true(icd_conflicts.icd10cm(as.icd10who("A00")))
+  expect_true(is.icd10who(icd10who("Z998")))
+  expect_true(is.icd10who(as.icd10who("Z998")))
+  expect_false(is.icd10who(as.icd10cm("Z998")))
+})
+test_that("icd10-fr France code classes look good", {
+  expect_false(icd_conflicts.icd10fr(as.icd10fr("A00")))
+  expect_true(icd_conflicts.icd10cm_pc(as.icd10("A00")))
+  expect_true(icd_conflicts.icd10fr(as.icd10cm_pc("001")))
+  expect_true(icd_conflicts.icd10fr(as.icd10who("A00")))
+  expect_true(is.icd10fr(icd10fr("Z998")))
+  expect_true(is.icd10fr(as.icd10fr("Z998")))
+  expect_false(is.icd10fr(as.icd10cm("Z998")))
+})
