+++ R/class.R
@@ -14,7 +14,7 @@ icd10_sub_classes <- c(icd10_dx_sub_classes,
 icd10_classes <- c(icd10_sub_classes, "icd10")
 # ICD-9 and ICD-10
 icd_dx_not_generic <- c(icd9_dx_sub_classes,
-                       icd10_dx_sub_classes)
+                        icd10_dx_sub_classes)
 icd_pc_not_generic <- c(icd9_pc_sub_classes,
                         icd10_pc_sub_classes)
 icd_version_classes <- c(icd9_classes,
@@ -64,7 +64,8 @@ icd_conflicts.icd9cm <- function(x, do_stop = FALSE) {
 icd_conflicts.icd9cm_pc <- function(x, do_stop = FALSE) {
   res <- icd_conflicts.icd9(x, do_stop) ||
     inherits(x, "icd9who") ||
-    inherits(x, "icd9cm")
+    inherits(x, "icd9cm") ||
+    inherits(x, "icd9")
   if (res && do_stop) stop_conflict(x)
   res
 }
@@ -94,7 +95,8 @@ icd_conflicts.icd10cm <- function(x, do_stop = FALSE) {
 icd_conflicts.icd10cm_pc <- function(x, do_stop = FALSE) {
   res <- icd_conflicts.icd10(x, do_stop) ||
     inherits(x, "icd10who") ||
-    inherits(x, "icd10cm")
+    inherits(x, "icd10cm") ||
+    inherits(x, "icd10")
   if (res && do_stop) stop_conflict(x)
   res
 }
@@ -102,7 +104,17 @@ icd_conflicts.icd10cm_pc <- function(x, do_stop = FALSE) {
 icd_conflicts.icd10who <- function(x, do_stop = FALSE) {
   res <- icd_conflicts.icd10(x) ||
     inherits(x, "icd10cm") ||
-    inherits(x, "icd10cm_pc")
+    inherits(x, "icd10cm_pc") ||
+    inherits(x, "icd10fr")
+  if (res && do_stop) stop_conflict(x)
+  res
+}
+
+icd_conflicts.icd10fr <- function(x, do_stop = FALSE) {
+  res <- icd_conflicts.icd10(x) ||
+    inherits(x, "icd10cm") ||
+    inherits(x, "icd10cm_pc") ||
+    inherits(x, "icd10who")
   if (res && do_stop) stop_conflict(x)
   res
 }
@@ -246,14 +258,12 @@ icd9cm_pc <- function(x) {
 as.icd9cm_pc <- function(x) {
   stopifnot(is.atomic(x))
   if ("icd10" %in% class(x))
-    stop("icd10 class already set on this variable")
-  if (any(c("icd9cm", "icd9who") %in% class(x)))
-    stop("and ICD-9 diagnostic code class is already set")
-  icd9_pos <- match("icd9", class(x))
-  if (!is.na(icd9_pos))
-    class(x) <- append(class(x), "icd9cm_pc", after = icd9_pos - 1)
-  else
-    class(x) <- append(class(x), c("icd9cm_pc", "icd9"), after = 0)
+    stop("An ICD-10 diagnostic class is already set on this variable")
+  if ("icd10cm_pc" %in% class(x))
+    stop("An ICD-10 procedure class is already set on this variable")
+  if (any(c("icd9", "icd9cm", "icd9who") %in% class(x)))
+    stop("An ICD-9 diagnostic code class is already set")
+  class(x) <- append(class(x), "icd9cm_pc", after = 0)
   icd_conflicts(x, do_stop = TRUE)
   x
 }
@@ -312,10 +322,7 @@ as.icd10cm <- function(x, short_code = NULL) {
[…]

[…]
