### 2019-02-16 f45644e
Order ICD-10-CM including the C7A, C7B and D3A quirks. Re-enable Catch C++ tests on all builds (for now), and make them all pass.
