+++ src/local.h
@@ -59,9 +59,9 @@ extern "C" {
 #endif
 
 #define ICD_CATCH
-#ifdef NDEBUG
-#undef ICD_CATCH
-#endif
+//#ifdef NDEBUG
+//#undef ICD_CATCH
+//#endif
 
 #if defined(ICD_VALGRIND) && defined(HAVE_VALGRIND_VALGRIND_H)
 #include <valgrind/callgrind.h>
