### 2019-02-16 e1f9c5b
Belgian ICD-10 classes, simple mechanism to skip slow tests on some platforms; more consistent C++ function names; quieter debug mode
