+++ tests/testthat/test-comorbid-maps-icd9.R
@@ -1,7 +1,8 @@
 context("comorbidity maps")
 
 test_that("try to induce c++ segfault bug", {
-  expect_error(icd9_comorbid(ahrq_test_dat, map = icd9_map_ahrq, short_code = TRUE),
+  expect_error(icd9_comorbid(ahrq_test_dat, map = icd9_map_ahrq,
+                             short_code = TRUE),
                regexp = NA)
 })
 
@@ -38,6 +39,7 @@ test_that("icd9 comorbidities correct, logical to binary ok", {
 })
 
 test_that("condense big lists of comorbidities", {
+  skip_slow("Skipping condense big lists of comorbidities")
   # this is a useful but slow (8s on my PC) test because the data weren't
   # generated by just expanding base ranges (which is how the condense works in
   # reverse)
