+++ tests/testthat/test-sort.R
@@ -26,7 +26,7 @@ test_that("sorting of icd10 object", {
 test_that("warn if NA when ordering ICD-9 codes", {
   expect_warning(order.icd9(c("a", NA)))
   expect_identical(
-    expect_warning(order.icd9(NA)), character())
+    expect_warning(order.icd9(NA)), integer())
 })
 
 test_that("order some icd-9 codes", {
