+++ tests/test-all.R
@@ -2,4 +2,5 @@
 library("icd")
 library("icd.data")
 library("testthat", warn.conflicts = FALSE, quietly = TRUE)
+# Set ICD_TEST_SLOW to yes/true to run slow tests
 testthat::test_check("icd")
