+++ NAMESPACE
@@ -1,354 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method("[",icd10)
-S3method("[",icd9)
-S3method("[[",comorbidity_map)
-S3method("[[",icd10)
-S3method("[[",icd9)
-S3method("[[",icd_comorbidity_map)
-S3method(c,icd10)
-S3method(c,icd9)
-S3method(charlson,data.frame)
-S3method(children,character)
-S3method(children,icd10)
-S3method(children,icd10cm)
-S3method(children,icd10who)
-S3method(children,icd9)
-S3method(children,icd9cm)
-S3method(children_defined,icd10cm)
-S3method(children_defined,icd10who)
-S3method(condense,character)
-S3method(condense,icd9)
-S3method(decimal_to_parts,character)
-S3method(decimal_to_parts,icd10)
-S3method(decimal_to_parts,icd9)
-S3method(decimal_to_short,default)
-S3method(decimal_to_short,icd10)
-S3method(decimal_to_short,icd10cm)
-S3method(decimal_to_short,icd9)
-S3method(diff_comorbid,list)
-S3method(expand_minor,icd10)
-S3method(expand_minor,icd9)
-S3method(expand_range,character)
-S3method(expand_range,icd10cm)
-S3method(expand_range,icd9)
-S3method(expand_range_major,default)
-S3method(expand_range_major,icd10cm)
-S3method(expand_range_major,icd9)
-S3method(explain_code,default)
-S3method(explain_code,icd10)
-S3method(explain_code,icd10be)
-S3method(explain_code,icd10cm)
-S3method(explain_code,icd10fr)
-S3method(explain_code,icd10who)
-S3method(explain_code,icd9)
-S3method(explain_code,icd9cm)
-S3method(explain_code,list)
-S3method(explain_table,default)
-S3method(explain_table,icd10)
-S3method(explain_table,icd10cm)
-S3method(explain_table,icd10who)
-S3method(explain_table,icd9)
-S3method(explain_table,icd9cm)
-S3method(get_billable,default)
-S3method(get_billable,icd10)
-S3method(get_billable,icd10cm)
-S3method(get_billable,icd9)
-S3method(get_billable,icd9cm)
-S3method(get_defined,default)
-S3method(get_defined,icd9)
-S3method(get_invalid,comorbidity_map)
-S3method(get_invalid,default)
-S3method(get_invalid,icd10)
-S3method(get_invalid,icd9)
-S3method(get_leaf,default)
-S3method(get_leaf,icd10)
-S3method(get_leaf,icd10cm)
-S3method(get_leaf,icd9)
-S3method(get_leaf,icd9cm)
-S3method(get_major,icd10)
-S3method(get_major,icd9)
-S3method(get_valid,character)
-S3method(get_valid,icd10)
-S3method(get_valid,icd10cm)
-S3method(get_valid,icd9)
-S3method(guess_version,character)
-S3method(guess_version,data.frame)
-S3method(guess_version,factor)
-S3method(guess_version,icd10)
-S3method(guess_version,icd10cm)
-S3method(guess_version,icd10who)
-S3method(guess_version,icd9)
-S3method(guess_version,icd9cm)
-S3method(is_billable,default)
-S3method(is_billable,icd10)
-S3method(is_billable,icd10cm)
-S3method(is_billable,icd9)
-S3method(is_billable,icd9cm)
-S3method(is_defined,default)
-S3method(is_defined,icd10)
-S3method(is_defined,icd10cm)
-S3method(is_defined,icd9)
-S3method(is_leaf,default)
[…]
