+++ man/logical_to_binary.Rd
@@ -20,7 +20,7 @@ When saving data as text files for distribution, printing large amounts of
 text containing \code{TRUE} and \code{FALSE} is inefficient. Converting to
 binary allows more compact output to screen. Most functions in \R will
 convert \code{TRUE} and \code{FALSE} logical flags to \code{1} and \code{0},
-respsectively.
+respectively.
 }
 \section{Functions}{
 \itemize{
