+++ man/explain_code.Rd
@@ -150,6 +150,15 @@ explain_code(icd9_map_ahrq$CHF[1:3], brief = TRUE)
 # The first three in the ICD-10 equivalent are a little different:
 explain_code(icd10_map_ahrq$CHF[1:3], brief = TRUE)
 # Belgian ICD-10 has three languages available
-explain_code(as.icd10be("A00"))
+\dontrun{
+explain_code(as.icd10be("C20"))
+# [1] "Malignant neoplasm of rectum"
+explain_code(as.icd10be("C20"), lang = "en")
+# [1] "Malignant neoplasm of rectum"
+explain_code(as.icd10be("C20"), lang = "fr")
+# [1] "néoplasme malin du rectum"
+explain_code(as.icd10be("C20"), lang = "nl")
+# [1] "maligne neoplasma van het rectum"
+}
 }
 \keyword{internal}
