+++ appveyor.yml
@@ -10,8 +10,9 @@ init:
 install:
   ps: Bootstrap
 
-cache:
-- C:\RLibrary
+# temporarily disable cache
+#cache:
+#- C:\RLibrary
 
 # Adapt as necessary starting from here
 
