### 2019-02-16 aa8fbba
spelling, and test fix for icd < 1.1
