+++ src/test-cpp.cpp
@@ -21,9 +21,9 @@ context("internal 'is' functions") {
     v.push_back("100.99");
     v.push_back("9999");
     v.push_back("1");
-    std::vector<bool> result = icd9_is_n_cpp(v);
-    std::vector<bool> result2 = icd9_is_v_cpp(v);
-    std::vector<bool> result3 = icd9_is_e_cpp(v);
+    std::vector<bool> result = icd9_is_n_rcpp(v);
+    std::vector<bool> result2 = icd9_is_v_rcpp(v);
+    std::vector<bool> result3 = icd9_is_e_rcpp(v);
     for (std::vector<bool>::size_type i = 0; i != result.size(); ++i) {
       expect_true(result[i]);
       expect_false(result2[i]);
@@ -39,9 +39,9 @@ context("internal 'is' functions") {
     v.push_back("V99.99");
     v.push_back("V9999");
     v.push_back("v1");
-    std::vector<bool> result2 = icd9_is_n_cpp(v);
-    std::vector<bool> result = icd9_is_v_cpp(v);
-    std::vector<bool> result3 = icd9_is_e_cpp(v);
+    std::vector<bool> result2 = icd9_is_n_rcpp(v);
+    std::vector<bool> result = icd9_is_v_rcpp(v);
+    std::vector<bool> result3 = icd9_is_e_rcpp(v);
     for (std::vector<bool>::size_type i = 0; i != result.size(); ++i) {
       expect_true(result[i]);
       expect_false(result2[i]);
@@ -56,9 +56,9 @@ context("internal 'is' functions") {
     v.push_back("E999.9");
     v.push_back("E9999");
     v.push_back("e000");
-    std::vector<bool> result3 = icd9_is_n_cpp(v);
-    std::vector<bool> result2 = icd9_is_v_cpp(v);
-    std::vector<bool> result = icd9_is_e_cpp(v);
+    std::vector<bool> result3 = icd9_is_n_rcpp(v);
+    std::vector<bool> result2 = icd9_is_v_rcpp(v);
+    std::vector<bool> result = icd9_is_e_rcpp(v);
     for (std::vector<bool>::size_type i = 0; i != result.size(); ++i) {
       expect_true(result[i]);
       expect_false(result2[i]);
