### 2019-02-17 de5c0c0
drop deprecated functions, clean up namespace for sort and order
