+++ R/z-deprecated.R
@@ -1,651 +1,0 @@
-#nocov start
-
-d <- function(fun, ...) {
-  message("'icd_' prefixed functions have been replaced with either ",
-          "ICD-9 or ICD-10 specific functions, which have an ",
-          "'icd9_' or 'icd10_' prefix, or lost their prefix altogether. ")
-}
-
-#' @export
-`[[.icd_comorbidity_map` <-
-  function(...) { d("[[.comorbidity_map"); `[[.comorbidity_map`(...) }
-
-#' @rdname charlson
-#' @template deprecated-icd3
-#' @export
-icd_charlson <-
-  function(...) { d("charlson"); charlson(...) }
-
-#' @rdname charlson
-#' @export
-icd_charlson.data.frame <-
-  function(...) { d("charlson.data.frame"); charlson.data.frame(...) }
-
-#' @rdname children
-#' @template deprecated-icd3
-#' @export
-icd_children.character <-
-  function(...) { d("children.character"); children.character(...) }
-
-#' @rdname children
-#' @export
-icd_children.icd10 <-
-  function(...) { d("children.icd10"); children.icd10(...) }
-
-#' @rdname children
-#' @export
-icd_children.icd10cm <-
-  function(...) { d("children.icd10cm"); children.icd10cm(...) }
-
-#' @rdname children
-#' @export
-icd_children.icd9 <-
-  function(...) { d("children.icd9"); children.icd9(...) }
-
-#' @rdname condense
-#' @template deprecated-icd3
-#' @export
-icd_condense.character <-
-  function(...) { d("condense.character"); condense.character(...) }
-
-#' @rdname condense
-#' @export
-icd_condense.icd9 <-
-  function(...) { d("condense.icd9"); condense.icd9(...) }
-
-icd_decimal_to_parts.icd9 <-
-  function(...) { d("decimal_to_parts.icd9"); decimal_to_parts.icd9(...) }
-
-#' @rdname decimal_to_short
-#' @export
-#' @keywords internal
-icd_decimal_to_short.default <-
-  function(...) { d("decimal_to_short.default"); decimal_to_short.default(...) }
-
-#' @rdname decimal_to_short
-#' @export
-#' @keywords internal
-icd_decimal_to_short.icd10 <-
-  function(...) { d("decimal_to_short.icd10"); decimal_to_short.icd10(...) }
-
-#' @rdname decimal_to_short
-#' @export
-#' @keywords internal
-icd_decimal_to_short.icd10cm <-
-  function(...) { d("decimal_to_short.icd10cm"); decimal_to_short.icd10cm(...) }
-
-#' @rdname decimal_to_short
-#' @export
-#' @keywords internal
-icd_decimal_to_short.icd9 <-
-  function(...) { d("decimal_to_short.icd9"); decimal_to_short.icd9(...) }
-
-#' @rdname diff_comorbid
-#' @template deprecated-icd3
-#' @export
-icd_diff_comorbid.list <-
-  function(...) { d("diff_comorbid.list"); diff_comorbid.list(...) }
-
-#' @rdname expand_range
-#' @template deprecated-icd3
-#' @export
-icd_expand_range.character <-
-  function(...) { d("expand_range.character"); expand_range.character(...) }
-
-#' @rdname expand_range
-#' @export
-icd_expand_range.icd10cm <-
-  function(...) { d("expand_range.icd10cm"); expand_range.icd10cm(...) }
-
-#' @rdname expand_range
-#' @export
-icd_expand_range.icd9 <-
[…]
