### 2019-02-18 d0c63e3
export 'chapters_to_map', more ICD-10 sorting quirks, and tests. Rebuild data.
