+++ src/test-cpp.cpp
@@ -386,47 +386,61 @@ context("ICD10 short to parts") {
 
 context("compare/sort/order ICD-10-CM") {
   test_that("compare quirks") {
-    expect_true(icd10cmCompare("C79", "C80"));
-    expect_true(icd10cmCompare("C79", "C7A"));
-    expect_true(icd10cmCompare("C80", "C7A"));
-    expect_true(icd10cmCompare("C80", "C7B"));
+    expect_true(icd10cmCompare("C74", "C75"));
+    expect_true(icd10cmCompare("C74", "C7A"));
+    expect_true(icd10cmCompare("C75", "C7A"));
+    expect_true(icd10cmCompare("C75", "C7B"));
     expect_true(icd10cmCompare("C7A", "C7B"));
-    expect_true(icd10cmCompare("C7B", "C81"));
-    expect_false(icd10cmCompare("C81", "C7A"));
+    expect_true(icd10cmCompare("C7B", "C76"));
+    expect_false(icd10cmCompare("C76", "C7A"));
     expect_false(icd10cmCompare("C7B", "C7A"));
     expect_false(icd10cmCompare("C7A", "C7A"));
     expect_false(icd10cmCompare("C7B", "C7B"));
-    expect_false(icd10cmCompare("C81", "C7B"));
-    expect_false(icd10cmCompare("C82", "C81"));
+    expect_false(icd10cmCompare("C76", "C7B"));
+    expect_false(icd10cmCompare("C77", "C76"));
   }
-  test_that("compare quirk sub-codes") {
-    expect_true(icd10cmCompare("C791", "C80"));
-    expect_true(icd10cmCompare("C79a", "C7A"));
-    expect_true(icd10cmCompare("C802", "C7A"));
-    expect_true(icd10cmCompare("C80b", "C7B"));
+  test_that("compare quirk sub-chap codes") {
+    expect_true(icd10cmCompare("C741", "C75"));
+    expect_true(icd10cmCompare("C74a", "C7A"));
+    expect_true(icd10cmCompare("C752", "C7A"));
+    expect_true(icd10cmCompare("C75b", "C7B"));
     expect_true(icd10cmCompare("C7A3", "C7B"));
-    expect_true(icd10cmCompare("C7Ac", "C81"));
+    expect_true(icd10cmCompare("C7Ac", "C76"));
     expect_false(icd10cmCompare("C7B4", "C7A"));
     expect_false(icd10cmCompare("C7Ad", "C7A"));
-    expect_false(icd10cmCompare("C815", "C7A"));
+    expect_false(icd10cmCompare("C765", "C7A"));
     expect_false(icd10cmCompare("C7Be", "C7B"));
-    expect_false(icd10cmCompare("C816", "C7B"));
-    expect_false(icd10cmCompare("C82v", "C81"));
-    expect_true(icd10cmCompare("C79", "C801"));
-    expect_true(icd10cmCompare("C79", "C7A2"));
-    expect_true(icd10cmCompare("C80", "C7A3"));
-    expect_true(icd10cmCompare("C80", "C7B4"));
+    expect_false(icd10cmCompare("C766", "C7B"));
+    expect_false(icd10cmCompare("C77v", "C76"));
+    expect_true(icd10cmCompare("C74", "C751"));
+    expect_true(icd10cmCompare("C74", "C7A2"));
+    expect_true(icd10cmCompare("C75", "C7A3"));
+    expect_true(icd10cmCompare("C75", "C7B4"));
     expect_true(icd10cmCompare("C7A", "C7B5"));
-    expect_true(icd10cmCompare("C7A", "C816"));
[…]
