+++ tests/test-all.R
@@ -3,4 +3,6 @@ library("icd")
 library("icd.data")
 library("testthat", warn.conflicts = FALSE, quietly = TRUE)
 # Set ICD_TEST_SLOW to yes/true to run slow tests
+#
+# Sys.setenv("ICD_TEST_SLOW" = "yes")
 testthat::test_check("icd")
