### 2019-02-21 f475ced
simpler include (preparation for armadillo), formatting
