+++ src/relevant.cpp
@@ -1,8 +1,8 @@
 #include "icd_types.h"
 #include "local.h"
 #include "fastIntToString.h"
-#include <algorithm>                   // for binary_search, copy
-#include <vector>                      // for vector, vector<>::const_iterator
+#include <algorithm> // for binary_search, copy
+#include <vector>
 #include <unordered_set>
 #include "refactor.h"
 #include <string>
