### 2019-02-21 a6c6ce0
clang tidy C++ code
