+++ src/relevant.cpp
@@ -1,48 +1,44 @@
+#include "relevant.h"
+#include "fastIntToString.h"
 #include "icd_types.h"
 #include "local.h"
-#include "fastIntToString.h"
-#include <algorithm> // for binary_search, copy
-#include <vector>
-#include <unordered_set>
 #include "refactor.h"
-#include <string>
+#include <algorithm> // for binary_search, copy
 #include <cstring>
-#include "relevant.h"
+#include <string>
+#include <unordered_set>
+#include <vector>
 
 using namespace Rcpp;
 
-void Relevant::buildCodeSetCV(const CV& codes) {
+void Relevant::buildCodeSetCV(const CV &codes) {
   // over-reserve (and maybe expand), target is unique number
   allCodesSet.reserve(allCodesSet.size() + codes.size());
   DEBUG_VEC(codes);
   for (String c : codes) {
-    if (c != NA_STRING) {
-      allCodesSet.insert(c.get_cstring());
-    }
+    if (c != NA_STRING) { allCodesSet.insert(c.get_cstring()); }
   }
 }
 
-void Relevant::buildCodeSetInt(const IntegerVector& codes) {
+void Relevant::buildCodeSetInt(const IntegerVector &codes) {
   // over-reserve (and maybe expand), target is unique number
   allCodesSet.reserve(allCodesSet.size() + codes.size());
   DEBUG_VEC(codes);
   for (R_xlen_t i = 0; i != codes.size(); ++i) {
-  auto ci = codes[i];
-  auto cs = std::to_string(ci);
-    if (!IntegerVector::is_na(ci)) {
-      allCodesSet.insert(cs);
-    }
+    auto ci = codes[i];
+    auto cs = std::to_string(ci);
+    if (!IntegerVector::is_na(ci)) { allCodesSet.insert(cs); }
   }
 }
 
-void Relevant::buildCodeSet(const SEXP& codes) {
+void Relevant::buildCodeSet(const SEXP &codes) {
   switch (TYPEOF(codes)) {
   case INTSXP: {
     if (!Rf_isFactor(codes)) {
-    buildCodeSetInt(codes);
-    break;
-  }
-    CV code_levs = ((IntegerVector) codes).attr("levels");
+      buildCodeSetInt(codes);
+      break;
+    }
+    CV code_levs = ((IntegerVector)codes).attr("levels");
     DEBUG_VEC(code_levs);
     buildCodeSetCV(code_levs);
     break;
@@ -52,14 +48,10 @@ void Relevant::buildCodeSet(const SEXP& codes) {
     break;
   }
   case VECSXP: {
-    for (SEXP listItem : (List) codes) {
-    buildCodeSet(listItem);
-  }
+    for (SEXP listItem : (List)codes) { buildCodeSet(listItem); }
     break;
   }
-  default: {
-    stop("Invalid type of codes to build set in Relevant");
-  }
+  default: { stop("Invalid type of codes to build set in Relevant"); }
   }
 }
 
@@ -69,35 +61,33 @@ CV Relevant::findRelevant() {
     for (String cmbCode : cmb) {
       if (allCodesSet.find(cmbCode.get_cstring()) != allCodesSet.end()) {
         TRACE("Pushing back" << cmbCode.get_cstring());
-        r.insert(((String) cmbCode).get_cstring());
+        r.insert(((String)cmbCode).get_cstring());
       }
     }
   }
-  return(wrap(r)); // or keep as STL container?
+  return (wrap(r)); // or keep as STL container?
 }
 
 // # nocov start
 
 // setup find based on a data frame, list or vector
-CV Relevant::findRelevant(const SEXP& codes) {
+CV Relevant::findRelevant(const SEXP &codes) {
   buildCodeSet(codes);
[…]
