### 2019-02-21 41a0fd6
const auto& when possible
