### 2019-02-22 d4d4e8c
tweaks for r cmd check and travis
