+++ vignettes/introduction.Rmd
@@ -319,10 +319,13 @@ system.time(
 This takes about five seconds on a 2016 8-core workstation. More detailed benchmarks can be seen comparing icd to other packages in the vignette article in this package.
 
 ### Use an arbitrary ICD-9 mapping
+
 The user can provide any ICD-9, ICD-10 or other code mapping to comorbidities they wish. Submissions of other peer-reviewed published mappings could be included in this package, if their license permits. Create an issue in [github](https://github.com/jackwasey/icd/issues) or email me at jack@jackwasey.com) Included in this package is a small data set called `icd9_chapters`, which lists the ICD-9-CM (and indeed ICD-9) Chapters. These can easily be expanded out and used as a mapping, so instead of a comorbidity, we see which patients have codes in each chapter of the ICD-9 defintion.
+
+See the [custom comorbidity map vignette](custom-maps.html).
 ```{r arbitrary Mapping}
 names(icd9_chapters)[c(1:5, 14)]
-my_map <- icd:::icd9_chapters_to_map(icd9_chapters[c(2, 5, 14)])
+my_map <- icd:::chapters_to_map(icd9_chapters[c(2, 5, 14)])
 icd9_comorbid(patients_icd9, my_map) # no positive 
 ```
 
