### 2019-02-23 60d6000
regenerate ICD-10 maps using codes from all available ICD-10-CM years.
