+++ man/icd10_sub_chapters.Rd
@@ -14,7 +14,7 @@ neoplasms" which itself is a subset of the chapter C00-D49 "Neoplasms"
 }
 \details{
 Format:  list with sub-chapter or major names stored in list names, each with
-  two element named character vector with start and end codes.
+two element named character vector with start and end codes.
 }
 \seealso{
 \code{\link[icd.data]{ls_icd_data}}
