### 2019-02-23 f02be03
bump to version 4.0 given huge number of improvements
