### 2019-02-23 f6d309f
cran comments before submission [ci skip]
