### 2019-02-25 159e5de
all minor comments addressed except first q on codes used in maps
