### 2019-02-25 bebed82
all minor comments addressed except first q on codes used in maps
