+++ R/generate_pccc.R
@@ -201,7 +201,7 @@ icd9_generate_map_pccc_pcs <- function(save_data = TRUE) {
 
 #' @rdname generate_maps_pccc
 #' @keywords internal
-icd10_generate_map_pccc_dx <- function(save_data) {
+icd10_generate_map_pccc_dx <- function(save_data, verbose = TRUE) {
   icd10_map_pccc_orig_dx <- list(
     neuromusc = c(
       "E75", "F71", "F72", "F73", "F842", "G111", "G112", "G114",
@@ -337,9 +337,10 @@ icd10_generate_map_pccc_dx <- function(save_data) {
       "T8642", "T86890", "T86891", "T86899", "T86850",
       "T86851", "T86859", "T865", "T8690", "T8691", "T8692", "T8699")
   )
-  icd10_map_pccc_dx <- lapply(
-    icd10_map_pccc_orig_dx,
-    children_defined.icd10cm, short_code = TRUE)
+  icd10_map_pccc_dx <- apply_over_icd10cm_vers(icd10_map_pccc_orig_dx,
+                                               verbose = verbose)
+  icd10_map_pccc_dx <- apply_over_icd10who_vers(icd10_map_pccc_orig_dx,
+                                               verbose = verbose)
   # add the fixed code, per JAMA Pediatrics Letter
   icd10_map_pccc_dx[["neuromusc"]] <-
     c(icd10_map_pccc_dx[["neuromusc"]], "G80")
