### 2019-02-27 b3e51a7
Get data from icd.data more carefully, add optional warning when getting children, and DEBUG (doesn't pass tests). Rebuild ICD-10 maps correctly
