### 2019-02-28 f7c92f2
update everything on icd.data and icd sides. unclear why a single byte changed in icdwords.rds, and four bytes are inserted in other data. Still serialization version 4. could be encoding. Will stick to linux for generating data from now on.
