### 2019-02-28 213682a
bump version for CRAN, script to compare datasets
