### 2019-03-16 173ca7d
clean up now data code is moved to icd.data; drop heavy dependency of linkingto testthat; dependencies badge; ICD-10 majors range code; use trimws instead of homegrown; billable changed to leaf in most cases; restyled code; drop obsolete onload; set icd.data options for testing; tests
