+++ tests/testthat/test-comorbid-calcs-icd10.R
@@ -1,33 +1,45 @@
 context("ICD-10 comorbidity calculations")
 test_that("simplest test case for a NA/factor interaction - factor", {
-  d <- data.frame(visit = c("1", "1"), icd10 = c("not_in_map", "D638"),
-                  stringsAsFactors = TRUE)
+  d <- data.frame(
+    visit = c("1", "1"), icd10 = c("not_in_map", "D638"),
+    stringsAsFactors = TRUE
+  )
   res <- icd10_comorbid_ahrq(d)
   expect_true(res[, "Anemia"])
 })
 test_that("simplest test case for a NA/factor interaction - str", {
-  d <- data.frame(visit = c("1", "1"), icd10 = c("not_in_map", "D638"),
-                  stringsAsFactors = FALSE)
+  d <- data.frame(
+    visit = c("1", "1"), icd10 = c("not_in_map", "D638"),
+    stringsAsFactors = FALSE
+  )
   res <- icd10_comorbid_ahrq(d)
   expect_true(res[, "Anemia"])
 })
 test_that("simplest test case for a NA/factor interaction - real", {
-  d <- data.frame(visit = c(1, 1), icd10 = c("not_in_map", "D638"),
-                  stringsAsFactors = FALSE)
+  d <- data.frame(
+    visit = c(1, 1), icd10 = c("not_in_map", "D638"),
+    stringsAsFactors = FALSE
+  )
   res <- icd10_comorbid_ahrq(d)
   expect_true(res[, "Anemia"])
 })
 test_that("simplest test case for a NA/factor interaction - int", {
-  d <- data.frame(visit = c(1L, 1L), icd10 = c("not_in_map", "D638"),
-                  stringsAsFactors = FALSE)
+  d <- data.frame(
+    visit = c(1L, 1L), icd10 = c("not_in_map", "D638"),
+    stringsAsFactors = FALSE
+  )
   res <- icd10_comorbid_ahrq(d)
   expect_true(res[, "Anemia"])
 })
 test_that("simplest test case for a NA in codes with matching code", {
-  d <- data.frame(visit = c(1L, 1L), icd10 = c("D638", NA),
-                  stringsAsFactors = FALSE)
-  expect_error(regexp = NA,
-               res <- icd10_comorbid_ahrq(d, return_df = TRUE))
+  d <- data.frame(
+    visit = c(1L, 1L), icd10 = c("D638", NA),
+    stringsAsFactors = FALSE
+  )
+  expect_error(
+    regexp = NA,
+    res <- icd10_comorbid_ahrq(d, return_df = TRUE)
+  )
   expect_true(res[, "Anemia"])
 })
 test_that("ahrq comorbidities found for test data", {
@@ -35,8 +47,10 @@ test_that("ahrq comorbidities found for test data", {
   test_two <- icd10_all_ahrq_one_pt
   test_two$icd10_code <- as.icd10cm(paste0(icd10_all_ahrq_one_pt$icd10_code, "J2"))
   test_three <- icd10_all_ahrq_one_pt
-  test_three$icd10_code <- paste0(icd10_all_ahrq_one_pt$icd10_code,
-                                  random_string(nrow(test_three), max_chars = 5))
+  test_three$icd10_code <- paste0(
+    icd10_all_ahrq_one_pt$icd10_code,
+    random_string(nrow(test_three), max_chars = 5)
+  )
   td <- named_list(icd10_all_ahrq_one_pt, test_two, test_three)
   for (test_name in names(td)) {
     expect_error(res <- comorbid(td[[test_name]], map = icd10_map_ahrq), regexp = NA, info = test_name)
@@ -46,7 +60,8 @@ test_that("ahrq comorbidities found for test data", {
     for (n in colnames(res))
[…]
