+++ R/as_char_no_warn.R
@@ -6,9 +6,11 @@ as_char_no_warn <- function(x) {
   if (is.character(x)) return(x)
   old <- options(warn = -1)
   on.exit(options(old))
-  if (is.integer(x))
+  if (is.integer(x)) {
     return(fastIntToStringRcpp(x))
-  if (is.factor(x))
+  }
+  if (is.factor(x)) {
     return(levels(x)[x])
+  }
   as.character(x)
 }

[…]
