### 2019-03-17 b91073f
all tests pass, fixing style, eliminated Catch tests
