### 2019-03-17 b8998bb
fix some check errors, and trying going back to import instead of depend on icd.data
