+++ vignettes/country-lang-vers.Rmd
@@ -64,7 +64,7 @@ if (icd:::icd_data_ver_ok()) {
 ICD-10-FR is significantly different from the WHO version, and is used by official bodies and health systems in France. As with all the features in this document, this is an early attempt at the complex problem of providing all the codes, transtatlanguages, 
 
 ```{r french, eval = FALSE}
-fr <- icd:::get_from_icd_data("icd10fr2019", lazy = TRUE)
+fr <- icd:::get_from_icd_data("icd10fr2019")
 fr[fr$code == "C43", ]
 ```
 ```
