+++ DESCRIPTION
@@ -43,11 +43,11 @@ License: GPL-3
 URL: https://jackwasey.github.io/icd/
 BugReports: https://github.com/jackwasey/icd/issues
 Depends:
-    R (>= 3.4),
-    icd.data
+    R (>= 3.4)
 Imports:
     methods,
-    Rcpp (>= 0.12.3)
+    Rcpp (>= 0.12.3),
+    icd.data
 Suggests:
     graphics,
     knitr,
