+++ R/children.R
@@ -13,18 +13,18 @@
 #' @family ICD-9 ranges
 #' @examples
 #' library(magrittr, warn.conflicts = FALSE, quietly = TRUE) # optional
-#'
+#' 
 #' # no children other than self
 #' children("10201", short_code = TRUE, defined = FALSE)
-#'
+#' 
 #' # guess it was ICD-9 and a short, not decimal code
 #' children("0032")
-#'
+#' 
 #' # empty because 102.01 is not meaningful
 #' children("10201", short_code = TRUE, defined = TRUE)
 #' x <- children("003", short_code = TRUE, defined = TRUE)
 #' explain_code(x, condense = FALSE, short_code = TRUE)
-#'
+#' 
 #' children(short_code = FALSE, "100.0")
 #' children(short_code = FALSE, "100.00")
 #' children(short_code = FALSE, "2.34")
@@ -167,15 +167,18 @@ children_defined.icd10cm <- function(x,
   if (!short_code) {
     x <- decimal_to_short.icd10cm(x)
   }
-  ver <- ifelse(icd_data_ver_ok(), icd.data::get_icd10cm_active_ver(), "2016")
+  ver <- ifelse(icd_data_ver_ok(),
+    get_from_icd_data("get_icd10cm_active_ver")(),
+    "2016"
+  )
   if (verbose) message("Using ICD-10-CM version: ", ver)
-  #dat <- get_from_icd_data("icd10cm_active", alt = icd.data::icd10cm2016)
-  get_fun <- get_from_icd_data("get_icd10cm_version")
-  if (is.null(get_fun))
+  # dat <- get_from_icd_data("icd10cm_active", alt = icd.data::icd10cm2016)
+  get_fun <- get_from_icd_data("get_icd10cm_active", must_work = FALSE)
+  if (is.null(get_fun)) {
     dat <- icd.data::icd10cm2016
-  else
+  } else {
     dat <- get_fun()
-  #dat <- icd.data::get_icd10cm_version()
+  }
   nc <- .icd10cm_get_nchars(ver)
   kids <- icd10_children_defined_rcpp(
     x = x,
