+++ tests/testthat/test-comorbid-maps-icd9.R
@@ -56,7 +56,7 @@ test_that("icd9 comorbidities correct, logical to binary ok", {
 })
 
 test_that("condense big lists of comorbidities", {
-  skip_slow("Skipping condense big lists of comorbidities")
+  skip_slow("Skipping slow condense big lists of comorbidities")
   # this is a useful but slow (8s on my PC) test because the data weren't
   # generated by just expanding base ranges (which is how the condense works in
   # reverse)
