### 2019-03-18 bcef874
only evaluate chunks in country-lang-vers vignette if icd.data is up-to-date. doc.
