+++ vignettes/country-lang-vers.Rmd
@@ -23,6 +23,10 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+if (is.null(icd:::get_from_icd_data("icd10who2016", must_work = FALSE))) {
+  message("Cannot build this vignette with older version of icd")
+  icd_eval = FALSE
+}
 ```
 
 # WHO
@@ -37,7 +41,7 @@ icd.data::icd10fr2019[1:5, ]
 library(icd.data)
 summary(icd10who2016)
 ```
-```{r quicklookeval, echo = FALSE}
+```{r quicklookeval, echo = FALSE, eval = icd_eval}
 if (icd:::icd_data_ver_ok()) {
   icd:::get_from_icd_data("icd10who2016")[1:5, ]
   icd:::get_from_icd_data("icd10fr2019")[1:5, ]
@@ -71,28 +75,24 @@ There is an initial beta-quality mechanism for selecting a particular version of
   # icd.data::set_icd10cm_active_ver("2015")
 ```
 The following code is wrapped in some protections in case the latest icd.data isn't available (notably for CRAN checks).
-```{r icd10cmoption}
+```{r icd10cmoption, eval = icd_eval}
 if (icd:::icd_data_ver_ok()) {
   # The code "C4311" goes from being a leaf in 2018, to a parent in 2019
   with_icd10cm_version(ver = "2018",
-                       lang = "en",
                        icd::is_leaf("C4311")
   )
   with_icd10cm_version(ver = "2019", 
-                       lang = "en", 
                        icd::is_leaf("C4311")
   )
   # In 2018 the following code was not even defined, but in 2019 was a new child
   # code of "C4311"s
   with_icd10cm_version(ver = "2018", 
-                       lang = "en",
                        {
                          print(icd::is_defined("C43111"))
                          print(icd::is_leaf("C43111"))
                        }
   )
   with_icd10cm_version(ver = "2019", 
-                       lang = "en", 
                        icd::is_leaf("C43111")
   )
 }
