### 2019-03-18 12eb44a
require icd.data >=1.1, drop all the complicated code to handle icd.data 1.0
