+++ vignettes/ICD-10.Rmd
@@ -113,7 +113,7 @@ The AHRQ publishes an annually updated categorization of ICD-10-CM procedure cod
 
 ```{r setuppcs, echo = FALSE}
 nm = "code"
-pcs <- icd::icd10cm2018_pc
+pcs <- icd.data::icd10cm2018_pc
 ```
 ```{r icd10pcs}
 set.seed(1441)
