+++ .travis.yml
@@ -71,6 +71,8 @@ before_script:
  - sed -i'' '/\.covrignore/d' .Rbuildignore
  - sed -i'' '/\.lintr/d' .Rbuildignore
  - sed -i'' '/\benchmark/d' .Rbuildignore
+ # otherwise, icd.data is installed from CRAN, and after the github install???
+ - echo 'Remotes: jackwasey/icd.data' >> DESCRIPTION
  - Rscript -e "Rcpp::compileAttributes()" # make sure Rcpp generated code is updated, ideally this is not be checked in repo...
 
 after_success:
