+++ .travis.yml
@@ -72,9 +72,7 @@ before_script:
  - sed -i'' '/\.lintr/d' .Rbuildignore
  - sed -i'' '/\benchmark/d' .Rbuildignore
  # otherwise, icd.data is installed from CRAN, and after the github install???
- - pwd
- - ls
- - echo 'Remotes: jackwasey/icd.data' >> DESCRIPTION
+ - echo "Remotes:\ jackwasey/icd.data" >> DESCRIPTION
  - Rscript -e "Rcpp::compileAttributes()" # make sure Rcpp generated code is updated, ideally this is not be checked in repo...
 
 after_success:
