### 2019-03-18 cee97fc
travis, don't lint as we now use styler; don't add -march, so check log is cleaner; shouldn't need to drop covrignore from rbuildignore as covr runs from source directory.
