### 2019-03-18 c4caed5
skip some tests that require dynamically downloaded data
