+++ tests/testthat/test-build-maps-icd10.R
@@ -1,14 +1,25 @@
 context("build icd10 maps")
 
+skip_multi <- function() {
+  skip_on_appveyor()
+  skip_on_travis()
+  skip_on_cran()
+  if (!isFALSE(getOption("icd.data.offline", default = TRUE)))
+    skip("Offline")
+}
+
 test_that("the icd-10 quan elix comorbidity map is reproduced", {
+  skip_multi()
   expect_equivalent(icd10_map_quan_elix, icd10_generate_map_quan_elix(save_data = FALSE))
 })
 
 test_that("the icd-10 quan deyo comorbidity map is reproduced", {
+  skip_multi()
   expect_equivalent(icd10_map_quan_deyo, icd10_generate_map_quan_deyo(save_data = FALSE))
 })
 
 test_that("the icd-10 elix comorbidity map is reproduced", {
+  skip_multi()
   expect_equivalent(icd10_map_elix, icd10_generate_map_elix(save_data = FALSE))
 })
 
