+++ DESCRIPTION
@@ -77,3 +77,4 @@ Language: en-US
 LazyData: true
 LazyDataCompression: xz
 RoxygenNote: 6.1.1
+Remotes: jackwasey/icd.data
