### 2019-03-22 a703d77
again, avoid dependency on icd.data >=1.1
