+++ R/generate_mappings.R
@@ -467,10 +467,10 @@ icd10_generate_map_quan_elix <- function(save_data = TRUE, verbose = FALSE) {
   # which are not present (?anymore) in the ICD-10-CM 2016 list. In particular,
   # see C43 in Tumor.
   icd10_map_quan_elix <- apply_over_icd10cm_vers(quan_elix_raw,
-    verbose = verbose
+                                                 verbose = verbose
   )
   icd10_map_quan_elix <- apply_over_icd10who_vers(icd10_map_quan_elix,
-    verbose = verbose
+                                                  verbose = verbose
   )
   icd10_map_quan_elix <- lapply(icd10_map_quan_elix, as.short_diag)
   icd10_map_quan_elix <- lapply(icd10_map_quan_elix, as.icd10)
@@ -574,10 +574,10 @@ icd10_generate_map_quan_deyo <- function(save_data = TRUE, verbose = FALSE) {
   # worth it, even for ICD-10-CM, because I do end up cutting it back down to
   # size based on the input data before comorbidity matching.
   icd10_map_quan_deyo <- apply_over_icd10cm_vers(quan_charl_raw,
-    verbose = verbose
+                                                 verbose = verbose
   )
   icd10_map_quan_deyo <- apply_over_icd10who_vers(icd10_map_quan_deyo,
-    verbose = verbose
+                                                  verbose = verbose
   )
   icd10_map_quan_deyo <- lapply(icd10_map_quan_deyo, as.short_diag)
   icd10_map_quan_deyo <- lapply(icd10_map_quan_deyo, as.icd10)
@@ -604,31 +604,33 @@ apply_over_icd10cm_vers <- function(raw, verbose = FALSE) {
   out <- raw
   for (yr in 2014:2019) {
     if (verbose) message("applying ICD-10-CM year: ", yr)
-    icd.data::with_icd10cm_version(as.character(yr), {
-      upd <- sapply(out,
-        FUN = .apply_over_ver_worker,
-        simplify = FALSE,
-        USE.NAMES = TRUE,
-        verbose = verbose
-      )
-      for (cmb in seq_along(out)) {
-        if (verbose) {
-          only_prev <- setdiff(out[[cmb]], upd[[cmb]])
-          only_this <- setdiff(upd[[cmb]], out[[cmb]])
-          if (length(only_prev)) {
-            if (verbose) message("Year/version = ", yr)
-            message("Only in previous for item ", cmb)
-            print(only_prev)
+    icd.data::with_icd10cm_version(
+      as.character(yr),
+      code = {
+        upd <- sapply(out,
+                      FUN = .apply_over_ver_worker,
+                      simplify = FALSE,
+                      USE.NAMES = TRUE,
+                      verbose = verbose
+        )
+        for (cmb in seq_along(out)) {
+          if (verbose) {
+            only_prev <- setdiff(out[[cmb]], upd[[cmb]])
+            only_this <- setdiff(upd[[cmb]], out[[cmb]])
+            if (length(only_prev)) {
+              if (verbose) message("Year/version = ", yr)
+              message("Only in previous for item ", cmb)
+              print(only_prev)
+            }
+            if (length(only_this)) {
+              if (verbose) message("Year/version = ", yr)
+              message("Only in current for item ", cmb)
+              print(only_this)
+            }
           }
-          if (length(only_this)) {
-            if (verbose) message("Year/version = ", yr)
-            message("Only in current for item ", cmb)
-            print(only_this)
-          }
+          out[[cmb]] <- sort(union(out[[cmb]], upd[[cmb]]))
         }
-        out[[cmb]] <- sort(union(out[[cmb]], upd[[cmb]]))
-      }
[…]
