+++ R/children.R
@@ -13,18 +13,18 @@
 #' @family ICD-9 ranges
 #' @examples
 #' library(magrittr, warn.conflicts = FALSE, quietly = TRUE) # optional
-#' 
+#'
 #' # no children other than self
 #' children("10201", short_code = TRUE, defined = FALSE)
-#' 
+#'
 #' # guess it was ICD-9 and a short, not decimal code
 #' children("0032")
-#' 
+#'
 #' # empty because 102.01 is not meaningful
 #' children("10201", short_code = TRUE, defined = TRUE)
 #' x <- children("003", short_code = TRUE, defined = TRUE)
 #' explain_code(x, condense = FALSE, short_code = TRUE)
-#' 
+#'
 #' children(short_code = FALSE, "100.0")
 #' children(short_code = FALSE, "100.00")
 #' children(short_code = FALSE, "2.34")
@@ -206,9 +206,9 @@ children_defined.icd10who <- function(x,
     x <- decimal_to_short.icd10cm(x)
   }
   d <- if (who_ver == "icd10who2008fr") {
-    icd.data::icd10who2008fr
+    .idget("icd10who2008fr")
   } else {
-    icd.data::icd10who2016
+    .idget("icd10who2016")
   }
   stopifnot(!is.null(d))
   stopifnot(!is.null(d[["code"]]))
