+++ R/util.R
@@ -430,3 +430,10 @@ with_icd10cm_version <- function(ver, lang = c("en", "fr"), code) {
   on.exit(options(old))
   force(code)
 }
+
+require_icd_data <- function() {
+  if (!requireNamespace("icd.data", quietly = TRUE)) {
+    stop("Package \"icd.data\" needed for this function to work. Please install it with install.packages(\"icd.data\")",
+         call. = FALSE)
+  }
+}
