### 2019-03-22 8c9c068
avoid newer R idioms for better old R compatiblity
