+++ tests/testthat/test-build-maps-icd10.R
@@ -4,7 +4,7 @@ skip_multi <- function() {
   skip_on_appveyor()
   skip_on_travis()
   skip_on_cran()
-  if (!isFALSE(getOption("icd.data.offline", default = TRUE)))
+  if (isTRUE(getOption("icd.data.offline", default = TRUE)))
     skip("Offline")
 }
 
