+++ vignettes/ICD-10.Rmd
@@ -112,19 +112,19 @@ icd10_map_quan_elix$HTNcx
 The AHRQ publishes an annually updated categorization of ICD-10-CM procedure codes into four classes, representing diagnostic and therapeutic procedures, each being either minor or major.
 
 ```{r setuppcs, echo = FALSE}
-nm = "code"
-pcs <- icd:::.idget("icd10cm2018_pc", alt = "icd10_pcs_2018")
-```
-```{r icd10pcs}
-set.seed(1441)
-pts <- data.frame(id = sample(LETTERS, 10),
-                  pc = sample(pcs[[nm]], 10))
-res <- icd10_comorbid(pts,
-                      map = icd10_map_ahrq_pcs, 
-                      icd_name = "pc",
-                      return_binary = TRUE)
-print(res)
-colSums(res)
+nm <- "code"
+pcs <- icd:::.idget("icd10cm2018_pc", alt = "icd10_pcs_2018", must_work = FALSE)
+if (!is.null(pcs)) {
+  set.seed(1441)
+  pts <- data.frame(id = sample(LETTERS, 10),
+                    pc = sample(pcs[[nm]], 10))
+  res <- icd10_comorbid(pts,
+                        map = icd10_map_ahrq_pcs, 
+                        icd_name = "pc",
+                        return_binary = TRUE)
+  print(res)
+  colSums(res)
+}
 ```
 
 For more information on working with ICD-10 codes, see the introduction vignette, and function examples. E.g.
