### 2019-03-22 48e7ff8
docs and vignette dependency fix
