+++ man/icd-package.Rd
@@ -109,14 +109,17 @@ Calculate comorbidities, medical risk scores, and
   \item{Explanation and decoding}{
 
   Use \code{\link{explain_code}} to convert a list of codes into
-  human-readable descriptions. This function can optionally reduce the codes
-  to a their top-level groups if all the child members of a group are
-  present. \code{\link{diff_comorbid}} allows summary of the differences
-  between comorbidity mappings, e.g. to find what has changed from
-  year-to-year or between revisions by different authors.
-  \code{\link{icd9cm_hierarchy}} is a \code{data.frame} containing the full
-  ICD-9 classification for each diagnosis. \code{\link{icd9_chapters}}
-  contains definitions of chapters, sub-chapters and three-digit groups.} }
+  human-readable descriptions. This (and a number of other functions in this
+  package) relies on the \code{icd.data} package, which is not installed by
+  default. This function can optionally reduce the codes to a their top-level
+  groups if all the child members of a group are present.
+  \code{\link{diff_comorbid}} allows summary of the differences between
+  comorbidity mappings, e.g. to find what has changed from year-to-year or
+  between revisions by different authors.
+  \code{\link[icd.data]{icd9cm_hierarchy}} is a \code{data.frame} containing
+  the full ICD-9 classification for each diagnosis.
+  \code{\link[icd.data]{icd9_chapters}} contains definitions of chapters,
+  sub-chapters and three-digit groups.} }
 }
 \seealso{
 Useful links:
