### 2019-03-22 cdeb71e
looks like active bindings are cached in memory, so we cannot update an active binding using an option. (icd.data now loses icd10cm_active). Also safer option setting.
