### 2019-03-22 99fa303
style and fix subtle maybe significant bug in calling C++ code for ICD-10 children
