+++ R/datadocs.R
@@ -260,4 +260,3 @@ NULL
 #' icd10_map_ahrq_pcs[["Major Diagnostic"]][1:5]
 #' icd10_map_ahrq_pcs[["Minor Therapeutic"]][1:5]
 NULL
-
