+++ R/comorbid.R
@@ -74,10 +74,10 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #' vd <- wide_to_long(vermont_dx)
 #' # get first few rows and columns of Charlson comorbidities using Quan mapping
 #' comorbid_quan_deyo(vd)[1:5, 1:14]
-#'
+#' 
 #' # get summary AHRQ (based on Elixhauser) comorbidities for the Uranium data:
 #' summary(comorbid_ahrq(uranium_pathology))
-#'
+#' 
 #' pts <- icd_long_data(
 #'   visit_name = c("2", "1", "2", "3", "3"),
 #'   icd9 = c("39891", "40110", "09322", "41514", "39891")
@@ -91,10 +91,10 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'     "2011-01-04", "2011-01-04"
 #'   ))
 #' )
-#'
+#' 
 #' pt_hccs <- comorbid_hcc(pts, date_name = "date")
 #' head(pt_hccs)
-#'
+#' 
 #' pts10 <- icd_long_data(
 #'   visit_name = c("a", "b", "c", "d", "e"),
 #'   icd_name = c("I058", NA, "T82817A", "", "I69369"),
@@ -102,13 +102,13 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'     c("2011-01-01", "2011-01-02", "2011-01-03", "2011-01-03", "2011-01-03")
 #'   )
 #' )
-#'
+#' 
 #' icd10_comorbid(pts10, map = icd10_map_ahrq)
 #' # or if library(icd) hasn't been called first:
 #' icd::icd10_comorbid(pts10, map = icd::icd10_map_ahrq)
 #' # or most simply:
 #' icd::icd10_comorbid_ahrq(pts10)
-#'
+#' 
 #' # specify a simple custom comorbidity map:
 #' my_map <- list(
 #'   "malady" = c("100", "2000"),
@@ -596,8 +596,9 @@ apply_hier_ahrq <- function(cbd, abbrev_names = TRUE, hierarchy = TRUE) {
 #' between tidy data frames and clean, row-named matrices
 #' @param x matrix or data.frame of comorbidities
 #' @keywords internal
-cr <- function(x)
+cr <- function(x) {
   seq(from = 1 + is.data.frame(x), to = ncol(x))
+}
 
 # pre-compute and store in an environment in the package namespace
 .icd10cm_get_nchars <- function(year) {
