+++ vignettes/ICD-10.Rmd
@@ -115,17 +115,15 @@ The AHRQ publishes an annually updated categorization of ICD-10-CM procedure cod
 n <- 10
 nm <- "code"
 pcs_sample <- sample(unname(unlist(icd10_map_ahrq_pcs)), n)
-if (!is.null(pcs)) {
-  set.seed(1441)
-  pts <- data.frame(id = sample(LETTERS, n),
-                    pc = pcs_sample)
-  res <- icd10_comorbid(pts,
-                        map = icd10_map_ahrq_pcs, 
-                        icd_name = "pc",
-                        return_binary = TRUE)
-  print(res)
-  colSums(res)
-}
+set.seed(1441)
+pts <- data.frame(id = sample(LETTERS, n),
+                  pc = pcs_sample)
+res <- icd10_comorbid(pts,
+                      map = icd10_map_ahrq_pcs, 
+                      icd_name = "pc",
+                      return_binary = TRUE)
+print(res)
+colSums(res)
 ```
 
 For more information on working with ICD-10 codes, see the introduction vignette, and function examples. E.g.
