### 2019-03-24 a369258
mode function for exists and get in .idget
