### 2019-03-24 8972050
don't error if we can't get data from (non-existent) icd.data 1.1
