### 2019-03-24 ff728be
fixed logic in icd1.1 compatibility
