+++ R/util-test.R
@@ -264,12 +264,12 @@ skip_missing_icd10who <- function() {
       offline = TRUE,
       with_interact(
         interact = FALSE, {
-          dat <- .idget("icd10who2016",
+          dat <- .idget("get_icd10who2016",
             must_work = FALSE,
             mode = "function"
           )()
           if (is.null(dat)) testthat::skip("No WHO ICD-10 2016 English data")
-          dat <- .idget("icd10who2008fr",
+          dat <- .idget("get_icd10who2008fr",
             must_work = FALSE,
             mode = "function"
           )()
@@ -287,7 +287,7 @@ skip_missing_icd10fr <- function() {
       offline = TRUE,
       with_interact(
         interact = FALSE, {
-          dat <- .idget("icd10fr2019", mode = "function")()
+          dat <- .idget("get_icd10fr2019", mode = "function")()
           if (is.null(dat)) testthat::skip("No ICD-10-FR 2019 French data")
         }
       )
