+++ R/class.R
@@ -203,11 +203,11 @@ classes_ordered <- function(x) {
 #' x
 #' attributes(x) <- list(icd_short_diag = NULL)
 #' x
-#' 
+#'
 #' y <- as.decimal_diag(as.icd10("A10.09"))
 #' y
 #' is.short_diag(y)
-#' 
+#'
 #' j <- as.short_diag(as.icd10(c("A11", "B2222")))
 #' j[2] <- "C33"
 #' stopifnot(is.short_diag(j))
@@ -500,13 +500,16 @@ icd10be <- function(x) {
 #'   class.
 #' @family ICD data types
 #' @examples
-#' class(icd.data::uranium_pathology)
-#' class(icd.data::vermont_dx)
-#' icd_wide_data(
+#' (w <- icd_wide_data(
 #'   id = c(1, 2, 3),
 #'   dx01 = c("100", "441", "V20"),
 #'   dx02 = c("E9981", "V10", "44004")
-#' )
+#' ))
+#' wide_to_long(w)
+#' if (requireNamespace("icd.data", quietly = TRUE)) {
+#'   class(icd.data::uranium_pathology)
+#'   class(icd.data::vermont_dx)
+#' }
 #' @seealso \code{\link{long_to_wide}} and \code{\link{wide_to_long}}
 NULL
 
@@ -713,7 +716,7 @@ c.icd10 <- function(..., warn = FALSE) {
 #' # preserving the ICD class
 #' stopifnot(!inherits(x[[1]], "list"))
 #' stopifnot(!inherits(x[[1]][2], "list"))
-#' 
+#'
 #' y <- as.icd10(c("A01", "B0234"))
 #' y[2]
 #' y[[2]]
@@ -856,8 +859,8 @@ print.icd9 <- function(x, verbose = FALSE, ...)
 
 #' @rdname print.icd9
 #' @examples
-#' u <- icd.data::uranium_pathology[1:10, "icd10"]
 #' \dontrun{
+#' u <- icd.data::uranium_pathology[1:10, "icd10"]
 #' print(u)
 #' print(u, verbose = TRUE)
 #' # as.character will unclass the 'icd' classes
