### 2019-03-27 36b9608
passes check without icd.data, with icd.data 1.0 and 1.1.2
