+++ tests/test-all.R
@@ -1,6 +1,8 @@
 # this is common code to all the tests, each of which runs test_check with a different filter:
 library("icd")
-library("icd.data")
+if (!require("icd.data")) message("Testing without icd.data installed")
+if (!requireNamespace("icd.data", versionCheck = ">= 1.1"))
+  message("Testing without icd.data >= 1.1 installed")
 library("testthat", warn.conflicts = FALSE, quietly = TRUE)
 # Set ICD_TEST_SLOW to yes/true to run slow tests
 if (identical(Sys.getenv("NOT_CRAN"), "false")) {
