+++ R/factor_nosort.R
@@ -24,7 +24,7 @@
 #' identical(icd:::factor_nosort(x), x)
 #' # unless the levels change:
 #' icd:::factor_nosort(x, levels = c("a", "z"))
-#' 
+#'
 #' # existing factor levels aren't re-ordered without also moving elements
 #' f <- factor(c("a", "b", "b", "c"))
 #' g <- icd:::factor_nosort(f, levels = c("a", "c", "b"))
@@ -45,7 +45,7 @@ factor_nosort <- function(x, levels) {
 
 #' @describeIn factor_nosort R wrapper to the \pkg{Rcpp} function. Will
 #'   re-factor a factor with new levels without converting to string vector.
-#' @param na.rm Logical, if `TRUE`, simple drop all NA values, i.e., values with
+#' @param na.rm Logical, if \code{TRUE}, simple drop all NA values, i.e., values with
 #'   no corresponding level.
 #' @md
 #' @keywords internal
@@ -68,18 +68,18 @@ factor_nosort_rcpp <- function(x, levels, na.rm = FALSE) {
 #' Refactor by integer matching levels in C++
 #'
 #' Slightly slower for small factors, three times faster for one hundred million
-#' elements with two million new levels. Three times faster for any `n > 1e6`.
-#' With `NA` values, margin is smaller, but still beats base `factor`.
-#' @param exclude_na Simpler equivalent to `base::factor` exclude. By default,
-#'   `refactor` will not count `NA` as a factor level if there are `NA` elements
-#'   in the input data. As with `base::factor`, if `exclude_na` is `TRUE`, an
-#'   `NA` level explicitly requested in the `levels` argument is still dropped.
-#' @param validate Single logical value, if `TRUE`, the input factor is checked
+#' elements with two million new levels. Three times faster for any \code{n > 1e6}.
+#' With \code{NA} values, margin is smaller, but still beats base \code{factor}.
+#' @param exclude_na Simpler equivalent to \code{base::factor} exclude. By default,
+#'   \code{refactor} will not count \code{NA} as a factor level if there are \code{NA} elements
+#'   in the input data. As with \code{base::factor}, if \code{exclude_na} is \code{TRUE}, an
+#'   \code{NA} level explicitly requested in the \code{levels} argument is still dropped.
+#' @param validate Single logical value, if \code{TRUE}, the input factor is checked
 #'   for consistency. This should almost never be a problem (i.e., for any
 #'   factors generated by R, or those carefully generated by the user). However,
 #'   invalid factor input has the possibility of breaking assumptions in the C++
-#'   code. The default is `FALSE`. For huge data sets, and known valid input,
-#'   skipping this check saves significant time. `icd::factor_is_valid` can be
+#'   code. The default is \code{FALSE}. For huge data sets, and known valid input,
+#'   skipping this check saves significant time. \code{icd::factor_is_valid} can be
 #'   used independently.
 #' @examples
 #' \dontrun{
