### 2019-03-27 fd019aa
appveyor, and better handling of possibly missing data from icd.data
