+++ R/util-test.R
@@ -101,8 +101,8 @@ random_string <- function(n, max_chars = 4) {
     sample(c(LETTERS, letters, 0:9, rep("", times = 50)), replace = TRUE, size = n)
 
   v <- vapply(1:max_chars,
-    FUN = function(x) rand_ch(),
-    FUN.VALUE = character(n)
+              FUN = function(x) rand_ch(),
+              FUN.VALUE = character(n)
   )
   apply(v, 1, paste0, collapse = "")
 }
@@ -260,43 +260,38 @@ with_absent_action <- function(absent_action, code) {
   on.exit(options(old), add = TRUE)
   force(code)
 }
-# workaround so icd.data 1.0 will not cause CRAN or user errors
-skip_missing_icd10who <- function() {
+
+is_missing_icd_data <- function(var_name) {
+  if (!icd_data_ver_ok()) return(TRUE)
   with_absent_action(
     absent_action = "silent",
     with_offline(
       offline = TRUE,
       with_interact(
         interact = FALSE, {
-          dat <- .idget("get_icd10who2016",
-            must_work = FALSE,
-            mode = "function"
-          )()
-          if (is.null(dat)) testthat::skip("No WHO ICD-10 2016 English data")
-          dat <- .idget("get_icd10who2008fr",
-            must_work = FALSE,
-            mode = "function"
-          )()
-          if (is.null(dat)) testthat::skip("No WHO ICD-10 2008 French data")
+          d <- .idget(var_name, must_work = FALSE)
+          i <- if (is.function(d))
+            d()
+          else
+            d
+          is.null(i)
         }
       )
     )
   )
 }
 
+# workaround so icd.data 1.0 will not cause CRAN or user errors
+skip_missing_icd10who <- function() {
+  if (is_missing_icd_data("icd10who2016"))
+    testthat::skip("No WHO ICD-10 2016 English data")
+  if (is_missing_icd_data("icd10who2008fr"))
+    testthat::skip("No WHO ICD-10 2008 French data")
+}
+
 skip_missing_icd10fr <- function() {
-  with_absent_action(
-    absent_action = "silent",
-    with_offline(
-      offline = TRUE,
-      with_interact(
-        interact = FALSE, {
-          dat <- .idget("get_icd10fr2019", mode = "function")
-          if (is.null(dat)) testthat::skip("No ICD-10-FR 2019 French data")
-        }
-      )
-    )
-  )
+  if (is_missing_icd_data("icd10fr2019"))
+    testthat::skip("No ICD-10-FR 2019 French data")
 }
 
 set_test_slow <- function(do_slow = TRUE) {
