### 2019-03-28 aea125d
fix a test that only fails on rhub/ubuntu-gcc-release, bump version, NEWS
