### 2019-03-28 278eda5
again fix a test which only fails (due to localization/ordering on one rhub image), re-styled, exported S3 internals, as recommended.
