+++ tests/testthat/test-valid-icd10.R
@@ -49,17 +49,17 @@ test_that("icd10 codes that are not billable but do exist", {
     test_that("icd.data required for billable/defined/leaf tests", {
       skip_if_not_installed("icd.data")
       expect_false(is_billable(x), info = x)
-    expect_false(is_billable(x, short_code = TRUE), info = x)
-    expect_false(is_billable.icd10(x), info = x)
-    expect_false(is_billable.icd10(x, short_code = TRUE), info = x)
-    expect_false(is_billable.icd10cm(x), info = x)
-    expect_false(is_billable.icd10cm(x, short_code = TRUE), info = x)
-    expect_true(is_defined(x), info = x)
-    expect_true(is_defined(x, short_code = TRUE), info = x)
-    expect_true(is_defined.icd10(x), info = x)
-    expect_true(is_defined.icd10(x, short_code = TRUE), info = x)
-    expect_true(is_defined.icd10cm(x), info = x)
-    expect_true(is_defined.icd10cm(x, short_code = TRUE), info = x)
+      expect_false(is_billable(x, short_code = TRUE), info = x)
+      expect_false(is_billable.icd10(x), info = x)
+      expect_false(is_billable.icd10(x, short_code = TRUE), info = x)
+      expect_false(is_billable.icd10cm(x), info = x)
+      expect_false(is_billable.icd10cm(x, short_code = TRUE), info = x)
+      expect_true(is_defined(x), info = x)
+      expect_true(is_defined(x, short_code = TRUE), info = x)
+      expect_true(is_defined.icd10(x), info = x)
+      expect_true(is_defined.icd10(x, short_code = TRUE), info = x)
+      expect_true(is_defined.icd10cm(x), info = x)
+      expect_true(is_defined.icd10cm(x, short_code = TRUE), info = x)
     })
   }
 })

[…]
