### 2019-03-28 1b93775
fix test with updated get_visit_name internal function
