+++ R/convert.R
@@ -263,7 +263,7 @@ comorbid_df_to_mat <- function(x, visit_name = get_visit_name(x),
 
 #' Convert ICD codes from short to decimal forms
 #'
-#' Convert codes between short and decimal forms
+#' Convert from short to decimal forms of ICD codes.
 #' @param x ICD codes
 #' @export
 #' @keywords manip
@@ -316,7 +316,7 @@ short_to_decimal.icd10cm <- function(x) {
 #' Convert Decimal format ICD codes to short format
 #'
 #' This usually just entails removing the decimal point, but also does some
-#' limited validation and tidying up.
+#' limited validation and tidying up. Missing leading zeroes will be added for correctness of the shortened codes.
 #'
 #' @param x ICD codes
 #' @export
