+++ man/comorbid_pccc_dx.Rd
@@ -33,10 +33,9 @@ icd10_comorbid_pccc_pcs(x, visit_name = get_visit_name(x), icd_name,
 }
 \arguments{
 \item{x}{\code{data.frame} containing a column of patient-visit identifiers
-and a column of ICD codes. The \code{data.frame} should be in \sQuote{long}
-format, like the example \code{\link[icd.data]{vermont_dx}} data. If it is
-in \sQuote{wide} format, it must be converted to \sQuote{long} using
-\code{\link{wide_to_long}} before calling any comorbidity functions.}
+and a column of ICD codes. The \code{data.frame} may be in \sQuote{long} or
+\sQuote{wide} format, like the example \code{\link[icd.data]{vermont_dx}}
+and \code{\link[icd.data]{uranium_pathology}} data.}
 
 \item{visit_name}{The name of the column in the data frame which contains the
 patient or visit identifier. Typically this is the visit identifier, since

[…]
