+++ man/generate_neds_pts.Rd
@@ -23,17 +23,17 @@ Generate simulated 'NEDS' data for 'PCCC' and bigger wide data testing
 }
 \examples{
 if (requireNamespace("icd.data", quietly = TRUE)) {
-summary(icd::comorbid_pccc_dx(icd:::generate_neds_pts()))
-neds <- icd:::generate_neds_pts(n = 100, ncol = 10L, icd10 = FALSE)
-stopifnot(dim(neds) == c(100L, 11L))
-summary(icd::comorbid_pccc_dx(neds))
+  summary(icd::comorbid_pccc_dx(icd:::generate_neds_pts()))
+  neds <- icd:::generate_neds_pts(n = 100, ncol = 10L, icd10 = FALSE)
+  stopifnot(dim(neds) == c(100L, 11L))
+  summary(icd::comorbid_pccc_dx(neds))
 }
 \dontrun{
 if (requireNamespace("icd.data", quietly = TRUE)) {
-# original size data for PCCC benchmarking:
-set.seed(1441)
-neds <- icd:::generate_neds_pts(28584301L)
-neds_comorbid <- icd::comorbid_pccc_dx(neds)
+  # original size data for PCCC benchmarking:
+  set.seed(1441)
+  neds <- icd:::generate_neds_pts(28584301L)
+  neds_comorbid <- icd::comorbid_pccc_dx(neds)
 }
 }
 }
