+++ man/categorize_simple.Rd
@@ -56,10 +56,10 @@ pure data manipulation and generalizable beyond medical data.
 }
 \examples{
 if (requireNamespace("icd.data", quietly = TRUE)) {
-u <- icd.data::uranium_pathology
-m <- icd10_map_ahrq
-u$icd10 <- decimal_to_short(u$icd10)
-j <- categorize_simple(u, m, id_name = "case", code_name = "icd10")
+  u <- icd.data::uranium_pathology
+  m <- icd10_map_ahrq
+  u$icd10 <- decimal_to_short(u$icd10)
+  j <- categorize_simple(u, m, id_name = "case", code_name = "icd10")
 }
 }
 \keyword{internal}
