+++ man/diff_comorbid.Rd
@@ -52,13 +52,13 @@ diff_comorbid(icd10_map_elix, icd10_map_ahrq, show = FALSE)[["CHF"]]
 \dontrun{
 # default is to show the results in a human readable manner:
 if (requireNamespace("icd.data", quietly = TRUE)) {
-diff_result <- diff_comorbid(icd9_map_elix, icd9_map_ahrq)[["CHF"]]
-# show differences for
-# give full report on all comorbidities for these mappings
-diff_result <- diff_comorbid(icd9_map_elix, icd9_map_ahrq, show = FALSE)
+  diff_result <- diff_comorbid(icd9_map_elix, icd9_map_ahrq)[["CHF"]]
+  # show differences for
+  # give full report on all comorbidities for these mappings
+  diff_result <- diff_comorbid(icd9_map_elix, icd9_map_ahrq, show = FALSE)
 
-# the following outputs a summary to the console:
-diff_comorbid(icd9_map_elix, icd9_map_ahrq)
+  # the following outputs a summary to the console:
+  diff_comorbid(icd9_map_elix, icd9_map_ahrq)
 }
 }
 }
