### 2019-03-28 d5a3a9a
rebased on 4.0.9 peri-cran submission
