+++ DESCRIPTION
@@ -75,4 +75,4 @@ Encoding: UTF-8
 Language: en-US
 LazyData: true
 LazyDataCompression: xz
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
