### 2019-03-29 7b08edb
docs, and we don't want to distribute all the raw data in the source package. Try to make a test more forwards compatible.
