### 2019-03-29 0e0c0c2
use .Rbuildignore to avoid two heavy dependencies by not building efficiency as part of package (but still include the actual vignette. Brought back PCCC vignette.
