+++ vignettes/PCCC.Rmd
@@ -1,0 +1,64 @@
+---
+title: "Pediatric Complex Chronic Conditions"
+author: "Jack O. Wasey"
+date: "`r Sys.Date()`"
+output:
+  html_document:
+    toc: yes
+vignette: >
+  %\VignetteIndexEntry{Pediatric Complex Chronic Conditions}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteDepends{icd.data}
+  %\VignetteEncoding{UTF-8}
+bibliography: icdpkg.bib
+---
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+library("icd")
+library("icd.data")
+n <- 10000
+set.seed(1441)
+nm <- "code"
+i <- if (icd:::icd_data_ver_ok()) {
+  icd:::get_from_icd_data("icd10cm2016_pc")
+} else {
+  nm <- "pcs"
+  icd:::get_from_icd_data("icd10_pcs_2016")
+}
+dat <- data.frame(
+  id = n + seq(n),
+  icd_dx1 = sample(icd.data::icd10cm2016$code, n, replace = TRUE),
+  icd_dx2 = sample(icd.data::icd10cm2016$code, n, replace = TRUE),
+  icd_pcs1 = sample(i[[nm]], n, replace = TRUE),
+  icd_pcs2 = sample(i[[nm]], n, replace = TRUE)
+)
+```
+
+# Pediatric Complex Chronic Conditions
+
+## Introduction
+
+The adult comorbidities defined by authors such as Charlson, Quan and Elixhuaser are not suitable for pediatric research. Feudtner and others produced comorbidity definitions suitable for infants and children, with mappings defined for both ICD-9 and ICD-10 codes [@Feinstein_packagepediatriccomplex_2018; Feudtner_Pediatriccomplexchronic_2014]. These are now included in the [icd package](https://cran.r-project.org/package=icd) for R, which enables [fast calculation of comorbidities](introduction.html) for PCCC and other ICD code to comorbidity maps.
+
+## Calculating PCCC comorbidities.
+
+The PCCC comorbidities may be derived from both diagnostic and procedure codes. The following simulated data has both:
+
+```{r headdat}
+head(dat)
+```
+The data is in 'wide' format. `icd` no longer needs the data to be in 'long' format. The PCCC comorbidity classes can be computed directly from the source data and summarized in a plot. Here we calculate the PCCC from ICD diagnostic codes and ICD procedure codes separately, and logically combine them to get the result.
+```{r calc}
+pccc_dx <- comorbid_pccc_dx(dat)
+pccc_pcs <- icd10_comorbid_pccc_pcs(dat,
+                                    icd_name = c("icd_pcs1", "icd_pcs2"))
+res <- pccc_dx | pccc_pcs
+```
+
+```{r pcccbarplot, fig.asp=0.5, echo=FALSE}
+barplot(sort(colSums(res), decreasing = TRUE),
+        ylim = c(0, signif(max(colSums(res)), 1)),
+        ylab = "count", las = 2)
+```
+
+# References
