### 2019-03-29 3f7fa05
re-introduce simple PCCC vignette, restyle, make PCCC comorbidity result column names more consistent and allow full or abbrevatied names.
