+++ tests/testthat/test-pccc.R
@@ -1,38 +1,29 @@
 context("PCCC")
-pccc_col_names <- c(
-  "neuromusc", "cvd", "respiratory", "renal", "gi",
-  "hemato_immu", "metabolic", "congeni_genetic",
-  "malignancy", "neonatal", "tech_dep", "transplant"
-)
-# Six random codes from each PCCC procedure code map. 'icd' will use
-# an heuristic to guess whether ICD-9 or ICD-10:
-pccc_pts <- data.frame(
-  encounters = c(10, 11, 12),
-  icd9_dx = c("31800", "41610", "27701"),
-  icd10_dx = c("K50", "B20", "E702"),
-  icd9_pcs = c("0152", "304", "0050"),
-  icd10_pcs = c("0B110Z4", "02YA0Z2", "031209D")
-)
+# pccc_col_names <- c(
+#   "neuromusc", "cvd", "respiratory", "renal", "gi",
+#   "hemato_immu", "metabolic", "congeni_genetic",
+#   "malignancy", "neonatal", "tech_dep", "transplant"
+# )
 
 test_that("procedure codes work", {
   res9 <- comorbid_pccc_pcs(pccc_pts,
     icd_name = "icd9_pcs",
     return_binary = FALSE
   )
-  expect_true(res9[1, "neuromusc"])
-  expect_true(res9[3, "cvd"])
-  expect_true(res9["11", "respiratory"])
-  expect_true(res9[3, "tech_dep"])
+  expect_true(res9[1, "Neuromusc"])
+  expect_true(res9[3, "CVD"])
+  expect_true(res9["11", "Respiratory"])
+  expect_true(res9[3, "TechDep"])
   expect_equal(sum(res9), 4)
   res0 <- comorbid_pccc_pcs(pccc_pts,
     icd_name = "icd10_pcs",
     return_binary = FALSE
   )
-  expect_true(res0["11", "cvd"])
-  expect_true(res0["10", "respiratory"])
+  expect_true(res0["11", "CVD"])
+  expect_true(res0["10", "Respiratory"])
   expect_true(res0[3, 11])
   expect_true(res0[3, 4])
-  expect_true(res0[2, "transplant"])
+  expect_true(res0[2, "Transplant"])
   expect_equal(sum(res0), 5)
 
   # All ICD-9 procedure codes are numeric, some ICD-10 procedure codes
@@ -48,7 +39,7 @@ test_that("PCCC dx works", {
     colSums(res[1:1000, ]),
     c(82, 270, 50, 119, 55, 39, 313, 30, 128, 7, 129, 21)
   )
-  expect_equal(colnames(res), pccc_col_names)
+  expect_equal(colnames(res), unlist(unname(names_pccc_abbrev)))
 })
 
 test_that("colnames same for both dx and procedure codes", {
