+++ R/comorbid_pccc.R
@@ -28,8 +28,10 @@ comorbid_pccc_dx <- function(x,
       icd9 = icd9_comorbid_pccc_dx,
       icd10 = icd10_comorbid_pccc_dx
     ),
-    visit_name = visit_name, icd_name = icd_name,
-    short_code = short_code, return_df = return_df,
+    visit_name = visit_name,
+    icd_name = icd_name,
+    short_code = short_code,
+    return_df = return_df,
     return_binary = return_binary, ...
   )
 }
@@ -54,45 +56,46 @@ comorbid_pccc_dx <- function(x,
 #' @family comorbidity computations
 #' @family comorbidities
 #' @export
-comorbid_pccc_pcs <- function(x,
-                              visit_name = get_visit_name(x),
-                              icd_name,
-                              return_df = FALSE,
-                              return_binary = FALSE,
-                              ...) {
-  stopifnot(visit_name %in% names(x), icd_name %in% names(x))
-  n <- min(10L, length(x[[icd_name]]))
-  test_some <- as.character(x[seq_len(n), icd_name])
-  nines_tens <- grepl(".*[A-Za-z].*", test_some, ignore.case = TRUE)
-  is_icd9 <- TRUE
-  threshold <- 0.7
-  if (sum(nines_tens) / n > threshold) {
-    is_icd9 <- FALSE
-  } else if (sum(nines_tens) / n > (1 - threshold)) {
-    warning(
-      "many invalid ICD-9 procedure codes, but not enough to ",
-      "determine that all the codes were ICD-10 codes. Please",
-      "check the input data. Assuming ICD-9."
-    )
+comorbid_pccc_pcs <-
+  function(x,
+             visit_name = get_visit_name(x),
+             icd_name,
+             return_df = FALSE,
+             return_binary = FALSE,
+             ...) {
+    stopifnot(visit_name %in% names(x), icd_name %in% names(x))
+    n <- min(10L, length(x[[icd_name]]))
+    test_some <- as.character(x[seq_len(n), icd_name])
+    nines_tens <- grepl(".*[A-Za-z].*", test_some, ignore.case = TRUE)
+    is_icd9 <- TRUE
+    threshold <- 0.7
+    if (sum(nines_tens) / n > threshold) {
+      is_icd9 <- FALSE
+    } else if (sum(nines_tens) / n > (1 - threshold)) {
+      warning(
+        "many invalid ICD-9 procedure codes, but not enough to ",
+        "determine that all the codes were ICD-10 codes. Please",
+        "check the input data. Assuming ICD-9."
+      )
+    }
+    if (is_icd9) {
+      icd9_comorbid_pccc_pcs(x,
+        visit_name = visit_name,
+        icd_name = icd_name,
+        return_df = return_df,
+        return_binary = return_binary,
+        ...
+      )
+    } else {
+      icd10_comorbid_pccc_pcs(x,
+        visit_name = visit_name,
+        icd_name = icd_name,
+        return_df = return_df,
+        return_binary = return_binary,
+        ...
+      )
+    }
   }
-  if (is_icd9) {
-    icd9_comorbid_pccc_pcs(x,
-      visit_name = visit_name,
-      icd_name = icd_name,
-      return_df = return_df,
-      return_binary = return_binary,
-      ...
-    )
-  } else {
-    icd10_comorbid_pccc_pcs(x,
-      visit_name = visit_name,
-      icd_name = icd_name,
-      return_df = return_df,
-      return_binary = return_binary,
-      ...
-    )
-  }
-}
 
 #' @describeIn comorbid_pccc_dx Calculate PCCC comorbidities from ICD-9
 #'   diagnosis codes
[…]
