### 2019-03-29 13e4835
better determination if icd.data has the requested data, let's PCCC vignette build
