+++ R/icd.data-1.0-compat.R
@@ -37,32 +37,47 @@ icd_data_get_icd10cm_active_ver <- function() {
   }
 }
 
+# alt is alternative data, not a variable name.
 .idget <- function(var_name, alt = NULL, must_work = is.null(alt), ...) {
   # this should only happen in weird R CMD check internals since we Import
   # icd.data...
-  if (!icd_data_ver_ok()) {
+  if (!.idexists_either(var_name)) {
     msg <- paste(
       "This data is not available with icd.data version < 1.1",
       "Upgrade icd.data using install.packages(\"icd.data\"."
     )
     if (must_work) stop(msg, " Cannot proceed.")
-    return(.idget(alt))
+    return(alt)
   }
-  # bindings are in the top level namespace
-  # lazy data will only be at top level if package is attached
+  .idget_either(var_name)
+}
+
+.idget_lazy <- function(var_name) {
   ns <- asNamespace("icd.data")
-  if (exists(var_name, ns, ...)) return(get(var_name, ns, ...))
   lz <- ns$.__NAMESPACE__.$lazydata
-  if (exists(var_name, lz, ...)) return(get(var_name, lz, ...))
-  if (!exists(var_name, ns, ...) && substring(var_name, 1, 4) != "get_") {
-    var_name <- paste0("get_", var_name)
-  }
-  g <- get(var_name, ns, ...)
-  if (is.function(g)) {
-    g()
-  } else {
-    g
-  }
+  get(var_name, lz)
+}
+
+.idexists_lazy <- function(var_name) {
+  ns <- asNamespace("icd.data")
+  lz <- ns$.__NAMESPACE__.$lazydata
+  exists(var_name, lz)
+}
+
+.idget_either <- function(var_name) {
+  ns <- asNamespace("icd.data")
+  if (.idexists_lazy(var_name)) return(.idget_lazy(var_name))
+  if (exists(var_name, ns)) return(get(var_name, ns))
+  var_name <- paste0("get_", var_name)
+  if (exists(var_name, ns)) return(get(var_name, ns, mode = "function")())
+  stop(var_name, " not available in icd.data regular or lazy data")
+}
+
+.idexists_either <- function(var_name) {
+  ns <- asNamespace("icd.data")
+  .idexists_lazy(var_name) ||
+    exists(var_name, ns) ||
+    exists(paste0("get_", var_name), ns, mode = "function")
 }
 
 with_icd10cm_version <- function(ver, code) {
