+++ vignettes/PCCC.Rmd
@@ -20,10 +20,10 @@ n <- 10000
 set.seed(1441)
 nm <- "code"
 i <- if (icd:::icd_data_ver_ok()) {
-  icd:::.idget("icd10cm2016_pc")
+  icd:::.idget("icd10cm2016_pc", must_work = FALSE)
 } else {
   nm <- "pcs"
-  icd:::.idget("icd10_pcs_2016")
+  icd:::.idget("icd10_pcs_2016", must_work = FALSE)
 }
 dat <- data.frame(
   id = n + seq(n),
