### 2019-03-30 b7a83e5
Drop internal documentation from package, and add a function to plot comorbid results, alongside plotting comorbidity results directly from patient data.
