+++ R/comorbid.R
@@ -67,7 +67,6 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'   and the underlying algorithm used. Usually this would be the order of the
 #'   first occurrence of each visit/patient identifier, but this is not
 #'   guaranteed unless \code{restore_id_order} is set to \code{TRUE}.
-#' @md
 #' @family comorbidity computations
 #' @family comorbidities
 #' @examples
@@ -78,17 +77,17 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'   # get first few rows and columns of Charlson comorbidities using Quan/Deyo
 #'   # mapping of ICD-9 or ICD-10 codes Charlson categories
 #'   comorbid_quan_deyo(icd.data::vermont_dx)[1:5, 1:14]
-#' 
+#'
 #'   # Note that the comorbidity calculations automatically finds the ICD code
 #'   # columns, and uses 'wide' or 'long' format data.
-#' 
+#'
 #'   stopifnot(
 #'     identical(
 #'       comorbid_quan_deyo(icd.data::vermont_dx),
 #'       comorbid_quan_deyo(wide_to_long(icd.data::vermont_dx))
 #'     )
 #'   )
-#' 
+#'
 #'   # get summary AHRQ (based on Elixhauser) comorbidities for the Uranium data:
 #'   summary(comorbid_ahrq(icd.data::uranium_pathology))
 #' }
@@ -105,10 +104,10 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'     "2011-01-04", "2011-01-04"
 #'   ))
 #' )
-#' 
+#'
 #' pt_hccs <- comorbid_hcc(pts, date_name = "date")
 #' head(pt_hccs)
-#' 
+#'
 #' pts10 <- icd_long_data(
 #'   visit_name = c("a", "b", "c", "d", "e"),
 #'   icd_name = c("I058", NA, "T82817A", "", "I69369"),
@@ -116,13 +115,13 @@ poa_choices <- c("yes", "no", "notYes", "notNo")
 #'     c("2011-01-01", "2011-01-02", "2011-01-03", "2011-01-03", "2011-01-03")
 #'   )
 #' )
-#' 
+#'
 #' icd10_comorbid(pts10, map = icd10_map_ahrq)
 #' # or if library(icd) hasn't been called first:
 #' icd::icd10_comorbid(pts10, map = icd::icd10_map_ahrq)
 #' # or most simply:
 #' icd::icd10_comorbid_ahrq(pts10)
-#' 
+#'
 #' # specify a simple custom comorbidity map:
 #' my_map <- list(
 #'   "malady" = c("100", "2000"),
