### 2019-03-30 1cf46b2
various tweaks and documentation
 - don't need all.bib workaround with rmarkdown >1.1
 - strip library on Mac and Linux for big reduction esp in latter
 - correct vignette references
 - update intermittently failing doi url.
