+++ vignettes/icdpkg.bib
@@ -105,7 +105,6 @@
 @article{Feinstein_packagepediatriccomplex_2018,
   author = {Feinstein, James A. and Russell, Seth and DeWitt, Peter E. and Feudtner, Chris and Dai, Dingwei and Bennett, Tellen D.},
   date = {2018-04-23},
-  doi = {10.1001/jamapediatrics.2018.0256},
   eprint = {29710063},
   eprinttype = {pmid},
   issn = {2168-6211},
@@ -116,6 +115,7 @@
   pmcid = {PMC5933455},
   shortjournal = {JAMA Pediatr},
   title = {R Package for Pediatric Complex Chronic Condition Classification},
+  url = {https://jamanetwork.com/journals/jamapediatrics/article-abstract/2677901},
   volume = {172}
 }
 % == BibLateX quality report for Feinstein_packagepediatriccomplex_2018:
