+++ vignettes/introduction.Rmd
@@ -5,6 +5,8 @@ date: "`r Sys.Date()`"
 output:
   html_document:
     toc: yes
+bibliography:
+  - icdpkg.bib
 vignette: >
   %\VignetteIndexEntry{An Introduction to icd}
   %\VignetteEngine{knitr::rmarkdown}
