+++ DESCRIPTION
@@ -53,7 +53,7 @@ Suggests:
     knitr,
     magrittr,
     nhds,
-    rmarkdown,
+    rmarkdown (>= 1.11),
     roxygen2 (>= 5.0.0),
     stats,
     testthat (>= 0.11.1),
