### 2019-03-30 e0db5c8
readme; don't push Rmd sources to the vignettes directory in source package. Makevars requires tabs! vignettes use html_vignette. For now, don't fix margins in plot_comorbid, because it messes up the html_vignette processing.
