+++ man/plot_comorbid.Rd
@@ -7,7 +7,7 @@
 \usage{
 plot_comorbid(x, comorbid_fun = icd::comorbid_ahrq, ...)
 
-plot_comorbid_results(x, sort = TRUE, fix_margin = TRUE, las = 2,
+plot_comorbid_results(x, sort = TRUE, fix_margin = FALSE, las = 2,
   cex.names = 0.75, ...)
 }
 \arguments{
