+++ vignettes/introduction.Rmd
@@ -3,7 +3,7 @@ title: 'An Introduction to icd'
 author: "Jack O. Wasey"
 date: "`r Sys.Date()`"
 output:
-  html_document:
+  html_vignette:
     toc: yes
 bibliography:
   - icdpkg.bib
