+++ vignettes/CMS-HCC.Rmd
@@ -5,7 +5,7 @@ author:
 - Jack Wasey (minor updates)
 date: "`r Sys.Date()`"
 output:
-  html_document:
+  html_vignette:
     toc: yes
 vignette: >
   %\VignetteIndexEntry{Using Hierarchical Condition Codes}
