+++ README.Rmd
@@ -1,5 +1,5 @@
 ---
-title: icd Readme
+title: icd README
 output:
   md_document:
     variant: gfm
@@ -51,7 +51,6 @@ patient_data <- data.frame(
 [![CRAN](https://www.r-pkg.org/badges/version/icd "CRAN")](https://cran.r-project.org/package=icd)
 [![lifecycle](https://img.shields.io/badge/lifecycle-stable-blue.svg)](https://www.tidyverse.org/lifecycle/#stable)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![GitHub](https://img.shields.io/badge/devel%20version-3.4-blue.svg?style=flat "GitHub")](https://github.com/jackwasey/icd)
 [![Travis](https://travis-ci.org/jackwasey/icd.svg?branch=master "Travis Build Status")](https://travis-ci.org/jackwasey/icd)
 [![Appveyor](https://ci.appveyor.com/api/projects/status/9ncfgxht3n5i8t60/branch/master?svg=true "Appveyor Build Status")](https://ci.appveyor.com/project/jackwasey/icd/branch/master)
 [![codecov.io](https://codecov.io/github/jackwasey/icd/coverage.svg?branch=master "Core Code Coverage")](https://codecov.io/github/jackwasey/icd?branch=master)
@@ -75,19 +74,16 @@ cat(packageDescription("icd")$Description)
 
  * find comorbidities of patients based on ICD-9 or ICD-10 codes, e.g. Cancer, Heart Disease
      + several standard mappings of ICD codes to comorbidities are included (Quan, Deyo, Elixhauser, AHRQ, PCCC)
-     + *very fast* assignment of ICD codes to comorbidities (using novel matrix multiplication algorithm and C++ internally)
- * use your existing data format, minimizing requirements for pre-processing
- * summarize groups of ICD codes in natural language
+     + *very fast* assignment of ICD codes to comorbidities (using novel matrix multiplication algorithm and C++ internally -- see 'efficiency' vignette for details)
+ * use your existing wide or long data format, icd can guess which columns are ICD-9 or ICD-10 codes.
+ * explain and summarize groups of ICD codes in natural language, using ICD editions from the WHO, USA, France and Belgium. Many different annual editions of these data are available, all via the 'icd.data' companion package.
  * Charlson and Van Walraven score calculations
  * Hierarchical Condition Codes (HCC) from CMS
  * Clinical Classifcations Software (CCS) comorbidities from AHRQ
  * Pediatric Complex Chronic Condition comorbidities
  * AHRQ ICD-10 procedure code classification
- * annual revisions of ICD-9-CM and ICD-10-CM
  * correct conversion between different representations of ICD codes, with and without a decimal points, leading and trailing characters (this is not trivial for ICD-9-CM). ICD-9 to ICD-10 cross-walk is not yet implemented
  * comprehensive test suite to increase confidence in accurate processing of ICD codes
- * all internal ICD and comorbidity data is extracted directly from public data or code, allowing end-to-end reproducibility
- * used, tested and benchmarked against other comorbidity calculators on hardware from laptops to big servers
 
 ## Examples
 
@@ -110,7 +106,9 @@ charlson(patient_data)
 plot_comorbid(icd.data::uranium_pathology)
 ```
 
-## Make "Table 1" summary data
+## Comorbodities example: make "Table 1" summary data
+
+A common requirement for medical research involving patients is determining new or existing comorbidities. This is often reported in *Table 1* of research papers to demonstrate the similarity or differences of groups of patients. This package is focussed on fast and accurate generation of this comorbidity information from raw lists of ICD-9 and ICD-10 codes.
 
[…]
