### 2019-04-01 67925a6
various
condense factor S3, tidy ignores, start to use S3 generic in tests (with view to migrating to icd.data when appropriate), vignette figure sizes better.
