+++ tests/testthat/test-ranges.R
@@ -38,7 +38,7 @@ test_that("expand icd9 range definition", {
     c("40100", "40101")
   )
   expect_equal_no_icd(
-    expand_range.icd9("40108", "40109", short_code = TRUE, defined = FALSE),
+    expand_range("40108", "40109", short_code = TRUE, defined = FALSE),
     c("40108", "40109")
   )
   expect_equal_no_icd(
@@ -66,13 +66,13 @@ test_that("expand icd9 range definition", {
 
   expect_equal(
     expand_range(short_code = TRUE, "401", "401", defined = FALSE),
-    sort(children.icd9("401", short_code = TRUE, defined = FALSE))
+    sort(children("401", short_code = TRUE, defined = FALSE))
   )
   # expand range should already be sorted. do i want to sort children by default
   # or with an option?
   expect_equal(
     expand_range(short_code = TRUE, "401", "402", defined = FALSE),
-    sort(children.icd9(c("401", "402"), short_code = TRUE, defined = FALSE))
+    sort(children(c("401", "402"), short_code = TRUE, defined = FALSE))
   )
   # the next two cases cover the HIV ranges in the co-morbidities, wherein the
   # final code is included, in which case the parent ("044" in this case) is
@@ -187,7 +187,7 @@ test_that("V code with ambiguous parent", {
   # descendants just to reach the specified codes, but not all the children of
   # the higher-level code.
   expect_equal_no_icd(
-    expand_range.icd9(
+    expand_range(
       short_code = TRUE, "V10", "V1001",
       defined = FALSE,
       ex_ambig_start = FALSE,
@@ -196,19 +196,19 @@ test_that("V code with ambiguous parent", {
     c("V10", "V100", "V1000", "V1001")
   )
   expect_equal_no_icd(
-    expand_range.icd9(short_code = TRUE, "V10", "V1001", defined = FALSE),
+    expand_range(short_code = TRUE, "V10", "V1001", defined = FALSE),
     c("V1000", "V1001")
   )
 })
 
 test_that("V code ranges", {
   expect_equal_no_icd(
-    expand_range.icd9(short_code = TRUE, "V1000", "V1002", defined = FALSE),
+    expand_range(short_code = TRUE, "V1000", "V1002", defined = FALSE),
     c("V1000", "V1001", "V1002")
   )
   # but we cap off the upper range correctly:
   expect_equal_no_icd(
-    expand_range.icd9(short_code = TRUE, "V1009", "V101", defined = FALSE),
+    expand_range(short_code = TRUE, "V1009", "V101", defined = FALSE),
     c(
       "V1009", "V101", "V1010", "V1011",
       "V1012", "V1013", "V1014", "V1015",
@@ -217,7 +217,7 @@ test_that("V code ranges", {
   )
   # and with narrower top end
   expect_equal_no_icd(
-    expand_range.icd9(
+    expand_range(
       short_code = TRUE, "V1009", "V1011",
       defined = FALSE, ex_ambig_start = TRUE,
       ex_ambig_end = TRUE
@@ -225,7 +225,7 @@ test_that("V code ranges", {
     c("V1009", "V1010", "V1011")
   )
   expect_equal_no_icd(
-    expand_range.icd9(
+    expand_range(
       short_code = TRUE, "V1009", "V1011",
       defined = FALSE, ex_ambig_start = FALSE,
       ex_ambig_end = FALSE
@@ -234,7 +234,7 @@ test_that("V code ranges", {
   )
   # but include those pesky parents when requested:
   expect_true(
[…]
