+++ tools/install_full.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 set -euo pipefail
 IFS=$'\n\t'
+${ICD_HOME:-$HOME/icd}/tools/build.sh
+R CMD INSTALL --no-clean-on-error \
+	--debug \
+       	--install-tests \
+	"$(ls -t icd_*.tar.gz | head -1)"
+# ${ICD_HOME:-$HOME/icd}
 
-#source ${ICD_HOME:-$HOME/icd}/tools/install_shared.sh
-R CMD INSTALL --no-clean-on-error -d --install-tests ${ICD_HOME:-$HOME/icd}
-#"$(ls -t icd*gz | head -1)"
