+++ man-roxygen/ver.R
@@ -1,0 +1,5 @@
+#' @param ver Character version of ICD-10-CM to use, corresponding to the suffix
+#'   of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use
+#'   \code{"icd10cm2019"} for Dutch 2014 ICD-CM translations, use
+#'   'icd10cm2014_nl"
+
