### 2019-04-03 fe45851
icd and icd.data reunion, and minor bug fixes and improvements
order/sort handle NA values in order/sort, and avoid possible memory corruption
tests all pass, check has issues
