### 2019-04-04 57434c9
make R CMD check not warn
beef up travis, including valgrind
fix ?last icd.data transition function names
drop offline argument now a global
document
