+++ tests/test-all.R
@@ -23,4 +23,3 @@ on.exit(options(old_interact), add = TRUE)
 icd:::.show_options()
 
 testthat::test_check("icd")
-
