+++ man/get_icd10cm2014_pc.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datadocs.R
+\docType{data}
+\name{get_icd10cm2014_pc}
+\alias{get_icd10cm2014_pc}
+\alias{get_icd10cm2015_pc}
+\alias{get_icd10cm2016_pc}
+\alias{get_icd10cm2017_pc}
+\alias{get_icd10cm2018_pc}
+\alias{get_icd10cm2019_pc}
+\title{ICD-10-CM Procedure codes for 2014}
+\description{
+ICD-10-CM Procedure codes for 2014
+}
+\keyword{datasets}
