+++ NAMESPACE
@@ -194,10 +194,15 @@ export(get_icd10be2014_pc)
 export(get_icd10be2017)
 export(get_icd10be2017_pc)
 export(get_icd10cm2014)
+export(get_icd10cm2014_pc)
 export(get_icd10cm2015)
+export(get_icd10cm2015_pc)
 export(get_icd10cm2016)
+export(get_icd10cm2016_pc)
 export(get_icd10cm2017)
+export(get_icd10cm2017_pc)
 export(get_icd10cm2018)
+export(get_icd10cm2018_pc)
 export(get_icd10cm2019)
 export(get_icd10cm2019_pc)
 export(get_icd10cm_active)
