+++ .travis.yml
@@ -3,7 +3,8 @@ language: r
 # https://ropensci.org/blog/2016/07/12/travis-osx
 # https://docs.travis-ci.com/user/ci-environment/
 
-fortran: false
+# for RcppEigen to compile:
+fortran: true
 
 os:
   - linux
@@ -16,6 +17,11 @@ sudo: false
 cache:
   - packages
   - ccache
+
+addons:
+  apt:
+    packages:
+      - valgrind
 
 r_packages:
   - covr
@@ -54,22 +60,6 @@ matrix:
   - r: devel
     os: linux
     dist: xenial
-  - r: devel
-    os: osx
-    osx_image: xcode10.2
-  - r: devel
-    os: osx
-    # as of April 2019, default osx image is 9.4
-  - r: release
-    os: linux
-    dist: precise
-  - r: release
-    os: osx
-    latex: false
-    r_build_args: --no-build-vignettes --no-manual --resave-data=no
-    r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
-    env:
-    - ICD_MINI_BENCH=true
   - r: oldrel
     os: linux
     dist: precise
@@ -86,6 +76,22 @@ matrix:
     - ICD_COVERAGE=true
     - ICD_TEST_SLOW=true
   allow_failures:
+  - r: devel
+    os: osx
+    osx_image: xcode10.2
+  - r: devel
+    os: osx
+    # as of April 2019, default osx image is 9.4
+  - r: release
+    os: linux
+    dist: precise
+  - r: release
+    os: osx
+    latex: false
+    r_build_args: --no-build-vignettes --no-manual --resave-data=no
+    r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
+    env:
+    - ICD_MINI_BENCH=true
   - r: oldrel
     latex: false
     r_build_args: --no-build-vignettes --no-manual --resave-data=no
