### 2019-04-04 676a6da
appveyor, fix PCCC again to tolerate no procedure codes, better functions to use sys env vars for options, passes tests.
