+++ tests/testthat/test-build-maps-icd10.R
@@ -25,6 +25,7 @@ test_that("the icd-10 elix comorbidity map is reproduced", {
 })
 
 test_that("icd-10 ahrq map is reproduced", {
+  skip_if_offline()
   if (is.null(icd10_fetch_ahrq_sas())) {
     skip("AHRQ ICD-10 SAS must be downloaded with icd10_fetch_ahrq_sas")
   }
