+++ R/update.R
@@ -53,7 +53,7 @@ update_everything <- function() {
   icd10_generate_map_quan_deyo(save_data = TRUE)
   icd10_generate_map_elix(save_data = TRUE)
   generate_maps_pccc(save_data = TRUE)
-  icd10_parse_ahrq_pcs(save_data = TRUE)
+  icd10_parse_map_ahrq_pc(save_pkg_data = TRUE)
   icd_parse_cc_hierarchy(save_data = TRUE)
 }
 
