+++ R/datadocs.R
@@ -273,9 +273,9 @@ NULL
 #'   columns, the first, named \code{code} is the procedure code; the second,
 #'   named \code{desc}, has the description.
 #' @name get_icd10cm2019_pc
-#' @aliases get_icd10cm2014_pc get_icd10cm2015_pc get_icd10cm2016_pc
-#'   get_icd10cm2017_pc get_icd10cm2018_pc
-#' @seealso
+#' @seealso \code{get_icd10cm2014_pc} \code{get_icd10cm2015_pc}
+#'   \code{get_icd10cm2016_pc} \code{get_icd10cm2017_pc}
+#'   \code{get_icd10cm2018_pc}
 #' \url{https://www.cms.gov/Medicare/Coding/ICD10/downloads/pcs_refman.pdf}
 #' @docType data
 #' @keywords datasets
@@ -284,8 +284,9 @@ NULL
 
 #' ICD-10-CM Procedure codes for 2014
 #' @name get_icd10cm2014_pc
-#' @aliases get_icd10cm2015_pc get_icd10cm2016_pc get_icd10cm2017_pc
-#'   get_icd10cm2018_pc get_icd10cm2019_pc
+#' @seealso \code{get_icd10cm2015_pc} \code{get_icd10cm2016_pc}
+#'   \code{get_icd10cm2017_pc} \code{get_icd10cm2018_pc}
+#'   \code{get_icd10cm2019_pc}
 #' @docType data
 #' @keywords datasets
 #' @export
@@ -293,8 +294,9 @@ NULL
 
 #' ICD-10-CM Procedure codes for 2015
 #' @name get_icd10cm2015_pc
-#' @aliases get_icd10cm2014_pc get_icd10cm2016_pc get_icd10cm2017_pc
-#'   get_icd10cm2018_pc get_icd10cm2019_pc
+#' @seealso \code{get_icd10cm2014_pc} \code{get_icd10cm2016_pc}
+#'   \code{get_icd10cm2017_pc} \code{get_icd10cm2018_pc}
+#'   \code{get_icd10cm2019_pc}
 #' @docType data
 #' @keywords datasets
 #' @export
@@ -302,8 +304,9 @@ NULL
 
 #' ICD-10-CM Procedure codes for 2016
 #' @name get_icd10cm2016_pc
-#' @aliases get_icd10cm2014_pc get_icd10cm2015_pc get_icd10cm2017_pc
-#'   get_icd10cm2018_pc get_icd10cm2019_pc
+#' @seealso \code{get_icd10cm2014_pc} \code{get_icd10cm2015_pc}
+#'   \code{get_icd10cm2017_pc} \code{get_icd10cm2018_pc}
+#'   \code{get_icd10cm2019_pc}
 #' @docType data
 #' @keywords datasets
 #' @export
@@ -311,8 +314,9 @@ NULL
 
 #' ICD-10-CM Procedure codes for 2017
 #' @name get_icd10cm2017_pc
-#' @aliases get_icd10cm2014_pc get_icd10cm2015_pc get_icd10cm2016_pc
-#'   get_icd10cm2018_pc get_icd10cm2019_pc
+#' @seealso \code{get_icd10cm2014_pc} \code{get_icd10cm2015_pc}
+#'   \code{get_icd10cm2016_pc} \code{get_icd10cm2018_pc}
+#'   \code{get_icd10cm2019_pc}
 #' @docType data
 #' @keywords datasets
 #' @export
@@ -320,8 +324,9 @@ NULL
 
 #' ICD-10-CM Procedure codes for 2018
 #' @name get_icd10cm2018_pc
[…]
