### 2019-04-04 0f1b393
travis sed for macos, and correctly allow testing of available data without stopping.
