### 2019-04-04 011b2d0
drop old vignettedepends (which r-devel picks up) bench refactor now Rmd; travis
