+++ benchmarks/bench-refactor.Rmd
@@ -1,9 +1,21 @@
+---
+title: "Benchmark refactoring"
+author: "Jack O. Wasey"
+date: "`r Sys.Date()`"
+output: html_document
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
 library(icd)
 library(microbenchmark)
-##########################
-# factoring and matching #
-##########################
+library(bench)
+library(stringr)
+```
 
+## match with Rcpp
+
+```{r}
 n = 1e7
 random_short_icd10_codes <- sample(unlist(icd::icd10_map_elix),
                                    replace = TRUE, size = n)
@@ -11,11 +23,10 @@ lookup <- unique(unname(unlist(icd10_map_quan_deyo)))
 microbenchmark(match(random_short_icd10_codes, lookup),
                icd:::match_rcpp(random_short_icd10_codes, lookup),
                times = 10)
-# yep, Rcpp is twice to thrice as fast for ten million items.
+```
 
-
-# benchmarking factor generation with known levels and no sorting:
-# getting unique levels is fast, 0.1 seconds for 1e7 codes
+## benchmarking factor generation with known levels and no sorting: getting unique levels is fast, 0.1 seconds for 1e7 codes
+```{r}
 system.time(lvls <- unique(random_short_icd10_codes))
 microbenchmark::microbenchmark(
   icd:::factor_nosort_rcpp(random_short_icd10_codes, lvls),
@@ -23,11 +34,11 @@ microbenchmark::microbenchmark(
   factor(random_short_icd10_codes, lvls),
   times = 5
 )
-# Rcpp is three times as fast.
+```
 
-# an even fast way might be to use the existing pointer values in the global
-# string cache directly...
+## factor_nosort
 
+```{r factor_nosort_basic_usage}
 x <- c("z", "a", "123")
 icd:::factor_nosort(x)
 # should return a factor without modification
@@ -35,46 +46,45 @@ x <- as.factor(x)
 identical(icd:::factor_nosort(x), x)
 # unless the levels change:
 icd:::factor_nosort(x, levels = c("a", "z"))
-
 # existing factor levels aren't re-ordered without also moving elements
 f <- factor(c("a", "b", "b", "c"))
 g <- icd:::factor_nosort(f, levels = c("a", "c", "b"))
 stopifnot(g[4] == "c")
-pts <- icd:::random_unordered_patients(1e6)
+pts <- icd:::generate_random_unordered_pts(1e6)
 u <- unique.default(pts$code)
-# this shows that stringr (which uses stringi) sort takes 50\% longer than
-# built-in R sort.
+```
+
+## stringr::str_sort vs base R sort
+
+```{r}
 microbenchmark::microbenchmark(sort(u), str_sort(u))
+```
 
-# this shows that \\code{factor_nosort} is about 50\% faster than \\code{factor}
-# for big vectors of strings
+## refactor
 
-# without sorting is much faster:
+```{r}
 microbenchmark::microbenchmark(factor(pts$code),
-                               # factor_(pts$code),
-                               factor_nosort(pts$code),
+                               icd:::factor_nosort(pts$code),
+                               icd:::factor_nosort_rcpp(pts$code),
                                times = 25)
+```
 
+## refactor complex
 
-################################
-# benchmark refactoring in C++ #
-################################
-
-# see tests-slow to show the results are identical
-
[…]
