### 2019-04-04 d9c44d1
simplify build matrix on travis, gcc on mac try [skip appveyor]
