+++ .travis.yml
@@ -3,7 +3,7 @@ language: r
 # https://ropensci.org/blog/2016/07/12/travis-osx
 # https://docs.travis-ci.com/user/ci-environment/
 
-# for RcppEigen to compile:
+# for RcppEigen to compile, (but disable if brew gcc on mac to avoid conflict)
 fortran: true
 
 # os:
@@ -22,6 +22,8 @@ fortran: true
 # and
 #
 # https://docs.travis-ci.com/user/languages/cpp/
+
+# TODO: "eval"" the benchmark step to error if it fails with bad return code.
 sudo: false
 cache:
   - packages
@@ -81,6 +83,7 @@ matrix:
     compiler: gcc
   - r: release
     os: osx
+    fortran: false
     # see before_install for use of MATRIX_EVAL
     env:
       - MATRIX_EVAL="brew install gcc && CC=gcc-8 && CXX=g++-8"
@@ -91,21 +94,9 @@ matrix:
   - r: devel
     os: osx
     # as of April 2019, default osx image is 9.4
-  - r: release
+  - r: oldrel
     os: linux
     dist: precise
-  - r: release
-    os: osx
-    # see before_install for use of MATRIX_EVAL
-    env:
-      - MATRIX_EVAL="brew install gcc && CC=gcc-8 && CXX=g++-8"
-      - ICD_MINI_BENCH=true
-  - r: devel
-    os: linux
-  - r: devel
-    os: osx
-    compiler: clang
-  - r: oldrel
     latex: false
     r_build_args: --no-build-vignettes --no-manual --resave-data=no
     r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
@@ -127,16 +118,12 @@ matrix:
     - ICD_MINI_BENCH=true
     - ICD_TEST_SLOW=true
   allow_failures:
+  # allow any osx with r-devel to fail for now: error prone tool chain
   - r: devel
     os: osx
-    osx_image: xcode10.2
-  - r: devel
-    os: osx
-    # as of April 2019, default osx image is 9.4
-  - r: release
+  - r: oldrel
     os: linux
     dist: precise
-  - r: oldrel
     latex: false
     r_build_args: --no-build-vignettes --no-manual --resave-data=no
     r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual
