### 2019-04-04 6f180e2
travis
should fail if mini-bench fails, put osx gcc in makevars, only install covr when needed [skip appveyor]

and note to self in JSS replication
