### 2019-04-04 bb2ab30
CXX11 for GCC on OSX in travis [skip appveyor]
