### 2019-04-04 8ce06ea
simplify travis build matrix, use gcc and clang on linux builds [skip appveyor]
