### 2019-04-04 fdda7da
travis and avoid cran note
