### 2019-04-05 37682c9
drop icd9cm_billable completely, and improve documentation formatting.
