### 2019-04-05 bb3145f
various
avoid (intended) fallthrough warning in a C++ switch
let test/check work with no options or environment variables set
styled
make icd9cm014_leaf package data, as it is used all over the place (formerly icd9cm_billable[["32"]]
(gcc-8 finds a memory access error in icd9MajMinToShort...)
data resaved more compressed using tools::
