+++ tests/test-all.R
@@ -1,15 +1,8 @@
-# this is common code to all the tests, each of which runs test_check with a different filter:
+# this is common code to all the tests, each of which runs test_check with a
+# different filter:
 library("icd")
 library("testthat", warn.conflicts = FALSE, quietly = TRUE)
-# Set ICD_TEST_SLOW to yes/true to run slow tests
-# if (identical(Sys.getenv("NOT_CRAN"), "false")) {
-#   options(icd.data.offline = TRUE)
-#   options(icd.data.absent_action = "stop")
-#   # ensure we don't write a temporary directory we don't clean up
-#   options(icd.data.resource = NA)
-# }
-
-# Don't download data on CRAN
+# Definitely don't download data on CRAN
 if (!icd:::.env_var_is_true("NOT_CRAN")) {
   old_offline <- options("icd.data.offline" = TRUE)
   on.exit(options(old_offline), add = TRUE)
@@ -21,5 +14,11 @@ if (icd:::.env_var_is_true("ICD_DATA_TEST_SLOW")) {
 old_interact <- options("icd.data.interact" = FALSE)
 on.exit(options(old_interact), add = TRUE)
 icd:::.show_options()
-
+stopifnot(
+  with_icd10cm_version(
+    ver = "2014",
+    code = icd:::get_icd10cm_active_ver() == "2014"
+  )
+)
+testthat::test_check("icd", filter = "icd10cm-ver", reporter = "summary")
 testthat::test_check("icd")
