### 2019-04-05 055c8e2
travis mac gcc, tidy up some C++ code, still have mem access error on gcc-8 macos (only)
