### 2019-04-06 cbeb1cc
travis gcc mac reinstall Rcpp, re-enable openmp
[skip appveyor]
