+++ .travis.yml
@@ -91,8 +91,9 @@ matrix:
     fortran: false
     # just to avoid matrix doing this twice, state gcc here
     compiler: gcc
+    # ?have to reinstall Rcpp with compiler and libc++ -> libstdc++ change
     env:
-      - MATRIX_EVAL="brew install gcc && mkdir -p ~/.R && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars"
+      - MATRIX_EVAL="brew install gcc && mkdir -p ~/.R && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars && Rscript -e 'install.packages(\"Rcpp\")'"
       - ICD_MINI_BENCH=true
   # r-devel currently failing to download for macos, so release for now.
   # also see https://docs.travis-ci.com/user/reference/osx
@@ -101,7 +102,7 @@ matrix:
     osx_image: xcode10.2
     compiler: gcc
     env:
-      - MATRIX_EVAL="brew install gcc && mkdir -p ~/.R && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars"
+      - MATRIX_EVAL="brew install gcc && mkdir -p ~/.R && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars && Rscript -e 'install.packages(\"Rcpp\")'"
   - r: release
     os: osx
     # as of April 2019, default osx image is 9.4
