### 2019-04-06 9f14542
rhub valrind avoid timeout by splitting testthat tests
and minor documentation change
[ci skip]
