### 2019-04-06 7b3b820
torture or valgrind script as functions [ci skip]
