### 2019-04-07 a7d7c67
configure, and roxygen2 version downgrade
configure script again for openmp macos, primarily, seems to work nicely
downgrade roxygen to 6.1.0 (which actually works with compiled code) https://github.com/klutometis/roxygen/issues/822
