+++ src/Makevars.in
@@ -3,9 +3,11 @@ CXX_STD=CXX11
 # maybe try without openmp, at least for MacOS testing, but this should work!
 # PKG_CPPFLAGS=-DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS
 
-PKG_CPPFLAGS=-DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS $(SHLIB_OPENMP_CXXFLAGS)
-PKG_LIBS=$(SHLIB_OPENMP_CXXFLAGS)
+#PKG_CPPFLAGS=-DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS $(SHLIB_OPENMP_CXXFLAGS)
+#PKG_LIBS=$(SHLIB_OPENMP_CXXFLAGS)
 
+PKG_CPPFLAGS=@ICD_OPENMP_FLAG@ -DEIGEN_PERMANENTLY_DISABLE_STUPID_WARNINGS
+PKG_LIBS=@ICD_OPENMP_FLAG@
 # size is about 10% of unstripped on Linux, about 90% on Mac. CRAN Windows DLL is already small.
 #
 # Don't forget to comment out for debugging...
