### 2019-04-08 3f42172
various
travis - try harder to make Rcpp with g++-8 for MacOS
cleanup script now we have configure again
new flags for better/quicker checking in tools.
