+++ tools/check_plus.sh
@@ -22,7 +22,7 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_CODE_DATA_INTO_GLOBALENV_=TRUE \
   _R_CHECK_CODE_USAGE_VIA_NAMESPACES_=TRUE \
   _R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE" \
-  _R_CHECK_COMPILATION_FLAGS_=TRUE \
+  _R_CHECK_COMPILATION_FLAGS_=FALSE \
   _R_CHECK_CRAN_INCOMING_=FALSE \
   _R_CHECK_CRAN_INCOMING_REMOTE_=FALSE \
   _R_CHECK_DEPRECATED_DEFUNCT_=TRUE \
@@ -33,6 +33,7 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_INSTALL_DEPENDS_=TRUE \
   _R_CHECK_NATIVE_ROUTINE_REGISTRATION_=TRUE \
   _R_CHECK_NO_RECOMMENDED_=TRUE \
+  _R_CHECK_NO_STOP_ON_TEST_ERROR_=TRUE \
   _R_CHECK_OVERWRITE_REGISTERED_S3_METHODS_=TRUE \
   _R_CHECK_PRAGMAS_=TRUE \
   _R_CHECK_R_DEPENDS_=warn \
