+++ tools/install_full.sh
@@ -2,9 +2,11 @@
 set -euo pipefail
 IFS=$'\n\t'
 ${ICD_HOME:-$HOME/icd}/tools/build.sh
-R CMD INSTALL --no-clean-on-error \
-	--debug \
-       	--install-tests \
-	"$(ls -t icd_*.tar.gz | head -1)"
-# ${ICD_HOME:-$HOME/icd}
-
+MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
+	R CMD INSTALL \
+		--no-clean-on-error \
+		--debug \
+		--install-tests \
+		--data-compress=none \
+		--no-resave_data \
+		"$(ls -t icd_*.tar.gz | head -1)"
