+++ tools/check_vquick.sh
@@ -20,10 +20,13 @@ R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
   _R_CHECK_CODE_ASSIGN_TO_GLOBALENV_=TRUE \
   _R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE" \
   _R_CHECK_COMPACT_DATA_=false \
+  _R_CHECK_COMPILATION_FLAGS_=FALSE \
   _R_CHECK_CRAN_INCOMING_=false \
   _R_CHECK_CRAN_INCOMING_REMOTE_=false \
   _R_CHECK_EXIT_ON_FIRST_ERROR_=TRUE \
   _R_CHECK_NATIVE_ROUTINE_REGISTRATION_=TRUE \
+  _R_CHECK_NO_STOP_ON_TEST_ERROR_=FALSE \
+  _R_CHECK_COMPILATION_FLAGS_=FALSE \
   _R_CHECK_RD_EXAMPLES_T_AND_F_=TRUE \
   _R_CHECK_RD_LINE_WIDTHS_=true \
   _R_CHECK_TESTS_NLINES_=0 \
@@ -34,6 +37,7 @@ R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
   _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
   R_MAKEVARS_USER=${HOME}/.R/Makevars.quick \
  R CMD check \
+   --no-vignettes \
    --no-build-vignettes \
    --no-manual \
    --ignore-vignettes \
