+++ tools/check_full.sh
@@ -21,10 +21,12 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_ALWAYS_LOG_VIGNETTE_OUTPUT_=TRUE \
   _R_CHECK_CODE_ASSIGN_TO_GLOBALENV_=TRUE \
   _R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE" \
+  _R_CHECK_COMPILATION_FLAGS_=FALSE \
   _R_CHECK_CRAN_INCOMING_=false \
   _R_CHECK_CRAN_INCOMING_REMOTE_=false \
   _R_CHECK_EXIT_ON_FIRST_ERROR_=TRUE \
   _R_CHECK_NATIVE_ROUTINE_REGISTRATION_=TRUE \
+  _R_CHECK_NO_STOP_ON_TEST_ERROR_=TRUE \
   _R_CHECK_RD_EXAMPLES_T_AND_F_=TRUE \
   _R_CHECK_RD_LINE_WIDTHS_=true \
   _R_CHECK_TESTS_NLINES_=0 \
