+++ tools/install_quick.sh
@@ -6,12 +6,14 @@ source ${ICD_HOME:-$HOME/rprojects/icd}/tools/build.sh
 R_MAKEVARS_USER="$HOME/.R/Makevars.quick" \
   MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   R CMD INSTALL -d \
+    --data-compress=none \
     --no-byte-compile \
     --no-resave-data \
     --no-clean-on-error \
-#    --library="$install_dir" \
     --install-tests \
     --no-docs \
     --no-build-vignettes \
+    --no-resave_data \
     "$(ls -t icd*.tar.gz | head -1)"
 
+    #    --library="$install_dir" \

[…]
