### 2019-04-08 2d07a99
travis
cran env
reduce omp threads
try skipping keeping of raw data
[skip appveyor]
