+++ .travis.yml
@@ -43,7 +43,9 @@ addons:
 #   - covr
 
 cran: https://cloud.r-project.org
-
+# by default, DON'T do --as-cran, as it stops us using environment variables
+# that conflict. Instead, state approximate CRAN check environment variables explicitly.
+r_check_args: ""
 warnings_are_errors: true
 
 # try to speed up compilation time
@@ -75,15 +77,29 @@ env:
     - _R_CHECK_LENGTH_1_CONDITION_="verbose,abort"
     - _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort"
     - _R_CHECK_NO_STOP_ON_TEST_ERROR_=FALSE
+    # the following is ignored if --as-cran is set
     - _R_CHECK_COMPILATION_FLAGS_=FALSE
+# the following are to simulate cran, and help debugging, esp if we didn't use --as-cran
+    - _R_CHECK_ALL_NON_ISO_C_=TRUE
+    - _R_CHECK_ALWAYS_LOG_VIGNETTE_OUTPUT_=TRUE
+    - _R_CHECK_CODE_ASSIGN_TO_GLOBALENV_=TRUE
+    - _R_CHECK_CODETOOLS_PROFILE_="suppressLocalUnused=FALSE"
+    - _R_CHECK_EXIT_ON_FIRST_ERROR_=TRUE
+    - _R_CHECK_NATIVE_ROUTINE_REGISTRATION_=TRUE
+    - _R_CHECK_NO_STOP_ON_TEST_ERROR_=TRUE
+    - _R_CHECK_RD_EXAMPLES_T_AND_F_=TRUE
+    - _R_CHECK_RD_LINE_WIDTHS_=true
+    - _R_CHECK_TESTS_NLINES_=0
+    - _R_CHECK_USE_INSTALL_LOG_=TRUE
+    - _R_CHECK_VIGNETTES_NLINES_=0
+    - _R_CHECK_PKG_SIZES_THRESHOLD_=50
     - ICD_COVERAGE=false
     - ICD_MINI_BENCH=false
     - ICD_DATA_VERBOSE=true
     - ICD_DATA_OFFLINE=true
     - ICD_DATA_INTERACT=false
-    # 4 is the official travis suggestion
-    - OMP_NUM_THREADS=4
-#    env: LINTR_COMMENT_BOT=true
+    # 4 is the official travis suggestion, but the docker images only have access to two cores/threads.
+    - OMP_NUM_THREADS=2
 
 matrix:
   fast_finish: false
@@ -157,8 +173,8 @@ matrix:
 
 before_script:
 # - rm .Rinstignore  # do install everything for testing (we have source tree, so no need)
- - if [ "$TRAVIS_OS_NAME" = "osx" ]; then sed -i '' '/.*[Dd]ata.*/d' .Rbuildignore; fi
- - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sed -i'' '/.*[Dd]ata.*/d' .Rbuildignore; fi
+# - if [ "$TRAVIS_OS_NAME" = "osx" ]; then sed -i '' '/.*[Dd]ata.*/d' .Rbuildignore; fi
+# - if [ "$TRAVIS_OS_NAME" = "linux" ]; then sed -i'' '/.*[Dd]ata.*/d' .Rbuildignore; fi
  # should fail and stop travis if the benchmarking fails.
  - if [ "ICD_MINI_BENCH" = "true" ]; then
      cd benchmarks/icd-JSS3447-replication;
@@ -169,8 +185,6 @@ before_script:
 after_success:
  - if [ "$ICD_COVERAGE" = "true" ]; then
[…]
