### 2019-04-08 cae0db3
gcc and clang should tolerate the warning supression on travis! [skip appveyor]
