### 2019-04-08 a3f1a00
travis tweak, and build should not have Makevars, just Makevars.in and configure script [skip appveyor]
