### 2019-04-09 9e1dd9a
various
workaround for travis/devtools/remotes https://github.com/r-lib/devtools/issues/2020#issuecomment-481219296
sort and refresh ignore files
roxygen2 github version (so it works)
cleanup exec permission
move m4 include to better location
check scripts ignore size, as CRAN mostly seems to
COPYRIGHTS include R/FSF openmp.m4
updated cran comments
strip binary to reduce package size (and probably improve performance)
