+++ tools/check_full.sh
@@ -27,13 +27,15 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_EXIT_ON_FIRST_ERROR_=TRUE \
   _R_CHECK_NATIVE_ROUTINE_REGISTRATION_=TRUE \
   _R_CHECK_NO_STOP_ON_TEST_ERROR_=TRUE \
+  _R_CHECK_PKG_SIZES_=false \
   _R_CHECK_RD_EXAMPLES_T_AND_F_=TRUE \
   _R_CHECK_RD_LINE_WIDTHS_=true \
   _R_CHECK_TESTS_NLINES_=0 \
   _R_CHECK_USE_INSTALL_LOG_=TRUE \
   _R_CHECK_VIGNETTES_NLINES_=0 \
-  _R_CHECK_PKG_SIZES_THRESHOLD_=50 \
+  _R_CHECK_PKG_SIZES_=false \
   _R_CHECK_LENGTH_1_CONDITION_="verbose,abort" \
   _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
  R CMD check "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
+#  _R_CHECK_PKG_SIZES_THRESHOLD_=50 \
