+++ tools/check_plus.sh
@@ -35,6 +35,7 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_NO_RECOMMENDED_=TRUE \
   _R_CHECK_NO_STOP_ON_TEST_ERROR_=TRUE \
   _R_CHECK_OVERWRITE_REGISTERED_S3_METHODS_=TRUE \
+  _R_CHECK_PKG_SIZES_=false \
   _R_CHECK_PRAGMAS_=TRUE \
   _R_CHECK_R_DEPENDS_=warn \
   _R_CHECK_R_ON_PATH_=TRUE \
