+++ DESCRIPTION
@@ -80,4 +80,4 @@ Encoding: UTF-8
 Language: en-US
 LazyData: true
 LazyDataCompression: xz
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.1.9000
