### 2019-04-09 3cd7bd8
actually confirm the benchmark code can run (but not benchmark) in travis [skip appveyor]
