### 2019-04-10 9b1c154
tweaks
travis should split tests so coverage doesn't time-out
fix https URL for AHRQ SAS code and rebuild the data
drop useless arguments
