+++ tests/testthat/helper-icd-data.R
@@ -65,7 +65,10 @@ skip_flat_icd9_all_avail <- function() {
 }
 
 skip_icd10cm_flat_avail <- function(year, dx = TRUE) {
-  msg <- "skipping test because flat file ICD-10-CM source not available"
+  msg <- paste(
+    "skipping test flat file ICD-10-CM source for ",
+    year, ", and dx = ", dx
+  )
   if (!.exists_in_cache(.get_icd10cm_name(year, dx))) {
     testthat::skip(msg)
   }
