+++ tools/split-tests.R
@@ -2,12 +2,17 @@
 
 # source from package root after setting `dry_run` to `FALSE`
 
-testthat_split <- function(base_name,
-                           dry_run,
-                           filter) {
+testthat_active_path <- function(base_name = "testthat") {
+  file.path("tests", paste0(base_name, ".R"))
+}
+
+testthat_disabled_path <- function(base_name = "testthat") {
+  file.path(paste0(base_name, "-disabled.R"))
+}
+
+testthat_split <- function(base_name = "testthat",
+                           filter = ".*") {
   prefix <- "testthat-split-"
-  test_all <- file.path("tests", paste0(base_name, ".R"))
-  test_all_disabled <- file.path(paste0(base_name, "-disabled.R"))
   test_files <- list.files(
     file.path("tests", "testthat"),
     pattern = "^test-"
@@ -17,7 +22,7 @@ testthat_split <- function(base_name,
       next
     test_name <- sub(".R", "", x = sub("test-", "", x = f))
     new_test <- file.path("tests", paste0(prefix, test_name, ".R"))
-    new_test_lines <- readLines(test_all)
+    new_test_lines <- readLines(testthat_active_path())
     test_check_line_num <- grep("^ *[^#]*(testthat::)?test_check.*",
                                 new_test_lines)
     test_check_line <- new_test_lines[test_check_line_num]
@@ -25,41 +30,43 @@ testthat_split <- function(base_name,
                            replacement = paste0(", filter = \"^", test_name, "$\")"),
                            test_check_line)
     new_test_lines[test_check_line_num] <- test_check_line
-    if (!dry_run) writeLines(new_test_lines, new_test)
+    writeLines(new_test_lines, new_test)
   }
-  if (!dry_run) ok <- file.rename(test_all, test_all_disabled)
+  ok <- file.rename(testthat_active_path(), testthat_disabled_path())
   stopifnot(ok)
+  invisible()
 }
 
 testthat_restore <- function(base_name = "testthat") {
-  test_all <- file.path("tests", paste0(base_name, ".R"))
-  test_all_disabled <- file.path(paste0(base_name, "-disabled.R"))
   prefix <- "testthat-split-"
-  file.rename(test_all_disabled, test_all)
+  file.rename(testthat_disabled_path(), testthat_active_path())
   ok <- all(file.remove(list.files("tests", pattern = paste0(prefix),
                                    full.names = TRUE)))
   stopifnot(ok)
+  invisible()
 }
 
 check_split_tests <- function(
   base_name = "testthat", # testthat is default from usethis_testthat
-  dry_run = TRUE,
-  restore = FALSE,
-  valgrind = TRUE,
+  valgrind = FALSE,
   torture = FALSE,
   filter = ".*",
-  check_args = c("--ignore-vignettes", "--no-vignettes", "--no-build-vignettes", "--no-manual", "--timings", "--no-codoc")) {
+  check_args = c("--ignore-vignettes",
+                 "--no-vignettes",
+                 "--no-build-vignettes",
+                 "--no-manual",
+                 "--timings",
+                 "--no-codoc")) {
   if (!requireNamespace("rcmdcheck")) install.packages("rcmdcheck")
   if (!requireNamespace("rhub")) install.packages("rhub")
-  if (restore) {
-    testthat_restore(base_name = base_name)
+  testthat_split(base_name = base_name,
+                 filter = filter)
+  if (valgrind) {
+    rhub::check_with_valgrind(check_args = check_args)
+  } else if (torture) {
+    rcmdcheck::rcmdcheck(args = c("--use-gct", check_args))
   } else {
-    testthat_split(base_name = base_name,
-                   dry_run = dry_run,
[…]
