+++ vignettes/country-lang-vers.Rmd
@@ -23,6 +23,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+library(icd)
 ```
 
 # WHO

[…]
