### 2019-04-10 8e297c7
travis
[skip appveyor]
and release qs
