+++ .travis.yml
@@ -111,9 +111,13 @@ matrix:
   - r: devel
     os: linux
     dist: xenial
+    env:
+      - ICD_MINI_BENCH=true
   - r: release
     os: linux
     dist: xenial
+    env:
+      - ICD_MINI_BENCH=true
   - r: release
     os: osx
     # this xcode so stuff like _stdio.h are available to GCC.
@@ -159,7 +163,7 @@ matrix:
     r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual --use-valgrind
     env:
     - ICD_COVERAGE=true
-    - ICD_MINI_BENCH=true
+    - ICD_MINI_BENCH=false
     - ICD_TEST_SLOW=true
   allow_failures:
   # allow any osx to fail for now: error prone tool chain
@@ -172,7 +176,7 @@ matrix:
     r_check_args: --no-build-vignettes --no-vignettes --ignore-vignettes --no-codoc --no-manual --use-valgrind
     env:
     - ICD_COVERAGE=true
-    - ICD_MINI_BENCH=true
+    - ICD_MINI_BENCH=false
     - ICD_TEST_SLOW=true
 
 before_script:
