### 2019-04-10 9baf870
update country-lang-vers vignette, rel qs [ci skip]
