+++ vignettes/country-lang-vers.Rmd
@@ -30,7 +30,7 @@ library(icd)
 
 Surprisingly, there are potential copyright restrictions on redistributing the WHO ICD-9 and ICD-10 definitions, which form the basis of ICD code systems around the world. These definitions are themselves also used as the internationally standard way to codify morbidity and mortality for public health.
 
-`icd` now includes the ability to work with these codes more easily via the companion `icd.data` package. `icd` asks the user whether they wish to download the data from the WHO web site, whereupon it is made available as functions, named similarly to existing data, e.g., `icd10who2016()`, which is the latest ICD-10 release from the WHO at the time of writing. For French codes, `icd10fr2016()`, with the localized synonym `cim10fr2016()` having internationalized column names, but this would duplicate the data making package storage and loading bigger, which is increasingly a problem as more data is included.
+`icd` now includes the ability to work with these codes more easily via the companion `icd.data` package. `icd` asks the user whether they wish to download the data from the WHO web site, whereupon it is made available as functions, named similarly to existing data, e.g., `get_icd10who2016()`, which is the latest ICD-10 release from the WHO at the time of writing. For French codes, `get_icd10fr2016()`, with the localized synonym `get_cim10fr2016()` having internationalized column names, but this would duplicate the data making package storage and loading bigger, which is increasingly a problem as more data is included.
 
 ```{r getstarted, eval = FALSE}
 # if not already done, set up the downloaded data cache:
@@ -68,10 +68,10 @@ There is an initial beta-quality mechanism for selecting a particular version of
   # get the active version string
   get_icd10cm_active_ver()
   # get the data itself:
-  devnull <- icd10cm_active()
+  devnull <- get_icd10cm_active()
   #set the active version, and check it is the one we expected
   set_icd10cm_active_ver("2015")
-  identical(icd10cm_active(), icd10cm2015())
+  identical(get_icd10cm_active(), get_icd10cm2015())
 ```
 
 The function `with_icd10cm_version` allows temporarily using a particular data set for a computation, analogous to the functions in the `withr` package:
@@ -79,20 +79,20 @@ The function `with_icd10cm_version` allows temporarily using a particular data s
 ```{r icd10cmoption}
   # The code "C4311" goes from being a leaf in 2018, to a parent in 2019
   with_icd10cm_version(ver = "2018",
-                                 icd::is_leaf("C4311")
+                                 is_leaf(as.icd10cm("C4311"))
   )
   with_icd10cm_version(ver = "2019", 
-                                 icd::is_leaf("C4311")
+                                 is_leaf(as.icd10cm("C4311"))
   )
   # In 2018 the following code was not even defined, but in 2019 was a new child
   # code of "C4311"s
   with_icd10cm_version(ver = "2018", 
                                  {
-                                   print(icd::is_defined("C43111"))
-                                   print(icd::is_leaf("C43111"))
+                                   print(is_defined(as.icd10cm("C4311")))
+                                   print(is_leaf(as.icd10cm("C4311")))
                                  }
   )
   with_icd10cm_version(ver = "2019", 
-                                 icd::is_leaf("C43111")
+                                 is_leaf("C43111")
   )
 ```
