### 2019-04-10 f67976f
[ci skip] fix country-lang-ver vignette
