### 2019-04-10 69e556e
fixes so country-lang-vers vignette builds [ci skip]
