### 2019-04-10 c6a7181
fix two dontrun examples [ci skip]
