+++ R/util-test.R
@@ -130,7 +130,8 @@ test_env <- function() {
 #' \dontrun{
 #' # original size data for PCCC benchmarking:
 #' set.seed(1441)
-#' neds <- icd:::generate_neds_pts(28584301L)
+#' # Large NEDS simulation: neds <- icd:::generate_neds_pts(28584301L)
+#' neds <- icd:::generate_neds_pts(2858L)
 #' neds_comorbid <- icd::comorbid_pccc_dx(neds)
 #' }
 #' @keywords internal
