+++ man/filter_poa.Rd
@@ -76,12 +76,15 @@ myData \%>\%
     icd_name = "diag", visit_name = "visit_id",
     short_code = TRUE
   )
+my_map <- head(icd9_map_elix)
 # can call the core comorbid() function with an arbitrary mapping
 myData \%>\%
   filter_poa_yes() \%>\%
-  comorbid_elix(
-    icd_name = "diag", visit_name = "visit_id",
-    short_mapping = TRUE
+  comorbid(
+    icd_name = "diag",
+    visit_name = "visit_id",
+    map = my_map,
+    short_map = TRUE
   )
 }
 }
