### 2019-04-11 d14dca5
update hcc raw data - fixes #174
 - Thanks to Anobel Odisho @anobel
 - fix incorrect labels for 2014/15 data fixes #140
 - add 2017-19 hierarchies and labels (unchanged from 2016 data)
 - fix tabs and inconsequential terminal characters in new HCC input data
 - save all data with serialization version 2, now R 3.6 defaults to version 3.
