### 2019-04-11 1986cdf
update hcc raw data
fix incorrect labels for 2014/15 data
add 2017-19 hierarchies and labels (unchanged from 2016 data)
