+++ data-raw/icd_hcc_rawdata/icd10/2017_icd10hcc.txt
@@ -1,0 +1,9505 @@
+A0103   115 
+A0104   39  
+A0105   39  
+A021    2   
+A0222   115 
+A0223   39  
+A0224   39  
+A065    115 
+A072    6   
+A202    115 
+A207    2   
+A212    115 
+A221    115 
+A227    2   
+A267    2   
+A310    6   
+A312    6   
+A327    2   
+A3681   85  
+A391    23  
+A392    2   
+A393    2   
+A394    2   
+A3983   39  
+A3984   39  
+A400    2   
+A401    2   
+A403    2   
+A408    2   
+A409    2   
+A4101   2   
+A4102   2   
+A411    2   
+A412    2   
+A413    2   
+A414    2   
+A4150   2   
+A4151   2   
+A4152   2   
+A4153   2   
+A4159   2   
+A4181   2   
+A4189   2   
+A419    2   
+A420    115 
+A427    2   
+A430    115 
+A480    106 
+A481    114 
+A483    2   
+A5055   39  
+A5440   39  
+A5441   39  
+A5442   39  
+A5443   39  
+A5449   39  
+A5484   115 
+A5485   33  
+A5486   2   
+A666    39  
+A6923   39  
+B007    2   
+B0082   72  
+B0112   72  
+B0224   72  
+B0682   39  
+B180    29  
+B181    29  
+B182    29  
+B188    29  
+B189    29  
+B20     1   
+B250    6   
+B251    6   
+B252    6   
+B258    6   
+B259    6   
+B2685   39  
+B3324   85  
+B371    6   
+B377    2   
+B377    6   D
+B3781   6   
+B380    115 
+B381    115 
+B382    115 
+B390    115 
+B391    115 
+B392    115 
+B400    115 
+B401    115 
+B402    115 
+B410    115 
+B4282   39  
+B440    6   
+B441    6   
+B442    6   
+B447    6   
+B4481   112 
+B4489   6   
+B449    6   
+B450    6   
+B451    6   
+B452    6   
+B453    6   
+B457    6   
+B458    6   
+B459    6   
+B460    6   
+B461    6   
+B462    6   
+B463    6   
+B464    6   
+B465    6   
+B468    6   
+B469    6   
+B484    6   
+B488    6   
+B582    6   
+B583    6   
+B59     6   
+B664    115 
+B671    115 
+B9735   1   
+C01     11  
+C020    11  
+C021    11  
+C022    11  
+C023    11  
+C024    11  
+C028    11  
+C029    11  
+C030    11  
+C031    11  
+C039    11  
+C040    11  
+C041    11  
+C048    11  
+C049    11  
+C050    11  
+C051    11  
+C052    11  
+C058    11  
+C059    11  
+C060    11  
+C061    11  
+C062    11  
+C0680   11  
+C0689   11  
+C069    11  
+C07     11  
+C080    11  
+C081    11  
+C089    11  
+C090    11  
+C091    11  
+C098    11  
+C099    11  
+C100    11  
+C101    11  
+C102    11  
+C103    11  
+C104    11  
+C108    11  
+C109    11  
+C110    11  
+C111    11  
+C112    11  
+C113    11  
+C118    11  
+C119    11  
+C12     11  
+C130    11  
+C131    11  
+C132    11  
+C138    11  
+C139    11  
+C140    11  
+C142    11  
+C148    11  
+C153    9   
+C154    9   
+C155    9   
+C158    9   
+C159    9   
+C160    9   
[…]
