### 2019-04-29 1ea7c5c
Complete transition back from icd.data to icd.
Squashed commit of the following:

commit 2c7b7da7ad5ea64e7b80f4d37bc8d9ba66b74a59
Author: Jack O. Wasey <jack@jackwasey.com>
Date:   Mon Apr 29 11:07:47 2019 -0400

    vectorize .exists, clarify others

commit 77ca2ad670cc793f496f5ea0b315bd9f592a4a79
Author: Jack O. Wasey <jack@jackwasey.com>
Date:   Mon Apr 29 11:00:32 2019 -0400

    various small

    final unicode
    vectorize .exists functions

commit a68e53e45345d40d2c4fee546cddd3d268fd5624
Author: Jack O. Wasey <jack@jackwasey.com>
Date:   Mon Apr 29 10:01:32 2019 -0400

    some data is
[…]
