+++ R/parse-icd10.R
@@ -1,0 +1,135 @@
+#' Get all ICD-10-CM codes
+#'
+#' Gets all ICD-10-CM codes from an archive on the CDC web site.
+#'
+#' The factor generation uses \code{sort} which is locale dependent. This meant
+#' a lot of time debugging a problem when white space was ignored for sorting on
+#' some platforms, but not others. For vectors of codes with an ICD class of
+#' some types, (e.g. ICD-9 and ICD-10-CM), a specific sort function and
+#' operators are defined.
+#' @param save_pkg_data Save specific year diagnostic codes in the package data
+#'   directory for CRAN package. All data retrieved is saved in the user cache
+#'   regardless.
+#' @source \url{https://www.cms.gov/Medicare/Coding/ICD10} also available from
+#'   \url{http://www.cdc.gov/nchs/data/icd/icd10cm/2016/ICD10CM_FY2016_code_descriptions.zip}.
+#'
+#' @references
+#' \href{https://www.cms.gov/Medicare/Coding/ICD10/downloads/icd-10quickrefer.pdf}{CMS
+#' ICD-10 Quick Reference}
+#' \href{https://www.cdc.gov/nchs/icd/icd10cm.htm#FY\%202019\%20release\%20of\%20ICD-10-CM}{CDC
+#' copy of ICD-10-CM for 2019}
+#' @keywords internal datagen
+#' @noRd
+.parse_icd10cm_all <- function(save_pkg_data = FALSE, ...) {
+  .msg("Parsing all ICD-10-CM diagnostic codes.")
+  yrs <- names(.icd10cm_sources)
+  out <- lapply(
+    yrs,
+    .parse_icd10cm_year,
+    ...
+  )
+  names(out) <- yrs
+  invisible(names(out))
+}
+
+.parse_icd10cm_year <- function(year,
+                                save_pkg_data = FALSE,
+                                dx = TRUE,
+                                ...) {
+  stopifnot(is.numeric(year) || is.character(year))
+  year <- as.character(year)
+  stopifnot(as.character(year) %in% names(.icd10cm_sources))
+  #
+  f_info <- .dl_icd10cm_year(year = year, dx = dx)
+  if (is.null(f_info)) {
+    .absent_action_switch(
+      paste("No path to downloaded data. Options are: "),
+      paste(names(.show_options()), unname(.show_options()),
+        collapse = ",",
+        sep = ":"
+      )
+    )
+    return()
+  }
+  .msg("Working on parsing ICD-10-CM year: ", year)
+  # readLines may muck up encoding, resulting in weird factor order generation
+  # later?
+  x <- readLines(con = f_info$file_path, encoding = "ASCII")
+  .msg("Got flat file for year: ", year)
+  stopifnot(all(Encoding(x) == "unknown"))
+  dat <- data.frame(
+    # id = substr(x, 1, 5),
+    code = trimws(substr(x, 7, 13)),
+    billable = trimws(substr(x, 14, 15)) == "1",
+    short_desc = trimws(substr(x, 16, 76)),
+    long_desc = trimws(substr(x, 77, stop = 1e5)),
+    stringsAsFactors = FALSE
+  )
+  dat[["code"]] <-
+    icd::as.short_diag(
+      if (dx) {
+        icd::as.icd10cm(dat[["code"]])
+      } else {
+        icd::as.icd10cm_pc(dat[["code"]])
+      }
+    )
+  dat[["three_digit"]] <- factor(get_major(dat[["code"]]))
+  # here we must re-factor so we don't have un-used levels in major
+  dat[["major"]] <- factor(
+    merge(
+      x = dat["three_digit"],
[…]
