### 2019-04-29 6147934
skip more tests when data will not be availble
