### 2019-05-01 5c87bbb
various
 - make sure all saved package data is in serialization version 2 for
   backward compatibility with R < 3.5
 - don't rely on library(icd) having been called for ranges and children
 - include efficiency in vignette index even though not built through
   vignettes process, same for country-lang-ver.Rmd
 - re-styled, added lint exceptions
 - docs tweaked to eliminate icd.data
