+++ R/datadocs.R
@@ -25,7 +25,9 @@
 #' @keywords datasets list category
 #' @docType data
 #' @source
+# nolint start
 #' \url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
+# nolint end
 #'
 #' @details Format: list with chapter/sub-chapter or major names stored in list
 #'   names, each with two element named character vector with start and end
@@ -78,10 +80,11 @@ NULL
 #' ICD-9-CM diagnosis codes including leaf nodes and branch names up to the
 #' three-digit codes.
 #'
-#' Unlike \code{\link{get_icd9cm2014_leaf}} and friends, these data frames contain
-#' the full structure of the ICD-9-CM scheme up to the three-digit codes. Unlike
-#' the equivalent ICD-10-CM data frames (e.g., \code{\link{icd10cm2019}}), they
-#' do not have columns indicating billable/leaf status, or chapter designations.
+#' Unlike \code{\link{get_icd9cm2014_leaf}} and friends, these data frames
+#' contain the full structure of the ICD-9-CM scheme up to the three-digit
+#' codes. Unlike the equivalent ICD-10-CM data frames (e.g.,
+#' \code{\link{icd10cm2019}}), they do not have columns indicating billable/leaf
+#' status, or chapter designations.
 #'
 #' \code{icd9cm_hierarchy} is the deprecated name, which currently points to the
 #' final ICD-9-CM 2014 release. 2011 -- 2014 are in fact identical. These are
@@ -94,7 +97,9 @@ NULL
 #' \url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9cm.asp}
 #' @source Rich text descriptions here:
 #'   \url{http://www.cdc.gov/nchs/icd/icd9cm.htm}
+# nolint start
 #'   \url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
+# nolint end
 #'    This page has versions 23 to 32 (2005 to 2014). At present, only the 2014
 #'   data is included in this package.
 #' @source
@@ -125,6 +130,7 @@ NULL
 #' @details Format:  data frame, with columns for code, leaf status (0 or 1),
 #'   short and long descriptions.
 #' @references \url{https://www.cms.gov/Medicare/Coding/ICD10/}
+# nolint start
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-PCS-Tables-And-Index.zip}
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-PCS-Order-File.zip}
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2017-PCS-Code-Tables.zip}
@@ -138,6 +144,7 @@ NULL
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-ICD10-Code-Tables-and-Index.zip}
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-Code-Tables-and-Index.zip}
 #'   \url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2014-PCS-long-and-abbreviated-titles.zip}
+# nolint end
 #' @name icd10cm2019
 NULL
 
@@ -209,8 +216,10 @@ NULL
 #' Définitions CIM-10-FR de l'OMS (WHO ICD-10 definitions)
 #' @source \url{https://www.atih.sante.fr/cim-10-fr-2018-usage-pmsi}
 #' @references
+# nolint start
[…]
