### 2019-05-02 3faeb92
Static vignettes attempt, resave final rda as version 2 serialization
