### 2019-05-02 ae28eb7
simpler static vignette solution
