+++ tests/testthat/test-spell.R
@@ -4,7 +4,7 @@ test_that("dictionary created", {
   skip_on_cran()
   skip_on_appveyor()
   skip_on_travis()
-  words <- generate_spelling(save_pkg_data = FALSE)
+  words <- .generate_spelling(save_pkg_data = FALSE)
   expect_character(words)
   expect_true(file.exists(system.file("WORDLIST", package = "icd")))
 })
