+++ R/update.R
@@ -8,7 +8,7 @@
 #' mappings are generated from transcribed codes.
 #' @keywords internal datagen
 #' @noRd
-update_everything <- function() {
+.update_everything <- function() {
   old_opt <- options(
     icd.data.offline = FALSE,
     icd.data.verbose = TRUE
@@ -22,9 +22,9 @@ update_everything <- function() {
   .set_icd_data_dir()
   .print_options()
   .icd10cm_extract_sub_chapters(save_pkg_data = TRUE)
-  generate_sysdata()
+  .generate_sysdata()
   load(file.path("R", "sysdata.rda"))
-  generate_spelling()
+  .generate_spelling()
   message("Parsing comorbidity mappings from SAS and text sources.
                        (Make sure lookup files are updated first.)
                        Depends on icd9cm_hierarchy being updated.")
@@ -58,45 +58,51 @@ update_everything <- function() {
   icd9cm_hierarchy <- get_icd9cm2014()
   names(icd9cm_hierarchy)[names(icd9cm_hierarchy) == "leaf"] <- "billable"
   .save_in_data_dir(icd9cm_hierarchy)
-  .generate_vigette_index
+  #.generate_vigette_index()
 }
-
-# this is sadly very complicated in order to pass R CMD check --as-cran
 #
-# Look in .Rinstignore, .Rbuildignore, /build/vignette.rds vignettes/ vignettes-prebuilt vignettes/.install_extras
+# # this is sadly very complicated in order to pass R CMD check --as-cran
+# #
+# # Look in .Rinstignore, .Rbuildignore, /build/vignette.rds vignettes/ vignettes-prebuilt vignettes/.install_extras
+# #
+# # https://stackoverflow.com/questions/51792384
+# #
+# # from R internals:
+# #
+# # File vignette.rds records a data frame with one row for each ‘vignette’ (.[RS]nw file in inst/doc) and with columns ‘File’ (the full file path in the sources), ‘Title’, ‘PDF’ (the pathless file name of the installed PDF version, if present), ‘Depends’, ‘Keywords’ and ‘R’ (the pathless file name of the installed R code, if present).
+# .generate_vigette_index <- function() {
+#   vi_col_names <- c("File", "Title", "PDF", "R", "Depends", "Keywords")
+#   vignette_index <- matrix(
+#       ncol = 4, byrow = TRUE,
+#       c(
+#         "charlson-scores.Rmd", "Charlson and Van Walraven scores", "charlson-scores.html", "charlson-scores.R",
 #
-# https://stackoverflow.com/questions/51792384
-.generate_vigette_index <- function() {
-  vi_col_names <- c("File", "Title", "PDF", "R", "Depends", "Keywords")
-  vignette_index <- matrix(
-      ncol = 4, byrow = TRUE,
-      c(
-        "charlson-scores.Rmd", "Charlson and Van Walraven scores", "charlson-scores.html", "charlson-scores.R",
-
-        "CMS-HCC.Rmd", "Using Hierarchical Condition Codes", "CMS-HCC.html", "CMS-HCC.R",
-
-        "compare-maps.Rmd", "Comparing Comorbidity Mappings", "compare-maps.html", "compare-maps.R",
-
-        "custom-maps.Rmd", "Quick custom comorbidity maps are helpful", "custom-maps.html", "custom-maps.R",
-
-        "ICD-10.Rmd", "ICD-10 comorbidities", "ICD-10.html", "ICD-10.R",
-
-        "introduction.Rmd", "An Introduction to icd", "introduction.html", "introduction.R",
-
[…]
