+++ vignettes/efficiency.Rmd
@@ -51,6 +51,7 @@ bibliography:
   - other.bib
   - icdjss.bib
 vignette: >
+  % The resuling PDF is included by vignettes/efficiency-vignette.Rnw which contains the Vignette info
   %\VignetteIndexEntry{icd: Efficient Computation of Comorbidities from ICD Codes Using Sparse Matrix Multiplication in R}
   %\VignetteDepends{rticles}
   %\VignetteEngine{knitr::rmarkdown}
