### 2019-05-02 788d4e6
cleaned-up and commented static vignette work
