+++ vignettes/efficiency.tex
@@ -1,1576 +1,0 @@
-\documentclass[article]{jss}
-\usepackage[utf8]{inputenc}
-
-\providecommand{\tightlist}{%
-  \setlength{\itemsep}{0pt}\setlength{\parskip}{0pt}}
-
-\author{
-Jack O. Wasey\\Children's Hospital of Philadelphia \And Steven M. Frank\\Johns Hopkins Medical Institutions \And Mohamed A. Rehman\\Johns Hopkins All Children's Hospital
-}
-\title{\pkg{icd}: Efficient Computation of Comorbidities\\
-from ICD Codes Using Sparse Matrix Multiplication in \proglang{R}}
-
-\Plainauthor{Jack O. Wasey, Steven M. Frank, Mohamed A. Rehman}
-\Plaintitle{icd: Efficient Computation of Comorbidities from ICD Codes Using Sparse
-Matrix Multiplication in R}
-\Shorttitle{\pkg{icd}: Efficient calculation of comorbidities from ICD codes}
-
-\Abstract{
-Medical research often describes the existing diseases (comorbidities)
-in its study populations, and retrospective research may rely on this
-information for risk adjustment. Sets of International Classification of
-Diseases (ICD) codes are often used for high-level classification of
-patients into comorbidities (e.g., `Cancer' or `Heart Disease') to help
-summarize characteristics of a population. Similar tools were not
-available in \proglang{R}, and do not scale to big data sets. The
-\pkg{icd} extension for \proglang{R} includes validated mappings of ICD
-codes to comorbidities, and uses an efficient, fast and scalable
-algorithm to apply these mappings to patient data. This enables
-reproducible workflows with hundreds of millions of patient records as
-are increasingly seen in large national and international databases.
-}
-
-\Keywords{medical informatics, administrative data, matrix algebra, \proglang{R}, \proglang{C++}, ICD, mapping}
-\Plainkeywords{biomedical informatics, administrative data, matrix algebra, R, C++, ICD, mapping}
-
-%% publication information
-%% \Volume{50}
-%% \Issue{9}
-%% \Month{June}
-%% \Year{2012}
-%% \Submitdate{}
-%% \Acceptdate{2012-06-04}
-
-\Address{
-    Jack O. Wasey\\
-  Children's Hospital of Philadelphia\\
-  3401 Civic Center Blvd. Philadelphia, PA 19102, USA email:
-  \email{waseyj@chop.edu} url:
-  \url{http://www.chop.edu/doctors/wasey-jack}\\
-  
-  
-      Steven M. Frank\\
-  Johns Hopkins Medical Institutions\\
-  1800 Orleans St.~Baltimore, MD 21287, USA email:
-  \href{mailto:sfrank3@jhmi.edu}{\nolinkurl{sfrank3@jhmi.edu}} url:
-  \url{https://www.hopkinsmedicine.org/profiles/results/directory/profile/1819685/steven-frank}\\
-  
-  
-      Mohamed A. Rehman\\
-  Johns Hopkins All Children's Hospital\\
-  501 6th Ave S. St Petersburg, FL 33701, USA email:
-  \href{mailto:rehman@jhmi.edu}{\nolinkurl{rehman@jhmi.edu}} url:
-  \url{http://www.jhmi.edu}\\
-  
-  
-  }
-
-% Pandoc header
-
-\usepackage{amsmath}
-\usepackage{longtable}
-\usepackage{booktabs}
-\usepackage{thumbpdf}
-\raggedbottom
-\shortcites{quan_coding_2005,quan_updating_2011, Pope_RiskadjustmentMedicare_2004,Lee_Predictingmortalitypatients_2003,frank_risk-adjusted_2014,elixhauser_comorbidity_1998}
[…]
