### 2019-05-02 4e5b589
travis [skip appveyor]
 - install a latex package enabling static pdfs to be built (from the source static pdfs)
 - fail if benchmarks fail
