### 2019-05-02 60357be
try harder to get pdfpages latex package in travis [skip appveyor]
