### 2019-05-02 b05d0e5
travis add ae latex package - maybe caused by tinytex changing path [skip appveyor]
