### 2019-05-02 8e39027
try really really hard and carefully to get a latex package on multiple different travis platforms [skip appveyor]
