+++ .travis.yml
@@ -50,7 +50,7 @@ before_install:
   #
   # Work-around r-lib/remotes bug which prevents travis building.
   # https://github.com/r-lib/devtools/issues/2020#issuecomment-481219296
-  - R -e 'source("https://install-github.me/r-lib/remotes")'
+  - Rscript -e 'source("https://install-github.me/r-lib/remotes")'
   # ready for Makevars
   - mkdir ~/.R ;
   - echo 'CXXFLAGS=-Wno-unused-parameter -Wno-unused-variable -Wno-ignored-attributes -Wno-cast-function-type -Wno-unknown-warning-option' >> ~/.R/Makevars ;
@@ -66,7 +66,7 @@ before_install:
   # for the pre-built static vignette stubs to include the actual vignettes, try:
   # - Rscript -e 'if (require("tinytex")) {tinytex::install_tinytex(); tinytex::tlmgr_install("pdfpages", "ae")}'
   # tlmgr --user-mode doesn't seem to be necessary
-  - tlmgr install pdfpages ae
+  - Rscript -e 'if (!require(\"tinytex\")) {quit(status=1)} else {tinytex::install_tinytex(); tinytex::tlmgr_install(\"pdfpages\", \"ae\")}' || hash tlmgr 2>/dev/null && { tlmgr install pdfpages ae; }
 
 env:
   global:
