### 2019-05-02 b57317a
cleaner macos logic in travis and no need for escaped quotes [skip appveyor]
