### 2019-05-03 cd5dee9
verbose, comment and tidy benchmark dep install [ci skip]
