### 2019-05-03 532b7c0
simpler not to try to build the static vignettes on travis [skip appveyor]
