+++ tools/split-tests.R
@@ -53,35 +53,3 @@ with_split_tests <- function(code, ...,
   on.exit(testthat_restore(base_name = base_name), add = TRUE)
   force(code)
 }
-
-check_local_gctorture <- function(filter = ".*") {
-  with_split_tests(
-      rcmdcheck::rcmdcheck(args = c("--use-gct", check_args))
-  )
-}
-
-# THIS DOESN'T WORK! Eigen spews out warnings when RcppEigen itself is installed, not when icd is installed.
-check_rhub_quiet <- function(
-  fun = rhub::check_with_valgrind,
-  filter = ".*") {
-  shutup <-
-    paste0(
-      "CXX11FLAGS=",
-      paste0(
-    "-Wno-",
-    c(
-      "unused-parameter",
-      "unused-variable",
-      "ignored-attributes",
-      "cast-function-type",
-      "unknown-pragmas",
-      "unknown-warning-option",
-      "unknown-warning"
-    ),
-    collapse = " "
-      )
-    )
-  with_split_tests(
-    fun(check_args = paste0("--no-vignettes --no-build-vignettes --ignore-vignettes --install-args='--configure-vars=\"", shutup, "\"'"))
-  )
-}
