+++ .travis.yml
@@ -29,10 +29,10 @@ addons:
   apt:
     packages:
       - valgrind
-      - r-cran-tinytex
+      #- r-cran-tinytex
 
-r_packages:
-  - tinytex
+#r_packages:
+#  - tinytex
 
 cran: https://cloud.r-project.org
 # by default, DON'T do --as-cran, as it stops us using environment variables
@@ -42,11 +42,12 @@ warnings_are_errors: true
 
 # try to speed up compilation time
 before_install:
-  # Set library directory so R CMD check finds packages installed by travis and by R itself (e.g. Rcpp rebuilt for MacOS builds)
+  # May need to set the R library directory so R CMD check finds packages
+  # installed by travis (automatically, and in my scripts) and by R itself
   # https://stackoverflow.com/questions/31879011/r-cant-find-packages-installed-by-travis
-  # cannot sudo in docker! - sudo mkdir -p /usr/lib/R/library
-  # - echo 'R_LIBS=/usr/lib/R/library:/usr/lib/R/site-library/' > ~/.Renviron
-  # - sudo chmod 2777 /usr/lib/R/library
+  # - mkdir -p "$HOME/R.icd.lib"
+  # - echo R_LIBS=$HOME/R.icd.lib:/usr/lib/R/site-library/" > ~/.Renviron
+  # - sudo chmod 2777 "$HOME/R.icd.lib"
   #
   # Work-around r-lib/remotes bug which prevents travis building.
   # https://github.com/r-lib/devtools/issues/2020#issuecomment-481219296
@@ -55,20 +56,27 @@ before_install:
   - mkdir ~/.R ;
   - echo 'CXXFLAGS=-Wno-unused-parameter -Wno-unused-variable -Wno-ignored-attributes -Wno-cast-function-type -Wno-unknown-warning-option' >> ~/.R/Makevars ;
   - echo 'CXX11FLAGS=-Wno-unused-parameter -Wno-unused-variable -Wno-ignored-attributes -Wno-cast-function-type -Wno-unknown-warning-option' >> ~/.R/Makevars ;
+  # ignore the stubs for the pre-built vignettes, as the latex environments on
+  # travis are heterogeneous!
+  - echo '^vignettes\/.*-vignette\.Rnw$' >> .Rbuildignore
   - if [[ "$TRAVIS_R_VERSION_STRING" = 'devel' ]]; then
       echo 'CFLAGS=-O0' >> ~/.R/Makevars ;
       echo 'CXXFLAGS=-O0' >> ~/.R/Makevars ;
       echo 'CXX11FLAGS=-O0' >> ~/.R/Makevars ;
       fi
+  # Homebrew GCC currently version 8.3 from
+  # https://github.com/Homebrew/homebrew-core/blob/master/Formula/gcc.rb
   - if [[ "$(uname -s)" = 'Darwin' ]]; then
-      command -v g++-8 >/dev/null 2>&1 || brew install gcc && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars && Rscript -e 'install.packages(\"Rcpp\", type = \"source\")';
+      command -v g++-8 >/dev/null 2>&1 || brew install gcc && echo 'CC=gcc-8' >> ~/.R/Makevars && echo 'CXX=g++-8' >> ~/.R/Makevars && echo 'CXX11=g++-8' >> ~/.R/Makevars && Rscript -e 'install.packages("Rcpp", type = "source")';
     fi
   - cat ~/.R/Makevars
+  - cat .Rbuildignore
   - Rscript -e '.libPaths()'
-  # for the pre-built static vignette stubs to include the actual vignettes, try:
[…]
