### 2019-05-03 2bec4f7
travis and tools [skip appveyor]
