### 2019-05-03 5c93b0f
add test and fix for certain WHO icd codes [ci skip]
