### 2019-05-03 3060838
don't run online tests on CI, and work again on travis for macos
