+++ tests/testthat/test-who.R
@@ -32,12 +32,25 @@ test_that("HIV is problematic for some reason", {
   expect_match(b21$ID, "^B21\\.[0-9]$")
 })
 
-test_that("Some Ixx codes are missed", {
+test_that("Some codes beginning with Roman numerals were missed", {
   test_that("we do download the relevant blocks", {
     skip_multi()
     skip_if_offline()
     expect_true(nrow(.dl_icd10who_children("I10-I15")) > 1)
     expect_true(nrow(.dl_icd10who_walk("I10-I15")) > 1)
   })
-  expect_true(length(explain_code(as.icd10who("I10"))) == 1)
+  skip_missing_icd10who()
+  expect_equal_no_icd(
+    explain_code(as.icd10who(c("V13", "X13", "I10"))),
+    c("Essential (primary) hypertension",
+      "Pedal cyclist injured in collision with car, pick-up truck or van",
+      "Contact with steam and hot vapours")
+  )
+  expect_equal_no_icd(
+    explain_code(as.icd10who(c("V13", "X13", "I10")), lang = "fr"),
+    c("Hypertension essentielle (primitive)",
+      "Cycliste bless\u00e9 dans une collision avec une automobile ou une camionnette",
+      "Contact avec de la vapeur d'eau et des vapeurs br\u00fblantes"
+    )
+  )
 })
