### 2019-05-03 c669a59
ignore travis-specific compilation flags [skip appveyor]
