+++ R/as_char_no_warn.R
@@ -4,7 +4,9 @@
 #' @keywords internal
 #' @noRd
 as_char_no_warn <- function(x) {
-  if (is.character(x)) return(x)
+  if (is.character(x)) {
+    return(x)
+  }
   old <- options(warn = -1)
   on.exit(options(old), add = TRUE)
   if (is.integer(x)) {
