+++ tests/testthat/test-build-maps-icd9.R
@@ -3,33 +3,37 @@ context("build icd9 maps")
 skip_slow("Skipping slow re-building of ICD-9 comorbidity maps")
 
 test_that("AHRQ children same as saved", {
-  for (i in icd9_map_ahrq)
+  for (i in icd9_map_ahrq) {
     expect_equal(
       children.icd9(i, defined = FALSE, short_code = TRUE),
       sort.icd9(i)
     )
+  }
 })
 
 test_that("Elixhauser children same as saved", {
-  for (i in icd9_map_quan_elix)
+  for (i in icd9_map_quan_elix) {
     expect_equal(
       children.icd9(i, defined = FALSE, short_code = TRUE),
       sort.icd9(i)
     )
+  }
 })
 
 test_that("Quan Charlson children same as saved", {
-  for (i in icd9_map_quan_deyo)
+  for (i in icd9_map_quan_deyo) {
     expect_equal_no_class_order(
       children.icd9(i, defined = FALSE, short_code = TRUE), sort.icd9(i)
     )
+  }
 })
 
 test_that("Quan Elixhauser children same as saved", {
-  for (i in icd9_map_quan_elix)
+  for (i in icd9_map_quan_elix) {
     expect_equal_no_class_order(
       children.icd9(i, defined = FALSE, short_code = TRUE), sort.icd9(i)
     )
+  }
 })
 
 
