+++ R/comorbid_pccc.R
@@ -47,7 +47,7 @@ comorbid_pccc_dx <- function(x,
 #' )
 #' comorbid_pccc_pcs(pts, icd_name = "icd9_pcs", return_binary = TRUE)
 #' comorbid_pccc_pcs(pts, icd_name = "icd10_pcs", return_binary = TRUE)
-#' 
+#'
 #' # All ICD-9 procedure codes are numeric, some ICD-10 procedure codes
 #' # are numeric, so best to call functions directly:
 #' pts <- data.frame(encounters = c(100), icd10_pcs = c("0016070"))
