### 2019-05-04 b32f7fd
various
minor tidy-up of diagnostic warnings in rstudio
fix minor Belgian code table issue
correctly use environment variable in testthat.r
allow old ICD-9-CM not to have long desc in extended tests
