### 2019-05-05 05ed474
tools change and CRAN comments [ci skip]
