+++ tools/check_full.sh
@@ -36,6 +36,7 @@ MAKEFLAGS=-j$(getconf _NPROCESSORS_ONLN) \
   _R_CHECK_PKG_SIZES_=false \
   _R_CHECK_LENGTH_1_CONDITION_="verbose,abort" \
   _R_CHECK_LENGTH_1_LOGIC2_="verbose,abort" \
+  ICD_TEST_SLOW=true \
  R CMD check "$(ls -t $tmpd/icd*.tar.gz | head -1)"
 popd
 #  _R_CHECK_PKG_SIZES_THRESHOLD_=50 \
