### 2019-05-05 5d7e08d
tiny CRAN fixes
remove doi URL which has TLS certificate errors on some Windows platforms, including CRAN.
only include the vignette-built efficiency in vignette in the packet, which is therefore indexed.
rename prebuilt vignettes more clearly
