+++ man/charlson.Rd
@@ -50,22 +50,22 @@ your system default or this setting.}
 e.g. \code{name}}
 }
 \description{
-Charlson score is calculated in the basis of the Quan revision of Deyo's
-ICD-9 mapping. (Peptic ulcer disease no longer warrants a point.) Quan
-published an updated set of scores, but it seems most people use the original
-scores for easier comparison between studies, even though Quan's were more
-predictive.
+Charlson score is calculated in the basis of the Quan revision
+  of Deyo's ICD-9 mapping. (Peptic ulcer disease no longer warrants a point.)
+  Quan published an updated set of scores, but it seems most people use the
+  original scores for easier comparison between studies, even though Quan's
+  were more predictive.
 }
 \details{
 When used, hierarchy is applied per Quan, \dQuote{The following
   comorbid conditions were mutually exclusive: diabetes with chronic
   complications and diabetes without chronic complications; mild liver
   disease and moderate or severe liver disease; and any malignancy and
-  metastatic solid tumor.} The Quan scoring weights come from the 2011 paper
-  (dx.doi.org/10.1093/aje/kwq433). The comorbidity weights were recalculated
-  using updated discharge data, and some changes, such as Myocardial
-  Infarction decreasing from 1 to 0, may reflect improved outcomes due to
-  advances in treatment since the original weights were determined in 1984.
+  metastatic solid tumor.} The Quan scoring weights come from the 2011 paper.
+  The comorbidity weights were recalculated using updated discharge data, and
+  some changes, such as Myocardial Infarction decreasing from 1 to 0, may
+  reflect improved outcomes due to advances in treatment since the original
+  weights were determined in 1984.
 }
 \section{Methods (by class)}{
 \itemize{
