+++ vignettes/ICD-10.Rmd
@@ -9,6 +9,7 @@ vignette: >
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteDepends{magrittr}
   %\VignetteEncoding{UTF-8}
+  %\VignetteKeyword{comorbidity, comorbidities, biomedical informatics, administrative data, matrix algebra, ICD-10, ICD-10-CM, WHO, France, Belgium}
 ---
 <!--
 Copyright (C) 2014 - 2019  Jack O. Wasey
@@ -76,7 +77,7 @@ comorbid_charlson(uranium_pathology, return_df = TRUE)[1:5, 1:5]
 
 ## Working with big data
 
-`icd` is carefully optimized to give accurate results as quickly as possible, and [shines](efficiency.html) with huge data sets. For users working with millions of rows of data and higher orders of magnitude, some options can improve throughput. 
+`icd` is carefully optimized to give accurate results as quickly as possible, and [shines](efficiency.html) with huge data sets. For users working with millions of rows of data and higher orders of magnitude, some options can improve throughput.
 
 In this example, we also return 'binary' numeric flags instead of a matrix of logical values. For very large data sets, matrices are faster both for `icd` and subsequent manipulation and analysis. Here we also show that we can request the 'id' field is return in the order the data was presented. Note that setting `restore_id_order` to `FALSE` does not sort: it simply returns the data as soon as it was calculated, and since parallel threads complete at different times, it is quicker, although not deterministic. Again, this is really for huge data sets, where the additional sorting and re-ordering may be very time consuming. Most users will not need to worry about this.
 ```{r big}
@@ -96,7 +97,7 @@ The ICD-10-CM mappings are recorded a bit differently from the ICD-9-CM mappings
 ```{r cholera}
 # create trivial comorbidity map:
 cholera_typhoid_map <- list(cholera = "A00", typhoid = "A01")
-patients <- data.frame(patient = c("0001", "0001", "0002"), 
+patients <- data.frame(patient = c("0001", "0001", "0002"),
                        code = c("A001234567", "A01", "A019"))
 comorbid(patients , map = cholera_typhoid_map)
 ```
@@ -117,7 +118,7 @@ set.seed(1441)
 pts <- data.frame(id = sample(LETTERS, n),
                   pc = pcs_sample)
 res <- icd10_comorbid(pts,
-                      map = icd10_map_ahrq_pcs, 
+                      map = icd10_map_ahrq_pcs,
                       icd_name = "pc",
                       return_binary = TRUE)
 print(res)
