+++ vignettes/CMS-HCC.Rmd
@@ -11,6 +11,7 @@ vignette: >
   %\VignetteIndexEntry{Using Hierarchical Condition Codes}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+  %\VignetteKeyword{hierarchical condition codes, CMS, HCC, comorbidity, comorbidities, biomedical informatics, administrative data, ICD-9, ICD-10, ICD-9-CM, ICD-10-CM}
 ---
 
 Hierarchical Condition Categories (HCC) is a method of categorizing ICD codes created by the Centers for Medicare and Medicaid Services (CMS), which is implemented in the [icd package](https://cran.r-project.org/package=icd) for R. The package provides [fast comorbidity calculations](introduction.html) for HCCs and other maps which translate ICD codes to comorbidities.
@@ -32,7 +33,7 @@ ICD also provides a method for mapping specific ICD codes to HCCs. The method `c
 
 ```{r pts}
 pts <- data.frame(patient_id = c("1", "2", "3", "4", "4"),
-                  icd_code = c("20084", "1742", "30410", "41514", "95893"), 
+                  icd_code = c("20084", "1742", "30410", "41514", "95893"),
                   date = as.Date(c("2011-01-01", "2011-01-02", "2011-01-03", "2011-01-04", "2011-01-04")))
 
 pts
