### 2019-05-05 3b52fef
Bump version, readme tweak, win64 versions of build files
