### 2019-05-06 8b6caf6
drop icd10cm2016 built-in data, so package is under 5mb and CRAN will accept
