+++ R/zzz.R
@@ -91,7 +91,6 @@ utils::globalVariables(c(
   "icd9_majors",
   "icd10_sub_chapters",
   "icd10_chapters",
-  "icd10cm2016",
   "icd10cm2019",
   "icd9cm_hierarchy"
 ))
