+++ R/update.R
@@ -50,15 +50,11 @@
   # icd9cm2014_leaf was formerly "icd9cm_billable[[version_number]]"
   icd9cm2014_leaf <- get_icd9cm2014_leaf()
   .save_in_data_dir(icd9cm2014_leaf)
-  # TODO: deprecate and remove icd10cm2016 in icd 4.1 or 4.2
-  icd10cm2016 <- .parse_icd10cm_year(2016)
-  .save_in_data_dir(icd10cm2016)
   icd10cm2019 <- .parse_icd10cm_year(2019)
   .save_in_data_dir(icd10cm2019)
   icd9cm_hierarchy <- get_icd9cm2014()
   names(icd9cm_hierarchy)[names(icd9cm_hierarchy) == "leaf"] <- "billable"
   .save_in_data_dir(icd9cm_hierarchy)
-  # .generate_vigette_index()
 }
 
 #' Generate \code{sysdata.rda}
