+++ README.md
@@ -35,13 +35,14 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 [![Appveyor](https://ci.appveyor.com/api/projects/status/9ncfgxht3n5i8t60/branch/master?svg=true
 "Appveyor Build Status")](https://ci.appveyor.com/project/jackwasey/icd/branch/master)
 [![codecov.io](https://codecov.io/github/jackwasey/icd/coverage.svg?branch=master
-"Core Code Coverage")](https://codecov.io/github/jackwasey/icd?branch=master)
+"Code Coverage")](https://codecov.io/github/jackwasey/icd?branch=master)
 [![CII Best
 Practices](https://bestpractices.coreinfrastructure.org/projects/2092/badge)](https://bestpractices.coreinfrastructure.org/projects/2092)
 [![Dependencies](https://tinyverse.netlify.com/badge/icd)](https://cran.r-project.org/package=icd)
 [![CRAN RStudio mirror downloads last calendar
 month](https://cranlogs.r-pkg.org/badges/icd
-"RStudio Mirror Downloads Last Calendar Month")](https://cran.r-project.org/package=icd)
+"RStudio Mirror Downloads")](https://cran.r-project.org/package=icd)
+[![](https://cranlogs.r-pkg.org/badges/icd)](https://cran.r-project.org/package=icd)
 <!-- badges: end -->
 <!-- [![CRAN RStudio mirror downloads last week](https://cranlogs.r-pkg.org/badges/last-week/icd "RStudio Mirror Downloads Last Week")](https://cran.r-project.org/package=icd) -->
 
