+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.2.1
-pkgdown: 1.3.0
-pkgdown_sha: ~
+pkgdown: 1.3.0.9000
+pkgdown_sha: 83f2afb5d2ec85abe29bfd8d67408cdf2bb55f03
 articles:
   CMS-HCC: CMS-HCC.html
   ICD-10: ICD-10.html
