### 2019-05-08 55865a0
pkgdown description update and remove search bar which messes up formatting [ci skip]
