+++ inst/_pkgdown.yml
@@ -3,21 +3,21 @@ url: https://jackwasey.github.io/icd/
 title: Comorbidity calculations and tools for ICD-9 and ICD-10 codes
 desc: >
     Calculate comorbidities, medical risk scores, and
-    work quickly and precisely with ICD-9 and ICD-10 codes. This package
-    enables a work flow from raw tables of ICD codes in hospital databases
-    to comorbidities. ICD-9 and ICD-10 comorbidity mappings from Quan
-    (Deyo and Elixhauser versions), Elixhauser and AHRQ included.  Common
-    ambiguities and code formats are handled. Comorbidity computation
-    includes Hierarchical Condition Codes, and a reimplementation of the
-    AHRQ Clinical Classifications Software. Risk scores include those of
-    Charlson and van Walraven.  US Clinical Modification, Word Health
-    Organization, Belgian and French ICD-10 codes are supported through
-    the 'icd.data' package, some of which are downloaded on demand.
-template:
-  params:
-    docsearch:
-      api_key: 797795ae271734faa8bc9b1f4aaae82b
-      index_name: icd
+    work very quickly and precisely with ICD-9 and ICD-10 codes. This
+    package enables a work flow from raw tables of ICD codes in hospital
+    databases to comorbidities. ICD-9 and ICD-10 comorbidity mappings from
+    Quan (Deyo and Elixhauser versions), Elixhauser and AHRQ included.
+    Common ambiguities and code formats are handled. Comorbidity
+    computation includes Hierarchical Condition Codes, and an
+    implementation of AHRQ Clinical Classifications. Risk scores include
+    those of Charlson and van Walraven.  US Clinical Modification, Word
+    Health Organization, Belgian and French ICD-10 codes are supported,
+    most of which are downloaded on demand.
+#template:
+#  params:
+#    docsearch:
+#      api_key: 797795ae271734faa8bc9b1f4aaae82b
+#      index_name: icd
 
 articles:
 - title: Journal of Statistical Software
