### 2019-05-12 5587b4f
minor doc updates [ci skip]
