+++ README.Rmd
@@ -57,7 +57,6 @@ patient_data <- data.frame(
 [![CII Best Practices](https://bestpractices.coreinfrastructure.org/projects/2092/badge)](https://bestpractices.coreinfrastructure.org/projects/2092)
 [![Dependencies](https://tinyverse.netlify.com/badge/icd)](https://cran.r-project.org/package=icd)
 [![CRAN RStudio mirror downloads last calendar month](https://cranlogs.r-pkg.org/badges/icd  "RStudio Mirror Downloads")](https://cran.r-project.org/package=icd)
-[![](https://cranlogs.r-pkg.org/badges/icd)](https://cran.r-project.org/package=icd)
 <!-- badges: end -->
 <!-- [![CRAN RStudio mirror downloads last week](https://cranlogs.r-pkg.org/badges/last-week/icd "RStudio Mirror Downloads Last Week")](https://cran.r-project.org/package=icd) -->
 
@@ -77,7 +76,7 @@ cat(packageDescription("icd")$Description)
      + several standard mappings of ICD codes to comorbidities are included (Quan, Deyo, Elixhauser, AHRQ, PCCC)
      + *very fast* assignment of ICD codes to comorbidities (using novel matrix multiplication algorithm and C++ internally -- see 'efficiency' vignette for details)
  * use your existing wide or long data format, icd can guess which columns are ICD-9 or ICD-10 codes.
- * explain and summarize groups of ICD codes in natural language, using ICD editions from the WHO, USA, France and Belgium. Many different annual editions of these data are available, all via the 'icd.data' companion package.
+ * explain and summarize groups of ICD codes in natural language, using ICD editions from the WHO, USA, France and Belgium. Many different annual editions of these data are available, and these may be downloaded automatically when used, or in bulk with `download_all_icd_data()`.
  * Charlson and Van Walraven score calculations
  * Hierarchical Condition Codes (HCC) from CMS
  * Clinical Classifications Software (CCS) comorbidities from AHRQ
