+++ R/parse-comorbid.R
@@ -142,9 +142,9 @@ icd9_parse_ahrq_sas <- function(save_pkg_data = FALSE) {
 # This is in some ways simpler than that ICD-9 equivalent because I make no
 # attempt to find all the child codes.
 icd10_parse_ahrq_sas <- function(save_pkg_data = FALSE,
-                                 offline = TRUE) {
+                                 offline = TRUE, ver = "2016") {
   assert_flag(save_pkg_data)
-  ahrq_info <- icd10_fetch_ahrq_sas()
+  ahrq_info <- icd10_fetch_ahrq_sas(ver = ver)
   ahrq_sas_lines <- readLines(ahrq_info$file_path, warn = FALSE)
   icd10_map_ahrq <- sas_format_extract_rcomfmt(ahrq_sas_lines)
   unun <- function(x) unname(unlist(x))
