### 2019-05-12 1793021
explain table improvements
 - new tests
 - don't sort in merge for massive speed improvement. This is especially so, now that ICD-9 and ICD-10-CM sorting is now done correctly, but using a slightly slower custom sort algorithm. R merge doesn't bulk sort, but makes many > comparisons in R code, and this is now avoided.
 - row names reset in output
 - valid rows were not always correct if merge re-ordered rows.
 - @edlee123
