### 2019-05-13 8904209
various
 - make get_ for 2016 and 2019 ICD-10-CM to be consistent with other years (and help pkgdown)
 - docs
