+++ man/get_icd10cm2016.Rd
@@ -1,0 +1,13 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datadocs.R
+\docType{data}
+\name{get_icd10cm2016}
+\alias{get_icd10cm2016}
+\title{ICD-10-CM 2016}
+\description{
+ICD-10-CM 2016
+}
+\seealso{
+\code{\link{icd10cm2019}}
+}
+\keyword{datasets}
