+++ NAMESPACE
@@ -206,11 +206,13 @@ export(get_icd10cm2014)
 export(get_icd10cm2014_pc)
 export(get_icd10cm2015)
 export(get_icd10cm2015_pc)
+export(get_icd10cm2016)
 export(get_icd10cm2016_pc)
 export(get_icd10cm2017)
 export(get_icd10cm2017_pc)
 export(get_icd10cm2018)
 export(get_icd10cm2018_pc)
+export(get_icd10cm2019)
 export(get_icd10cm2019_pc)
 export(get_icd10cm_active)
 export(get_icd10cm_active_year)

[…]
