### 2019-05-14 0387306
various doc fixes
