+++ man/get_icd10cm2019.Rd
@@ -1,15 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/resource.R
+% Please edit documentation in R/datadocs.R
 \docType{data}
 \name{get_icd10cm2019}
 \alias{get_icd10cm2019}
-\title{ICD-10-CM 2018}
-\usage{
-get_icd10cm2019(alt = NULL, must_work = is.null(alt),
-  msg = paste("Unable to find", var_name))
-}
+\title{ICD-10-CM 2019}
 \description{
-ICD-10-CM 2018
+ICD-10-CM 2019
 }
 \seealso{
 \code{\link{icd10cm2019}}
