### 2019-05-14 765620a
simpler option handling
