+++ R/options.R
@@ -7,11 +7,13 @@
   "interact"
 )
 
-.opt_full_name <- function(opt_name) paste0("icd.data.", opt_name)
+.opt_full_name <- function(opt_name) {
+  paste0("icd.", opt_name)
+}
 
 .show_options <- function() {
   o <- options()
-  o[grepl("^icd\\.data", names(o))]
+  o[grepl("^icd\\.", names(o))]
 }
 
 .print_options <- function() {
@@ -28,57 +30,37 @@
 
 #' Set initial options for the package
 #'
-#' \code{icd.data.offline} - default is \code{TRUE}, unless the system
+#' \code{icd.offline} - default is \code{TRUE}, unless the system
 #' environment vairable \code{ICD_DATA_OFFLINE} is \sQuote{false} or
 #' \sQuote{no}. This will only ever be turned on with explicit user
 #' authorization (or by directly setting it). Turning this on also results in
 #' data being saved in the data directory. See below.
 #'
-#' \code{icd.data.interact} - default is based on interactive mode of R, as
+#' \code{icd.interact} - default is based on interactive mode of R, as
 #' given by \code{base::interactive()}, but can be overridden, e.g. to simulate
 #' non-interactive testing in an interactive environment.
 #'
-#' \code{icd.data.resource} - default is ~/.icd.data but won't write unless user
+#' \code{icd.resource} - default is ~/.icd.data but won't write unless user
 #' gives permission, e.g., using \code{\link{set_icd_data_dir}}
 #'
-#' \code{icd.data.absent_action} - what to do if data is missing, \sQuote{stop},
+#' \code{icd.absent_action} - what to do if data is missing, \sQuote{stop},
 #' \sQuote{warning}, \sQuote{message"}, or \sQuote{silent}.
 #'
-#' \code{icd.data.icd10cm_active_year} - which ICD-10-CM version is currently
+#' \code{icd.icd10cm_active_year} - which ICD-10-CM version is currently
 #' active. Default is \sQuote{2019}.
 #'
-#' See also \code{.show_options} \code{.clear_options} \code{.set_dev_options}
+#' See also \code{.show_options} \code{.clear_options}
 #' @keywords internal
 #' @noRd
 NULL
 
 # the whole point of this is to have 'enum' like behavior, so I can't mistype an
 # option name string elsewhere.
-.set_opt <- function(..., overwrite = FALSE) {
+.set_opt <- function(...) {
   f <- list(...)
-  invisible(
-    sapply(
-      simplify = FALSE,
-      USE.NAMES = TRUE,
-      names(f),
-      function(o) {
-        if (!(o %in% .opt_names)) {
-          stop("Option ", sQuote(o), " is not a valid icd option.")
-        }
-        if (overwrite || is.null(.get_opt(o))) {
-          .dbg("Writing option ", sQuote(o))
-          if (overwrite && .verbose() > 1) {
-            message("overwriting the previous value: ", getOption(o))
-          }
-          args <- list(f[[o]])
-          names(args) <- .opt_full_name(o)
-          do.call(options, args = args)
-        } else {
-          options(o)
-        }
-      }
-    )
-  )
+  stopifnot(!any(grepl(names(f), pattern = "^icd\\.")))
+  names(f) <- paste0("icd.", names(f))
+  options(f)
 }
 
 # only get options we know about, to avoid typo giving a NULL
@@ -99,10 +81,10 @@ NULL
     length(x) == 1L &&
     !is.na(x)) {
     if (is.numeric(x)) x <- as.integer(x)
-    options("icd.data.verbose" = x)
[…]
