### 2019-05-14 e5be672
travis, lastest macos homebrew gcc is 9 [skip appveyor]
