### 2019-05-14 b1cc7ee
split up slower test files so they don't time-out on travis [skip appveyor]
