+++ R/util-str.R
@@ -88,8 +88,8 @@ strip <- function(x, pattern = " ", use_bytes = TRUE) {
 #' \dontrun{
 #' sapply(icd9cm_hierarchy, get_non_ascii)
 #' get_encodings(icd9cm_hierarchy)
-#' sapply(icd9cm_leaf_v32, get_non_ascii)
-#' sapply(icd9cm_leaf_v32, get_encodings)
+#' sapply(icd9cm2014_leaf, get_non_ascii)
+#' sapply(icd9cm2014_leaf, get_encodings)
 #' }
 #' @noRd
 #' @keywords internal
